Cyclops: Cyclic Coordinate Descent for Logistic, Poisson and Survival Analysis

This model fitting tool incorporates cyclic coordinate descent and majorization-minimization approaches to fit a variety of regression models found in large-scale observational healthcare data. Implementations focus on computational optimization and fine-scale parallelization to yield efficient inference in massive datasets. Please see: Suchard, Simpson, Zorych, Ryan and Madigan (2013) doi:10.1145/2414416.2414791 .

See also

Author

Maintainer: Marc A. Suchard msuchard@ucla.edu

Authors:

Martijn J. Schuemie

Trevor R. Shaddox

Yuxi Tian

Jianxiao Yang

Eric Kawaguchi

Other contributors:

Sushil Mittal [contributor]

Observational Health Data Sciences and Informatics [copyright holder]

Marcus Geelnard (provided the TinyThread library) [copyright holder, contributor]

Rutgers University (provided the HParSearch routine) [copyright holder, contributor]

R Development Core Team (provided the ZeroIn routine) [copyright holder, contributor]