Multitype {Cyclops}	R Documentation

Create a multitype outcome object

Description

Multitype creates a multitype outcome object, usually used as a response variable in a hierarchical Cyclops model fit.

Usage

Multitype(y, type)

Arguments

y
Numeric: Response count(s)

type
Numeric or factor: Response type

Value

An object of class Multitype with length equal to the length of y and type.

Examples

Multitype(c(0,1,0), as.factor(c("A","A","B")))

[Package Cyclops version 3.2.0 Index]