appendSqlCyclopsData

appendSqlCyclopsData appends data to an OHDSI data object.

appendSqlCyclopsData(
  object,
  oStratumId,
  oRowId,
  oY,
  oTime,
  cRowId,
  cCovariateId,
  cCovariateValue
)

Arguments

object

OHDSI Cyclops data object to append entries

oStratumId

Integer vector (optional): non-unique stratum identifier for each row in outcomes table

oRowId

Integer vector: unique row identifier for each row in outcomes table

oY

Numeric vector: model outcome variable for each row in outcomes table

oTime

Numeric vector (optional): exposure interval or censoring time for each row in outcomes table

cRowId

Integer vector: non-unique row identifier for each row in covariates table that matches a single outcomes table entry

cCovariateId

Integer vector: covariate identifier

cCovariateValue

Numeric vector: covariate value

Details

Append data using two tables. The outcomes table is dense and contains ... The covariates table is sparse and contains ... All entries in the outcome table must be sorted in increasing order by (oStratumId, oRowId). All entries in the covariate table must be sorted in increasing order by (cRowId). Each cRowId value must match exactly one oRowId value.