coef.cyclopsFit {Cyclops}	R Documentation

Extract model coefficients

Description

coef.cyclopsFit extracts model coefficients from an Cyclops model fit object

Usage

## S3 method for class 'cyclopsFit'
coef(object, rescale = FALSE, ignoreConvergence = FALSE, ...)

Arguments

object
Cyclops model fit object

rescale
Boolean: rescale coefficients for unnormalized covariate values

ignoreConvergence
Boolean: return coefficients even if fit object did not converge

...
Other arguments

Value

Named numeric vector of model coefficients.

[Package Cyclops version 3.2.0 Index]