Confidence intervals for Cyclops model parameters

confinit.cyclopsFit profiles the data likelihood to construct confidence intervals of arbitrary level. Usually it only makes sense to do this for variables that have not been regularized.

# S3 method for cyclopsFit
confint(
  object,
  parm,
  level = 0.95,
  overrideNoRegularization = FALSE,
  includePenalty = TRUE,
  rescale = FALSE,
  ...
)

Arguments

object
A fitted Cyclops model object

parm
A specification of which parameters require confidence intervals, either a vector of numbers of covariateId names

level
Numeric: confidence level required

overrideNoRegularization
Logical: Enable confidence interval estimation for regularized parameters

includePenalty
Logical: Include regularized covariate penalty in profile

rescale
Boolean: rescale coefficients for unnormalized covariate values

...
Additional argument(s) for methods

Value

A matrix with columns reporting lower and upper confidence limits for each parameter. These columns are labelled as (1-level) / 2 and 1 - (1 - level) / 2 in percent (by default 2.5 percent and 97.5 percent)

Examples

#Generate some simulated data:
sim <- simulateCyclopsData(nstrata = 1, nrows = 1000, ncovars = 2, eCovarsPerRow = 0.5, 
                           model = "poisson")

#> Sparseness = 74.45 %

cyclopsData <- convertToCyclopsData(sim$outcomes, sim$covariates, modelType = "pr", 
                                    addIntercept = TRUE)

#> Sorting covariates by covariateId and rowId

#Define the prior and control objects to use cross-validation for finding the 
#optimal hyperparameter:
prior <- createPrior("laplace", exclude = 0, useCrossValidation = TRUE)
control <- createControl(cvType = "auto", noiseLevel = "quiet")

#Fit the model
fit <- fitCyclopsModel(cyclopsData,prior = prior, control = control)  

#> Using cross-validation selector type byRow
#> Performing 10-fold cross-validation [seed = 1623936644] with data partitions of sizes 100 100 100 100 100 100 100 100 100 100
#> Using 1 thread(s)
#> Starting var = 0.2555 (default)
#> Running at Laplace(2.79782) None  Grid-point #1 at 0.2555 	Fold #1 Rep #1 pred log like = 375.98
#> Running at Laplace(2.79782) None  Grid-point #1 at 0.2555 	Fold #2 Rep #1 pred log like = 358.921
#> Running at Laplace(2.79782) None  Grid-point #1 at 0.2555 	Fold #3 Rep #1 pred log like = 402.461
#> Running at Laplace(2.79782) None  Grid-point #1 at 0.2555 	Fold #4 Rep #1 pred log like = 224.91
#> Running at Laplace(2.79782) None  Grid-point #1 at 0.2555 	Fold #5 Rep #1 pred log like = 416.062
#> Running at Laplace(2.79782) None  Grid-point #1 at 0.2555 	Fold #6 Rep #1 pred log like = 381.177
#> Running at Laplace(2.79782) None  Grid-point #1 at 0.2555 	Fold #7 Rep #1 pred log like = 410.934
#> Running at Laplace(2.79782) None  Grid-point #1 at 0.2555 	Fold #8 Rep #1 pred log like = 363.449
#> Running at Laplace(2.79782) None  Grid-point #1 at 0.2555 	Fold #9 Rep #1 pred log like = 380.333
#> Running at Laplace(2.79782) None  Grid-point #1 at 0.2555 	Fold #10 Rep #1 pred log like = 355.956
#> AvgPred = 367.018 with stdev = 51.4233
#> Completed at 0.2555
#> Next point at 2.555 with value 0 and continue = 1
#> search[ 0.2555 ] = 367.018(51.4233)
#> 
#> Running at Laplace(0.884748) None  Grid-point #2 at 2.555 	Fold #1 Rep #1 pred log like = 375.99
#> Running at Laplace(0.884748) None  Grid-point #2 at 2.555 	Fold #2 Rep #1 pred log like = 358.892
#> Running at Laplace(0.884748) None  Grid-point #2 at 2.555 	Fold #3 Rep #1 pred log like = 402.439
#> Running at Laplace(0.884748) None  Grid-point #2 at 2.555 	Fold #4 Rep #1 pred log like = 224.885
#> Running at Laplace(0.884748) None  Grid-point #2 at 2.555 	Fold #5 Rep #1 pred log like = 416.043
#> Running at Laplace(0.884748) None  Grid-point #2 at 2.555 	Fold #6 Rep #1 pred log like = 381.179
#> Running at Laplace(0.884748) None  Grid-point #2 at 2.555 	Fold #7 Rep #1 pred log like = 410.904
#> Running at Laplace(0.884748) None  Grid-point #2 at 2.555 	Fold #8 Rep #1 pred log like = 363.425
#> Running at Laplace(0.884748) None  Grid-point #2 at 2.555 	Fold #9 Rep #1 pred log like = 380.301
#> Running at Laplace(0.884748) None  Grid-point #2 at 2.555 	Fold #10 Rep #1 pred log like = 355.924
#> AvgPred = 366.998 with stdev = 51.4251
#> Completed at 2.555
#> Next point at 0.02555 with value 0 and continue = 1
#> search[ 0.2555 ] = 367.018(51.4233)
#> search[ 2.555 ] = 366.998(51.4251)
#> 
#> Running at Laplace(8.84748) None  Grid-point #3 at 0.02555 	Fold #1 Rep #1 pred log like = 375.97
#> Running at Laplace(8.84748) None  Grid-point #3 at 0.02555 	Fold #2 Rep #1 pred log like = 358.995
#> Running at Laplace(8.84748) None  Grid-point #3 at 0.02555 	Fold #3 Rep #1 pred log like = 402.483
#> Running at Laplace(8.84748) None  Grid-point #3 at 0.02555 	Fold #4 Rep #1 pred log like = 224.985
#> Running at Laplace(8.84748) None  Grid-point #3 at 0.02555 	Fold #5 Rep #1 pred log like = 416.12
#> Running at Laplace(8.84748) None  Grid-point #3 at 0.02555 	Fold #6 Rep #1 pred log like = 381.169
#> Running at Laplace(8.84748) None  Grid-point #3 at 0.02555 	Fold #7 Rep #1 pred log like = 411.069
#> Running at Laplace(8.84748) None  Grid-point #3 at 0.02555 	Fold #8 Rep #1 pred log like = 363.513
#> Running at Laplace(8.84748) None  Grid-point #3 at 0.02555 	Fold #9 Rep #1 pred log like = 380.385
#> Running at Laplace(8.84748) None  Grid-point #3 at 0.02555 	Fold #10 Rep #1 pred log like = 356.053
#> AvgPred = 367.074 with stdev = 51.4186
#> Completed at 0.02555
#> Next point at 0.002555 with value 0 and continue = 1
#> search[ 0.02555 ] = 367.074(51.4186)
#> search[ 0.2555 ] = 367.018(51.4233)
#> search[ 2.555 ] = 366.998(51.4251)
#> 
#> Running at Laplace(27.9782) None  Grid-point #4 at 0.002555 	Fold #1 Rep #1 pred log like = 375.97
#> Running at Laplace(27.9782) None  Grid-point #4 at 0.002555 	Fold #2 Rep #1 pred log like = 359.05
#> Running at Laplace(27.9782) None  Grid-point #4 at 0.002555 	Fold #3 Rep #1 pred log like = 402.485
#> Running at Laplace(27.9782) None  Grid-point #4 at 0.002555 	Fold #4 Rep #1 pred log like = 225.084
#> Running at Laplace(27.9782) None  Grid-point #4 at 0.002555 	Fold #5 Rep #1 pred log like = 416.178
#> Running at Laplace(27.9782) None  Grid-point #4 at 0.002555 	Fold #6 Rep #1 pred log like = 381.169
#> Running at Laplace(27.9782) None  Grid-point #4 at 0.002555 	Fold #7 Rep #1 pred log like = 411.373
#> Running at Laplace(27.9782) None  Grid-point #4 at 0.002555 	Fold #8 Rep #1 pred log like = 363.543
#> Running at Laplace(27.9782) None  Grid-point #4 at 0.002555 	Fold #9 Rep #1 pred log like = 380.399
#> Running at Laplace(27.9782) None  Grid-point #4 at 0.002555 	Fold #10 Rep #1 pred log like = 356.206
#> AvgPred = 367.146 with stdev = 51.4189
#> Completed at 0.002555
#> Next point at 0.0002555 with value 0 and continue = 1
#> search[ 0.002555 ] = 367.146(51.4189)
#> search[ 0.02555 ] = 367.074(51.4186)
#> search[ 0.2555 ] = 367.018(51.4233)
#> search[ 2.555 ] = 366.998(51.4251)
#> 
#> Running at Laplace(88.4748) None  Grid-point #5 at 0.0002555 	Fold #1 Rep #1 pred log like = 375.97
#> Running at Laplace(88.4748) None  Grid-point #5 at 0.0002555 	Fold #2 Rep #1 pred log like = 359.05
#> Running at Laplace(88.4748) None  Grid-point #5 at 0.0002555 	Fold #3 Rep #1 pred log like = 402.485
#> Running at Laplace(88.4748) None  Grid-point #5 at 0.0002555 	Fold #4 Rep #1 pred log like = 225.084
#> Running at Laplace(88.4748) None  Grid-point #5 at 0.0002555 	Fold #5 Rep #1 pred log like = 416.178
#> Running at Laplace(88.4748) None  Grid-point #5 at 0.0002555 	Fold #6 Rep #1 pred log like = 381.169
#> Running at Laplace(88.4748) None  Grid-point #5 at 0.0002555 	Fold #7 Rep #1 pred log like = 411.373
#> Running at Laplace(88.4748) None  Grid-point #5 at 0.0002555 	Fold #8 Rep #1 pred log like = 363.543
#> Running at Laplace(88.4748) None  Grid-point #5 at 0.0002555 	Fold #9 Rep #1 pred log like = 380.399
#> Running at Laplace(88.4748) None  Grid-point #5 at 0.0002555 	Fold #10 Rep #1 pred log like = 356.206
#> AvgPred = 367.146 with stdev = 51.4189
#> Completed at 0.0002555
#> Next point at 2.16648e-14 with value 367.335 and continue = 0
#> search[ 0.0002555 ] = 367.146(51.4189)
#> search[ 0.002555 ] = 367.146(51.4189)
#> search[ 0.02555 ] = 367.074(51.4186)
#> search[ 0.2555 ] = 367.018(51.4233)
#> search[ 2.555 ] = 366.998(51.4251)
#> 
#> 
#> Maximum predicted log likelihood (367.335) estimated at:
#> 	2.16648e-14 (variance)
#> 	9.6081e+06 (lambda)
#> 
#> Fitting model at optimal hyperparameter
#> Using prior: Laplace(9.6081e+06) None 

#Find out what the optimal hyperparameter was:
getHyperParameter(fit)

#> [1] 2.166483e-14

#Extract the current log-likelihood, and coefficients
logLik(fit)

#> 'log Lik.' -1973.343 (df=3)

coef(fit)

#> (Intercept)           1           2 
#>   -3.937029    0.000000    0.000000 

#We can only retrieve the confidence interval for unregularized coefficients:
confint(fit, c(0))

#> Using 1 thread(s)

#>      covariate     2.5 %    97.5 % evaluations
#> [1,]         0 -3.965463 -3.908849          24