Create a Cyclops control object

createControl creates a Cyclops control object for use with fitCyclopsModel.

createControl(
  maxIterations = 1000,
  tolerance = 1e-06,
  convergenceType = "gradient",
  cvType = "auto",
  fold = 10,
  lowerLimit = 0.01,
  upperLimit = 20,
  gridSteps = 10,
  cvRepetitions = 1,
  minCVData = 100,
  noiseLevel = "silent",
  threads = 1,
  seed = NULL,
  resetCoefficients = FALSE,
  startingVariance = -1,
  useKKTSwindle = FALSE,
  tuneSwindle = 10,
  selectorType = "auto",
  initialBound = 2,
  maxBoundCount = 5,
  algorithm = "ccd"
)

Arguments

maxIterations
Integer: maximum iterations of Cyclops to attempt before returning a failed-to-converge error

tolerance
Numeric: maximum relative change in convergence criterion from successive iterations to achieve convergence

convergenceType
String: name of convergence criterion to employ (described in more detail below)

cvType
String: name of cross validation search. Option "auto" selects an auto-search following BBR. Option "grid" selects a grid-search cross validation

fold
Numeric: Number of random folds to employ in cross validation

lowerLimit
Numeric: Lower prior variance limit for grid-search

upperLimit
Numeric: Upper prior variance limit for grid-search

gridSteps
Numeric: Number of steps in grid-search

cvRepetitions
Numeric: Number of repetitions of X-fold cross validation

minCVData
Numeric: Minimum number of data for cross validation

noiseLevel
String: level of Cyclops screen output ("silent", "quiet", "noisy")

threads
Numeric: Specify number of CPU threads to employ in cross-validation; default = 1 (auto = -1)

seed
Numeric: Specify random number generator seed. A null value sets seed via Sys.time.

resetCoefficients
Logical: Reset all coefficients to 0 between model fits under cross-validation

startingVariance
Numeric: Starting variance for auto-search cross-validation; default = -1 (use estimate based on data)

useKKTSwindle
Logical: Use the Karush-Kuhn-Tucker conditions to limit search

tuneSwindle
Numeric: Size multiplier for active set

selectorType
String: name of exchangeable sampling unit. Option "byPid" selects entire strata. Option "byRow" selects single rows. If set to "auto", "byRow" will be used for all models except conditional models where the average number of rows per stratum is smaller than the number of strata.

initialBound
Numeric: Starting trust-region size

maxBoundCount
Numeric: Maximum number of tries to decrease initial trust-region size

algorithm
String: name of fitting algorithm to employ; default is `ccd`

Todo: Describe convegence types

Value

A Cyclops control object of class inheriting from "cyclopsControl" for use with fitCyclopsModel.

Examples

#Generate some simulated data:
sim <- simulateCyclopsData(nstrata = 1, nrows = 1000, ncovars = 2, eCovarsPerRow = 0.5,
                           model = "poisson")

#> Sparseness = 75.2 %

cyclopsData <- convertToCyclopsData(sim$outcomes, sim$covariates, modelType = "pr",
                                    addIntercept = TRUE)

#> Sorting covariates by covariateId and rowId

#Define the prior and control objects to use cross-validation for finding the 
#optimal hyperparameter:
prior <- createPrior("laplace", exclude = 0, useCrossValidation = TRUE)
control <- createControl(cvType = "auto", noiseLevel = "quiet")

#Fit the model
fit <- fitCyclopsModel(cyclopsData,prior = prior, control = control)

#> Using cross-validation selector type byRow
#> Performing 10-fold cross-validation [seed = 1591596882] with data partitions of sizes 100 100 100 100 100 100 100 100 100 100
#> Using 1 thread(s)
#> Starting var = 0.248 (default)
#> Running at Laplace(2.83981) None  Grid-point #1 at 0.248 	Fold #1 Rep #1 pred log like = 500.971
#> Running at Laplace(2.83981) None  Grid-point #1 at 0.248 	Fold #2 Rep #1 pred log like = 553.783
#> Running at Laplace(2.83981) None  Grid-point #1 at 0.248 	Fold #3 Rep #1 pred log like = 507.299
#> Running at Laplace(2.83981) None  Grid-point #1 at 0.248 	Fold #4 Rep #1 pred log like = 548.896
#> Running at Laplace(2.83981) None  Grid-point #1 at 0.248 	Fold #5 Rep #1 pred log like = 468.404
#> Running at Laplace(2.83981) None  Grid-point #1 at 0.248 	Fold #6 Rep #1 pred log like = 422.213
#> Running at Laplace(2.83981) None  Grid-point #1 at 0.248 	Fold #7 Rep #1 pred log like = 415.364
#> Running at Laplace(2.83981) None  Grid-point #1 at 0.248 	Fold #8 Rep #1 pred log like = 442.488
#> Running at Laplace(2.83981) None  Grid-point #1 at 0.248 	Fold #9 Rep #1 pred log like = 498.304
#> Running at Laplace(2.83981) None  Grid-point #1 at 0.248 	Fold #10 Rep #1 pred log like = 367.703
#> AvgPred = 472.543 with stdev = 57.2922
#> Completed at 0.248
#> Next point at 2.48 with value 0 and continue = 1
#> search[ 0.248 ] = 472.543(57.2922)
#> 
#> Running at Laplace(0.898027) None  Grid-point #2 at 2.48 	Fold #1 Rep #1 pred log like = 500.963
#> Running at Laplace(0.898027) None  Grid-point #2 at 2.48 	Fold #2 Rep #1 pred log like = 553.766
#> Running at Laplace(0.898027) None  Grid-point #2 at 2.48 	Fold #3 Rep #1 pred log like = 507.277
#> Running at Laplace(0.898027) None  Grid-point #2 at 2.48 	Fold #4 Rep #1 pred log like = 548.862
#> Running at Laplace(0.898027) None  Grid-point #2 at 2.48 	Fold #5 Rep #1 pred log like = 468.383
#> Running at Laplace(0.898027) None  Grid-point #2 at 2.48 	Fold #6 Rep #1 pred log like = 422.186
#> Running at Laplace(0.898027) None  Grid-point #2 at 2.48 	Fold #7 Rep #1 pred log like = 415.387
#> Running at Laplace(0.898027) None  Grid-point #2 at 2.48 	Fold #8 Rep #1 pred log like = 442.459
#> Running at Laplace(0.898027) None  Grid-point #2 at 2.48 	Fold #9 Rep #1 pred log like = 498.287
#> Running at Laplace(0.898027) None  Grid-point #2 at 2.48 	Fold #10 Rep #1 pred log like = 367.701
#> AvgPred = 472.527 with stdev = 57.2847
#> Completed at 2.48
#> Next point at 0.0248 with value 0 and continue = 1
#> search[ 0.248 ] = 472.543(57.2922)
#> search[ 2.48 ] = 472.527(57.2847)
#> 
#> Running at Laplace(8.98027) None  Grid-point #3 at 0.0248 	Fold #1 Rep #1 pred log like = 500.995
#> Running at Laplace(8.98027) None  Grid-point #3 at 0.0248 	Fold #2 Rep #1 pred log like = 553.831
#> Running at Laplace(8.98027) None  Grid-point #3 at 0.0248 	Fold #3 Rep #1 pred log like = 507.355
#> Running at Laplace(8.98027) None  Grid-point #3 at 0.0248 	Fold #4 Rep #1 pred log like = 549.002
#> Running at Laplace(8.98027) None  Grid-point #3 at 0.0248 	Fold #5 Rep #1 pred log like = 468.394
#> Running at Laplace(8.98027) None  Grid-point #3 at 0.0248 	Fold #6 Rep #1 pred log like = 422.293
#> Running at Laplace(8.98027) None  Grid-point #3 at 0.0248 	Fold #7 Rep #1 pred log like = 415.288
#> Running at Laplace(8.98027) None  Grid-point #3 at 0.0248 	Fold #8 Rep #1 pred log like = 442.575
#> Running at Laplace(8.98027) None  Grid-point #3 at 0.0248 	Fold #9 Rep #1 pred log like = 498.352
#> Running at Laplace(8.98027) None  Grid-point #3 at 0.0248 	Fold #10 Rep #1 pred log like = 367.704
#> AvgPred = 472.579 with stdev = 57.3156
#> Completed at 0.0248
#> Next point at 0.00248 with value 0 and continue = 1
#> search[ 0.0248 ] = 472.579(57.3156)
#> search[ 0.248 ] = 472.543(57.2922)
#> search[ 2.48 ] = 472.527(57.2847)
#> 
#> Running at Laplace(28.3981) None  Grid-point #4 at 0.00248 	Fold #1 Rep #1 pred log like = 501.054
#> Running at Laplace(28.3981) None  Grid-point #4 at 0.00248 	Fold #2 Rep #1 pred log like = 553.863
#> Running at Laplace(28.3981) None  Grid-point #4 at 0.00248 	Fold #3 Rep #1 pred log like = 507.074
#> Running at Laplace(28.3981) None  Grid-point #4 at 0.00248 	Fold #4 Rep #1 pred log like = 549.294
#> Running at Laplace(28.3981) None  Grid-point #4 at 0.00248 	Fold #5 Rep #1 pred log like = 468.291
#> Running at Laplace(28.3981) None  Grid-point #4 at 0.00248 	Fold #6 Rep #1 pred log like = 422.434
#> Running at Laplace(28.3981) None  Grid-point #4 at 0.00248 	Fold #7 Rep #1 pred log like = 415.017
#> Running at Laplace(28.3981) None  Grid-point #4 at 0.00248 	Fold #8 Rep #1 pred log like = 442.724
#> Running at Laplace(28.3981) None  Grid-point #4 at 0.00248 	Fold #9 Rep #1 pred log like = 498.374
#> Running at Laplace(28.3981) None  Grid-point #4 at 0.00248 	Fold #10 Rep #1 pred log like = 367.617
#> AvgPred = 472.574 with stdev = 57.3698
#> Completed at 0.00248
#> Next point at 0.00134154 with value 472.578 and continue = 0
#> search[ 0.00248 ] = 472.574(57.3698)
#> search[ 0.0248 ] = 472.579(57.3156)
#> search[ 0.248 ] = 472.543(57.2922)
#> search[ 2.48 ] = 472.527(57.2847)
#> 
#> 
#> Maximum predicted log likelihood (472.578) estimated at:
#> 	0.00134154 (variance)
#> 	38.6112 (lambda)
#> 
#> Fitting model at optimal hyperparameter
#> Using prior: Laplace(38.6112) None 

#Find out what the optimal hyperparameter was:
getHyperParameter(fit)

#> [1] 0.001341541

#Extract the current log-likelihood, and coefficients
logLik(fit)

#> 'log Lik.' -2055.199 (df=3)

coef(fit)

#> (Intercept)           1           2 
#> -3.84526823  0.00000000 -0.02136532 

#We can only retrieve the confidence interval for unregularized coefficients:
confint(fit, c(0))

#> Using 1 thread(s)

#>      covariate     2.5 %    97.5 % evaluations
#> [1,]         0 -3.875853 -3.814564          23