createNonSeparablePrior {Cyclops}	R Documentation

Create a Cyclops prior object that returns the MLE of non-separable coefficients

Description

createNonSeparablePrior creates a Cyclops prior object for use with fitCyclopsModel.

Usage

createNonSeparablePrior(maxIterations = 10, ...)

Arguments

maxIterations
Numeric: maximum iterations to achieve convergence

...
Additional argument(s) for fitCyclopsModel

Value

A Cyclops prior object of class inheriting from "cyclopsPrior" for use with fitCyclopsModel.

Examples

prior <- createNonSeparablePrior()

[Package Cyclops version 3.2.0 Index]