finalizeSqlCyclopsData {Cyclops}	R Documentation

finalizeSqlCyclopsData

Description

finalizeSqlCyclopsData finalizes a Cyclops data object

Usage

finalizeSqlCyclopsData(
  object,
  addIntercept = FALSE,
  useOffsetCovariate = NULL,
  offsetAlreadyOnLogScale = FALSE,
  sortCovariates = FALSE,
  makeCovariatesDense = NULL
)

Arguments

object
Cyclops data object

addIntercept
Add an intercept covariate if one was not imported through SQL

useOffsetCovariate
Specify is a covariate should be used as an offset (fixed coefficient = 1). Set option to -1 to specify the time-to-event column, otherwise include a single numeric or character covariate name.

offsetAlreadyOnLogScale
Set to TRUE to indicate that offsets were log-transformed before importing into Cyclops data object.

sortCovariates
Sort covariates in numeric-order with intercept first if it exists.

makeCovariatesDense
List of numeric or character covariates names to densely represent in Cyclops data object. For efficiency, we suggest making at least the intercept dense.

[Package Cyclops version 3.2.1 Index]