fitCyclopsSimulation {Cyclops}	R Documentation

Fit simulated data

Description

fitCyclopsSimulation fits simulated Cyclops data using Cyclops or a standard routine. This function is useful for simulation studies comparing the performance of Cyclops when considering large, sparse datasets.

Usage

fitCyclopsSimulation(
  sim,
  useCyclops = TRUE,
  model = "logistic",
  coverage = TRUE,
  includePenalty = FALSE
)

Arguments

sim
A simulated Cyclops dataset generated via simulateCyclopsData

useCyclops
Logical: use Cyclops or a standard routine

model
String: Fitted regression model type

coverage
Logical: report coverage statistics

includePenalty
Logical: include regularized regression penalty in computing profile likelihood based confidence intervals

[Package Cyclops version 3.2.1 Index]