getCyclopsPredictiveLogLikelihood {Cyclops}	R Documentation

Compute predictive log-likelihood from a Cyclops model fit

Description

getCyclopsPredictiveLogLikelihood returns the log-likelihood of a subset of the data in a Cyclops model fit object.

Usage

getCyclopsPredictiveLogLikelihood(object, weights)

Arguments

object
A Cyclops model fit object

weights
Numeric vector: vector of 0/1 identifying subset (=1) of rows from object to use in computing the log-likelihood

Value

The predictive log-likelihood

[Package Cyclops version 3.2.1 Index]