Get hyperparameter

getHyperParameter returns the current hyper parameter in a Cyclops model fit object

getHyperParameter(object)

Arguments

object

A Cyclops model fit object

Examples

#Generate some simulated data:
sim <- simulateCyclopsData(nstrata = 1, nrows = 1000, ncovars = 2, eCovarsPerRow = 0.5, 
                           model = "poisson")
#> Sparseness = 76.45 %
cyclopsData <- convertToCyclopsData(sim$outcomes, sim$covariates, modelType = "pr", 
                                    addIntercept = TRUE)
#> Sorting covariates by covariateId and rowId

#Define the prior and control objects to use cross-validation for finding the 
#optimal hyperparameter:
prior <- createPrior("laplace", exclude = 0, useCrossValidation = TRUE)
control <- createControl(cvType = "auto", noiseLevel = "quiet")

#Fit the model
fit <- fitCyclopsModel(cyclopsData,prior = prior, control = control)  
#> Using cross-validation selector type byRow
#> Performing 10-fold cross-validation [seed = 1698789251] with data partitions of sizes 100 100 100 100 100 100 100 100 100 100
#> Using 1 thread(s)
#> Starting var = 0.2355 (default)
#> Running at Laplace(2.9142) None  Grid-point #1 at 0.2355 	Fold #1 Rep #1 pred log like = 1071.67
#> Running at Laplace(2.9142) None  Grid-point #1 at 0.2355 	Fold #2 Rep #1 pred log like = 887.72
#> Running at Laplace(2.9142) None  Grid-point #1 at 0.2355 	Fold #3 Rep #1 pred log like = 989.777
#> Running at Laplace(2.9142) None  Grid-point #1 at 0.2355 	Fold #4 Rep #1 pred log like = 1248.9
#> Running at Laplace(2.9142) None  Grid-point #1 at 0.2355 	Fold #5 Rep #1 pred log like = 1083.64
#> Running at Laplace(2.9142) None  Grid-point #1 at 0.2355 	Fold #6 Rep #1 pred log like = 1084.42
#> Running at Laplace(2.9142) None  Grid-point #1 at 0.2355 	Fold #7 Rep #1 pred log like = 1009.42
#> Running at Laplace(2.9142) None  Grid-point #1 at 0.2355 	Fold #8 Rep #1 pred log like = 927.842
#> Running at Laplace(2.9142) None  Grid-point #1 at 0.2355 	Fold #9 Rep #1 pred log like = 1098.02
#> Running at Laplace(2.9142) None  Grid-point #1 at 0.2355 	Fold #10 Rep #1 pred log like = 1089.69
#> AvgPred = 1049.11 with stdev = 96.3088
#> Completed at 0.2355
#> Next point at 2.355 with value 0 and continue = 1
#> search[ 0.2355 ] = 1049.11(96.3088)
#> 
#> Running at Laplace(0.921551) None  Grid-point #2 at 2.355 	Fold #1 Rep #1 pred log like = 1071.67
#> Running at Laplace(0.921551) None  Grid-point #2 at 2.355 	Fold #2 Rep #1 pred log like = 887.711
#> Running at Laplace(0.921551) None  Grid-point #2 at 2.355 	Fold #3 Rep #1 pred log like = 989.76
#> Running at Laplace(0.921551) None  Grid-point #2 at 2.355 	Fold #4 Rep #1 pred log like = 1248.86
#> Running at Laplace(0.921551) None  Grid-point #2 at 2.355 	Fold #5 Rep #1 pred log like = 1083.66
#> Running at Laplace(0.921551) None  Grid-point #2 at 2.355 	Fold #6 Rep #1 pred log like = 1084.42
#> Running at Laplace(0.921551) None  Grid-point #2 at 2.355 	Fold #7 Rep #1 pred log like = 1009.42
#> Running at Laplace(0.921551) None  Grid-point #2 at 2.355 	Fold #8 Rep #1 pred log like = 927.82
#> Running at Laplace(0.921551) None  Grid-point #2 at 2.355 	Fold #9 Rep #1 pred log like = 1098
#> Running at Laplace(0.921551) None  Grid-point #2 at 2.355 	Fold #10 Rep #1 pred log like = 1089.79
#> AvgPred = 1049.11 with stdev = 96.3073
#> Completed at 2.355
#> Next point at 23.55 with value 0 and continue = 1
#> search[ 0.2355 ] = 1049.11(96.3088)
#> search[ 2.355 ] = 1049.11(96.3073)
#> 
#> Running at Laplace(0.29142) None  Grid-point #3 at 23.55 	Fold #1 Rep #1 pred log like = 1071.68
#> Running at Laplace(0.29142) None  Grid-point #3 at 23.55 	Fold #2 Rep #1 pred log like = 887.709
#> Running at Laplace(0.29142) None  Grid-point #3 at 23.55 	Fold #3 Rep #1 pred log like = 989.755
#> Running at Laplace(0.29142) None  Grid-point #3 at 23.55 	Fold #4 Rep #1 pred log like = 1248.84
#> Running at Laplace(0.29142) None  Grid-point #3 at 23.55 	Fold #5 Rep #1 pred log like = 1083.66
#> Running at Laplace(0.29142) None  Grid-point #3 at 23.55 	Fold #6 Rep #1 pred log like = 1084.42
#> Running at Laplace(0.29142) None  Grid-point #3 at 23.55 	Fold #7 Rep #1 pred log like = 1009.43
#> Running at Laplace(0.29142) None  Grid-point #3 at 23.55 	Fold #8 Rep #1 pred log like = 927.813
#> Running at Laplace(0.29142) None  Grid-point #3 at 23.55 	Fold #9 Rep #1 pred log like = 1097.99
#> Running at Laplace(0.29142) None  Grid-point #3 at 23.55 	Fold #10 Rep #1 pred log like = 1089.83
#> AvgPred = 1049.11 with stdev = 96.3071
#> Completed at 23.55
#> Next point at 235.5 with value 0 and continue = 1
#> search[ 0.2355 ] = 1049.11(96.3088)
#> search[ 2.355 ] = 1049.11(96.3073)
#> search[ 23.55 ] = 1049.11(96.3071)
#> 
#> Running at Laplace(0.0921551) None  Grid-point #4 at 235.5 	Fold #1 Rep #1 pred log like = 1071.68
#> Running at Laplace(0.0921551) None  Grid-point #4 at 235.5 	Fold #2 Rep #1 pred log like = 887.708
#> Running at Laplace(0.0921551) None  Grid-point #4 at 235.5 	Fold #3 Rep #1 pred log like = 989.753
#> Running at Laplace(0.0921551) None  Grid-point #4 at 235.5 	Fold #4 Rep #1 pred log like = 1248.84
#> Running at Laplace(0.0921551) None  Grid-point #4 at 235.5 	Fold #5 Rep #1 pred log like = 1083.66
#> Running at Laplace(0.0921551) None  Grid-point #4 at 235.5 	Fold #6 Rep #1 pred log like = 1084.42
#> Running at Laplace(0.0921551) None  Grid-point #4 at 235.5 	Fold #7 Rep #1 pred log like = 1009.43
#> Running at Laplace(0.0921551) None  Grid-point #4 at 235.5 	Fold #8 Rep #1 pred log like = 927.81
#> Running at Laplace(0.0921551) None  Grid-point #4 at 235.5 	Fold #9 Rep #1 pred log like = 1097.98
#> Running at Laplace(0.0921551) None  Grid-point #4 at 235.5 	Fold #10 Rep #1 pred log like = 1089.84
#> AvgPred = 1049.11 with stdev = 96.307
#> Completed at 235.5
#> Next point at 2355 with value 0 and continue = 1
#> search[ 0.2355 ] = 1049.11(96.3088)
#> search[ 2.355 ] = 1049.11(96.3073)
#> search[ 23.55 ] = 1049.11(96.3071)
#> search[ 235.5 ] = 1049.11(96.307)
#> 
#> Running at Laplace(0.029142) None  Grid-point #5 at 2355 	Fold #1 Rep #1 pred log like = 1071.68
#> Running at Laplace(0.029142) None  Grid-point #5 at 2355 	Fold #2 Rep #1 pred log like = 887.707
#> Running at Laplace(0.029142) None  Grid-point #5 at 2355 	Fold #3 Rep #1 pred log like = 989.752
#> Running at Laplace(0.029142) None  Grid-point #5 at 2355 	Fold #4 Rep #1 pred log like = 1248.83
#> Running at Laplace(0.029142) None  Grid-point #5 at 2355 	Fold #5 Rep #1 pred log like = 1083.66
#> Running at Laplace(0.029142) None  Grid-point #5 at 2355 	Fold #6 Rep #1 pred log like = 1084.42
#> Running at Laplace(0.029142) None  Grid-point #5 at 2355 	Fold #7 Rep #1 pred log like = 1009.43
#> Running at Laplace(0.029142) None  Grid-point #5 at 2355 	Fold #8 Rep #1 pred log like = 927.81
#> Running at Laplace(0.029142) None  Grid-point #5 at 2355 	Fold #9 Rep #1 pred log like = 1097.98
#> Running at Laplace(0.029142) None  Grid-point #5 at 2355 	Fold #10 Rep #1 pred log like = 1089.84
#> AvgPred = 1049.11 with stdev = 96.307
#> Completed at 2355
#> Next point at 23550 with value 0 and continue = 1
#> search[ 0.2355 ] = 1049.11(96.3088)
#> search[ 2.355 ] = 1049.11(96.3073)
#> search[ 23.55 ] = 1049.11(96.3071)
#> search[ 235.5 ] = 1049.11(96.307)
#> search[ 2355 ] = 1049.11(96.307)
#> 
#> Running at Laplace(0.00921551) None  Grid-point #6 at 23550 	Fold #1 Rep #1 pred log like = 1071.68
#> Running at Laplace(0.00921551) None  Grid-point #6 at 23550 	Fold #2 Rep #1 pred log like = 887.707
#> Running at Laplace(0.00921551) None  Grid-point #6 at 23550 	Fold #3 Rep #1 pred log like = 989.752
#> Running at Laplace(0.00921551) None  Grid-point #6 at 23550 	Fold #4 Rep #1 pred log like = 1248.83
#> Running at Laplace(0.00921551) None  Grid-point #6 at 23550 	Fold #5 Rep #1 pred log like = 1083.66
#> Running at Laplace(0.00921551) None  Grid-point #6 at 23550 	Fold #6 Rep #1 pred log like = 1084.42
#> Running at Laplace(0.00921551) None  Grid-point #6 at 23550 	Fold #7 Rep #1 pred log like = 1009.43
#> Running at Laplace(0.00921551) None  Grid-point #6 at 23550 	Fold #8 Rep #1 pred log like = 927.809
#> Running at Laplace(0.00921551) None  Grid-point #6 at 23550 	Fold #9 Rep #1 pred log like = 1097.98
#> Running at Laplace(0.00921551) None  Grid-point #6 at 23550 	Fold #10 Rep #1 pred log like = 1089.84
#> AvgPred = 1049.11 with stdev = 96.307
#> Completed at 23550
#> Next point at 235500 with value 0 and continue = 1
#> search[ 0.2355 ] = 1049.11(96.3088)
#> search[ 2.355 ] = 1049.11(96.3073)
#> search[ 23.55 ] = 1049.11(96.3071)
#> search[ 235.5 ] = 1049.11(96.307)
#> search[ 2355 ] = 1049.11(96.307)
#> search[ 23550 ] = 1049.11(96.307)
#> 
#> Running at Laplace(0.0029142) None  Grid-point #7 at 235500 	Fold #1 Rep #1 pred log like = 1071.68
#> Running at Laplace(0.0029142) None  Grid-point #7 at 235500 	Fold #2 Rep #1 pred log like = 887.707
#> Running at Laplace(0.0029142) None  Grid-point #7 at 235500 	Fold #3 Rep #1 pred log like = 989.752
#> Running at Laplace(0.0029142) None  Grid-point #7 at 235500 	Fold #4 Rep #1 pred log like = 1248.83
#> Running at Laplace(0.0029142) None  Grid-point #7 at 235500 	Fold #5 Rep #1 pred log like = 1083.66
#> Running at Laplace(0.0029142) None  Grid-point #7 at 235500 	Fold #6 Rep #1 pred log like = 1084.42
#> Running at Laplace(0.0029142) None  Grid-point #7 at 235500 	Fold #7 Rep #1 pred log like = 1009.43
#> Running at Laplace(0.0029142) None  Grid-point #7 at 235500 	Fold #8 Rep #1 pred log like = 927.809
#> Running at Laplace(0.0029142) None  Grid-point #7 at 235500 	Fold #9 Rep #1 pred log like = 1097.98
#> Running at Laplace(0.0029142) None  Grid-point #7 at 235500 	Fold #10 Rep #1 pred log like = 1089.84
#> AvgPred = 1049.11 with stdev = 96.307
#> Completed at 235500
#> Next point at 2.355e+06 with value 0 and continue = 1
#> search[ 0.2355 ] = 1049.11(96.3088)
#> search[ 2.355 ] = 1049.11(96.3073)
#> search[ 23.55 ] = 1049.11(96.3071)
#> search[ 235.5 ] = 1049.11(96.307)
#> search[ 2355 ] = 1049.11(96.307)
#> search[ 23550 ] = 1049.11(96.307)
#> search[ 235500 ] = 1049.11(96.307)
#> 
#> Running at Laplace(0.000921551) None  Grid-point #8 at 2.355e+06 	Fold #1 Rep #1 pred log like = 1071.68
#> Running at Laplace(0.000921551) None  Grid-point #8 at 2.355e+06 	Fold #2 Rep #1 pred log like = 887.707
#> Running at Laplace(0.000921551) None  Grid-point #8 at 2.355e+06 	Fold #3 Rep #1 pred log like = 989.752
#> Running at Laplace(0.000921551) None  Grid-point #8 at 2.355e+06 	Fold #4 Rep #1 pred log like = 1248.83
#> Running at Laplace(0.000921551) None  Grid-point #8 at 2.355e+06 	Fold #5 Rep #1 pred log like = 1083.66
#> Running at Laplace(0.000921551) None  Grid-point #8 at 2.355e+06 	Fold #6 Rep #1 pred log like = 1084.42
#> Running at Laplace(0.000921551) None  Grid-point #8 at 2.355e+06 	Fold #7 Rep #1 pred log like = 1009.43
#> Running at Laplace(0.000921551) None  Grid-point #8 at 2.355e+06 	Fold #8 Rep #1 pred log like = 927.809
#> Running at Laplace(0.000921551) None  Grid-point #8 at 2.355e+06 	Fold #9 Rep #1 pred log like = 1097.98
#> Running at Laplace(0.000921551) None  Grid-point #8 at 2.355e+06 	Fold #10 Rep #1 pred log like = 1089.84
#> AvgPred = 1049.11 with stdev = 96.307
#> Completed at 2.355e+06
#> Next point at 2.355e+07 with value 0 and continue = 1
#> search[ 0.2355 ] = 1049.11(96.3088)
#> search[ 2.355 ] = 1049.11(96.3073)
#> search[ 23.55 ] = 1049.11(96.3071)
#> search[ 235.5 ] = 1049.11(96.307)
#> search[ 2355 ] = 1049.11(96.307)
#> search[ 23550 ] = 1049.11(96.307)
#> search[ 235500 ] = 1049.11(96.307)
#> search[ 2.355e+06 ] = 1049.11(96.307)
#> 
#> Running at Laplace(0.00029142) None  Grid-point #9 at 2.355e+07 	Fold #1 Rep #1 pred log like = 1071.68
#> Running at Laplace(0.00029142) None  Grid-point #9 at 2.355e+07 	Fold #2 Rep #1 pred log like = 887.707
#> Running at Laplace(0.00029142) None  Grid-point #9 at 2.355e+07 	Fold #3 Rep #1 pred log like = 989.752
#> Running at Laplace(0.00029142) None  Grid-point #9 at 2.355e+07 	Fold #4 Rep #1 pred log like = 1248.83
#> Running at Laplace(0.00029142) None  Grid-point #9 at 2.355e+07 	Fold #5 Rep #1 pred log like = 1083.66
#> Running at Laplace(0.00029142) None  Grid-point #9 at 2.355e+07 	Fold #6 Rep #1 pred log like = 1084.42
#> Running at Laplace(0.00029142) None  Grid-point #9 at 2.355e+07 	Fold #7 Rep #1 pred log like = 1009.43
#> Running at Laplace(0.00029142) None  Grid-point #9 at 2.355e+07 	Fold #8 Rep #1 pred log like = 927.809
#> Running at Laplace(0.00029142) None  Grid-point #9 at 2.355e+07 	Fold #9 Rep #1 pred log like = 1097.98
#> Running at Laplace(0.00029142) None  Grid-point #9 at 2.355e+07 	Fold #10 Rep #1 pred log like = 1089.84
#> AvgPred = 1049.11 with stdev = 96.307
#> Completed at 2.355e+07
#> Next point at 2.355e+08 with value 0 and continue = 1
#> search[ 0.2355 ] = 1049.11(96.3088)
#> search[ 2.355 ] = 1049.11(96.3073)
#> search[ 23.55 ] = 1049.11(96.3071)
#> search[ 235.5 ] = 1049.11(96.307)
#> search[ 2355 ] = 1049.11(96.307)
#> search[ 23550 ] = 1049.11(96.307)
#> search[ 235500 ] = 1049.11(96.307)
#> search[ 2.355e+06 ] = 1049.11(96.307)
#> search[ 2.355e+07 ] = 1049.11(96.307)
#> 
#> Running at Laplace(9.21551e-05) None  Grid-point #10 at 2.355e+08 	Fold #1 Rep #1 pred log like = 1071.68
#> Running at Laplace(9.21551e-05) None  Grid-point #10 at 2.355e+08 	Fold #2 Rep #1 pred log like = 887.707
#> Running at Laplace(9.21551e-05) None  Grid-point #10 at 2.355e+08 	Fold #3 Rep #1 pred log like = 989.752
#> Running at Laplace(9.21551e-05) None  Grid-point #10 at 2.355e+08 	Fold #4 Rep #1 pred log like = 1248.83
#> Running at Laplace(9.21551e-05) None  Grid-point #10 at 2.355e+08 	Fold #5 Rep #1 pred log like = 1083.66
#> Running at Laplace(9.21551e-05) None  Grid-point #10 at 2.355e+08 	Fold #6 Rep #1 pred log like = 1084.42
#> Running at Laplace(9.21551e-05) None  Grid-point #10 at 2.355e+08 	Fold #7 Rep #1 pred log like = 1009.43
#> Running at Laplace(9.21551e-05) None  Grid-point #10 at 2.355e+08 	Fold #8 Rep #1 pred log like = 927.809
#> Running at Laplace(9.21551e-05) None  Grid-point #10 at 2.355e+08 	Fold #9 Rep #1 pred log like = 1097.98
#> Running at Laplace(9.21551e-05) None  Grid-point #10 at 2.355e+08 	Fold #10 Rep #1 pred log like = 1089.84
#> AvgPred = 1049.11 with stdev = 96.307
#> Completed at 2.355e+08
#> Next point at 2.355e+09 with value 0 and continue = 1
#> search[ 0.2355 ] = 1049.11(96.3088)
#> search[ 2.355 ] = 1049.11(96.3073)
#> search[ 23.55 ] = 1049.11(96.3071)
#> search[ 235.5 ] = 1049.11(96.307)
#> search[ 2355 ] = 1049.11(96.307)
#> search[ 23550 ] = 1049.11(96.307)
#> search[ 235500 ] = 1049.11(96.307)
#> search[ 2.355e+06 ] = 1049.11(96.307)
#> search[ 2.355e+07 ] = 1049.11(96.307)
#> search[ 2.355e+08 ] = 1049.11(96.307)
#> 
#> Running at Laplace(2.9142e-05) None  Grid-point #11 at 2.355e+09 	Fold #1 Rep #1 pred log like = 1071.68
#> Running at Laplace(2.9142e-05) None  Grid-point #11 at 2.355e+09 	Fold #2 Rep #1 pred log like = 887.707
#> Running at Laplace(2.9142e-05) None  Grid-point #11 at 2.355e+09 	Fold #3 Rep #1 pred log like = 989.752
#> Running at Laplace(2.9142e-05) None  Grid-point #11 at 2.355e+09 	Fold #4 Rep #1 pred log like = 1248.83
#> Running at Laplace(2.9142e-05) None  Grid-point #11 at 2.355e+09 	Fold #5 Rep #1 pred log like = 1083.66
#> Running at Laplace(2.9142e-05) None  Grid-point #11 at 2.355e+09 	Fold #6 Rep #1 pred log like = 1084.42
#> Running at Laplace(2.9142e-05) None  Grid-point #11 at 2.355e+09 	Fold #7 Rep #1 pred log like = 1009.43
#> Running at Laplace(2.9142e-05) None  Grid-point #11 at 2.355e+09 	Fold #8 Rep #1 pred log like = 927.809
#> Running at Laplace(2.9142e-05) None  Grid-point #11 at 2.355e+09 	Fold #9 Rep #1 pred log like = 1097.98
#> Running at Laplace(2.9142e-05) None  Grid-point #11 at 2.355e+09 	Fold #10 Rep #1 pred log like = 1089.84
#> AvgPred = 1049.11 with stdev = 96.307
#> Completed at 2.355e+09
#> Next point at 2.355e+10 with value 0 and continue = 1
#> search[ 0.2355 ] = 1049.11(96.3088)
#> search[ 2.355 ] = 1049.11(96.3073)
#> search[ 23.55 ] = 1049.11(96.3071)
#> search[ 235.5 ] = 1049.11(96.307)
#> search[ 2355 ] = 1049.11(96.307)
#> search[ 23550 ] = 1049.11(96.307)
#> search[ 235500 ] = 1049.11(96.307)
#> search[ 2.355e+06 ] = 1049.11(96.307)
#> search[ 2.355e+07 ] = 1049.11(96.307)
#> search[ 2.355e+08 ] = 1049.11(96.307)
#> search[ 2.355e+09 ] = 1049.11(96.307)
#> 
#> Running at Laplace(9.21551e-06) None  Grid-point #12 at 2.355e+10 	Fold #1 Rep #1 pred log like = 1071.68
#> Running at Laplace(9.21551e-06) None  Grid-point #12 at 2.355e+10 	Fold #2 Rep #1 pred log like = 887.707
#> Running at Laplace(9.21551e-06) None  Grid-point #12 at 2.355e+10 	Fold #3 Rep #1 pred log like = 989.752
#> Running at Laplace(9.21551e-06) None  Grid-point #12 at 2.355e+10 	Fold #4 Rep #1 pred log like = 1248.83
#> Running at Laplace(9.21551e-06) None  Grid-point #12 at 2.355e+10 	Fold #5 Rep #1 pred log like = 1083.66
#> Running at Laplace(9.21551e-06) None  Grid-point #12 at 2.355e+10 	Fold #6 Rep #1 pred log like = 1084.42
#> Running at Laplace(9.21551e-06) None  Grid-point #12 at 2.355e+10 	Fold #7 Rep #1 pred log like = 1009.43
#> Running at Laplace(9.21551e-06) None  Grid-point #12 at 2.355e+10 	Fold #8 Rep #1 pred log like = 927.809
#> Running at Laplace(9.21551e-06) None  Grid-point #12 at 2.355e+10 	Fold #9 Rep #1 pred log like = 1097.98
#> Running at Laplace(9.21551e-06) None  Grid-point #12 at 2.355e+10 	Fold #10 Rep #1 pred log like = 1089.84
#> AvgPred = 1049.11 with stdev = 96.307
#> Completed at 2.355e+10
#> Next point at 2.355e+11 with value 0 and continue = 1
#> search[ 0.2355 ] = 1049.11(96.3088)
#> search[ 2.355 ] = 1049.11(96.3073)
#> search[ 23.55 ] = 1049.11(96.3071)
#> search[ 235.5 ] = 1049.11(96.307)
#> search[ 2355 ] = 1049.11(96.307)
#> search[ 23550 ] = 1049.11(96.307)
#> search[ 235500 ] = 1049.11(96.307)
#> search[ 2.355e+06 ] = 1049.11(96.307)
#> search[ 2.355e+07 ] = 1049.11(96.307)
#> search[ 2.355e+08 ] = 1049.11(96.307)
#> search[ 2.355e+09 ] = 1049.11(96.307)
#> search[ 2.355e+10 ] = 1049.11(96.307)
#> 
#> Running at Laplace(2.9142e-06) None  Grid-point #13 at 2.355e+11 	Fold #1 Rep #1 pred log like = 1071.68
#> Running at Laplace(2.9142e-06) None  Grid-point #13 at 2.355e+11 	Fold #2 Rep #1 pred log like = 887.707
#> Running at Laplace(2.9142e-06) None  Grid-point #13 at 2.355e+11 	Fold #3 Rep #1 pred log like = 989.752
#> Running at Laplace(2.9142e-06) None  Grid-point #13 at 2.355e+11 	Fold #4 Rep #1 pred log like = 1248.83
#> Running at Laplace(2.9142e-06) None  Grid-point #13 at 2.355e+11 	Fold #5 Rep #1 pred log like = 1083.66
#> Running at Laplace(2.9142e-06) None  Grid-point #13 at 2.355e+11 	Fold #6 Rep #1 pred log like = 1084.42
#> Running at Laplace(2.9142e-06) None  Grid-point #13 at 2.355e+11 	Fold #7 Rep #1 pred log like = 1009.43
#> Running at Laplace(2.9142e-06) None  Grid-point #13 at 2.355e+11 	Fold #8 Rep #1 pred log like = 927.809
#> Running at Laplace(2.9142e-06) None  Grid-point #13 at 2.355e+11 	Fold #9 Rep #1 pred log like = 1097.98
#> Running at Laplace(2.9142e-06) None  Grid-point #13 at 2.355e+11 	Fold #10 Rep #1 pred log like = 1089.84
#> AvgPred = 1049.11 with stdev = 96.307
#> Completed at 2.355e+11
#> Next point at 2.355e+12 with value 0 and continue = 1
#> search[ 0.2355 ] = 1049.11(96.3088)
#> search[ 2.355 ] = 1049.11(96.3073)
#> search[ 23.55 ] = 1049.11(96.3071)
#> search[ 235.5 ] = 1049.11(96.307)
#> search[ 2355 ] = 1049.11(96.307)
#> search[ 23550 ] = 1049.11(96.307)
#> search[ 235500 ] = 1049.11(96.307)
#> search[ 2.355e+06 ] = 1049.11(96.307)
#> search[ 2.355e+07 ] = 1049.11(96.307)
#> search[ 2.355e+08 ] = 1049.11(96.307)
#> search[ 2.355e+09 ] = 1049.11(96.307)
#> search[ 2.355e+10 ] = 1049.11(96.307)
#> search[ 2.355e+11 ] = 1049.11(96.307)
#> 
#> Running at Laplace(9.21551e-07) None  Grid-point #14 at 2.355e+12 	Fold #1 Rep #1 pred log like = 1071.68
#> Running at Laplace(9.21551e-07) None  Grid-point #14 at 2.355e+12 	Fold #2 Rep #1 pred log like = 887.707
#> Running at Laplace(9.21551e-07) None  Grid-point #14 at 2.355e+12 	Fold #3 Rep #1 pred log like = 989.752
#> Running at Laplace(9.21551e-07) None  Grid-point #14 at 2.355e+12 	Fold #4 Rep #1 pred log like = 1248.83
#> Running at Laplace(9.21551e-07) None  Grid-point #14 at 2.355e+12 	Fold #5 Rep #1 pred log like = 1083.66
#> Running at Laplace(9.21551e-07) None  Grid-point #14 at 2.355e+12 	Fold #6 Rep #1 pred log like = 1084.42
#> Running at Laplace(9.21551e-07) None  Grid-point #14 at 2.355e+12 	Fold #7 Rep #1 pred log like = 1009.43
#> Running at Laplace(9.21551e-07) None  Grid-point #14 at 2.355e+12 	Fold #8 Rep #1 pred log like = 927.809
#> Running at Laplace(9.21551e-07) None  Grid-point #14 at 2.355e+12 	Fold #9 Rep #1 pred log like = 1097.98
#> Running at Laplace(9.21551e-07) None  Grid-point #14 at 2.355e+12 	Fold #10 Rep #1 pred log like = 1089.84
#> AvgPred = 1049.11 with stdev = 96.307
#> Completed at 2.355e+12
#> Next point at 2.355e+13 with value 0 and continue = 1
#> search[ 0.2355 ] = 1049.11(96.3088)
#> search[ 2.355 ] = 1049.11(96.3073)
#> search[ 23.55 ] = 1049.11(96.3071)
#> search[ 235.5 ] = 1049.11(96.307)
#> search[ 2355 ] = 1049.11(96.307)
#> search[ 23550 ] = 1049.11(96.307)
#> search[ 235500 ] = 1049.11(96.307)
#> search[ 2.355e+06 ] = 1049.11(96.307)
#> search[ 2.355e+07 ] = 1049.11(96.307)
#> search[ 2.355e+08 ] = 1049.11(96.307)
#> search[ 2.355e+09 ] = 1049.11(96.307)
#> search[ 2.355e+10 ] = 1049.11(96.307)
#> search[ 2.355e+11 ] = 1049.11(96.307)
#> search[ 2.355e+12 ] = 1049.11(96.307)
#> 
#> Running at Laplace(2.9142e-07) None  Grid-point #15 at 2.355e+13 	Fold #1 Rep #1 pred log like = 1071.68
#> Running at Laplace(2.9142e-07) None  Grid-point #15 at 2.355e+13 	Fold #2 Rep #1 pred log like = 887.707
#> Running at Laplace(2.9142e-07) None  Grid-point #15 at 2.355e+13 	Fold #3 Rep #1 pred log like = 989.752
#> Running at Laplace(2.9142e-07) None  Grid-point #15 at 2.355e+13 	Fold #4 Rep #1 pred log like = 1248.83
#> Running at Laplace(2.9142e-07) None  Grid-point #15 at 2.355e+13 	Fold #5 Rep #1 pred log like = 1083.66
#> Running at Laplace(2.9142e-07) None  Grid-point #15 at 2.355e+13 	Fold #6 Rep #1 pred log like = 1084.42
#> Running at Laplace(2.9142e-07) None  Grid-point #15 at 2.355e+13 	Fold #7 Rep #1 pred log like = 1009.43
#> Running at Laplace(2.9142e-07) None  Grid-point #15 at 2.355e+13 	Fold #8 Rep #1 pred log like = 927.809
#> Running at Laplace(2.9142e-07) None  Grid-point #15 at 2.355e+13 	Fold #9 Rep #1 pred log like = 1097.98
#> Running at Laplace(2.9142e-07) None  Grid-point #15 at 2.355e+13 	Fold #10 Rep #1 pred log like = 1089.84
#> AvgPred = 1049.11 with stdev = 96.307
#> Completed at 2.355e+13
#> Next point at 2.355e+14 with value 0 and continue = 1
#> search[ 0.2355 ] = 1049.11(96.3088)
#> search[ 2.355 ] = 1049.11(96.3073)
#> search[ 23.55 ] = 1049.11(96.3071)
#> search[ 235.5 ] = 1049.11(96.307)
#> search[ 2355 ] = 1049.11(96.307)
#> search[ 23550 ] = 1049.11(96.307)
#> search[ 235500 ] = 1049.11(96.307)
#> search[ 2.355e+06 ] = 1049.11(96.307)
#> search[ 2.355e+07 ] = 1049.11(96.307)
#> search[ 2.355e+08 ] = 1049.11(96.307)
#> search[ 2.355e+09 ] = 1049.11(96.307)
#> search[ 2.355e+10 ] = 1049.11(96.307)
#> search[ 2.355e+11 ] = 1049.11(96.307)
#> search[ 2.355e+12 ] = 1049.11(96.307)
#> search[ 2.355e+13 ] = 1049.11(96.307)
#> 
#> Running at Laplace(9.21551e-08) None  Grid-point #16 at 2.355e+14 	Fold #1 Rep #1 pred log like = 1071.68
#> Running at Laplace(9.21551e-08) None  Grid-point #16 at 2.355e+14 	Fold #2 Rep #1 pred log like = 887.707
#> Running at Laplace(9.21551e-08) None  Grid-point #16 at 2.355e+14 	Fold #3 Rep #1 pred log like = 989.752
#> Running at Laplace(9.21551e-08) None  Grid-point #16 at 2.355e+14 	Fold #4 Rep #1 pred log like = 1248.83
#> Running at Laplace(9.21551e-08) None  Grid-point #16 at 2.355e+14 	Fold #5 Rep #1 pred log like = 1083.66
#> Running at Laplace(9.21551e-08) None  Grid-point #16 at 2.355e+14 	Fold #6 Rep #1 pred log like = 1084.42
#> Running at Laplace(9.21551e-08) None  Grid-point #16 at 2.355e+14 	Fold #7 Rep #1 pred log like = 1009.43
#> Running at Laplace(9.21551e-08) None  Grid-point #16 at 2.355e+14 	Fold #8 Rep #1 pred log like = 927.809
#> Running at Laplace(9.21551e-08) None  Grid-point #16 at 2.355e+14 	Fold #9 Rep #1 pred log like = 1097.98
#> Running at Laplace(9.21551e-08) None  Grid-point #16 at 2.355e+14 	Fold #10 Rep #1 pred log like = 1089.84
#> AvgPred = 1049.11 with stdev = 96.307
#> Completed at 2.355e+14
#> Next point at 2.355e+15 with value 0 and continue = 1
#> search[ 0.2355 ] = 1049.11(96.3088)
#> search[ 2.355 ] = 1049.11(96.3073)
#> search[ 23.55 ] = 1049.11(96.3071)
#> search[ 235.5 ] = 1049.11(96.307)
#> search[ 2355 ] = 1049.11(96.307)
#> search[ 23550 ] = 1049.11(96.307)
#> search[ 235500 ] = 1049.11(96.307)
#> search[ 2.355e+06 ] = 1049.11(96.307)
#> search[ 2.355e+07 ] = 1049.11(96.307)
#> search[ 2.355e+08 ] = 1049.11(96.307)
#> search[ 2.355e+09 ] = 1049.11(96.307)
#> search[ 2.355e+10 ] = 1049.11(96.307)
#> search[ 2.355e+11 ] = 1049.11(96.307)
#> search[ 2.355e+12 ] = 1049.11(96.307)
#> search[ 2.355e+13 ] = 1049.11(96.307)
#> search[ 2.355e+14 ] = 1049.11(96.307)
#> 
#> Running at Laplace(2.9142e-08) None  Grid-point #17 at 2.355e+15 	Fold #1 Rep #1 pred log like = 1071.68
#> Running at Laplace(2.9142e-08) None  Grid-point #17 at 2.355e+15 	Fold #2 Rep #1 pred log like = 887.707
#> Running at Laplace(2.9142e-08) None  Grid-point #17 at 2.355e+15 	Fold #3 Rep #1 pred log like = 989.752
#> Running at Laplace(2.9142e-08) None  Grid-point #17 at 2.355e+15 	Fold #4 Rep #1 pred log like = 1248.83
#> Running at Laplace(2.9142e-08) None  Grid-point #17 at 2.355e+15 	Fold #5 Rep #1 pred log like = 1083.66
#> Running at Laplace(2.9142e-08) None  Grid-point #17 at 2.355e+15 	Fold #6 Rep #1 pred log like = 1084.42
#> Running at Laplace(2.9142e-08) None  Grid-point #17 at 2.355e+15 	Fold #7 Rep #1 pred log like = 1009.43
#> Running at Laplace(2.9142e-08) None  Grid-point #17 at 2.355e+15 	Fold #8 Rep #1 pred log like = 927.809
#> Running at Laplace(2.9142e-08) None  Grid-point #17 at 2.355e+15 	Fold #9 Rep #1 pred log like = 1097.98
#> Running at Laplace(2.9142e-08) None  Grid-point #17 at 2.355e+15 	Fold #10 Rep #1 pred log like = 1089.84
#> AvgPred = 1049.11 with stdev = 96.307
#> Completed at 2.355e+15
#> Next point at 2.355e+16 with value 0 and continue = 1
#> search[ 0.2355 ] = 1049.11(96.3088)
#> search[ 2.355 ] = 1049.11(96.3073)
#> search[ 23.55 ] = 1049.11(96.3071)
#> search[ 235.5 ] = 1049.11(96.307)
#> search[ 2355 ] = 1049.11(96.307)
#> search[ 23550 ] = 1049.11(96.307)
#> search[ 235500 ] = 1049.11(96.307)
#> search[ 2.355e+06 ] = 1049.11(96.307)
#> search[ 2.355e+07 ] = 1049.11(96.307)
#> search[ 2.355e+08 ] = 1049.11(96.307)
#> search[ 2.355e+09 ] = 1049.11(96.307)
#> search[ 2.355e+10 ] = 1049.11(96.307)
#> search[ 2.355e+11 ] = 1049.11(96.307)
#> search[ 2.355e+12 ] = 1049.11(96.307)
#> search[ 2.355e+13 ] = 1049.11(96.307)
#> search[ 2.355e+14 ] = 1049.11(96.307)
#> search[ 2.355e+15 ] = 1049.11(96.307)
#> 
#> Running at Laplace(9.21551e-09) None  Grid-point #18 at 2.355e+16 	Fold #1 Rep #1 pred log like = 1071.68
#> Running at Laplace(9.21551e-09) None  Grid-point #18 at 2.355e+16 	Fold #2 Rep #1 pred log like = 887.707
#> Running at Laplace(9.21551e-09) None  Grid-point #18 at 2.355e+16 	Fold #3 Rep #1 pred log like = 989.752
#> Running at Laplace(9.21551e-09) None  Grid-point #18 at 2.355e+16 	Fold #4 Rep #1 pred log like = 1248.83
#> Running at Laplace(9.21551e-09) None  Grid-point #18 at 2.355e+16 	Fold #5 Rep #1 pred log like = 1083.66
#> Running at Laplace(9.21551e-09) None  Grid-point #18 at 2.355e+16 	Fold #6 Rep #1 pred log like = 1084.42
#> Running at Laplace(9.21551e-09) None  Grid-point #18 at 2.355e+16 	Fold #7 Rep #1 pred log like = 1009.43
#> Running at Laplace(9.21551e-09) None  Grid-point #18 at 2.355e+16 	Fold #8 Rep #1 pred log like = 927.809
#> Running at Laplace(9.21551e-09) None  Grid-point #18 at 2.355e+16 	Fold #9 Rep #1 pred log like = 1097.98
#> Running at Laplace(9.21551e-09) None  Grid-point #18 at 2.355e+16 	Fold #10 Rep #1 pred log like = 1089.84
#> AvgPred = 1049.11 with stdev = 96.307
#> Completed at 2.355e+16
#> Next point at 2.355e+17 with value 0 and continue = 1
#> search[ 0.2355 ] = 1049.11(96.3088)
#> search[ 2.355 ] = 1049.11(96.3073)
#> search[ 23.55 ] = 1049.11(96.3071)
#> search[ 235.5 ] = 1049.11(96.307)
#> search[ 2355 ] = 1049.11(96.307)
#> search[ 23550 ] = 1049.11(96.307)
#> search[ 235500 ] = 1049.11(96.307)
#> search[ 2.355e+06 ] = 1049.11(96.307)
#> search[ 2.355e+07 ] = 1049.11(96.307)
#> search[ 2.355e+08 ] = 1049.11(96.307)
#> search[ 2.355e+09 ] = 1049.11(96.307)
#> search[ 2.355e+10 ] = 1049.11(96.307)
#> search[ 2.355e+11 ] = 1049.11(96.307)
#> search[ 2.355e+12 ] = 1049.11(96.307)
#> search[ 2.355e+13 ] = 1049.11(96.307)
#> search[ 2.355e+14 ] = 1049.11(96.307)
#> search[ 2.355e+15 ] = 1049.11(96.307)
#> search[ 2.355e+16 ] = 1049.11(96.307)
#> 
#> Running at Laplace(2.9142e-09) None  Grid-point #19 at 2.355e+17 	Fold #1 Rep #1 pred log like = 1071.68
#> Running at Laplace(2.9142e-09) None  Grid-point #19 at 2.355e+17 	Fold #2 Rep #1 pred log like = 887.707
#> Running at Laplace(2.9142e-09) None  Grid-point #19 at 2.355e+17 	Fold #3 Rep #1 pred log like = 989.752
#> Running at Laplace(2.9142e-09) None  Grid-point #19 at 2.355e+17 	Fold #4 Rep #1 pred log like = 1248.83
#> Running at Laplace(2.9142e-09) None  Grid-point #19 at 2.355e+17 	Fold #5 Rep #1 pred log like = 1083.66
#> Running at Laplace(2.9142e-09) None  Grid-point #19 at 2.355e+17 	Fold #6 Rep #1 pred log like = 1084.42
#> Running at Laplace(2.9142e-09) None  Grid-point #19 at 2.355e+17 	Fold #7 Rep #1 pred log like = 1009.43
#> Running at Laplace(2.9142e-09) None  Grid-point #19 at 2.355e+17 	Fold #8 Rep #1 pred log like = 927.809
#> Running at Laplace(2.9142e-09) None  Grid-point #19 at 2.355e+17 	Fold #9 Rep #1 pred log like = 1097.98
#> Running at Laplace(2.9142e-09) None  Grid-point #19 at 2.355e+17 	Fold #10 Rep #1 pred log like = 1089.84
#> AvgPred = 1049.11 with stdev = 96.307
#> Completed at 2.355e+17
#> Next point at 2.355e+18 with value 0 and continue = 1
#> search[ 0.2355 ] = 1049.11(96.3088)
#> search[ 2.355 ] = 1049.11(96.3073)
#> search[ 23.55 ] = 1049.11(96.3071)
#> search[ 235.5 ] = 1049.11(96.307)
#> search[ 2355 ] = 1049.11(96.307)
#> search[ 23550 ] = 1049.11(96.307)
#> search[ 235500 ] = 1049.11(96.307)
#> search[ 2.355e+06 ] = 1049.11(96.307)
#> search[ 2.355e+07 ] = 1049.11(96.307)
#> search[ 2.355e+08 ] = 1049.11(96.307)
#> search[ 2.355e+09 ] = 1049.11(96.307)
#> search[ 2.355e+10 ] = 1049.11(96.307)
#> search[ 2.355e+11 ] = 1049.11(96.307)
#> search[ 2.355e+12 ] = 1049.11(96.307)
#> search[ 2.355e+13 ] = 1049.11(96.307)
#> search[ 2.355e+14 ] = 1049.11(96.307)
#> search[ 2.355e+15 ] = 1049.11(96.307)
#> search[ 2.355e+16 ] = 1049.11(96.307)
#> search[ 2.355e+17 ] = 1049.11(96.307)
#> 
#> Running at Laplace(9.21551e-10) None  Grid-point #20 at 2.355e+18 	Fold #1 Rep #1 pred log like = 1071.68
#> Running at Laplace(9.21551e-10) None  Grid-point #20 at 2.355e+18 	Fold #2 Rep #1 pred log like = 887.707
#> Running at Laplace(9.21551e-10) None  Grid-point #20 at 2.355e+18 	Fold #3 Rep #1 pred log like = 989.752
#> Running at Laplace(9.21551e-10) None  Grid-point #20 at 2.355e+18 	Fold #4 Rep #1 pred log like = 1248.83
#> Running at Laplace(9.21551e-10) None  Grid-point #20 at 2.355e+18 	Fold #5 Rep #1 pred log like = 1083.66
#> Running at Laplace(9.21551e-10) None  Grid-point #20 at 2.355e+18 	Fold #6 Rep #1 pred log like = 1084.42
#> Running at Laplace(9.21551e-10) None  Grid-point #20 at 2.355e+18 	Fold #7 Rep #1 pred log like = 1009.43
#> Running at Laplace(9.21551e-10) None  Grid-point #20 at 2.355e+18 	Fold #8 Rep #1 pred log like = 927.809
#> Running at Laplace(9.21551e-10) None  Grid-point #20 at 2.355e+18 	Fold #9 Rep #1 pred log like = 1097.98
#> Running at Laplace(9.21551e-10) None  Grid-point #20 at 2.355e+18 	Fold #10 Rep #1 pred log like = 1089.84
#> AvgPred = 1049.11 with stdev = 96.307
#> Completed at 2.355e+18
#> Next point at 2.355e+19 with value 0 and continue = 1
#> search[ 0.2355 ] = 1049.11(96.3088)
#> search[ 2.355 ] = 1049.11(96.3073)
#> search[ 23.55 ] = 1049.11(96.3071)
#> search[ 235.5 ] = 1049.11(96.307)
#> search[ 2355 ] = 1049.11(96.307)
#> search[ 23550 ] = 1049.11(96.307)
#> search[ 235500 ] = 1049.11(96.307)
#> search[ 2.355e+06 ] = 1049.11(96.307)
#> search[ 2.355e+07 ] = 1049.11(96.307)
#> search[ 2.355e+08 ] = 1049.11(96.307)
#> search[ 2.355e+09 ] = 1049.11(96.307)
#> search[ 2.355e+10 ] = 1049.11(96.307)
#> search[ 2.355e+11 ] = 1049.11(96.307)
#> search[ 2.355e+12 ] = 1049.11(96.307)
#> search[ 2.355e+13 ] = 1049.11(96.307)
#> search[ 2.355e+14 ] = 1049.11(96.307)
#> search[ 2.355e+15 ] = 1049.11(96.307)
#> search[ 2.355e+16 ] = 1049.11(96.307)
#> search[ 2.355e+17 ] = 1049.11(96.307)
#> search[ 2.355e+18 ] = 1049.11(96.307)
#> 
#> Running at Laplace(2.9142e-10) None  Grid-point #21 at 2.355e+19 	Fold #1 Rep #1 pred log like = 1071.68
#> Running at Laplace(2.9142e-10) None  Grid-point #21 at 2.355e+19 	Fold #2 Rep #1 pred log like = 887.707
#> Running at Laplace(2.9142e-10) None  Grid-point #21 at 2.355e+19 	Fold #3 Rep #1 pred log like = 989.752
#> Running at Laplace(2.9142e-10) None  Grid-point #21 at 2.355e+19 	Fold #4 Rep #1 pred log like = 1248.83
#> Running at Laplace(2.9142e-10) None  Grid-point #21 at 2.355e+19 	Fold #5 Rep #1 pred log like = 1083.66
#> Running at Laplace(2.9142e-10) None  Grid-point #21 at 2.355e+19 	Fold #6 Rep #1 pred log like = 1084.42
#> Running at Laplace(2.9142e-10) None  Grid-point #21 at 2.355e+19 	Fold #7 Rep #1 pred log like = 1009.43
#> Running at Laplace(2.9142e-10) None  Grid-point #21 at 2.355e+19 	Fold #8 Rep #1 pred log like = 927.809
#> Running at Laplace(2.9142e-10) None  Grid-point #21 at 2.355e+19 	Fold #9 Rep #1 pred log like = 1097.98
#> Running at Laplace(2.9142e-10) None  Grid-point #21 at 2.355e+19 	Fold #10 Rep #1 pred log like = 1089.84
#> AvgPred = 1049.11 with stdev = 96.307
#> Completed at 2.355e+19
#> Next point at 2.355e+20 with value 0 and continue = 1
#> search[ 0.2355 ] = 1049.11(96.3088)
#> search[ 2.355 ] = 1049.11(96.3073)
#> search[ 23.55 ] = 1049.11(96.3071)
#> search[ 235.5 ] = 1049.11(96.307)
#> search[ 2355 ] = 1049.11(96.307)
#> search[ 23550 ] = 1049.11(96.307)
#> search[ 235500 ] = 1049.11(96.307)
#> search[ 2.355e+06 ] = 1049.11(96.307)
#> search[ 2.355e+07 ] = 1049.11(96.307)
#> search[ 2.355e+08 ] = 1049.11(96.307)
#> search[ 2.355e+09 ] = 1049.11(96.307)
#> search[ 2.355e+10 ] = 1049.11(96.307)
#> search[ 2.355e+11 ] = 1049.11(96.307)
#> search[ 2.355e+12 ] = 1049.11(96.307)
#> search[ 2.355e+13 ] = 1049.11(96.307)
#> search[ 2.355e+14 ] = 1049.11(96.307)
#> search[ 2.355e+15 ] = 1049.11(96.307)
#> search[ 2.355e+16 ] = 1049.11(96.307)
#> search[ 2.355e+17 ] = 1049.11(96.307)
#> search[ 2.355e+18 ] = 1049.11(96.307)
#> search[ 2.355e+19 ] = 1049.11(96.307)
#> 
#> Running at Laplace(9.21551e-11) None  Grid-point #22 at 2.355e+20 	Fold #1 Rep #1 pred log like = 1071.68
#> Running at Laplace(9.21551e-11) None  Grid-point #22 at 2.355e+20 	Fold #2 Rep #1 pred log like = 887.707
#> Running at Laplace(9.21551e-11) None  Grid-point #22 at 2.355e+20 	Fold #3 Rep #1 pred log like = 989.752
#> Running at Laplace(9.21551e-11) None  Grid-point #22 at 2.355e+20 	Fold #4 Rep #1 pred log like = 1248.83
#> Running at Laplace(9.21551e-11) None  Grid-point #22 at 2.355e+20 	Fold #5 Rep #1 pred log like = 1083.66
#> Running at Laplace(9.21551e-11) None  Grid-point #22 at 2.355e+20 	Fold #6 Rep #1 pred log like = 1084.42
#> Running at Laplace(9.21551e-11) None  Grid-point #22 at 2.355e+20 	Fold #7 Rep #1 pred log like = 1009.43
#> Running at Laplace(9.21551e-11) None  Grid-point #22 at 2.355e+20 	Fold #8 Rep #1 pred log like = 927.809
#> Running at Laplace(9.21551e-11) None  Grid-point #22 at 2.355e+20 	Fold #9 Rep #1 pred log like = 1097.98
#> Running at Laplace(9.21551e-11) None  Grid-point #22 at 2.355e+20 	Fold #10 Rep #1 pred log like = 1089.84
#> AvgPred = 1049.11 with stdev = 96.307
#> Completed at 2.355e+20
#> Next point at 5.19595e+12 with value 1049.11 and continue = 0
#> search[ 0.2355 ] = 1049.11(96.3088)
#> search[ 2.355 ] = 1049.11(96.3073)
#> search[ 23.55 ] = 1049.11(96.3071)
#> search[ 235.5 ] = 1049.11(96.307)
#> search[ 2355 ] = 1049.11(96.307)
#> search[ 23550 ] = 1049.11(96.307)
#> search[ 235500 ] = 1049.11(96.307)
#> search[ 2.355e+06 ] = 1049.11(96.307)
#> search[ 2.355e+07 ] = 1049.11(96.307)
#> search[ 2.355e+08 ] = 1049.11(96.307)
#> search[ 2.355e+09 ] = 1049.11(96.307)
#> search[ 2.355e+10 ] = 1049.11(96.307)
#> search[ 2.355e+11 ] = 1049.11(96.307)
#> search[ 2.355e+12 ] = 1049.11(96.307)
#> search[ 2.355e+13 ] = 1049.11(96.307)
#> search[ 2.355e+14 ] = 1049.11(96.307)
#> search[ 2.355e+15 ] = 1049.11(96.307)
#> search[ 2.355e+16 ] = 1049.11(96.307)
#> search[ 2.355e+17 ] = 1049.11(96.307)
#> search[ 2.355e+18 ] = 1049.11(96.307)
#> search[ 2.355e+19 ] = 1049.11(96.307)
#> search[ 2.355e+20 ] = 1049.11(96.307)
#> 
#> 
#> Maximum predicted log likelihood (1049.11) estimated at:
#> 	5.19595e+12 (variance)
#> 	6.20415e-07 (lambda)
#> 
#> Fitting model at optimal hyperparameter
#> Using prior: Laplace(6.20415e-07) None 

#Find out what the optimal hyperparameter was:
getHyperParameter(fit)
#> [1] 5.19595e+12

#Extract the current log-likelihood, and coefficients
logLik(fit)
#> 'log Lik.' -2307.763 (df=3)
coef(fit)
#> (Intercept)           1           2 
#> -3.54550102  0.31433174 -0.03279651 

#We can only retrieve the confidence interval for unregularized coefficients:
confint(fit, c(0))
#> Using 1 thread(s)
#>             covariate     2.5 %    97.5 % evaluations
#> (Intercept)         0 -3.573019 -3.518211          22