getSEs {Cyclops}	R Documentation

Extract standard errors

Description

getSEs extracts asymptotic standard errors for specific covariates from a Cyclops model fit object.

Usage

getSEs(object, covariates)

Arguments

object
A Cyclops model fit object

covariates
Integer or string vector: list of covariates for which asymptotic standard errors are wanted

Details

This function first computes the (partial) Fisher information matrix for just the requested covariates and then returns the square root of the diagonal elements of the inverse of the Fisher information matrix. These are the asymptotic standard errors when all possible covariates are included. When the requested covariates do not equate to all coefficients in the model, then interpretation is more challenging.

Value

Vector of standard error estimates

[Package Cyclops version 3.2.0 Index]