getUnivariableSeparability {Cyclops}	R Documentation

Get univariable linear separability

Description

getUnivariableSeparability reports covariates that are univariably separable with the outcome

Usage

getUnivariableSeparability(cyclopsData, covariates = NULL)

Arguments

cyclopsData
A Cyclops data object

covariates
Integer or string vector: list of covariates to report; default (NULL) implies all covariates

Value

A list of covariates that are univariably separable with the outcome

[Package Cyclops version 3.2.1 Index]