Extract log-likelihood

logLik returns the current log-likelihood of the fit in a Cyclops model fit object

# S3 method for cyclopsFit
logLik(object, ...)

Arguments

object
A Cyclops model fit object

...
Additional arguments

Examples

#Generate some simulated data:
sim <- simulateCyclopsData(nstrata = 1, nrows = 1000, ncovars = 2, eCovarsPerRow = 0.5,
                           model = "poisson")

#> Sparseness = 76.3 %

cyclopsData <- convertToCyclopsData(sim$outcomes, sim$covariates, modelType = "pr",
                                    addIntercept = TRUE)

#> Sorting covariates by covariateId and rowId

#Define the prior and control objects to use cross-validation for finding the 
#optimal hyperparameter:
prior <- createPrior("laplace", exclude = 0, useCrossValidation = TRUE)
control <- createControl(cvType = "auto", noiseLevel = "quiet")

#Fit the model
fit <- fitCyclopsModel(cyclopsData,prior = prior, control = control)

#> Using cross-validation selector type byRow
#> Performing 10-fold cross-validation [seed = 1591596886] with data partitions of sizes 100 100 100 100 100 100 100 100 100 100
#> Using 1 thread(s)
#> Starting var = 0.237 (default)
#> Running at Laplace(2.90496) None  Grid-point #1 at 0.237 	Fold #1 Rep #1 pred log like = 19.6151
#> Running at Laplace(2.90496) None  Grid-point #1 at 0.237 	Fold #2 Rep #1 pred log like = 49.6982
#> Running at Laplace(2.90496) None  Grid-point #1 at 0.237 	Fold #3 Rep #1 pred log like = 82.0621
#> Running at Laplace(2.90496) None  Grid-point #1 at 0.237 	Fold #4 Rep #1 pred log like = 75.7658
#> Running at Laplace(2.90496) None  Grid-point #1 at 0.237 	Fold #5 Rep #1 pred log like = 86.7556
#> Running at Laplace(2.90496) None  Grid-point #1 at 0.237 	Fold #6 Rep #1 pred log like = 19.6117
#> Running at Laplace(2.90496) None  Grid-point #1 at 0.237 	Fold #7 Rep #1 pred log like = 61.5565
#> Running at Laplace(2.90496) None  Grid-point #1 at 0.237 	Fold #8 Rep #1 pred log like = 74.3982
#> Running at Laplace(2.90496) None  Grid-point #1 at 0.237 	Fold #9 Rep #1 pred log like = 103.109
#> Running at Laplace(2.90496) None  Grid-point #1 at 0.237 	Fold #10 Rep #1 pred log like = 51.8211
#> AvgPred = 62.4393 with stdev = 26.2582
#> Completed at 0.237
#> Next point at 2.37 with value 0 and continue = 1
#> search[ 0.237 ] = 62.4393(26.2582)
#> 
#> Running at Laplace(0.91863) None  Grid-point #2 at 2.37 	Fold #1 Rep #1 pred log like = 19.6126
#> Running at Laplace(0.91863) None  Grid-point #2 at 2.37 	Fold #2 Rep #1 pred log like = 49.6607
#> Running at Laplace(0.91863) None  Grid-point #2 at 2.37 	Fold #3 Rep #1 pred log like = 82.0524
#> Running at Laplace(0.91863) None  Grid-point #2 at 2.37 	Fold #4 Rep #1 pred log like = 75.7476
#> Running at Laplace(0.91863) None  Grid-point #2 at 2.37 	Fold #5 Rep #1 pred log like = 86.7789
#> Running at Laplace(0.91863) None  Grid-point #2 at 2.37 	Fold #6 Rep #1 pred log like = 19.6625
#> Running at Laplace(0.91863) None  Grid-point #2 at 2.37 	Fold #7 Rep #1 pred log like = 61.5363
#> Running at Laplace(0.91863) None  Grid-point #2 at 2.37 	Fold #8 Rep #1 pred log like = 74.4037
#> Running at Laplace(0.91863) None  Grid-point #2 at 2.37 	Fold #9 Rep #1 pred log like = 103.155
#> Running at Laplace(0.91863) None  Grid-point #2 at 2.37 	Fold #10 Rep #1 pred log like = 51.8059
#> AvgPred = 62.4416 with stdev = 26.2607
#> Completed at 2.37
#> Next point at 23.7 with value 0 and continue = 1
#> search[ 0.237 ] = 62.4393(26.2582)
#> search[ 2.37 ] = 62.4416(26.2607)
#> 
#> Running at Laplace(0.290496) None  Grid-point #3 at 23.7 	Fold #1 Rep #1 pred log like = 19.6113
#> Running at Laplace(0.290496) None  Grid-point #3 at 23.7 	Fold #2 Rep #1 pred log like = 49.6481
#> Running at Laplace(0.290496) None  Grid-point #3 at 23.7 	Fold #3 Rep #1 pred log like = 82.049
#> Running at Laplace(0.290496) None  Grid-point #3 at 23.7 	Fold #4 Rep #1 pred log like = 75.7411
#> Running at Laplace(0.290496) None  Grid-point #3 at 23.7 	Fold #5 Rep #1 pred log like = 86.7858
#> Running at Laplace(0.290496) None  Grid-point #3 at 23.7 	Fold #6 Rep #1 pred log like = 19.6783
#> Running at Laplace(0.290496) None  Grid-point #3 at 23.7 	Fold #7 Rep #1 pred log like = 61.5294
#> Running at Laplace(0.290496) None  Grid-point #3 at 23.7 	Fold #8 Rep #1 pred log like = 74.405
#> Running at Laplace(0.290496) None  Grid-point #3 at 23.7 	Fold #9 Rep #1 pred log like = 103.169
#> Running at Laplace(0.290496) None  Grid-point #3 at 23.7 	Fold #10 Rep #1 pred log like = 51.8008
#> AvgPred = 62.4418 with stdev = 26.2615
#> Completed at 23.7
#> Next point at 237 with value 0 and continue = 1
#> search[ 0.237 ] = 62.4393(26.2582)
#> search[ 2.37 ] = 62.4416(26.2607)
#> search[ 23.7 ] = 62.4418(26.2615)
#> 
#> Running at Laplace(0.091863) None  Grid-point #4 at 237 	Fold #1 Rep #1 pred log like = 19.6109
#> Running at Laplace(0.091863) None  Grid-point #4 at 237 	Fold #2 Rep #1 pred log like = 49.644
#> Running at Laplace(0.091863) None  Grid-point #4 at 237 	Fold #3 Rep #1 pred log like = 82.0479
#> Running at Laplace(0.091863) None  Grid-point #4 at 237 	Fold #4 Rep #1 pred log like = 75.739
#> Running at Laplace(0.091863) None  Grid-point #4 at 237 	Fold #5 Rep #1 pred log like = 86.788
#> Running at Laplace(0.091863) None  Grid-point #4 at 237 	Fold #6 Rep #1 pred log like = 19.6833
#> Running at Laplace(0.091863) None  Grid-point #4 at 237 	Fold #7 Rep #1 pred log like = 61.5272
#> Running at Laplace(0.091863) None  Grid-point #4 at 237 	Fold #8 Rep #1 pred log like = 74.4054
#> Running at Laplace(0.091863) None  Grid-point #4 at 237 	Fold #9 Rep #1 pred log like = 103.173
#> Running at Laplace(0.091863) None  Grid-point #4 at 237 	Fold #10 Rep #1 pred log like = 51.7991
#> AvgPred = 62.4418 with stdev = 26.2618
#> Completed at 237
#> Next point at 2370 with value 0 and continue = 1
#> search[ 0.237 ] = 62.4393(26.2582)
#> search[ 2.37 ] = 62.4416(26.2607)
#> search[ 23.7 ] = 62.4418(26.2615)
#> search[ 237 ] = 62.4418(26.2618)
#> 
#> Running at Laplace(0.0290496) None  Grid-point #5 at 2370 	Fold #1 Rep #1 pred log like = 19.6108
#> Running at Laplace(0.0290496) None  Grid-point #5 at 2370 	Fold #2 Rep #1 pred log like = 49.6427
#> Running at Laplace(0.0290496) None  Grid-point #5 at 2370 	Fold #3 Rep #1 pred log like = 82.0475
#> Running at Laplace(0.0290496) None  Grid-point #5 at 2370 	Fold #4 Rep #1 pred log like = 75.7383
#> Running at Laplace(0.0290496) None  Grid-point #5 at 2370 	Fold #5 Rep #1 pred log like = 86.7887
#> Running at Laplace(0.0290496) None  Grid-point #5 at 2370 	Fold #6 Rep #1 pred log like = 19.6849
#> Running at Laplace(0.0290496) None  Grid-point #5 at 2370 	Fold #7 Rep #1 pred log like = 61.5265
#> Running at Laplace(0.0290496) None  Grid-point #5 at 2370 	Fold #8 Rep #1 pred log like = 74.4055
#> Running at Laplace(0.0290496) None  Grid-point #5 at 2370 	Fold #9 Rep #1 pred log like = 103.175
#> Running at Laplace(0.0290496) None  Grid-point #5 at 2370 	Fold #10 Rep #1 pred log like = 51.7986
#> AvgPred = 62.4418 with stdev = 26.2618
#> Completed at 2370
#> Next point at 23700 with value 0 and continue = 1
#> search[ 0.237 ] = 62.4393(26.2582)
#> search[ 2.37 ] = 62.4416(26.2607)
#> search[ 23.7 ] = 62.4418(26.2615)
#> search[ 237 ] = 62.4418(26.2618)
#> search[ 2370 ] = 62.4418(26.2618)
#> 
#> Running at Laplace(0.0091863) None  Grid-point #6 at 23700 	Fold #1 Rep #1 pred log like = 19.6107
#> Running at Laplace(0.0091863) None  Grid-point #6 at 23700 	Fold #2 Rep #1 pred log like = 49.6423
#> Running at Laplace(0.0091863) None  Grid-point #6 at 23700 	Fold #3 Rep #1 pred log like = 82.0474
#> Running at Laplace(0.0091863) None  Grid-point #6 at 23700 	Fold #4 Rep #1 pred log like = 75.7381
#> Running at Laplace(0.0091863) None  Grid-point #6 at 23700 	Fold #5 Rep #1 pred log like = 86.7889
#> Running at Laplace(0.0091863) None  Grid-point #6 at 23700 	Fold #6 Rep #1 pred log like = 19.6854
#> Running at Laplace(0.0091863) None  Grid-point #6 at 23700 	Fold #7 Rep #1 pred log like = 61.5263
#> Running at Laplace(0.0091863) None  Grid-point #6 at 23700 	Fold #8 Rep #1 pred log like = 74.4055
#> Running at Laplace(0.0091863) None  Grid-point #6 at 23700 	Fold #9 Rep #1 pred log like = 103.175
#> Running at Laplace(0.0091863) None  Grid-point #6 at 23700 	Fold #10 Rep #1 pred log like = 51.7985
#> AvgPred = 62.4418 with stdev = 26.2619
#> Completed at 23700
#> Next point at 237000 with value 0 and continue = 1
#> search[ 0.237 ] = 62.4393(26.2582)
#> search[ 2.37 ] = 62.4416(26.2607)
#> search[ 23.7 ] = 62.4418(26.2615)
#> search[ 237 ] = 62.4418(26.2618)
#> search[ 2370 ] = 62.4418(26.2618)
#> search[ 23700 ] = 62.4418(26.2619)
#> 
#> Running at Laplace(0.00290496) None  Grid-point #7 at 237000 	Fold #1 Rep #1 pred log like = 19.6107
#> Running at Laplace(0.00290496) None  Grid-point #7 at 237000 	Fold #2 Rep #1 pred log like = 49.6421
#> Running at Laplace(0.00290496) None  Grid-point #7 at 237000 	Fold #3 Rep #1 pred log like = 82.0474
#> Running at Laplace(0.00290496) None  Grid-point #7 at 237000 	Fold #4 Rep #1 pred log like = 75.7381
#> Running at Laplace(0.00290496) None  Grid-point #7 at 237000 	Fold #5 Rep #1 pred log like = 86.7889
#> Running at Laplace(0.00290496) None  Grid-point #7 at 237000 	Fold #6 Rep #1 pred log like = 19.6855
#> Running at Laplace(0.00290496) None  Grid-point #7 at 237000 	Fold #7 Rep #1 pred log like = 61.5262
#> Running at Laplace(0.00290496) None  Grid-point #7 at 237000 	Fold #8 Rep #1 pred log like = 74.4055
#> Running at Laplace(0.00290496) None  Grid-point #7 at 237000 	Fold #9 Rep #1 pred log like = 103.175
#> Running at Laplace(0.00290496) None  Grid-point #7 at 237000 	Fold #10 Rep #1 pred log like = 51.7984
#> AvgPred = 62.4418 with stdev = 26.2619
#> Completed at 237000
#> Next point at 2.37e+06 with value 0 and continue = 1
#> search[ 0.237 ] = 62.4393(26.2582)
#> search[ 2.37 ] = 62.4416(26.2607)
#> search[ 23.7 ] = 62.4418(26.2615)
#> search[ 237 ] = 62.4418(26.2618)
#> search[ 2370 ] = 62.4418(26.2618)
#> search[ 23700 ] = 62.4418(26.2619)
#> search[ 237000 ] = 62.4418(26.2619)
#> 
#> Running at Laplace(0.00091863) None  Grid-point #8 at 2.37e+06 	Fold #1 Rep #1 pred log like = 19.6107
#> Running at Laplace(0.00091863) None  Grid-point #8 at 2.37e+06 	Fold #2 Rep #1 pred log like = 49.6421
#> Running at Laplace(0.00091863) None  Grid-point #8 at 2.37e+06 	Fold #3 Rep #1 pred log like = 82.0474
#> Running at Laplace(0.00091863) None  Grid-point #8 at 2.37e+06 	Fold #4 Rep #1 pred log like = 75.738
#> Running at Laplace(0.00091863) None  Grid-point #8 at 2.37e+06 	Fold #5 Rep #1 pred log like = 86.789
#> Running at Laplace(0.00091863) None  Grid-point #8 at 2.37e+06 	Fold #6 Rep #1 pred log like = 19.6856
#> Running at Laplace(0.00091863) None  Grid-point #8 at 2.37e+06 	Fold #7 Rep #1 pred log like = 61.5262
#> Running at Laplace(0.00091863) None  Grid-point #8 at 2.37e+06 	Fold #8 Rep #1 pred log like = 74.4055
#> Running at Laplace(0.00091863) None  Grid-point #8 at 2.37e+06 	Fold #9 Rep #1 pred log like = 103.175
#> Running at Laplace(0.00091863) None  Grid-point #8 at 2.37e+06 	Fold #10 Rep #1 pred log like = 51.7984
#> AvgPred = 62.4418 with stdev = 26.2619
#> Completed at 2.37e+06
#> Next point at 2.37e+07 with value 0 and continue = 1
#> search[ 0.237 ] = 62.4393(26.2582)
#> search[ 2.37 ] = 62.4416(26.2607)
#> search[ 23.7 ] = 62.4418(26.2615)
#> search[ 237 ] = 62.4418(26.2618)
#> search[ 2370 ] = 62.4418(26.2618)
#> search[ 23700 ] = 62.4418(26.2619)
#> search[ 237000 ] = 62.4418(26.2619)
#> search[ 2.37e+06 ] = 62.4418(26.2619)
#> 
#> Running at Laplace(0.000290496) None  Grid-point #9 at 2.37e+07 	Fold #1 Rep #1 pred log like = 19.6107
#> Running at Laplace(0.000290496) None  Grid-point #9 at 2.37e+07 	Fold #2 Rep #1 pred log like = 49.6421
#> Running at Laplace(0.000290496) None  Grid-point #9 at 2.37e+07 	Fold #3 Rep #1 pred log like = 82.0473
#> Running at Laplace(0.000290496) None  Grid-point #9 at 2.37e+07 	Fold #4 Rep #1 pred log like = 75.738
#> Running at Laplace(0.000290496) None  Grid-point #9 at 2.37e+07 	Fold #5 Rep #1 pred log like = 86.789
#> Running at Laplace(0.000290496) None  Grid-point #9 at 2.37e+07 	Fold #6 Rep #1 pred log like = 19.6856
#> Running at Laplace(0.000290496) None  Grid-point #9 at 2.37e+07 	Fold #7 Rep #1 pred log like = 61.5262
#> Running at Laplace(0.000290496) None  Grid-point #9 at 2.37e+07 	Fold #8 Rep #1 pred log like = 74.4055
#> Running at Laplace(0.000290496) None  Grid-point #9 at 2.37e+07 	Fold #9 Rep #1 pred log like = 103.175
#> Running at Laplace(0.000290496) None  Grid-point #9 at 2.37e+07 	Fold #10 Rep #1 pred log like = 51.7984
#> AvgPred = 62.4418 with stdev = 26.2619
#> Completed at 2.37e+07
#> Next point at 2.37e+08 with value 0 and continue = 1
#> search[ 0.237 ] = 62.4393(26.2582)
#> search[ 2.37 ] = 62.4416(26.2607)
#> search[ 23.7 ] = 62.4418(26.2615)
#> search[ 237 ] = 62.4418(26.2618)
#> search[ 2370 ] = 62.4418(26.2618)
#> search[ 23700 ] = 62.4418(26.2619)
#> search[ 237000 ] = 62.4418(26.2619)
#> search[ 2.37e+06 ] = 62.4418(26.2619)
#> search[ 2.37e+07 ] = 62.4418(26.2619)
#> 
#> Running at Laplace(9.1863e-05) None  Grid-point #10 at 2.37e+08 	Fold #1 Rep #1 pred log like = 19.6107
#> Running at Laplace(9.1863e-05) None  Grid-point #10 at 2.37e+08 	Fold #2 Rep #1 pred log like = 49.6421
#> Running at Laplace(9.1863e-05) None  Grid-point #10 at 2.37e+08 	Fold #3 Rep #1 pred log like = 82.0473
#> Running at Laplace(9.1863e-05) None  Grid-point #10 at 2.37e+08 	Fold #4 Rep #1 pred log like = 75.738
#> Running at Laplace(9.1863e-05) None  Grid-point #10 at 2.37e+08 	Fold #5 Rep #1 pred log like = 86.789
#> Running at Laplace(9.1863e-05) None  Grid-point #10 at 2.37e+08 	Fold #6 Rep #1 pred log like = 19.6856
#> Running at Laplace(9.1863e-05) None  Grid-point #10 at 2.37e+08 	Fold #7 Rep #1 pred log like = 61.5262
#> Running at Laplace(9.1863e-05) None  Grid-point #10 at 2.37e+08 	Fold #8 Rep #1 pred log like = 74.4055
#> Running at Laplace(9.1863e-05) None  Grid-point #10 at 2.37e+08 	Fold #9 Rep #1 pred log like = 103.175
#> Running at Laplace(9.1863e-05) None  Grid-point #10 at 2.37e+08 	Fold #10 Rep #1 pred log like = 51.7984
#> AvgPred = 62.4418 with stdev = 26.2619
#> Completed at 2.37e+08
#> Next point at 2.37e+09 with value 0 and continue = 1
#> search[ 0.237 ] = 62.4393(26.2582)
#> search[ 2.37 ] = 62.4416(26.2607)
#> search[ 23.7 ] = 62.4418(26.2615)
#> search[ 237 ] = 62.4418(26.2618)
#> search[ 2370 ] = 62.4418(26.2618)
#> search[ 23700 ] = 62.4418(26.2619)
#> search[ 237000 ] = 62.4418(26.2619)
#> search[ 2.37e+06 ] = 62.4418(26.2619)
#> search[ 2.37e+07 ] = 62.4418(26.2619)
#> search[ 2.37e+08 ] = 62.4418(26.2619)
#> 
#> Running at Laplace(2.90496e-05) None  Grid-point #11 at 2.37e+09 	Fold #1 Rep #1 pred log like = 19.6107
#> Running at Laplace(2.90496e-05) None  Grid-point #11 at 2.37e+09 	Fold #2 Rep #1 pred log like = 49.6421
#> Running at Laplace(2.90496e-05) None  Grid-point #11 at 2.37e+09 	Fold #3 Rep #1 pred log like = 82.0473
#> Running at Laplace(2.90496e-05) None  Grid-point #11 at 2.37e+09 	Fold #4 Rep #1 pred log like = 75.738
#> Running at Laplace(2.90496e-05) None  Grid-point #11 at 2.37e+09 	Fold #5 Rep #1 pred log like = 86.789
#> Running at Laplace(2.90496e-05) None  Grid-point #11 at 2.37e+09 	Fold #6 Rep #1 pred log like = 19.6856
#> Running at Laplace(2.90496e-05) None  Grid-point #11 at 2.37e+09 	Fold #7 Rep #1 pred log like = 61.5262
#> Running at Laplace(2.90496e-05) None  Grid-point #11 at 2.37e+09 	Fold #8 Rep #1 pred log like = 74.4055
#> Running at Laplace(2.90496e-05) None  Grid-point #11 at 2.37e+09 	Fold #9 Rep #1 pred log like = 103.175
#> Running at Laplace(2.90496e-05) None  Grid-point #11 at 2.37e+09 	Fold #10 Rep #1 pred log like = 51.7984
#> AvgPred = 62.4418 with stdev = 26.2619
#> Completed at 2.37e+09
#> Next point at 2.37e+10 with value 0 and continue = 1
#> search[ 0.237 ] = 62.4393(26.2582)
#> search[ 2.37 ] = 62.4416(26.2607)
#> search[ 23.7 ] = 62.4418(26.2615)
#> search[ 237 ] = 62.4418(26.2618)
#> search[ 2370 ] = 62.4418(26.2618)
#> search[ 23700 ] = 62.4418(26.2619)
#> search[ 237000 ] = 62.4418(26.2619)
#> search[ 2.37e+06 ] = 62.4418(26.2619)
#> search[ 2.37e+07 ] = 62.4418(26.2619)
#> search[ 2.37e+08 ] = 62.4418(26.2619)
#> search[ 2.37e+09 ] = 62.4418(26.2619)
#> 
#> Running at Laplace(9.1863e-06) None  Grid-point #12 at 2.37e+10 	Fold #1 Rep #1 pred log like = 19.6107
#> Running at Laplace(9.1863e-06) None  Grid-point #12 at 2.37e+10 	Fold #2 Rep #1 pred log like = 49.6421
#> Running at Laplace(9.1863e-06) None  Grid-point #12 at 2.37e+10 	Fold #3 Rep #1 pred log like = 82.0473
#> Running at Laplace(9.1863e-06) None  Grid-point #12 at 2.37e+10 	Fold #4 Rep #1 pred log like = 75.738
#> Running at Laplace(9.1863e-06) None  Grid-point #12 at 2.37e+10 	Fold #5 Rep #1 pred log like = 86.789
#> Running at Laplace(9.1863e-06) None  Grid-point #12 at 2.37e+10 	Fold #6 Rep #1 pred log like = 19.6856
#> Running at Laplace(9.1863e-06) None  Grid-point #12 at 2.37e+10 	Fold #7 Rep #1 pred log like = 61.5262
#> Running at Laplace(9.1863e-06) None  Grid-point #12 at 2.37e+10 	Fold #8 Rep #1 pred log like = 74.4055
#> Running at Laplace(9.1863e-06) None  Grid-point #12 at 2.37e+10 	Fold #9 Rep #1 pred log like = 103.175
#> Running at Laplace(9.1863e-06) None  Grid-point #12 at 2.37e+10 	Fold #10 Rep #1 pred log like = 51.7984
#> AvgPred = 62.4418 with stdev = 26.2619
#> Completed at 2.37e+10
#> Next point at 2.37e+11 with value 0 and continue = 1
#> search[ 0.237 ] = 62.4393(26.2582)
#> search[ 2.37 ] = 62.4416(26.2607)
#> search[ 23.7 ] = 62.4418(26.2615)
#> search[ 237 ] = 62.4418(26.2618)
#> search[ 2370 ] = 62.4418(26.2618)
#> search[ 23700 ] = 62.4418(26.2619)
#> search[ 237000 ] = 62.4418(26.2619)
#> search[ 2.37e+06 ] = 62.4418(26.2619)
#> search[ 2.37e+07 ] = 62.4418(26.2619)
#> search[ 2.37e+08 ] = 62.4418(26.2619)
#> search[ 2.37e+09 ] = 62.4418(26.2619)
#> search[ 2.37e+10 ] = 62.4418(26.2619)
#> 
#> Running at Laplace(2.90496e-06) None  Grid-point #13 at 2.37e+11 	Fold #1 Rep #1 pred log like = 19.6107
#> Running at Laplace(2.90496e-06) None  Grid-point #13 at 2.37e+11 	Fold #2 Rep #1 pred log like = 49.6421
#> Running at Laplace(2.90496e-06) None  Grid-point #13 at 2.37e+11 	Fold #3 Rep #1 pred log like = 82.0473
#> Running at Laplace(2.90496e-06) None  Grid-point #13 at 2.37e+11 	Fold #4 Rep #1 pred log like = 75.738
#> Running at Laplace(2.90496e-06) None  Grid-point #13 at 2.37e+11 	Fold #5 Rep #1 pred log like = 86.789
#> Running at Laplace(2.90496e-06) None  Grid-point #13 at 2.37e+11 	Fold #6 Rep #1 pred log like = 19.6856
#> Running at Laplace(2.90496e-06) None  Grid-point #13 at 2.37e+11 	Fold #7 Rep #1 pred log like = 61.5262
#> Running at Laplace(2.90496e-06) None  Grid-point #13 at 2.37e+11 	Fold #8 Rep #1 pred log like = 74.4055
#> Running at Laplace(2.90496e-06) None  Grid-point #13 at 2.37e+11 	Fold #9 Rep #1 pred log like = 103.175
#> Running at Laplace(2.90496e-06) None  Grid-point #13 at 2.37e+11 	Fold #10 Rep #1 pred log like = 51.7984
#> AvgPred = 62.4418 with stdev = 26.2619
#> Completed at 2.37e+11
#> Next point at 2.37e+12 with value 0 and continue = 1
#> search[ 0.237 ] = 62.4393(26.2582)
#> search[ 2.37 ] = 62.4416(26.2607)
#> search[ 23.7 ] = 62.4418(26.2615)
#> search[ 237 ] = 62.4418(26.2618)
#> search[ 2370 ] = 62.4418(26.2618)
#> search[ 23700 ] = 62.4418(26.2619)
#> search[ 237000 ] = 62.4418(26.2619)
#> search[ 2.37e+06 ] = 62.4418(26.2619)
#> search[ 2.37e+07 ] = 62.4418(26.2619)
#> search[ 2.37e+08 ] = 62.4418(26.2619)
#> search[ 2.37e+09 ] = 62.4418(26.2619)
#> search[ 2.37e+10 ] = 62.4418(26.2619)
#> search[ 2.37e+11 ] = 62.4418(26.2619)
#> 
#> Running at Laplace(9.1863e-07) None  Grid-point #14 at 2.37e+12 	Fold #1 Rep #1 pred log like = 19.6107
#> Running at Laplace(9.1863e-07) None  Grid-point #14 at 2.37e+12 	Fold #2 Rep #1 pred log like = 49.6421
#> Running at Laplace(9.1863e-07) None  Grid-point #14 at 2.37e+12 	Fold #3 Rep #1 pred log like = 82.0473
#> Running at Laplace(9.1863e-07) None  Grid-point #14 at 2.37e+12 	Fold #4 Rep #1 pred log like = 75.738
#> Running at Laplace(9.1863e-07) None  Grid-point #14 at 2.37e+12 	Fold #5 Rep #1 pred log like = 86.789
#> Running at Laplace(9.1863e-07) None  Grid-point #14 at 2.37e+12 	Fold #6 Rep #1 pred log like = 19.6856
#> Running at Laplace(9.1863e-07) None  Grid-point #14 at 2.37e+12 	Fold #7 Rep #1 pred log like = 61.5262
#> Running at Laplace(9.1863e-07) None  Grid-point #14 at 2.37e+12 	Fold #8 Rep #1 pred log like = 74.4055
#> Running at Laplace(9.1863e-07) None  Grid-point #14 at 2.37e+12 	Fold #9 Rep #1 pred log like = 103.175
#> Running at Laplace(9.1863e-07) None  Grid-point #14 at 2.37e+12 	Fold #10 Rep #1 pred log like = 51.7984
#> AvgPred = 62.4418 with stdev = 26.2619
#> Completed at 2.37e+12
#> Next point at 2.37e+13 with value 0 and continue = 1
#> search[ 0.237 ] = 62.4393(26.2582)
#> search[ 2.37 ] = 62.4416(26.2607)
#> search[ 23.7 ] = 62.4418(26.2615)
#> search[ 237 ] = 62.4418(26.2618)
#> search[ 2370 ] = 62.4418(26.2618)
#> search[ 23700 ] = 62.4418(26.2619)
#> search[ 237000 ] = 62.4418(26.2619)
#> search[ 2.37e+06 ] = 62.4418(26.2619)
#> search[ 2.37e+07 ] = 62.4418(26.2619)
#> search[ 2.37e+08 ] = 62.4418(26.2619)
#> search[ 2.37e+09 ] = 62.4418(26.2619)
#> search[ 2.37e+10 ] = 62.4418(26.2619)
#> search[ 2.37e+11 ] = 62.4418(26.2619)
#> search[ 2.37e+12 ] = 62.4418(26.2619)
#> 
#> Running at Laplace(2.90496e-07) None  Grid-point #15 at 2.37e+13 	Fold #1 Rep #1 pred log like = 19.6107
#> Running at Laplace(2.90496e-07) None  Grid-point #15 at 2.37e+13 	Fold #2 Rep #1 pred log like = 49.6421
#> Running at Laplace(2.90496e-07) None  Grid-point #15 at 2.37e+13 	Fold #3 Rep #1 pred log like = 82.0473
#> Running at Laplace(2.90496e-07) None  Grid-point #15 at 2.37e+13 	Fold #4 Rep #1 pred log like = 75.738
#> Running at Laplace(2.90496e-07) None  Grid-point #15 at 2.37e+13 	Fold #5 Rep #1 pred log like = 86.789
#> Running at Laplace(2.90496e-07) None  Grid-point #15 at 2.37e+13 	Fold #6 Rep #1 pred log like = 19.6856
#> Running at Laplace(2.90496e-07) None  Grid-point #15 at 2.37e+13 	Fold #7 Rep #1 pred log like = 61.5262
#> Running at Laplace(2.90496e-07) None  Grid-point #15 at 2.37e+13 	Fold #8 Rep #1 pred log like = 74.4055
#> Running at Laplace(2.90496e-07) None  Grid-point #15 at 2.37e+13 	Fold #9 Rep #1 pred log like = 103.175
#> Running at Laplace(2.90496e-07) None  Grid-point #15 at 2.37e+13 	Fold #10 Rep #1 pred log like = 51.7984
#> AvgPred = 62.4418 with stdev = 26.2619
#> Completed at 2.37e+13
#> Next point at 2.37e+14 with value 0 and continue = 1
#> search[ 0.237 ] = 62.4393(26.2582)
#> search[ 2.37 ] = 62.4416(26.2607)
#> search[ 23.7 ] = 62.4418(26.2615)
#> search[ 237 ] = 62.4418(26.2618)
#> search[ 2370 ] = 62.4418(26.2618)
#> search[ 23700 ] = 62.4418(26.2619)
#> search[ 237000 ] = 62.4418(26.2619)
#> search[ 2.37e+06 ] = 62.4418(26.2619)
#> search[ 2.37e+07 ] = 62.4418(26.2619)
#> search[ 2.37e+08 ] = 62.4418(26.2619)
#> search[ 2.37e+09 ] = 62.4418(26.2619)
#> search[ 2.37e+10 ] = 62.4418(26.2619)
#> search[ 2.37e+11 ] = 62.4418(26.2619)
#> search[ 2.37e+12 ] = 62.4418(26.2619)
#> search[ 2.37e+13 ] = 62.4418(26.2619)
#> 
#> Running at Laplace(9.1863e-08) None  Grid-point #16 at 2.37e+14 	Fold #1 Rep #1 pred log like = 19.6107
#> Running at Laplace(9.1863e-08) None  Grid-point #16 at 2.37e+14 	Fold #2 Rep #1 pred log like = 49.6421
#> Running at Laplace(9.1863e-08) None  Grid-point #16 at 2.37e+14 	Fold #3 Rep #1 pred log like = 82.0473
#> Running at Laplace(9.1863e-08) None  Grid-point #16 at 2.37e+14 	Fold #4 Rep #1 pred log like = 75.738
#> Running at Laplace(9.1863e-08) None  Grid-point #16 at 2.37e+14 	Fold #5 Rep #1 pred log like = 86.789
#> Running at Laplace(9.1863e-08) None  Grid-point #16 at 2.37e+14 	Fold #6 Rep #1 pred log like = 19.6856
#> Running at Laplace(9.1863e-08) None  Grid-point #16 at 2.37e+14 	Fold #7 Rep #1 pred log like = 61.5262
#> Running at Laplace(9.1863e-08) None  Grid-point #16 at 2.37e+14 	Fold #8 Rep #1 pred log like = 74.4055
#> Running at Laplace(9.1863e-08) None  Grid-point #16 at 2.37e+14 	Fold #9 Rep #1 pred log like = 103.175
#> Running at Laplace(9.1863e-08) None  Grid-point #16 at 2.37e+14 	Fold #10 Rep #1 pred log like = 51.7984
#> AvgPred = 62.4418 with stdev = 26.2619
#> Completed at 2.37e+14
#> Next point at 2.37e+15 with value 0 and continue = 1
#> search[ 0.237 ] = 62.4393(26.2582)
#> search[ 2.37 ] = 62.4416(26.2607)
#> search[ 23.7 ] = 62.4418(26.2615)
#> search[ 237 ] = 62.4418(26.2618)
#> search[ 2370 ] = 62.4418(26.2618)
#> search[ 23700 ] = 62.4418(26.2619)
#> search[ 237000 ] = 62.4418(26.2619)
#> search[ 2.37e+06 ] = 62.4418(26.2619)
#> search[ 2.37e+07 ] = 62.4418(26.2619)
#> search[ 2.37e+08 ] = 62.4418(26.2619)
#> search[ 2.37e+09 ] = 62.4418(26.2619)
#> search[ 2.37e+10 ] = 62.4418(26.2619)
#> search[ 2.37e+11 ] = 62.4418(26.2619)
#> search[ 2.37e+12 ] = 62.4418(26.2619)
#> search[ 2.37e+13 ] = 62.4418(26.2619)
#> search[ 2.37e+14 ] = 62.4418(26.2619)
#> 
#> Running at Laplace(2.90496e-08) None  Grid-point #17 at 2.37e+15 	Fold #1 Rep #1 pred log like = 19.6107
#> Running at Laplace(2.90496e-08) None  Grid-point #17 at 2.37e+15 	Fold #2 Rep #1 pred log like = 49.6421
#> Running at Laplace(2.90496e-08) None  Grid-point #17 at 2.37e+15 	Fold #3 Rep #1 pred log like = 82.0473
#> Running at Laplace(2.90496e-08) None  Grid-point #17 at 2.37e+15 	Fold #4 Rep #1 pred log like = 75.738
#> Running at Laplace(2.90496e-08) None  Grid-point #17 at 2.37e+15 	Fold #5 Rep #1 pred log like = 86.789
#> Running at Laplace(2.90496e-08) None  Grid-point #17 at 2.37e+15 	Fold #6 Rep #1 pred log like = 19.6856
#> Running at Laplace(2.90496e-08) None  Grid-point #17 at 2.37e+15 	Fold #7 Rep #1 pred log like = 61.5262
#> Running at Laplace(2.90496e-08) None  Grid-point #17 at 2.37e+15 	Fold #8 Rep #1 pred log like = 74.4055
#> Running at Laplace(2.90496e-08) None  Grid-point #17 at 2.37e+15 	Fold #9 Rep #1 pred log like = 103.175
#> Running at Laplace(2.90496e-08) None  Grid-point #17 at 2.37e+15 	Fold #10 Rep #1 pred log like = 51.7984
#> AvgPred = 62.4418 with stdev = 26.2619
#> Completed at 2.37e+15
#> Next point at 2.37e+16 with value 0 and continue = 1
#> search[ 0.237 ] = 62.4393(26.2582)
#> search[ 2.37 ] = 62.4416(26.2607)
#> search[ 23.7 ] = 62.4418(26.2615)
#> search[ 237 ] = 62.4418(26.2618)
#> search[ 2370 ] = 62.4418(26.2618)
#> search[ 23700 ] = 62.4418(26.2619)
#> search[ 237000 ] = 62.4418(26.2619)
#> search[ 2.37e+06 ] = 62.4418(26.2619)
#> search[ 2.37e+07 ] = 62.4418(26.2619)
#> search[ 2.37e+08 ] = 62.4418(26.2619)
#> search[ 2.37e+09 ] = 62.4418(26.2619)
#> search[ 2.37e+10 ] = 62.4418(26.2619)
#> search[ 2.37e+11 ] = 62.4418(26.2619)
#> search[ 2.37e+12 ] = 62.4418(26.2619)
#> search[ 2.37e+13 ] = 62.4418(26.2619)
#> search[ 2.37e+14 ] = 62.4418(26.2619)
#> search[ 2.37e+15 ] = 62.4418(26.2619)
#> 
#> Running at Laplace(9.1863e-09) None  Grid-point #18 at 2.37e+16 	Fold #1 Rep #1 pred log like = 19.6107
#> Running at Laplace(9.1863e-09) None  Grid-point #18 at 2.37e+16 	Fold #2 Rep #1 pred log like = 49.6421
#> Running at Laplace(9.1863e-09) None  Grid-point #18 at 2.37e+16 	Fold #3 Rep #1 pred log like = 82.0473
#> Running at Laplace(9.1863e-09) None  Grid-point #18 at 2.37e+16 	Fold #4 Rep #1 pred log like = 75.738
#> Running at Laplace(9.1863e-09) None  Grid-point #18 at 2.37e+16 	Fold #5 Rep #1 pred log like = 86.789
#> Running at Laplace(9.1863e-09) None  Grid-point #18 at 2.37e+16 	Fold #6 Rep #1 pred log like = 19.6856
#> Running at Laplace(9.1863e-09) None  Grid-point #18 at 2.37e+16 	Fold #7 Rep #1 pred log like = 61.5262
#> Running at Laplace(9.1863e-09) None  Grid-point #18 at 2.37e+16 	Fold #8 Rep #1 pred log like = 74.4055
#> Running at Laplace(9.1863e-09) None  Grid-point #18 at 2.37e+16 	Fold #9 Rep #1 pred log like = 103.175
#> Running at Laplace(9.1863e-09) None  Grid-point #18 at 2.37e+16 	Fold #10 Rep #1 pred log like = 51.7984
#> AvgPred = 62.4418 with stdev = 26.2619
#> Completed at 2.37e+16
#> Next point at 2.37e+17 with value 0 and continue = 1
#> search[ 0.237 ] = 62.4393(26.2582)
#> search[ 2.37 ] = 62.4416(26.2607)
#> search[ 23.7 ] = 62.4418(26.2615)
#> search[ 237 ] = 62.4418(26.2618)
#> search[ 2370 ] = 62.4418(26.2618)
#> search[ 23700 ] = 62.4418(26.2619)
#> search[ 237000 ] = 62.4418(26.2619)
#> search[ 2.37e+06 ] = 62.4418(26.2619)
#> search[ 2.37e+07 ] = 62.4418(26.2619)
#> search[ 2.37e+08 ] = 62.4418(26.2619)
#> search[ 2.37e+09 ] = 62.4418(26.2619)
#> search[ 2.37e+10 ] = 62.4418(26.2619)
#> search[ 2.37e+11 ] = 62.4418(26.2619)
#> search[ 2.37e+12 ] = 62.4418(26.2619)
#> search[ 2.37e+13 ] = 62.4418(26.2619)
#> search[ 2.37e+14 ] = 62.4418(26.2619)
#> search[ 2.37e+15 ] = 62.4418(26.2619)
#> search[ 2.37e+16 ] = 62.4418(26.2619)
#> 
#> Running at Laplace(2.90496e-09) None  Grid-point #19 at 2.37e+17 	Fold #1 Rep #1 pred log like = 19.6107
#> Running at Laplace(2.90496e-09) None  Grid-point #19 at 2.37e+17 	Fold #2 Rep #1 pred log like = 49.6421
#> Running at Laplace(2.90496e-09) None  Grid-point #19 at 2.37e+17 	Fold #3 Rep #1 pred log like = 82.0473
#> Running at Laplace(2.90496e-09) None  Grid-point #19 at 2.37e+17 	Fold #4 Rep #1 pred log like = 75.738
#> Running at Laplace(2.90496e-09) None  Grid-point #19 at 2.37e+17 	Fold #5 Rep #1 pred log like = 86.789
#> Running at Laplace(2.90496e-09) None  Grid-point #19 at 2.37e+17 	Fold #6 Rep #1 pred log like = 19.6856
#> Running at Laplace(2.90496e-09) None  Grid-point #19 at 2.37e+17 	Fold #7 Rep #1 pred log like = 61.5262
#> Running at Laplace(2.90496e-09) None  Grid-point #19 at 2.37e+17 	Fold #8 Rep #1 pred log like = 74.4055
#> Running at Laplace(2.90496e-09) None  Grid-point #19 at 2.37e+17 	Fold #9 Rep #1 pred log like = 103.175
#> Running at Laplace(2.90496e-09) None  Grid-point #19 at 2.37e+17 	Fold #10 Rep #1 pred log like = 51.7984
#> AvgPred = 62.4418 with stdev = 26.2619
#> Completed at 2.37e+17
#> Next point at 2.37e+18 with value 0 and continue = 1
#> search[ 0.237 ] = 62.4393(26.2582)
#> search[ 2.37 ] = 62.4416(26.2607)
#> search[ 23.7 ] = 62.4418(26.2615)
#> search[ 237 ] = 62.4418(26.2618)
#> search[ 2370 ] = 62.4418(26.2618)
#> search[ 23700 ] = 62.4418(26.2619)
#> search[ 237000 ] = 62.4418(26.2619)
#> search[ 2.37e+06 ] = 62.4418(26.2619)
#> search[ 2.37e+07 ] = 62.4418(26.2619)
#> search[ 2.37e+08 ] = 62.4418(26.2619)
#> search[ 2.37e+09 ] = 62.4418(26.2619)
#> search[ 2.37e+10 ] = 62.4418(26.2619)
#> search[ 2.37e+11 ] = 62.4418(26.2619)
#> search[ 2.37e+12 ] = 62.4418(26.2619)
#> search[ 2.37e+13 ] = 62.4418(26.2619)
#> search[ 2.37e+14 ] = 62.4418(26.2619)
#> search[ 2.37e+15 ] = 62.4418(26.2619)
#> search[ 2.37e+16 ] = 62.4418(26.2619)
#> search[ 2.37e+17 ] = 62.4418(26.2619)
#> 
#> Running at Laplace(9.1863e-10) None  Grid-point #20 at 2.37e+18 	Fold #1 Rep #1 pred log like = 19.6107
#> Running at Laplace(9.1863e-10) None  Grid-point #20 at 2.37e+18 	Fold #2 Rep #1 pred log like = 49.6421
#> Running at Laplace(9.1863e-10) None  Grid-point #20 at 2.37e+18 	Fold #3 Rep #1 pred log like = 82.0473
#> Running at Laplace(9.1863e-10) None  Grid-point #20 at 2.37e+18 	Fold #4 Rep #1 pred log like = 75.738
#> Running at Laplace(9.1863e-10) None  Grid-point #20 at 2.37e+18 	Fold #5 Rep #1 pred log like = 86.789
#> Running at Laplace(9.1863e-10) None  Grid-point #20 at 2.37e+18 	Fold #6 Rep #1 pred log like = 19.6856
#> Running at Laplace(9.1863e-10) None  Grid-point #20 at 2.37e+18 	Fold #7 Rep #1 pred log like = 61.5262
#> Running at Laplace(9.1863e-10) None  Grid-point #20 at 2.37e+18 	Fold #8 Rep #1 pred log like = 74.4055
#> Running at Laplace(9.1863e-10) None  Grid-point #20 at 2.37e+18 	Fold #9 Rep #1 pred log like = 103.175
#> Running at Laplace(9.1863e-10) None  Grid-point #20 at 2.37e+18 	Fold #10 Rep #1 pred log like = 51.7984
#> AvgPred = 62.4418 with stdev = 26.2619
#> Completed at 2.37e+18
#> Next point at 2.37e+19 with value 0 and continue = 1
#> search[ 0.237 ] = 62.4393(26.2582)
#> search[ 2.37 ] = 62.4416(26.2607)
#> search[ 23.7 ] = 62.4418(26.2615)
#> search[ 237 ] = 62.4418(26.2618)
#> search[ 2370 ] = 62.4418(26.2618)
#> search[ 23700 ] = 62.4418(26.2619)
#> search[ 237000 ] = 62.4418(26.2619)
#> search[ 2.37e+06 ] = 62.4418(26.2619)
#> search[ 2.37e+07 ] = 62.4418(26.2619)
#> search[ 2.37e+08 ] = 62.4418(26.2619)
#> search[ 2.37e+09 ] = 62.4418(26.2619)
#> search[ 2.37e+10 ] = 62.4418(26.2619)
#> search[ 2.37e+11 ] = 62.4418(26.2619)
#> search[ 2.37e+12 ] = 62.4418(26.2619)
#> search[ 2.37e+13 ] = 62.4418(26.2619)
#> search[ 2.37e+14 ] = 62.4418(26.2619)
#> search[ 2.37e+15 ] = 62.4418(26.2619)
#> search[ 2.37e+16 ] = 62.4418(26.2619)
#> search[ 2.37e+17 ] = 62.4418(26.2619)
#> search[ 2.37e+18 ] = 62.4418(26.2619)
#> 
#> Running at Laplace(2.90496e-10) None  Grid-point #21 at 2.37e+19 	Fold #1 Rep #1 pred log like = 19.6107
#> Running at Laplace(2.90496e-10) None  Grid-point #21 at 2.37e+19 	Fold #2 Rep #1 pred log like = 49.6421
#> Running at Laplace(2.90496e-10) None  Grid-point #21 at 2.37e+19 	Fold #3 Rep #1 pred log like = 82.0473
#> Running at Laplace(2.90496e-10) None  Grid-point #21 at 2.37e+19 	Fold #4 Rep #1 pred log like = 75.738
#> Running at Laplace(2.90496e-10) None  Grid-point #21 at 2.37e+19 	Fold #5 Rep #1 pred log like = 86.789
#> Running at Laplace(2.90496e-10) None  Grid-point #21 at 2.37e+19 	Fold #6 Rep #1 pred log like = 19.6856
#> Running at Laplace(2.90496e-10) None  Grid-point #21 at 2.37e+19 	Fold #7 Rep #1 pred log like = 61.5262
#> Running at Laplace(2.90496e-10) None  Grid-point #21 at 2.37e+19 	Fold #8 Rep #1 pred log like = 74.4055
#> Running at Laplace(2.90496e-10) None  Grid-point #21 at 2.37e+19 	Fold #9 Rep #1 pred log like = 103.175
#> Running at Laplace(2.90496e-10) None  Grid-point #21 at 2.37e+19 	Fold #10 Rep #1 pred log like = 51.7984
#> AvgPred = 62.4418 with stdev = 26.2619
#> Completed at 2.37e+19
#> Next point at 2.37e+20 with value 0 and continue = 1
#> search[ 0.237 ] = 62.4393(26.2582)
#> search[ 2.37 ] = 62.4416(26.2607)
#> search[ 23.7 ] = 62.4418(26.2615)
#> search[ 237 ] = 62.4418(26.2618)
#> search[ 2370 ] = 62.4418(26.2618)
#> search[ 23700 ] = 62.4418(26.2619)
#> search[ 237000 ] = 62.4418(26.2619)
#> search[ 2.37e+06 ] = 62.4418(26.2619)
#> search[ 2.37e+07 ] = 62.4418(26.2619)
#> search[ 2.37e+08 ] = 62.4418(26.2619)
#> search[ 2.37e+09 ] = 62.4418(26.2619)
#> search[ 2.37e+10 ] = 62.4418(26.2619)
#> search[ 2.37e+11 ] = 62.4418(26.2619)
#> search[ 2.37e+12 ] = 62.4418(26.2619)
#> search[ 2.37e+13 ] = 62.4418(26.2619)
#> search[ 2.37e+14 ] = 62.4418(26.2619)
#> search[ 2.37e+15 ] = 62.4418(26.2619)
#> search[ 2.37e+16 ] = 62.4418(26.2619)
#> search[ 2.37e+17 ] = 62.4418(26.2619)
#> search[ 2.37e+18 ] = 62.4418(26.2619)
#> search[ 2.37e+19 ] = 62.4418(26.2619)
#> 
#> Running at Laplace(9.1863e-11) None  Grid-point #22 at 2.37e+20 	Fold #1 Rep #1 pred log like = 19.6107
#> Running at Laplace(9.1863e-11) None  Grid-point #22 at 2.37e+20 	Fold #2 Rep #1 pred log like = 49.6421
#> Running at Laplace(9.1863e-11) None  Grid-point #22 at 2.37e+20 	Fold #3 Rep #1 pred log like = 82.0473
#> Running at Laplace(9.1863e-11) None  Grid-point #22 at 2.37e+20 	Fold #4 Rep #1 pred log like = 75.738
#> Running at Laplace(9.1863e-11) None  Grid-point #22 at 2.37e+20 	Fold #5 Rep #1 pred log like = 86.789
#> Running at Laplace(9.1863e-11) None  Grid-point #22 at 2.37e+20 	Fold #6 Rep #1 pred log like = 19.6856
#> Running at Laplace(9.1863e-11) None  Grid-point #22 at 2.37e+20 	Fold #7 Rep #1 pred log like = 61.5262
#> Running at Laplace(9.1863e-11) None  Grid-point #22 at 2.37e+20 	Fold #8 Rep #1 pred log like = 74.4055
#> Running at Laplace(9.1863e-11) None  Grid-point #22 at 2.37e+20 	Fold #9 Rep #1 pred log like = 103.175
#> Running at Laplace(9.1863e-11) None  Grid-point #22 at 2.37e+20 	Fold #10 Rep #1 pred log like = 51.7984
#> AvgPred = 62.4418 with stdev = 26.2619
#> Completed at 2.37e+20
#> Next point at 2.37e+21 with value 0 and continue = 1
#> search[ 0.237 ] = 62.4393(26.2582)
#> search[ 2.37 ] = 62.4416(26.2607)
#> search[ 23.7 ] = 62.4418(26.2615)
#> search[ 237 ] = 62.4418(26.2618)
#> search[ 2370 ] = 62.4418(26.2618)
#> search[ 23700 ] = 62.4418(26.2619)
#> search[ 237000 ] = 62.4418(26.2619)
#> search[ 2.37e+06 ] = 62.4418(26.2619)
#> search[ 2.37e+07 ] = 62.4418(26.2619)
#> search[ 2.37e+08 ] = 62.4418(26.2619)
#> search[ 2.37e+09 ] = 62.4418(26.2619)
#> search[ 2.37e+10 ] = 62.4418(26.2619)
#> search[ 2.37e+11 ] = 62.4418(26.2619)
#> search[ 2.37e+12 ] = 62.4418(26.2619)
#> search[ 2.37e+13 ] = 62.4418(26.2619)
#> search[ 2.37e+14 ] = 62.4418(26.2619)
#> search[ 2.37e+15 ] = 62.4418(26.2619)
#> search[ 2.37e+16 ] = 62.4418(26.2619)
#> search[ 2.37e+17 ] = 62.4418(26.2619)
#> search[ 2.37e+18 ] = 62.4418(26.2619)
#> search[ 2.37e+19 ] = 62.4418(26.2619)
#> search[ 2.37e+20 ] = 62.4418(26.2619)
#> 
#> Running at Laplace(2.90496e-11) None  Grid-point #23 at 2.37e+21 	Fold #1 Rep #1 pred log like = 19.6107
#> Running at Laplace(2.90496e-11) None  Grid-point #23 at 2.37e+21 	Fold #2 Rep #1 pred log like = 49.6421
#> Running at Laplace(2.90496e-11) None  Grid-point #23 at 2.37e+21 	Fold #3 Rep #1 pred log like = 82.0473
#> Running at Laplace(2.90496e-11) None  Grid-point #23 at 2.37e+21 	Fold #4 Rep #1 pred log like = 75.738
#> Running at Laplace(2.90496e-11) None  Grid-point #23 at 2.37e+21 	Fold #5 Rep #1 pred log like = 86.789
#> Running at Laplace(2.90496e-11) None  Grid-point #23 at 2.37e+21 	Fold #6 Rep #1 pred log like = 19.6856
#> Running at Laplace(2.90496e-11) None  Grid-point #23 at 2.37e+21 	Fold #7 Rep #1 pred log like = 61.5262
#> Running at Laplace(2.90496e-11) None  Grid-point #23 at 2.37e+21 	Fold #8 Rep #1 pred log like = 74.4055
#> Running at Laplace(2.90496e-11) None  Grid-point #23 at 2.37e+21 	Fold #9 Rep #1 pred log like = 103.175
#> Running at Laplace(2.90496e-11) None  Grid-point #23 at 2.37e+21 	Fold #10 Rep #1 pred log like = 51.7984
#> AvgPred = 62.4418 with stdev = 26.2619
#> Completed at 2.37e+21
#> Next point at 2.37e+22 with value 0 and continue = 1
#> search[ 0.237 ] = 62.4393(26.2582)
#> search[ 2.37 ] = 62.4416(26.2607)
#> search[ 23.7 ] = 62.4418(26.2615)
#> search[ 237 ] = 62.4418(26.2618)
#> search[ 2370 ] = 62.4418(26.2618)
#> search[ 23700 ] = 62.4418(26.2619)
#> search[ 237000 ] = 62.4418(26.2619)
#> search[ 2.37e+06 ] = 62.4418(26.2619)
#> search[ 2.37e+07 ] = 62.4418(26.2619)
#> search[ 2.37e+08 ] = 62.4418(26.2619)
#> search[ 2.37e+09 ] = 62.4418(26.2619)
#> search[ 2.37e+10 ] = 62.4418(26.2619)
#> search[ 2.37e+11 ] = 62.4418(26.2619)
#> search[ 2.37e+12 ] = 62.4418(26.2619)
#> search[ 2.37e+13 ] = 62.4418(26.2619)
#> search[ 2.37e+14 ] = 62.4418(26.2619)
#> search[ 2.37e+15 ] = 62.4418(26.2619)
#> search[ 2.37e+16 ] = 62.4418(26.2619)
#> search[ 2.37e+17 ] = 62.4418(26.2619)
#> search[ 2.37e+18 ] = 62.4418(26.2619)
#> search[ 2.37e+19 ] = 62.4418(26.2619)
#> search[ 2.37e+20 ] = 62.4418(26.2619)
#> search[ 2.37e+21 ] = 62.4418(26.2619)
#> 
#> Running at Laplace(9.1863e-12) None  Grid-point #24 at 2.37e+22 	Fold #1 Rep #1 pred log like = 19.6107
#> Running at Laplace(9.1863e-12) None  Grid-point #24 at 2.37e+22 	Fold #2 Rep #1 pred log like = 49.6421
#> Running at Laplace(9.1863e-12) None  Grid-point #24 at 2.37e+22 	Fold #3 Rep #1 pred log like = 82.0473
#> Running at Laplace(9.1863e-12) None  Grid-point #24 at 2.37e+22 	Fold #4 Rep #1 pred log like = 75.738
#> Running at Laplace(9.1863e-12) None  Grid-point #24 at 2.37e+22 	Fold #5 Rep #1 pred log like = 86.789
#> Running at Laplace(9.1863e-12) None  Grid-point #24 at 2.37e+22 	Fold #6 Rep #1 pred log like = 19.6856
#> Running at Laplace(9.1863e-12) None  Grid-point #24 at 2.37e+22 	Fold #7 Rep #1 pred log like = 61.5262
#> Running at Laplace(9.1863e-12) None  Grid-point #24 at 2.37e+22 	Fold #8 Rep #1 pred log like = 74.4055
#> Running at Laplace(9.1863e-12) None  Grid-point #24 at 2.37e+22 	Fold #9 Rep #1 pred log like = 103.175
#> Running at Laplace(9.1863e-12) None  Grid-point #24 at 2.37e+22 	Fold #10 Rep #1 pred log like = 51.7984
#> AvgPred = 62.4418 with stdev = 26.2619
#> Completed at 2.37e+22
#> Next point at 3.41956e+13 with value 62.442 and continue = 0
#> search[ 0.237 ] = 62.4393(26.2582)
#> search[ 2.37 ] = 62.4416(26.2607)
#> search[ 23.7 ] = 62.4418(26.2615)
#> search[ 237 ] = 62.4418(26.2618)
#> search[ 2370 ] = 62.4418(26.2618)
#> search[ 23700 ] = 62.4418(26.2619)
#> search[ 237000 ] = 62.4418(26.2619)
#> search[ 2.37e+06 ] = 62.4418(26.2619)
#> search[ 2.37e+07 ] = 62.4418(26.2619)
#> search[ 2.37e+08 ] = 62.4418(26.2619)
#> search[ 2.37e+09 ] = 62.4418(26.2619)
#> search[ 2.37e+10 ] = 62.4418(26.2619)
#> search[ 2.37e+11 ] = 62.4418(26.2619)
#> search[ 2.37e+12 ] = 62.4418(26.2619)
#> search[ 2.37e+13 ] = 62.4418(26.2619)
#> search[ 2.37e+14 ] = 62.4418(26.2619)
#> search[ 2.37e+15 ] = 62.4418(26.2619)
#> search[ 2.37e+16 ] = 62.4418(26.2619)
#> search[ 2.37e+17 ] = 62.4418(26.2619)
#> search[ 2.37e+18 ] = 62.4418(26.2619)
#> search[ 2.37e+19 ] = 62.4418(26.2619)
#> search[ 2.37e+20 ] = 62.4418(26.2619)
#> search[ 2.37e+21 ] = 62.4418(26.2619)
#> search[ 2.37e+22 ] = 62.4418(26.2619)
#> 
#> 
#> Maximum predicted log likelihood (62.442) estimated at:
#> 	3.41956e+13 (variance)
#> 	2.41841e-07 (lambda)
#> 
#> Fitting model at optimal hyperparameter
#> Using prior: Laplace(2.41841e-07) None 

#Find out what the optimal hyperparameter was:
getHyperParameter(fit)

#> [1] 3.419565e+13

#Extract the current log-likelihood, and coefficients
logLik(fit)

#> 'log Lik.' -1691.248 (df=3)

coef(fit)

#> (Intercept)           1           2 
#> -4.39464620 -0.04762332 -0.55968738 

#We can only retrieve the confidence interval for unregularized coefficients:
confint(fit, c(0))

#> Using 1 thread(s)

#>      covariate     2.5 %    97.5 % evaluations
#> [1,]         0 -4.439161 -4.350773          20