predict.cyclopsFit {Cyclops}	R Documentation

Model predictions

Description

predict.cyclopsFit computes model response-scale predictive values for all data rows

Usage

## S3 method for class 'cyclopsFit'
predict(object, newOutcomes, newCovariates, ...)

Arguments

object
A Cyclops model fit object

newOutcomes
An optional data frame or Andromeda table object, similar to the object used in convertToCyclopsData.

newCovariates
An optional data frame or Andromeda table object, similar to the object used in convertToCyclopsData.

...
Additional arguments

[Package Cyclops version 3.2.0 Index]