Simulation Cyclops dataset

simulateCyclopsData generates a simulated large, sparse data set for use by fitCyclopsSimulation.

simulateCyclopsData(
  nstrata = 200,
  nrows = 10000,
  ncovars = 20,
  effectSizeSd = 1,
  zeroEffectSizeProp = 0.9,
  eCovarsPerRow = ncovars/100,
  model = "survival"
)

Arguments

nstrata
Numeric: Number of strata

nrows
Numeric: Number of observation rows

ncovars
Numeric: Number of covariates

effectSizeSd
Numeric: Standard derivation of the non-zero simulated regression coefficients

zeroEffectSizeProp
Numeric: Expected proportion of zero effect size

eCovarsPerRow
Number: Effective number of non-zero covariates per data row

model
String: Simulation model. Choices are: logistic, poisson or survival

Value

A simulated data set

Examples

#Generate some simulated data:
sim <- simulateCyclopsData(nstrata = 1, nrows = 1000, ncovars = 2, eCovarsPerRow = 0.5, 
                           model = "poisson")

#> Sparseness = 77.4 %

cyclopsData <- convertToCyclopsData(sim$outcomes, sim$covariates, modelType = "pr", 
                                    addIntercept = TRUE)

#> Sorting covariates by covariateId and rowId

#Define the prior and control objects to use cross-validation for finding the 
#optimal hyperparameter:
prior <- createPrior("laplace", exclude = 0, useCrossValidation = TRUE)
control <- createControl(cvType = "auto", noiseLevel = "quiet")

#Fit the model
fit <- fitCyclopsModel(cyclopsData,prior = prior, control = control)  

#> Using cross-validation selector type byRow
#> Performing 10-fold cross-validation [seed = 1604638988] with data partitions of sizes 100 100 100 100 100 100 100 100 100 100
#> Using 1 thread(s)
#> Starting var = 0.226 (default)
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #1 Rep #1 pred log like = 792.844
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #2 Rep #1 pred log like = 617.557
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #3 Rep #1 pred log like = 616.862
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #4 Rep #1 pred log like = 590.672
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #5 Rep #1 pred log like = 677.662
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #6 Rep #1 pred log like = 553.036
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #7 Rep #1 pred log like = 698.08
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #8 Rep #1 pred log like = 737.317
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #9 Rep #1 pred log like = 807.911
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #10 Rep #1 pred log like = 576.817
#> AvgPred = 666.876 with stdev = 85.803
#> Completed at 0.226
#> Next point at 2.26 with value 0 and continue = 1
#> search[ 0.226 ] = 666.876(85.803)
#> 
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #1 Rep #1 pred log like = 792.826
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #2 Rep #1 pred log like = 617.55
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #3 Rep #1 pred log like = 616.879
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #4 Rep #1 pred log like = 590.679
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #5 Rep #1 pred log like = 677.644
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #6 Rep #1 pred log like = 553.048
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #7 Rep #1 pred log like = 698.06
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #8 Rep #1 pred log like = 737.304
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #9 Rep #1 pred log like = 807.94
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #10 Rep #1 pred log like = 576.831
#> AvgPred = 666.876 with stdev = 85.7988
#> Completed at 2.26
#> Next point at 22.6 with value 0 and continue = 1
#> search[ 0.226 ] = 666.876(85.803)
#> search[ 2.26 ] = 666.876(85.7988)
#> 
#> Running at Laplace(0.297482) None  Grid-point #3 at 22.6 	Fold #1 Rep #1 pred log like = 792.82
#> Running at Laplace(0.297482) None  Grid-point #3 at 22.6 	Fold #2 Rep #1 pred log like = 617.548
#> Running at Laplace(0.297482) None  Grid-point #3 at 22.6 	Fold #3 Rep #1 pred log like = 616.885
#> Running at Laplace(0.297482) None  Grid-point #3 at 22.6 	Fold #4 Rep #1 pred log like = 590.682
#> Running at Laplace(0.297482) None  Grid-point #3 at 22.6 	Fold #5 Rep #1 pred log like = 677.639
#> Running at Laplace(0.297482) None  Grid-point #3 at 22.6 	Fold #6 Rep #1 pred log like = 553.052
#> Running at Laplace(0.297482) None  Grid-point #3 at 22.6 	Fold #7 Rep #1 pred log like = 698.054
#> Running at Laplace(0.297482) None  Grid-point #3 at 22.6 	Fold #8 Rep #1 pred log like = 737.299
#> Running at Laplace(0.297482) None  Grid-point #3 at 22.6 	Fold #9 Rep #1 pred log like = 807.949
#> Running at Laplace(0.297482) None  Grid-point #3 at 22.6 	Fold #10 Rep #1 pred log like = 576.835
#> AvgPred = 666.876 with stdev = 85.7974
#> Completed at 22.6
#> Next point at 6.20929 with value 666.876 and continue = 0
#> search[ 0.226 ] = 666.876(85.803)
#> search[ 2.26 ] = 666.876(85.7988)
#> search[ 22.6 ] = 666.876(85.7974)
#> 
#> 
#> Maximum predicted log likelihood (666.876) estimated at:
#> 	6.20929 (variance)
#> 	0.567537 (lambda)
#> 
#> Fitting model at optimal hyperparameter
#> Using prior: Laplace(0.567537) None 

#Find out what the optimal hyperparameter was:
getHyperParameter(fit)

#> [1] 6.209294

#Extract the current log-likelihood, and coefficients
logLik(fit)

#> 'log Lik.' -2175.875 (df=3)

coef(fit)

#> (Intercept)           1           2 
#> -3.65827000 -0.02978819 -0.01326344 

#We can only retrieve the confidence interval for unregularized coefficients:
confint(fit, c(0))

#> Using 1 thread(s)

#>      covariate     2.5 %    97.5 % evaluations
#> [1,]         0 -3.688468 -3.628274          22