summary.cyclopsData {Cyclops}	R Documentation

Cyclops data object summary

Description

summary.cyclopsData summarizes the data held in an Cyclops data object.

Usage

## S3 method for class 'cyclopsData'
summary(object, ...)

Arguments

object
A Cyclops data object

...
Additional arguments

Value

Returns a data.frame that reports simply summarize statistics for each covariate in a Cyclops data object.

[Package Cyclops version 3.2.1 Index]