aconfint {Cyclops}	R Documentation

Asymptotic confidence intervals for a fitted Cyclops model object

Description

aconfinit constructs confidence intervals of arbitrary level using asymptotic standard error estimates.

Usage

aconfint(
  object,
  parm,
  level = 0.95,
  control,
  overrideNoRegularization = FALSE,
  ...
)

Arguments

object
A fitted Cyclops model object

parm
A specification of which parameters require confidence intervals, either a vector of numbers of covariateId names

level
Numeric: confidence level required

control
A "cyclopsControl" object constructed by createControl

overrideNoRegularization
Logical: Enable confidence interval estimation for regularized parameters

...
Additional argument(s) for methods

Value

A matrix with columns reporting lower and upper confidence limits for each parameter. These columns are labelled as (1-level) / 2 and 1 - (1 - level) / 2 in (by default 2.5

[Package Cyclops version 3.2.0 Index]