Confidence intervals for Cyclops model parameters

confinit.cyclopsFit profiles the data likelihood to construct confidence intervals of arbitrary level. Usually it only makes sense to do this for variables that have not been regularized.

# S3 method for cyclopsFit
confint(
  object,
  parm,
  level = 0.95,
  overrideNoRegularization = FALSE,
  includePenalty = TRUE,
  rescale = FALSE,
  ...
)

Arguments

object
A fitted Cyclops model object

parm
A specification of which parameters require confidence intervals, either a vector of numbers of covariateId names

level
Numeric: confidence level required

overrideNoRegularization
Logical: Enable confidence interval estimation for regularized parameters

includePenalty
Logical: Include regularized covariate penalty in profile

rescale
Boolean: rescale coefficients for unnormalized covariate values

...
Additional argument(s) for methods

Value

A matrix with columns reporting lower and upper confidence limits for each parameter. These columns are labelled as (1-level) / 2 and 1 - (1 - level) / 2 in percent (by default 2.5 percent and 97.5 percent)

Examples

#Generate some simulated data:
sim <- simulateCyclopsData(nstrata = 1, nrows = 1000, ncovars = 2, eCovarsPerRow = 0.5,
                           model = "poisson")

#> Sparseness = 74.45 %

cyclopsData <- convertToCyclopsData(sim$outcomes, sim$covariates, modelType = "pr",
                                    addIntercept = TRUE)

#> Sorting covariates by covariateId and rowId

#Define the prior and control objects to use cross-validation for finding the 
#optimal hyperparameter:
prior <- createPrior("laplace", exclude = 0, useCrossValidation = TRUE)
control <- createControl(cvType = "auto", noiseLevel = "quiet")

#Fit the model
fit <- fitCyclopsModel(cyclopsData,prior = prior, control = control)

#> Using cross-validation selector type byRow
#> Performing 10-fold cross-validation [seed = 1591596881] with data partitions of sizes 100 100 100 100 100 100 100 100 100 100
#> Using 1 thread(s)
#> Starting var = 0.2555 (default)
#> Running at Laplace(2.79782) None  Grid-point #1 at 0.2555 	Fold #1 Rep #1 pred log like = 290.553
#> Running at Laplace(2.79782) None  Grid-point #1 at 0.2555 	Fold #2 Rep #1 pred log like = 395.938
#> Running at Laplace(2.79782) None  Grid-point #1 at 0.2555 	Fold #3 Rep #1 pred log like = 434.852
#> Running at Laplace(2.79782) None  Grid-point #1 at 0.2555 	Fold #4 Rep #1 pred log like = 438.312
#> Running at Laplace(2.79782) None  Grid-point #1 at 0.2555 	Fold #5 Rep #1 pred log like = 297.902
#> Running at Laplace(2.79782) None  Grid-point #1 at 0.2555 	Fold #6 Rep #1 pred log like = 360.762
#> Running at Laplace(2.79782) None  Grid-point #1 at 0.2555 	Fold #7 Rep #1 pred log like = 323.776
#> Running at Laplace(2.79782) None  Grid-point #1 at 0.2555 	Fold #8 Rep #1 pred log like = 407.344
#> Running at Laplace(2.79782) None  Grid-point #1 at 0.2555 	Fold #9 Rep #1 pred log like = 339.028
#> Running at Laplace(2.79782) None  Grid-point #1 at 0.2555 	Fold #10 Rep #1 pred log like = 382.03
#> AvgPred = 367.05 with stdev = 50.684
#> Completed at 0.2555
#> Next point at 2.555 with value 0 and continue = 1
#> search[ 0.2555 ] = 367.05(50.684)
#> 
#> Running at Laplace(0.884748) None  Grid-point #2 at 2.555 	Fold #1 Rep #1 pred log like = 290.545
#> Running at Laplace(0.884748) None  Grid-point #2 at 2.555 	Fold #2 Rep #1 pred log like = 395.925
#> Running at Laplace(0.884748) None  Grid-point #2 at 2.555 	Fold #3 Rep #1 pred log like = 434.849
#> Running at Laplace(0.884748) None  Grid-point #2 at 2.555 	Fold #4 Rep #1 pred log like = 438.302
#> Running at Laplace(0.884748) None  Grid-point #2 at 2.555 	Fold #5 Rep #1 pred log like = 297.877
#> Running at Laplace(0.884748) None  Grid-point #2 at 2.555 	Fold #6 Rep #1 pred log like = 360.712
#> Running at Laplace(0.884748) None  Grid-point #2 at 2.555 	Fold #7 Rep #1 pred log like = 323.78
#> Running at Laplace(0.884748) None  Grid-point #2 at 2.555 	Fold #8 Rep #1 pred log like = 407.278
#> Running at Laplace(0.884748) None  Grid-point #2 at 2.555 	Fold #9 Rep #1 pred log like = 338.976
#> Running at Laplace(0.884748) None  Grid-point #2 at 2.555 	Fold #10 Rep #1 pred log like = 382.002
#> AvgPred = 367.025 with stdev = 50.6831
#> Completed at 2.555
#> Next point at 0.02555 with value 0 and continue = 1
#> search[ 0.2555 ] = 367.05(50.684)
#> search[ 2.555 ] = 367.025(50.6831)
#> 
#> Running at Laplace(8.84748) None  Grid-point #3 at 0.02555 	Fold #1 Rep #1 pred log like = 290.576
#> Running at Laplace(8.84748) None  Grid-point #3 at 0.02555 	Fold #2 Rep #1 pred log like = 395.968
#> Running at Laplace(8.84748) None  Grid-point #3 at 0.02555 	Fold #3 Rep #1 pred log like = 434.884
#> Running at Laplace(8.84748) None  Grid-point #3 at 0.02555 	Fold #4 Rep #1 pred log like = 438.369
#> Running at Laplace(8.84748) None  Grid-point #3 at 0.02555 	Fold #5 Rep #1 pred log like = 297.997
#> Running at Laplace(8.84748) None  Grid-point #3 at 0.02555 	Fold #6 Rep #1 pred log like = 360.911
#> Running at Laplace(8.84748) None  Grid-point #3 at 0.02555 	Fold #7 Rep #1 pred log like = 323.779
#> Running at Laplace(8.84748) None  Grid-point #3 at 0.02555 	Fold #8 Rep #1 pred log like = 407.546
#> Running at Laplace(8.84748) None  Grid-point #3 at 0.02555 	Fold #9 Rep #1 pred log like = 339.163
#> Running at Laplace(8.84748) None  Grid-point #3 at 0.02555 	Fold #10 Rep #1 pred log like = 382.062
#> AvgPred = 367.125 with stdev = 50.6888
#> Completed at 0.02555
#> Next point at 0.002555 with value 0 and continue = 1
#> search[ 0.02555 ] = 367.125(50.6888)
#> search[ 0.2555 ] = 367.05(50.684)
#> search[ 2.555 ] = 367.025(50.6831)
#> 
#> Running at Laplace(27.9782) None  Grid-point #4 at 0.002555 	Fold #1 Rep #1 pred log like = 290.576
#> Running at Laplace(27.9782) None  Grid-point #4 at 0.002555 	Fold #2 Rep #1 pred log like = 395.977
#> Running at Laplace(27.9782) None  Grid-point #4 at 0.002555 	Fold #3 Rep #1 pred log like = 434.884
#> Running at Laplace(27.9782) None  Grid-point #4 at 0.002555 	Fold #4 Rep #1 pred log like = 438.41
#> Running at Laplace(27.9782) None  Grid-point #4 at 0.002555 	Fold #5 Rep #1 pred log like = 298.153
#> Running at Laplace(27.9782) None  Grid-point #4 at 0.002555 	Fold #6 Rep #1 pred log like = 361.114
#> Running at Laplace(27.9782) None  Grid-point #4 at 0.002555 	Fold #7 Rep #1 pred log like = 323.779
#> Running at Laplace(27.9782) None  Grid-point #4 at 0.002555 	Fold #8 Rep #1 pred log like = 407.78
#> Running at Laplace(27.9782) None  Grid-point #4 at 0.002555 	Fold #9 Rep #1 pred log like = 339.167
#> Running at Laplace(27.9782) None  Grid-point #4 at 0.002555 	Fold #10 Rep #1 pred log like = 382.076
#> AvgPred = 367.192 with stdev = 50.6904
#> Completed at 0.002555
#> Next point at 0.0002555 with value 0 and continue = 1
#> search[ 0.002555 ] = 367.192(50.6904)
#> search[ 0.02555 ] = 367.125(50.6888)
#> search[ 0.2555 ] = 367.05(50.684)
#> search[ 2.555 ] = 367.025(50.6831)
#> 
#> Running at Laplace(88.4748) None  Grid-point #5 at 0.0002555 	Fold #1 Rep #1 pred log like = 290.576
#> Running at Laplace(88.4748) None  Grid-point #5 at 0.0002555 	Fold #2 Rep #1 pred log like = 395.977
#> Running at Laplace(88.4748) None  Grid-point #5 at 0.0002555 	Fold #3 Rep #1 pred log like = 434.884
#> Running at Laplace(88.4748) None  Grid-point #5 at 0.0002555 	Fold #4 Rep #1 pred log like = 438.41
#> Running at Laplace(88.4748) None  Grid-point #5 at 0.0002555 	Fold #5 Rep #1 pred log like = 298.153
#> Running at Laplace(88.4748) None  Grid-point #5 at 0.0002555 	Fold #6 Rep #1 pred log like = 361.114
#> Running at Laplace(88.4748) None  Grid-point #5 at 0.0002555 	Fold #7 Rep #1 pred log like = 323.779
#> Running at Laplace(88.4748) None  Grid-point #5 at 0.0002555 	Fold #8 Rep #1 pred log like = 407.78
#> Running at Laplace(88.4748) None  Grid-point #5 at 0.0002555 	Fold #9 Rep #1 pred log like = 339.167
#> Running at Laplace(88.4748) None  Grid-point #5 at 0.0002555 	Fold #10 Rep #1 pred log like = 382.076
#> AvgPred = 367.192 with stdev = 50.6904
#> Completed at 0.0002555
#> Next point at 6.39478e-08 with value 367.258 and continue = 0
#> search[ 0.0002555 ] = 367.192(50.6904)
#> search[ 0.002555 ] = 367.192(50.6904)
#> search[ 0.02555 ] = 367.125(50.6888)
#> search[ 0.2555 ] = 367.05(50.684)
#> search[ 2.555 ] = 367.025(50.6831)
#> 
#> 
#> Maximum predicted log likelihood (367.258) estimated at:
#> 	6.39478e-08 (variance)
#> 	5592.45 (lambda)
#> 
#> Fitting model at optimal hyperparameter
#> Using prior: Laplace(5592.45) None 

#Find out what the optimal hyperparameter was:
getHyperParameter(fit)

#> [1] 6.39478e-08

#Extract the current log-likelihood, and coefficients
logLik(fit)

#> 'log Lik.' -1973.343 (df=3)

coef(fit)

#> (Intercept)           1           2 
#>   -3.937029    0.000000    0.000000 

#We can only retrieve the confidence interval for unregularized coefficients:
confint(fit, c(0))

#> Using 1 thread(s)

#>      covariate     2.5 %    97.5 % evaluations
#> [1,]         0 -3.965463 -3.908849          24