Confidence intervals for Cyclops model parameters

confinit.cyclopsFit profiles the data likelihood to construct confidence intervals of arbitrary level. Usually it only makes sense to do this for variables that have not been regularized.

# S3 method for cyclopsFit
confint(
  object,
  parm,
  level = 0.95,
  overrideNoRegularization = FALSE,
  includePenalty = TRUE,
  rescale = FALSE,
  ...
)

Arguments

object

A fitted Cyclops model object

parm

A specification of which parameters require confidence intervals, either a vector of numbers of covariateId names

level

Numeric: confidence level required

overrideNoRegularization

Logical: Enable confidence interval estimation for regularized parameters

includePenalty

Logical: Include regularized covariate penalty in profile

rescale

Boolean: rescale coefficients for unnormalized covariate values

...

Additional argument(s) for methods

Value

A matrix with columns reporting lower and upper confidence limits for each parameter. These columns are labelled as (1-level) / 2 and 1 - (1 - level) / 2 in percent (by default 2.5 percent and 97.5 percent)

Examples

#Generate some simulated data:
sim <- simulateCyclopsData(nstrata = 1, nrows = 1000, ncovars = 2, eCovarsPerRow = 0.5, 
                           model = "poisson")
#> Sparseness = 77 %
cyclopsData <- convertToCyclopsData(sim$outcomes, sim$covariates, modelType = "pr", 
                                    addIntercept = TRUE)
#> Sorting covariates by covariateId and rowId

#Define the prior and control objects to use cross-validation for finding the 
#optimal hyperparameter:
prior <- createPrior("laplace", exclude = 0, useCrossValidation = TRUE)
control <- createControl(cvType = "auto", noiseLevel = "quiet")

#Fit the model
fit <- fitCyclopsModel(cyclopsData,prior = prior, control = control)  
#> Using cross-validation selector type byRow
#> Performing 10-fold cross-validation [seed = 1698757931] with data partitions of sizes 100 100 100 100 100 100 100 100 100 100
#> Using 1 thread(s)
#> Starting var = 0.23 (default)
#> Running at Laplace(2.94884) None  Grid-point #1 at 0.23 	Fold #1 Rep #1 pred log like = 267.858
#> Running at Laplace(2.94884) None  Grid-point #1 at 0.23 	Fold #2 Rep #1 pred log like = 210.218
#> Running at Laplace(2.94884) None  Grid-point #1 at 0.23 	Fold #3 Rep #1 pred log like = 247.642
#> Running at Laplace(2.94884) None  Grid-point #1 at 0.23 	Fold #4 Rep #1 pred log like = 268.625
#> Running at Laplace(2.94884) None  Grid-point #1 at 0.23 	Fold #5 Rep #1 pred log like = 234.92
#> Running at Laplace(2.94884) None  Grid-point #1 at 0.23 	Fold #6 Rep #1 pred log like = 256.935
#> Running at Laplace(2.94884) None  Grid-point #1 at 0.23 	Fold #7 Rep #1 pred log like = 237.206
#> Running at Laplace(2.94884) None  Grid-point #1 at 0.23 	Fold #8 Rep #1 pred log like = 203.575
#> Running at Laplace(2.94884) None  Grid-point #1 at 0.23 	Fold #9 Rep #1 pred log like = 268.979
#> Running at Laplace(2.94884) None  Grid-point #1 at 0.23 	Fold #10 Rep #1 pred log like = 313.499
#> AvgPred = 250.946 with stdev = 30.3841
#> Completed at 0.23
#> Next point at 2.3 with value 0 and continue = 1
#> search[ 0.23 ] = 250.946(30.3841)
#> 
#> Running at Laplace(0.932505) None  Grid-point #2 at 2.3 	Fold #1 Rep #1 pred log like = 267.817
#> Running at Laplace(0.932505) None  Grid-point #2 at 2.3 	Fold #2 Rep #1 pred log like = 210.18
#> Running at Laplace(0.932505) None  Grid-point #2 at 2.3 	Fold #3 Rep #1 pred log like = 247.667
#> Running at Laplace(0.932505) None  Grid-point #2 at 2.3 	Fold #4 Rep #1 pred log like = 268.65
#> Running at Laplace(0.932505) None  Grid-point #2 at 2.3 	Fold #5 Rep #1 pred log like = 234.965
#> Running at Laplace(0.932505) None  Grid-point #2 at 2.3 	Fold #6 Rep #1 pred log like = 256.945
#> Running at Laplace(0.932505) None  Grid-point #2 at 2.3 	Fold #7 Rep #1 pred log like = 237.151
#> Running at Laplace(0.932505) None  Grid-point #2 at 2.3 	Fold #8 Rep #1 pred log like = 203.606
#> Running at Laplace(0.932505) None  Grid-point #2 at 2.3 	Fold #9 Rep #1 pred log like = 268.987
#> Running at Laplace(0.932505) None  Grid-point #2 at 2.3 	Fold #10 Rep #1 pred log like = 313.505
#> AvgPred = 250.947 with stdev = 30.3849
#> Completed at 2.3
#> Next point at 23 with value 0 and continue = 1
#> search[ 0.23 ] = 250.946(30.3841)
#> search[ 2.3 ] = 250.947(30.3849)
#> 
#> Running at Laplace(0.294884) None  Grid-point #3 at 23 	Fold #1 Rep #1 pred log like = 267.803
#> Running at Laplace(0.294884) None  Grid-point #3 at 23 	Fold #2 Rep #1 pred log like = 210.168
#> Running at Laplace(0.294884) None  Grid-point #3 at 23 	Fold #3 Rep #1 pred log like = 247.674
#> Running at Laplace(0.294884) None  Grid-point #3 at 23 	Fold #4 Rep #1 pred log like = 268.657
#> Running at Laplace(0.294884) None  Grid-point #3 at 23 	Fold #5 Rep #1 pred log like = 234.98
#> Running at Laplace(0.294884) None  Grid-point #3 at 23 	Fold #6 Rep #1 pred log like = 256.948
#> Running at Laplace(0.294884) None  Grid-point #3 at 23 	Fold #7 Rep #1 pred log like = 237.134
#> Running at Laplace(0.294884) None  Grid-point #3 at 23 	Fold #8 Rep #1 pred log like = 203.616
#> Running at Laplace(0.294884) None  Grid-point #3 at 23 	Fold #9 Rep #1 pred log like = 268.988
#> Running at Laplace(0.294884) None  Grid-point #3 at 23 	Fold #10 Rep #1 pred log like = 313.506
#> AvgPred = 250.947 with stdev = 30.3852
#> Completed at 23
#> Next point at 230 with value 0 and continue = 1
#> search[ 0.23 ] = 250.946(30.3841)
#> search[ 2.3 ] = 250.947(30.3849)
#> search[ 23 ] = 250.947(30.3852)
#> 
#> Running at Laplace(0.0932505) None  Grid-point #4 at 230 	Fold #1 Rep #1 pred log like = 267.799
#> Running at Laplace(0.0932505) None  Grid-point #4 at 230 	Fold #2 Rep #1 pred log like = 210.165
#> Running at Laplace(0.0932505) None  Grid-point #4 at 230 	Fold #3 Rep #1 pred log like = 247.677
#> Running at Laplace(0.0932505) None  Grid-point #4 at 230 	Fold #4 Rep #1 pred log like = 268.66
#> Running at Laplace(0.0932505) None  Grid-point #4 at 230 	Fold #5 Rep #1 pred log like = 234.984
#> Running at Laplace(0.0932505) None  Grid-point #4 at 230 	Fold #6 Rep #1 pred log like = 256.948
#> Running at Laplace(0.0932505) None  Grid-point #4 at 230 	Fold #7 Rep #1 pred log like = 237.128
#> Running at Laplace(0.0932505) None  Grid-point #4 at 230 	Fold #8 Rep #1 pred log like = 203.619
#> Running at Laplace(0.0932505) None  Grid-point #4 at 230 	Fold #9 Rep #1 pred log like = 268.989
#> Running at Laplace(0.0932505) None  Grid-point #4 at 230 	Fold #10 Rep #1 pred log like = 313.507
#> AvgPred = 250.948 with stdev = 30.3853
#> Completed at 230
#> Next point at 2300 with value 0 and continue = 1
#> search[ 0.23 ] = 250.946(30.3841)
#> search[ 2.3 ] = 250.947(30.3849)
#> search[ 23 ] = 250.947(30.3852)
#> search[ 230 ] = 250.948(30.3853)
#> 
#> Running at Laplace(0.0294884) None  Grid-point #5 at 2300 	Fold #1 Rep #1 pred log like = 267.798
#> Running at Laplace(0.0294884) None  Grid-point #5 at 2300 	Fold #2 Rep #1 pred log like = 210.163
#> Running at Laplace(0.0294884) None  Grid-point #5 at 2300 	Fold #3 Rep #1 pred log like = 247.678
#> Running at Laplace(0.0294884) None  Grid-point #5 at 2300 	Fold #4 Rep #1 pred log like = 268.66
#> Running at Laplace(0.0294884) None  Grid-point #5 at 2300 	Fold #5 Rep #1 pred log like = 234.985
#> Running at Laplace(0.0294884) None  Grid-point #5 at 2300 	Fold #6 Rep #1 pred log like = 256.949
#> Running at Laplace(0.0294884) None  Grid-point #5 at 2300 	Fold #7 Rep #1 pred log like = 237.126
#> Running at Laplace(0.0294884) None  Grid-point #5 at 2300 	Fold #8 Rep #1 pred log like = 203.62
#> Running at Laplace(0.0294884) None  Grid-point #5 at 2300 	Fold #9 Rep #1 pred log like = 268.989
#> Running at Laplace(0.0294884) None  Grid-point #5 at 2300 	Fold #10 Rep #1 pred log like = 313.507
#> AvgPred = 250.948 with stdev = 30.3853
#> Completed at 2300
#> Next point at 23000 with value 0 and continue = 1
#> search[ 0.23 ] = 250.946(30.3841)
#> search[ 2.3 ] = 250.947(30.3849)
#> search[ 23 ] = 250.947(30.3852)
#> search[ 230 ] = 250.948(30.3853)
#> search[ 2300 ] = 250.948(30.3853)
#> 
#> Running at Laplace(0.00932505) None  Grid-point #6 at 23000 	Fold #1 Rep #1 pred log like = 267.797
#> Running at Laplace(0.00932505) None  Grid-point #6 at 23000 	Fold #2 Rep #1 pred log like = 210.163
#> Running at Laplace(0.00932505) None  Grid-point #6 at 23000 	Fold #3 Rep #1 pred log like = 247.678
#> Running at Laplace(0.00932505) None  Grid-point #6 at 23000 	Fold #4 Rep #1 pred log like = 268.661
#> Running at Laplace(0.00932505) None  Grid-point #6 at 23000 	Fold #5 Rep #1 pred log like = 234.986
#> Running at Laplace(0.00932505) None  Grid-point #6 at 23000 	Fold #6 Rep #1 pred log like = 256.949
#> Running at Laplace(0.00932505) None  Grid-point #6 at 23000 	Fold #7 Rep #1 pred log like = 237.126
#> Running at Laplace(0.00932505) None  Grid-point #6 at 23000 	Fold #8 Rep #1 pred log like = 203.62
#> Running at Laplace(0.00932505) None  Grid-point #6 at 23000 	Fold #9 Rep #1 pred log like = 268.989
#> Running at Laplace(0.00932505) None  Grid-point #6 at 23000 	Fold #10 Rep #1 pred log like = 313.507
#> AvgPred = 250.948 with stdev = 30.3853
#> Completed at 23000
#> Next point at 230000 with value 0 and continue = 1
#> search[ 0.23 ] = 250.946(30.3841)
#> search[ 2.3 ] = 250.947(30.3849)
#> search[ 23 ] = 250.947(30.3852)
#> search[ 230 ] = 250.948(30.3853)
#> search[ 2300 ] = 250.948(30.3853)
#> search[ 23000 ] = 250.948(30.3853)
#> 
#> Running at Laplace(0.00294884) None  Grid-point #7 at 230000 	Fold #1 Rep #1 pred log like = 267.797
#> Running at Laplace(0.00294884) None  Grid-point #7 at 230000 	Fold #2 Rep #1 pred log like = 210.163
#> Running at Laplace(0.00294884) None  Grid-point #7 at 230000 	Fold #3 Rep #1 pred log like = 247.678
#> Running at Laplace(0.00294884) None  Grid-point #7 at 230000 	Fold #4 Rep #1 pred log like = 268.661
#> Running at Laplace(0.00294884) None  Grid-point #7 at 230000 	Fold #5 Rep #1 pred log like = 234.986
#> Running at Laplace(0.00294884) None  Grid-point #7 at 230000 	Fold #6 Rep #1 pred log like = 256.949
#> Running at Laplace(0.00294884) None  Grid-point #7 at 230000 	Fold #7 Rep #1 pred log like = 237.126
#> Running at Laplace(0.00294884) None  Grid-point #7 at 230000 	Fold #8 Rep #1 pred log like = 203.62
#> Running at Laplace(0.00294884) None  Grid-point #7 at 230000 	Fold #9 Rep #1 pred log like = 268.989
#> Running at Laplace(0.00294884) None  Grid-point #7 at 230000 	Fold #10 Rep #1 pred log like = 313.507
#> AvgPred = 250.948 with stdev = 30.3853
#> Completed at 230000
#> Next point at 2.3e+06 with value 0 and continue = 1
#> search[ 0.23 ] = 250.946(30.3841)
#> search[ 2.3 ] = 250.947(30.3849)
#> search[ 23 ] = 250.947(30.3852)
#> search[ 230 ] = 250.948(30.3853)
#> search[ 2300 ] = 250.948(30.3853)
#> search[ 23000 ] = 250.948(30.3853)
#> search[ 230000 ] = 250.948(30.3853)
#> 
#> Running at Laplace(0.000932505) None  Grid-point #8 at 2.3e+06 	Fold #1 Rep #1 pred log like = 267.797
#> Running at Laplace(0.000932505) None  Grid-point #8 at 2.3e+06 	Fold #2 Rep #1 pred log like = 210.163
#> Running at Laplace(0.000932505) None  Grid-point #8 at 2.3e+06 	Fold #3 Rep #1 pred log like = 247.678
#> Running at Laplace(0.000932505) None  Grid-point #8 at 2.3e+06 	Fold #4 Rep #1 pred log like = 268.661
#> Running at Laplace(0.000932505) None  Grid-point #8 at 2.3e+06 	Fold #5 Rep #1 pred log like = 234.986
#> Running at Laplace(0.000932505) None  Grid-point #8 at 2.3e+06 	Fold #6 Rep #1 pred log like = 256.949
#> Running at Laplace(0.000932505) None  Grid-point #8 at 2.3e+06 	Fold #7 Rep #1 pred log like = 237.125
#> Running at Laplace(0.000932505) None  Grid-point #8 at 2.3e+06 	Fold #8 Rep #1 pred log like = 203.62
#> Running at Laplace(0.000932505) None  Grid-point #8 at 2.3e+06 	Fold #9 Rep #1 pred log like = 268.989
#> Running at Laplace(0.000932505) None  Grid-point #8 at 2.3e+06 	Fold #10 Rep #1 pred log like = 313.507
#> AvgPred = 250.948 with stdev = 30.3853
#> Completed at 2.3e+06
#> Next point at 2.3e+07 with value 0 and continue = 1
#> search[ 0.23 ] = 250.946(30.3841)
#> search[ 2.3 ] = 250.947(30.3849)
#> search[ 23 ] = 250.947(30.3852)
#> search[ 230 ] = 250.948(30.3853)
#> search[ 2300 ] = 250.948(30.3853)
#> search[ 23000 ] = 250.948(30.3853)
#> search[ 230000 ] = 250.948(30.3853)
#> search[ 2.3e+06 ] = 250.948(30.3853)
#> 
#> Running at Laplace(0.000294884) None  Grid-point #9 at 2.3e+07 	Fold #1 Rep #1 pred log like = 267.797
#> Running at Laplace(0.000294884) None  Grid-point #9 at 2.3e+07 	Fold #2 Rep #1 pred log like = 210.163
#> Running at Laplace(0.000294884) None  Grid-point #9 at 2.3e+07 	Fold #3 Rep #1 pred log like = 247.678
#> Running at Laplace(0.000294884) None  Grid-point #9 at 2.3e+07 	Fold #4 Rep #1 pred log like = 268.661
#> Running at Laplace(0.000294884) None  Grid-point #9 at 2.3e+07 	Fold #5 Rep #1 pred log like = 234.986
#> Running at Laplace(0.000294884) None  Grid-point #9 at 2.3e+07 	Fold #6 Rep #1 pred log like = 256.949
#> Running at Laplace(0.000294884) None  Grid-point #9 at 2.3e+07 	Fold #7 Rep #1 pred log like = 237.125
#> Running at Laplace(0.000294884) None  Grid-point #9 at 2.3e+07 	Fold #8 Rep #1 pred log like = 203.62
#> Running at Laplace(0.000294884) None  Grid-point #9 at 2.3e+07 	Fold #9 Rep #1 pred log like = 268.989
#> Running at Laplace(0.000294884) None  Grid-point #9 at 2.3e+07 	Fold #10 Rep #1 pred log like = 313.507
#> AvgPred = 250.948 with stdev = 30.3853
#> Completed at 2.3e+07
#> Next point at 2.3e+08 with value 0 and continue = 1
#> search[ 0.23 ] = 250.946(30.3841)
#> search[ 2.3 ] = 250.947(30.3849)
#> search[ 23 ] = 250.947(30.3852)
#> search[ 230 ] = 250.948(30.3853)
#> search[ 2300 ] = 250.948(30.3853)
#> search[ 23000 ] = 250.948(30.3853)
#> search[ 230000 ] = 250.948(30.3853)
#> search[ 2.3e+06 ] = 250.948(30.3853)
#> search[ 2.3e+07 ] = 250.948(30.3853)
#> 
#> Running at Laplace(9.32505e-05) None  Grid-point #10 at 2.3e+08 	Fold #1 Rep #1 pred log like = 267.797
#> Running at Laplace(9.32505e-05) None  Grid-point #10 at 2.3e+08 	Fold #2 Rep #1 pred log like = 210.163
#> Running at Laplace(9.32505e-05) None  Grid-point #10 at 2.3e+08 	Fold #3 Rep #1 pred log like = 247.678
#> Running at Laplace(9.32505e-05) None  Grid-point #10 at 2.3e+08 	Fold #4 Rep #1 pred log like = 268.661
#> Running at Laplace(9.32505e-05) None  Grid-point #10 at 2.3e+08 	Fold #5 Rep #1 pred log like = 234.986
#> Running at Laplace(9.32505e-05) None  Grid-point #10 at 2.3e+08 	Fold #6 Rep #1 pred log like = 256.949
#> Running at Laplace(9.32505e-05) None  Grid-point #10 at 2.3e+08 	Fold #7 Rep #1 pred log like = 237.125
#> Running at Laplace(9.32505e-05) None  Grid-point #10 at 2.3e+08 	Fold #8 Rep #1 pred log like = 203.62
#> Running at Laplace(9.32505e-05) None  Grid-point #10 at 2.3e+08 	Fold #9 Rep #1 pred log like = 268.989
#> Running at Laplace(9.32505e-05) None  Grid-point #10 at 2.3e+08 	Fold #10 Rep #1 pred log like = 313.507
#> AvgPred = 250.948 with stdev = 30.3853
#> Completed at 2.3e+08
#> Next point at 2.3e+09 with value 0 and continue = 1
#> search[ 0.23 ] = 250.946(30.3841)
#> search[ 2.3 ] = 250.947(30.3849)
#> search[ 23 ] = 250.947(30.3852)
#> search[ 230 ] = 250.948(30.3853)
#> search[ 2300 ] = 250.948(30.3853)
#> search[ 23000 ] = 250.948(30.3853)
#> search[ 230000 ] = 250.948(30.3853)
#> search[ 2.3e+06 ] = 250.948(30.3853)
#> search[ 2.3e+07 ] = 250.948(30.3853)
#> search[ 2.3e+08 ] = 250.948(30.3853)
#> 
#> Running at Laplace(2.94884e-05) None  Grid-point #11 at 2.3e+09 	Fold #1 Rep #1 pred log like = 267.797
#> Running at Laplace(2.94884e-05) None  Grid-point #11 at 2.3e+09 	Fold #2 Rep #1 pred log like = 210.163
#> Running at Laplace(2.94884e-05) None  Grid-point #11 at 2.3e+09 	Fold #3 Rep #1 pred log like = 247.678
#> Running at Laplace(2.94884e-05) None  Grid-point #11 at 2.3e+09 	Fold #4 Rep #1 pred log like = 268.661
#> Running at Laplace(2.94884e-05) None  Grid-point #11 at 2.3e+09 	Fold #5 Rep #1 pred log like = 234.986
#> Running at Laplace(2.94884e-05) None  Grid-point #11 at 2.3e+09 	Fold #6 Rep #1 pred log like = 256.949
#> Running at Laplace(2.94884e-05) None  Grid-point #11 at 2.3e+09 	Fold #7 Rep #1 pred log like = 237.125
#> Running at Laplace(2.94884e-05) None  Grid-point #11 at 2.3e+09 	Fold #8 Rep #1 pred log like = 203.62
#> Running at Laplace(2.94884e-05) None  Grid-point #11 at 2.3e+09 	Fold #9 Rep #1 pred log like = 268.989
#> Running at Laplace(2.94884e-05) None  Grid-point #11 at 2.3e+09 	Fold #10 Rep #1 pred log like = 313.507
#> AvgPred = 250.948 with stdev = 30.3853
#> Completed at 2.3e+09
#> Next point at 2.3e+10 with value 0 and continue = 1
#> search[ 0.23 ] = 250.946(30.3841)
#> search[ 2.3 ] = 250.947(30.3849)
#> search[ 23 ] = 250.947(30.3852)
#> search[ 230 ] = 250.948(30.3853)
#> search[ 2300 ] = 250.948(30.3853)
#> search[ 23000 ] = 250.948(30.3853)
#> search[ 230000 ] = 250.948(30.3853)
#> search[ 2.3e+06 ] = 250.948(30.3853)
#> search[ 2.3e+07 ] = 250.948(30.3853)
#> search[ 2.3e+08 ] = 250.948(30.3853)
#> search[ 2.3e+09 ] = 250.948(30.3853)
#> 
#> Running at Laplace(9.32505e-06) None  Grid-point #12 at 2.3e+10 	Fold #1 Rep #1 pred log like = 267.797
#> Running at Laplace(9.32505e-06) None  Grid-point #12 at 2.3e+10 	Fold #2 Rep #1 pred log like = 210.163
#> Running at Laplace(9.32505e-06) None  Grid-point #12 at 2.3e+10 	Fold #3 Rep #1 pred log like = 247.678
#> Running at Laplace(9.32505e-06) None  Grid-point #12 at 2.3e+10 	Fold #4 Rep #1 pred log like = 268.661
#> Running at Laplace(9.32505e-06) None  Grid-point #12 at 2.3e+10 	Fold #5 Rep #1 pred log like = 234.986
#> Running at Laplace(9.32505e-06) None  Grid-point #12 at 2.3e+10 	Fold #6 Rep #1 pred log like = 256.949
#> Running at Laplace(9.32505e-06) None  Grid-point #12 at 2.3e+10 	Fold #7 Rep #1 pred log like = 237.125
#> Running at Laplace(9.32505e-06) None  Grid-point #12 at 2.3e+10 	Fold #8 Rep #1 pred log like = 203.62
#> Running at Laplace(9.32505e-06) None  Grid-point #12 at 2.3e+10 	Fold #9 Rep #1 pred log like = 268.989
#> Running at Laplace(9.32505e-06) None  Grid-point #12 at 2.3e+10 	Fold #10 Rep #1 pred log like = 313.507
#> AvgPred = 250.948 with stdev = 30.3853
#> Completed at 2.3e+10
#> Next point at 2.3e+11 with value 0 and continue = 1
#> search[ 0.23 ] = 250.946(30.3841)
#> search[ 2.3 ] = 250.947(30.3849)
#> search[ 23 ] = 250.947(30.3852)
#> search[ 230 ] = 250.948(30.3853)
#> search[ 2300 ] = 250.948(30.3853)
#> search[ 23000 ] = 250.948(30.3853)
#> search[ 230000 ] = 250.948(30.3853)
#> search[ 2.3e+06 ] = 250.948(30.3853)
#> search[ 2.3e+07 ] = 250.948(30.3853)
#> search[ 2.3e+08 ] = 250.948(30.3853)
#> search[ 2.3e+09 ] = 250.948(30.3853)
#> search[ 2.3e+10 ] = 250.948(30.3853)
#> 
#> Running at Laplace(2.94884e-06) None  Grid-point #13 at 2.3e+11 	Fold #1 Rep #1 pred log like = 267.797
#> Running at Laplace(2.94884e-06) None  Grid-point #13 at 2.3e+11 	Fold #2 Rep #1 pred log like = 210.163
#> Running at Laplace(2.94884e-06) None  Grid-point #13 at 2.3e+11 	Fold #3 Rep #1 pred log like = 247.678
#> Running at Laplace(2.94884e-06) None  Grid-point #13 at 2.3e+11 	Fold #4 Rep #1 pred log like = 268.661
#> Running at Laplace(2.94884e-06) None  Grid-point #13 at 2.3e+11 	Fold #5 Rep #1 pred log like = 234.986
#> Running at Laplace(2.94884e-06) None  Grid-point #13 at 2.3e+11 	Fold #6 Rep #1 pred log like = 256.949
#> Running at Laplace(2.94884e-06) None  Grid-point #13 at 2.3e+11 	Fold #7 Rep #1 pred log like = 237.125
#> Running at Laplace(2.94884e-06) None  Grid-point #13 at 2.3e+11 	Fold #8 Rep #1 pred log like = 203.62
#> Running at Laplace(2.94884e-06) None  Grid-point #13 at 2.3e+11 	Fold #9 Rep #1 pred log like = 268.989
#> Running at Laplace(2.94884e-06) None  Grid-point #13 at 2.3e+11 	Fold #10 Rep #1 pred log like = 313.507
#> AvgPred = 250.948 with stdev = 30.3853
#> Completed at 2.3e+11
#> Next point at 2.3e+12 with value 0 and continue = 1
#> search[ 0.23 ] = 250.946(30.3841)
#> search[ 2.3 ] = 250.947(30.3849)
#> search[ 23 ] = 250.947(30.3852)
#> search[ 230 ] = 250.948(30.3853)
#> search[ 2300 ] = 250.948(30.3853)
#> search[ 23000 ] = 250.948(30.3853)
#> search[ 230000 ] = 250.948(30.3853)
#> search[ 2.3e+06 ] = 250.948(30.3853)
#> search[ 2.3e+07 ] = 250.948(30.3853)
#> search[ 2.3e+08 ] = 250.948(30.3853)
#> search[ 2.3e+09 ] = 250.948(30.3853)
#> search[ 2.3e+10 ] = 250.948(30.3853)
#> search[ 2.3e+11 ] = 250.948(30.3853)
#> 
#> Running at Laplace(9.32505e-07) None  Grid-point #14 at 2.3e+12 	Fold #1 Rep #1 pred log like = 267.797
#> Running at Laplace(9.32505e-07) None  Grid-point #14 at 2.3e+12 	Fold #2 Rep #1 pred log like = 210.163
#> Running at Laplace(9.32505e-07) None  Grid-point #14 at 2.3e+12 	Fold #3 Rep #1 pred log like = 247.678
#> Running at Laplace(9.32505e-07) None  Grid-point #14 at 2.3e+12 	Fold #4 Rep #1 pred log like = 268.661
#> Running at Laplace(9.32505e-07) None  Grid-point #14 at 2.3e+12 	Fold #5 Rep #1 pred log like = 234.986
#> Running at Laplace(9.32505e-07) None  Grid-point #14 at 2.3e+12 	Fold #6 Rep #1 pred log like = 256.949
#> Running at Laplace(9.32505e-07) None  Grid-point #14 at 2.3e+12 	Fold #7 Rep #1 pred log like = 237.125
#> Running at Laplace(9.32505e-07) None  Grid-point #14 at 2.3e+12 	Fold #8 Rep #1 pred log like = 203.62
#> Running at Laplace(9.32505e-07) None  Grid-point #14 at 2.3e+12 	Fold #9 Rep #1 pred log like = 268.989
#> Running at Laplace(9.32505e-07) None  Grid-point #14 at 2.3e+12 	Fold #10 Rep #1 pred log like = 313.507
#> AvgPred = 250.948 with stdev = 30.3853
#> Completed at 2.3e+12
#> Next point at 2.3e+13 with value 0 and continue = 1
#> search[ 0.23 ] = 250.946(30.3841)
#> search[ 2.3 ] = 250.947(30.3849)
#> search[ 23 ] = 250.947(30.3852)
#> search[ 230 ] = 250.948(30.3853)
#> search[ 2300 ] = 250.948(30.3853)
#> search[ 23000 ] = 250.948(30.3853)
#> search[ 230000 ] = 250.948(30.3853)
#> search[ 2.3e+06 ] = 250.948(30.3853)
#> search[ 2.3e+07 ] = 250.948(30.3853)
#> search[ 2.3e+08 ] = 250.948(30.3853)
#> search[ 2.3e+09 ] = 250.948(30.3853)
#> search[ 2.3e+10 ] = 250.948(30.3853)
#> search[ 2.3e+11 ] = 250.948(30.3853)
#> search[ 2.3e+12 ] = 250.948(30.3853)
#> 
#> Running at Laplace(2.94884e-07) None  Grid-point #15 at 2.3e+13 	Fold #1 Rep #1 pred log like = 267.797
#> Running at Laplace(2.94884e-07) None  Grid-point #15 at 2.3e+13 	Fold #2 Rep #1 pred log like = 210.163
#> Running at Laplace(2.94884e-07) None  Grid-point #15 at 2.3e+13 	Fold #3 Rep #1 pred log like = 247.678
#> Running at Laplace(2.94884e-07) None  Grid-point #15 at 2.3e+13 	Fold #4 Rep #1 pred log like = 268.661
#> Running at Laplace(2.94884e-07) None  Grid-point #15 at 2.3e+13 	Fold #5 Rep #1 pred log like = 234.986
#> Running at Laplace(2.94884e-07) None  Grid-point #15 at 2.3e+13 	Fold #6 Rep #1 pred log like = 256.949
#> Running at Laplace(2.94884e-07) None  Grid-point #15 at 2.3e+13 	Fold #7 Rep #1 pred log like = 237.125
#> Running at Laplace(2.94884e-07) None  Grid-point #15 at 2.3e+13 	Fold #8 Rep #1 pred log like = 203.62
#> Running at Laplace(2.94884e-07) None  Grid-point #15 at 2.3e+13 	Fold #9 Rep #1 pred log like = 268.989
#> Running at Laplace(2.94884e-07) None  Grid-point #15 at 2.3e+13 	Fold #10 Rep #1 pred log like = 313.507
#> AvgPred = 250.948 with stdev = 30.3853
#> Completed at 2.3e+13
#> Next point at 2.3e+14 with value 0 and continue = 1
#> search[ 0.23 ] = 250.946(30.3841)
#> search[ 2.3 ] = 250.947(30.3849)
#> search[ 23 ] = 250.947(30.3852)
#> search[ 230 ] = 250.948(30.3853)
#> search[ 2300 ] = 250.948(30.3853)
#> search[ 23000 ] = 250.948(30.3853)
#> search[ 230000 ] = 250.948(30.3853)
#> search[ 2.3e+06 ] = 250.948(30.3853)
#> search[ 2.3e+07 ] = 250.948(30.3853)
#> search[ 2.3e+08 ] = 250.948(30.3853)
#> search[ 2.3e+09 ] = 250.948(30.3853)
#> search[ 2.3e+10 ] = 250.948(30.3853)
#> search[ 2.3e+11 ] = 250.948(30.3853)
#> search[ 2.3e+12 ] = 250.948(30.3853)
#> search[ 2.3e+13 ] = 250.948(30.3853)
#> 
#> Running at Laplace(9.32505e-08) None  Grid-point #16 at 2.3e+14 	Fold #1 Rep #1 pred log like = 267.797
#> Running at Laplace(9.32505e-08) None  Grid-point #16 at 2.3e+14 	Fold #2 Rep #1 pred log like = 210.163
#> Running at Laplace(9.32505e-08) None  Grid-point #16 at 2.3e+14 	Fold #3 Rep #1 pred log like = 247.678
#> Running at Laplace(9.32505e-08) None  Grid-point #16 at 2.3e+14 	Fold #4 Rep #1 pred log like = 268.661
#> Running at Laplace(9.32505e-08) None  Grid-point #16 at 2.3e+14 	Fold #5 Rep #1 pred log like = 234.986
#> Running at Laplace(9.32505e-08) None  Grid-point #16 at 2.3e+14 	Fold #6 Rep #1 pred log like = 256.949
#> Running at Laplace(9.32505e-08) None  Grid-point #16 at 2.3e+14 	Fold #7 Rep #1 pred log like = 237.125
#> Running at Laplace(9.32505e-08) None  Grid-point #16 at 2.3e+14 	Fold #8 Rep #1 pred log like = 203.62
#> Running at Laplace(9.32505e-08) None  Grid-point #16 at 2.3e+14 	Fold #9 Rep #1 pred log like = 268.989
#> Running at Laplace(9.32505e-08) None  Grid-point #16 at 2.3e+14 	Fold #10 Rep #1 pred log like = 313.507
#> AvgPred = 250.948 with stdev = 30.3853
#> Completed at 2.3e+14
#> Next point at 2.3e+15 with value 0 and continue = 1
#> search[ 0.23 ] = 250.946(30.3841)
#> search[ 2.3 ] = 250.947(30.3849)
#> search[ 23 ] = 250.947(30.3852)
#> search[ 230 ] = 250.948(30.3853)
#> search[ 2300 ] = 250.948(30.3853)
#> search[ 23000 ] = 250.948(30.3853)
#> search[ 230000 ] = 250.948(30.3853)
#> search[ 2.3e+06 ] = 250.948(30.3853)
#> search[ 2.3e+07 ] = 250.948(30.3853)
#> search[ 2.3e+08 ] = 250.948(30.3853)
#> search[ 2.3e+09 ] = 250.948(30.3853)
#> search[ 2.3e+10 ] = 250.948(30.3853)
#> search[ 2.3e+11 ] = 250.948(30.3853)
#> search[ 2.3e+12 ] = 250.948(30.3853)
#> search[ 2.3e+13 ] = 250.948(30.3853)
#> search[ 2.3e+14 ] = 250.948(30.3853)
#> 
#> Running at Laplace(2.94884e-08) None  Grid-point #17 at 2.3e+15 	Fold #1 Rep #1 pred log like = 267.797
#> Running at Laplace(2.94884e-08) None  Grid-point #17 at 2.3e+15 	Fold #2 Rep #1 pred log like = 210.163
#> Running at Laplace(2.94884e-08) None  Grid-point #17 at 2.3e+15 	Fold #3 Rep #1 pred log like = 247.678
#> Running at Laplace(2.94884e-08) None  Grid-point #17 at 2.3e+15 	Fold #4 Rep #1 pred log like = 268.661
#> Running at Laplace(2.94884e-08) None  Grid-point #17 at 2.3e+15 	Fold #5 Rep #1 pred log like = 234.986
#> Running at Laplace(2.94884e-08) None  Grid-point #17 at 2.3e+15 	Fold #6 Rep #1 pred log like = 256.949
#> Running at Laplace(2.94884e-08) None  Grid-point #17 at 2.3e+15 	Fold #7 Rep #1 pred log like = 237.125
#> Running at Laplace(2.94884e-08) None  Grid-point #17 at 2.3e+15 	Fold #8 Rep #1 pred log like = 203.62
#> Running at Laplace(2.94884e-08) None  Grid-point #17 at 2.3e+15 	Fold #9 Rep #1 pred log like = 268.989
#> Running at Laplace(2.94884e-08) None  Grid-point #17 at 2.3e+15 	Fold #10 Rep #1 pred log like = 313.507
#> AvgPred = 250.948 with stdev = 30.3853
#> Completed at 2.3e+15
#> Next point at 2.3e+16 with value 0 and continue = 1
#> search[ 0.23 ] = 250.946(30.3841)
#> search[ 2.3 ] = 250.947(30.3849)
#> search[ 23 ] = 250.947(30.3852)
#> search[ 230 ] = 250.948(30.3853)
#> search[ 2300 ] = 250.948(30.3853)
#> search[ 23000 ] = 250.948(30.3853)
#> search[ 230000 ] = 250.948(30.3853)
#> search[ 2.3e+06 ] = 250.948(30.3853)
#> search[ 2.3e+07 ] = 250.948(30.3853)
#> search[ 2.3e+08 ] = 250.948(30.3853)
#> search[ 2.3e+09 ] = 250.948(30.3853)
#> search[ 2.3e+10 ] = 250.948(30.3853)
#> search[ 2.3e+11 ] = 250.948(30.3853)
#> search[ 2.3e+12 ] = 250.948(30.3853)
#> search[ 2.3e+13 ] = 250.948(30.3853)
#> search[ 2.3e+14 ] = 250.948(30.3853)
#> search[ 2.3e+15 ] = 250.948(30.3853)
#> 
#> Running at Laplace(9.32505e-09) None  Grid-point #18 at 2.3e+16 	Fold #1 Rep #1 pred log like = 267.797
#> Running at Laplace(9.32505e-09) None  Grid-point #18 at 2.3e+16 	Fold #2 Rep #1 pred log like = 210.163
#> Running at Laplace(9.32505e-09) None  Grid-point #18 at 2.3e+16 	Fold #3 Rep #1 pred log like = 247.678
#> Running at Laplace(9.32505e-09) None  Grid-point #18 at 2.3e+16 	Fold #4 Rep #1 pred log like = 268.661
#> Running at Laplace(9.32505e-09) None  Grid-point #18 at 2.3e+16 	Fold #5 Rep #1 pred log like = 234.986
#> Running at Laplace(9.32505e-09) None  Grid-point #18 at 2.3e+16 	Fold #6 Rep #1 pred log like = 256.949
#> Running at Laplace(9.32505e-09) None  Grid-point #18 at 2.3e+16 	Fold #7 Rep #1 pred log like = 237.125
#> Running at Laplace(9.32505e-09) None  Grid-point #18 at 2.3e+16 	Fold #8 Rep #1 pred log like = 203.62
#> Running at Laplace(9.32505e-09) None  Grid-point #18 at 2.3e+16 	Fold #9 Rep #1 pred log like = 268.989
#> Running at Laplace(9.32505e-09) None  Grid-point #18 at 2.3e+16 	Fold #10 Rep #1 pred log like = 313.507
#> AvgPred = 250.948 with stdev = 30.3853
#> Completed at 2.3e+16
#> Next point at 2.3e+17 with value 0 and continue = 1
#> search[ 0.23 ] = 250.946(30.3841)
#> search[ 2.3 ] = 250.947(30.3849)
#> search[ 23 ] = 250.947(30.3852)
#> search[ 230 ] = 250.948(30.3853)
#> search[ 2300 ] = 250.948(30.3853)
#> search[ 23000 ] = 250.948(30.3853)
#> search[ 230000 ] = 250.948(30.3853)
#> search[ 2.3e+06 ] = 250.948(30.3853)
#> search[ 2.3e+07 ] = 250.948(30.3853)
#> search[ 2.3e+08 ] = 250.948(30.3853)
#> search[ 2.3e+09 ] = 250.948(30.3853)
#> search[ 2.3e+10 ] = 250.948(30.3853)
#> search[ 2.3e+11 ] = 250.948(30.3853)
#> search[ 2.3e+12 ] = 250.948(30.3853)
#> search[ 2.3e+13 ] = 250.948(30.3853)
#> search[ 2.3e+14 ] = 250.948(30.3853)
#> search[ 2.3e+15 ] = 250.948(30.3853)
#> search[ 2.3e+16 ] = 250.948(30.3853)
#> 
#> Running at Laplace(2.94884e-09) None  Grid-point #19 at 2.3e+17 	Fold #1 Rep #1 pred log like = 267.797
#> Running at Laplace(2.94884e-09) None  Grid-point #19 at 2.3e+17 	Fold #2 Rep #1 pred log like = 210.163
#> Running at Laplace(2.94884e-09) None  Grid-point #19 at 2.3e+17 	Fold #3 Rep #1 pred log like = 247.678
#> Running at Laplace(2.94884e-09) None  Grid-point #19 at 2.3e+17 	Fold #4 Rep #1 pred log like = 268.661
#> Running at Laplace(2.94884e-09) None  Grid-point #19 at 2.3e+17 	Fold #5 Rep #1 pred log like = 234.986
#> Running at Laplace(2.94884e-09) None  Grid-point #19 at 2.3e+17 	Fold #6 Rep #1 pred log like = 256.949
#> Running at Laplace(2.94884e-09) None  Grid-point #19 at 2.3e+17 	Fold #7 Rep #1 pred log like = 237.125
#> Running at Laplace(2.94884e-09) None  Grid-point #19 at 2.3e+17 	Fold #8 Rep #1 pred log like = 203.62
#> Running at Laplace(2.94884e-09) None  Grid-point #19 at 2.3e+17 	Fold #9 Rep #1 pred log like = 268.989
#> Running at Laplace(2.94884e-09) None  Grid-point #19 at 2.3e+17 	Fold #10 Rep #1 pred log like = 313.507
#> AvgPred = 250.948 with stdev = 30.3853
#> Completed at 2.3e+17
#> Next point at 2.3e+18 with value 0 and continue = 1
#> search[ 0.23 ] = 250.946(30.3841)
#> search[ 2.3 ] = 250.947(30.3849)
#> search[ 23 ] = 250.947(30.3852)
#> search[ 230 ] = 250.948(30.3853)
#> search[ 2300 ] = 250.948(30.3853)
#> search[ 23000 ] = 250.948(30.3853)
#> search[ 230000 ] = 250.948(30.3853)
#> search[ 2.3e+06 ] = 250.948(30.3853)
#> search[ 2.3e+07 ] = 250.948(30.3853)
#> search[ 2.3e+08 ] = 250.948(30.3853)
#> search[ 2.3e+09 ] = 250.948(30.3853)
#> search[ 2.3e+10 ] = 250.948(30.3853)
#> search[ 2.3e+11 ] = 250.948(30.3853)
#> search[ 2.3e+12 ] = 250.948(30.3853)
#> search[ 2.3e+13 ] = 250.948(30.3853)
#> search[ 2.3e+14 ] = 250.948(30.3853)
#> search[ 2.3e+15 ] = 250.948(30.3853)
#> search[ 2.3e+16 ] = 250.948(30.3853)
#> search[ 2.3e+17 ] = 250.948(30.3853)
#> 
#> Running at Laplace(9.32505e-10) None  Grid-point #20 at 2.3e+18 	Fold #1 Rep #1 pred log like = 267.797
#> Running at Laplace(9.32505e-10) None  Grid-point #20 at 2.3e+18 	Fold #2 Rep #1 pred log like = 210.163
#> Running at Laplace(9.32505e-10) None  Grid-point #20 at 2.3e+18 	Fold #3 Rep #1 pred log like = 247.678
#> Running at Laplace(9.32505e-10) None  Grid-point #20 at 2.3e+18 	Fold #4 Rep #1 pred log like = 268.661
#> Running at Laplace(9.32505e-10) None  Grid-point #20 at 2.3e+18 	Fold #5 Rep #1 pred log like = 234.986
#> Running at Laplace(9.32505e-10) None  Grid-point #20 at 2.3e+18 	Fold #6 Rep #1 pred log like = 256.949
#> Running at Laplace(9.32505e-10) None  Grid-point #20 at 2.3e+18 	Fold #7 Rep #1 pred log like = 237.125
#> Running at Laplace(9.32505e-10) None  Grid-point #20 at 2.3e+18 	Fold #8 Rep #1 pred log like = 203.62
#> Running at Laplace(9.32505e-10) None  Grid-point #20 at 2.3e+18 	Fold #9 Rep #1 pred log like = 268.989
#> Running at Laplace(9.32505e-10) None  Grid-point #20 at 2.3e+18 	Fold #10 Rep #1 pred log like = 313.507
#> AvgPred = 250.948 with stdev = 30.3853
#> Completed at 2.3e+18
#> Next point at 2.3e+19 with value 0 and continue = 1
#> search[ 0.23 ] = 250.946(30.3841)
#> search[ 2.3 ] = 250.947(30.3849)
#> search[ 23 ] = 250.947(30.3852)
#> search[ 230 ] = 250.948(30.3853)
#> search[ 2300 ] = 250.948(30.3853)
#> search[ 23000 ] = 250.948(30.3853)
#> search[ 230000 ] = 250.948(30.3853)
#> search[ 2.3e+06 ] = 250.948(30.3853)
#> search[ 2.3e+07 ] = 250.948(30.3853)
#> search[ 2.3e+08 ] = 250.948(30.3853)
#> search[ 2.3e+09 ] = 250.948(30.3853)
#> search[ 2.3e+10 ] = 250.948(30.3853)
#> search[ 2.3e+11 ] = 250.948(30.3853)
#> search[ 2.3e+12 ] = 250.948(30.3853)
#> search[ 2.3e+13 ] = 250.948(30.3853)
#> search[ 2.3e+14 ] = 250.948(30.3853)
#> search[ 2.3e+15 ] = 250.948(30.3853)
#> search[ 2.3e+16 ] = 250.948(30.3853)
#> search[ 2.3e+17 ] = 250.948(30.3853)
#> search[ 2.3e+18 ] = 250.948(30.3853)
#> 
#> Running at Laplace(2.94884e-10) None  Grid-point #21 at 2.3e+19 	Fold #1 Rep #1 pred log like = 267.797
#> Running at Laplace(2.94884e-10) None  Grid-point #21 at 2.3e+19 	Fold #2 Rep #1 pred log like = 210.163
#> Running at Laplace(2.94884e-10) None  Grid-point #21 at 2.3e+19 	Fold #3 Rep #1 pred log like = 247.678
#> Running at Laplace(2.94884e-10) None  Grid-point #21 at 2.3e+19 	Fold #4 Rep #1 pred log like = 268.661
#> Running at Laplace(2.94884e-10) None  Grid-point #21 at 2.3e+19 	Fold #5 Rep #1 pred log like = 234.986
#> Running at Laplace(2.94884e-10) None  Grid-point #21 at 2.3e+19 	Fold #6 Rep #1 pred log like = 256.949
#> Running at Laplace(2.94884e-10) None  Grid-point #21 at 2.3e+19 	Fold #7 Rep #1 pred log like = 237.125
#> Running at Laplace(2.94884e-10) None  Grid-point #21 at 2.3e+19 	Fold #8 Rep #1 pred log like = 203.62
#> Running at Laplace(2.94884e-10) None  Grid-point #21 at 2.3e+19 	Fold #9 Rep #1 pred log like = 268.989
#> Running at Laplace(2.94884e-10) None  Grid-point #21 at 2.3e+19 	Fold #10 Rep #1 pred log like = 313.507
#> AvgPred = 250.948 with stdev = 30.3853
#> Completed at 2.3e+19
#> Next point at 2.3e+20 with value 0 and continue = 1
#> search[ 0.23 ] = 250.946(30.3841)
#> search[ 2.3 ] = 250.947(30.3849)
#> search[ 23 ] = 250.947(30.3852)
#> search[ 230 ] = 250.948(30.3853)
#> search[ 2300 ] = 250.948(30.3853)
#> search[ 23000 ] = 250.948(30.3853)
#> search[ 230000 ] = 250.948(30.3853)
#> search[ 2.3e+06 ] = 250.948(30.3853)
#> search[ 2.3e+07 ] = 250.948(30.3853)
#> search[ 2.3e+08 ] = 250.948(30.3853)
#> search[ 2.3e+09 ] = 250.948(30.3853)
#> search[ 2.3e+10 ] = 250.948(30.3853)
#> search[ 2.3e+11 ] = 250.948(30.3853)
#> search[ 2.3e+12 ] = 250.948(30.3853)
#> search[ 2.3e+13 ] = 250.948(30.3853)
#> search[ 2.3e+14 ] = 250.948(30.3853)
#> search[ 2.3e+15 ] = 250.948(30.3853)
#> search[ 2.3e+16 ] = 250.948(30.3853)
#> search[ 2.3e+17 ] = 250.948(30.3853)
#> search[ 2.3e+18 ] = 250.948(30.3853)
#> search[ 2.3e+19 ] = 250.948(30.3853)
#> 
#> Running at Laplace(9.32505e-11) None  Grid-point #22 at 2.3e+20 	Fold #1 Rep #1 pred log like = 267.797
#> Running at Laplace(9.32505e-11) None  Grid-point #22 at 2.3e+20 	Fold #2 Rep #1 pred log like = 210.163
#> Running at Laplace(9.32505e-11) None  Grid-point #22 at 2.3e+20 	Fold #3 Rep #1 pred log like = 247.678
#> Running at Laplace(9.32505e-11) None  Grid-point #22 at 2.3e+20 	Fold #4 Rep #1 pred log like = 268.661
#> Running at Laplace(9.32505e-11) None  Grid-point #22 at 2.3e+20 	Fold #5 Rep #1 pred log like = 234.986
#> Running at Laplace(9.32505e-11) None  Grid-point #22 at 2.3e+20 	Fold #6 Rep #1 pred log like = 256.949
#> Running at Laplace(9.32505e-11) None  Grid-point #22 at 2.3e+20 	Fold #7 Rep #1 pred log like = 237.125
#> Running at Laplace(9.32505e-11) None  Grid-point #22 at 2.3e+20 	Fold #8 Rep #1 pred log like = 203.62
#> Running at Laplace(9.32505e-11) None  Grid-point #22 at 2.3e+20 	Fold #9 Rep #1 pred log like = 268.989
#> Running at Laplace(9.32505e-11) None  Grid-point #22 at 2.3e+20 	Fold #10 Rep #1 pred log like = 313.507
#> AvgPred = 250.948 with stdev = 30.3853
#> Completed at 2.3e+20
#> Next point at 2.41831e+12 with value 250.948 and continue = 0
#> search[ 0.23 ] = 250.946(30.3841)
#> search[ 2.3 ] = 250.947(30.3849)
#> search[ 23 ] = 250.947(30.3852)
#> search[ 230 ] = 250.948(30.3853)
#> search[ 2300 ] = 250.948(30.3853)
#> search[ 23000 ] = 250.948(30.3853)
#> search[ 230000 ] = 250.948(30.3853)
#> search[ 2.3e+06 ] = 250.948(30.3853)
#> search[ 2.3e+07 ] = 250.948(30.3853)
#> search[ 2.3e+08 ] = 250.948(30.3853)
#> search[ 2.3e+09 ] = 250.948(30.3853)
#> search[ 2.3e+10 ] = 250.948(30.3853)
#> search[ 2.3e+11 ] = 250.948(30.3853)
#> search[ 2.3e+12 ] = 250.948(30.3853)
#> search[ 2.3e+13 ] = 250.948(30.3853)
#> search[ 2.3e+14 ] = 250.948(30.3853)
#> search[ 2.3e+15 ] = 250.948(30.3853)
#> search[ 2.3e+16 ] = 250.948(30.3853)
#> search[ 2.3e+17 ] = 250.948(30.3853)
#> search[ 2.3e+18 ] = 250.948(30.3853)
#> search[ 2.3e+19 ] = 250.948(30.3853)
#> search[ 2.3e+20 ] = 250.948(30.3853)
#> 
#> 
#> Maximum predicted log likelihood (250.948) estimated at:
#> 	2.41831e+12 (variance)
#> 	9.09409e-07 (lambda)
#> 
#> Fitting model at optimal hyperparameter
#> Using prior: Laplace(9.09409e-07) None 

#Find out what the optimal hyperparameter was:
getHyperParameter(fit)
#> [1] 2.418308e+12

#Extract the current log-likelihood, and coefficients
logLik(fit)
#> 'log Lik.' -1953.765 (df=3)
coef(fit)
#> (Intercept)           1           2 
#> -4.19783727  0.30746087  0.02797239 

#We can only retrieve the confidence interval for unregularized coefficients:
confint(fit, c(0))
#> Using 1 thread(s)
#>             covariate     2.5 %    97.5 % evaluations
#> (Intercept)         0 -4.236526 -4.159602          22