Confidence intervals for Cyclops model parameters

confinit.cyclopsFit profiles the data likelihood to construct confidence intervals of arbitrary level. Usually it only makes sense to do this for variables that have not been regularized TODO: Profile data likelihood or joint distribution of remaining parameters.

# S3 method for cyclopsFit
confint(object, parm, level = 0.95,
  overrideNoRegularization = FALSE, includePenalty = TRUE,
  rescale = FALSE, ...)

Arguments

object
A fitted Cyclops model object

parm
A specification of which parameters require confidence intervals, either a vector of numbers of covariateId names

level
Numeric: confidence level required

overrideNoRegularization
Logical: Enable confidence interval estimation for regularized parameters

includePenalty
Logical: Include regularized covariate penalty in profile

rescale
Boolean: rescale coefficients for unnormalized covariate values

...
Additional argument(s) for methods

Value

A matrix with columns reporting lower and upper confidence limits for each parameter. These columns are labelled as (1-level) / 2 and 1 - (1 - level) / 2 in percent (by default 2.5 percent and 97.5 percent)

Examples

#Generate some simulated data:
sim <- simulateCyclopsData(nstrata = 1, nrows = 1000, ncovars = 2, eCovarsPerRow = 0.5,
                           model = "poisson")

#> Sparseness = 74.45 %

cyclopsData <- convertToCyclopsData(sim$outcomes, sim$covariates, modelType = "pr",
                                    addIntercept = TRUE)

#> Sorting covariates by covariateId and rowId

#Define the prior and control objects to use cross-validation for finding the 
#optimal hyperparameter:
prior <- createPrior("laplace", exclude = 0, useCrossValidation = TRUE)
control <- createControl(cvType = "auto", noiseLevel = "quiet")

#Fit the model
fit <- fitCyclopsModel(cyclopsData,prior = prior, control = control)

#> Using cross-validation selector type byRow
#> Performing 10-fold cross-validation [seed = 1543260547] with data partitions of sizes 100 100 100 100 100 100 100 100 100 100
#> Using 1 thread(s)
#> Starting var = 0.2555 (default)
#> Running at Laplace(2.79782) None  Grid-point #1 at 0.2555 	Fold #1 Rep #1 pred log like = -206.637
#> Running at Laplace(2.79782) None  Grid-point #1 at 0.2555 	Fold #2 Rep #1 pred log like = -223.818
#> Running at Laplace(2.79782) None  Grid-point #1 at 0.2555 	Fold #3 Rep #1 pred log like = -197.02
#> Running at Laplace(2.79782) None  Grid-point #1 at 0.2555 	Fold #4 Rep #1 pred log like = -190.623
#> Running at Laplace(2.79782) None  Grid-point #1 at 0.2555 	Fold #5 Rep #1 pred log like = -201.684
#> Running at Laplace(2.79782) None  Grid-point #1 at 0.2555 	Fold #6 Rep #1 pred log like = -198.136
#> Running at Laplace(2.79782) None  Grid-point #1 at 0.2555 	Fold #7 Rep #1 pred log like = -179.472
#> Running at Laplace(2.79782) None  Grid-point #1 at 0.2555 	Fold #8 Rep #1 pred log like = -185.297
#> Running at Laplace(2.79782) None  Grid-point #1 at 0.2555 	Fold #9 Rep #1 pred log like = -198.634
#> Running at Laplace(2.79782) None  Grid-point #1 at 0.2555 	Fold #10 Rep #1 pred log like = -193.643
#> AvgPred = -197.497 with stdev = 11.5396
#> Completed at 0.2555
#> Next point at 2.555 with value 0 and continue = 1
#> search[ 0.2555 ] = -197.497(11.5396)
#> 
#> Running at Laplace(0.884748) None  Grid-point #2 at 2.555 	Fold #1 Rep #1 pred log like = -206.631
#> Running at Laplace(0.884748) None  Grid-point #2 at 2.555 	Fold #2 Rep #1 pred log like = -223.872
#> Running at Laplace(0.884748) None  Grid-point #2 at 2.555 	Fold #3 Rep #1 pred log like = -197.047
#> Running at Laplace(0.884748) None  Grid-point #2 at 2.555 	Fold #4 Rep #1 pred log like = -190.624
#> Running at Laplace(0.884748) None  Grid-point #2 at 2.555 	Fold #5 Rep #1 pred log like = -201.693
#> Running at Laplace(0.884748) None  Grid-point #2 at 2.555 	Fold #6 Rep #1 pred log like = -198.173
#> Running at Laplace(0.884748) None  Grid-point #2 at 2.555 	Fold #7 Rep #1 pred log like = -179.506
#> Running at Laplace(0.884748) None  Grid-point #2 at 2.555 	Fold #8 Rep #1 pred log like = -185.349
#> Running at Laplace(0.884748) None  Grid-point #2 at 2.555 	Fold #9 Rep #1 pred log like = -198.65
#> Running at Laplace(0.884748) None  Grid-point #2 at 2.555 	Fold #10 Rep #1 pred log like = -193.65
#> AvgPred = -197.519 with stdev = 11.541
#> Completed at 2.555
#> Next point at 0.02555 with value 0 and continue = 1
#> search[ 0.2555 ] = -197.497(11.5396)
#> search[ 2.555 ] = -197.519(11.541)
#> 
#> Running at Laplace(8.84748) None  Grid-point #3 at 0.02555 	Fold #1 Rep #1 pred log like = -206.645
#> Running at Laplace(8.84748) None  Grid-point #3 at 0.02555 	Fold #2 Rep #1 pred log like = -223.655
#> Running at Laplace(8.84748) None  Grid-point #3 at 0.02555 	Fold #3 Rep #1 pred log like = -196.938
#> Running at Laplace(8.84748) None  Grid-point #3 at 0.02555 	Fold #4 Rep #1 pred log like = -190.618
#> Running at Laplace(8.84748) None  Grid-point #3 at 0.02555 	Fold #5 Rep #1 pred log like = -201.676
#> Running at Laplace(8.84748) None  Grid-point #3 at 0.02555 	Fold #6 Rep #1 pred log like = -198.089
#> Running at Laplace(8.84748) None  Grid-point #3 at 0.02555 	Fold #7 Rep #1 pred log like = -179.393
#> Running at Laplace(8.84748) None  Grid-point #3 at 0.02555 	Fold #8 Rep #1 pred log like = -185.198
#> Running at Laplace(8.84748) None  Grid-point #3 at 0.02555 	Fold #9 Rep #1 pred log like = -198.588
#> Running at Laplace(8.84748) None  Grid-point #3 at 0.02555 	Fold #10 Rep #1 pred log like = -193.643
#> AvgPred = -197.444 with stdev = 11.5256
#> Completed at 0.02555
#> Next point at 0.002555 with value 0 and continue = 1
#> search[ 0.02555 ] = -197.444(11.5256)
#> search[ 0.2555 ] = -197.497(11.5396)
#> search[ 2.555 ] = -197.519(11.541)
#> 
#> Running at Laplace(27.9782) None  Grid-point #4 at 0.002555 	Fold #1 Rep #1 pred log like = -206.645
#> Running at Laplace(27.9782) None  Grid-point #4 at 0.002555 	Fold #2 Rep #1 pred log like = -223.525
#> Running at Laplace(27.9782) None  Grid-point #4 at 0.002555 	Fold #3 Rep #1 pred log like = -196.843
#> Running at Laplace(27.9782) None  Grid-point #4 at 0.002555 	Fold #4 Rep #1 pred log like = -190.618
#> Running at Laplace(27.9782) None  Grid-point #4 at 0.002555 	Fold #5 Rep #1 pred log like = -201.676
#> Running at Laplace(27.9782) None  Grid-point #4 at 0.002555 	Fold #6 Rep #1 pred log like = -198.089
#> Running at Laplace(27.9782) None  Grid-point #4 at 0.002555 	Fold #7 Rep #1 pred log like = -179.376
#> Running at Laplace(27.9782) None  Grid-point #4 at 0.002555 	Fold #8 Rep #1 pred log like = -185.157
#> Running at Laplace(27.9782) None  Grid-point #4 at 0.002555 	Fold #9 Rep #1 pred log like = -198.56
#> Running at Laplace(27.9782) None  Grid-point #4 at 0.002555 	Fold #10 Rep #1 pred log like = -193.643
#> AvgPred = -197.413 with stdev = 11.5032
#> Completed at 0.002555
#> Next point at 0.0002555 with value 0 and continue = 1
#> search[ 0.002555 ] = -197.413(11.5032)
#> search[ 0.02555 ] = -197.444(11.5256)
#> search[ 0.2555 ] = -197.497(11.5396)
#> search[ 2.555 ] = -197.519(11.541)
#> 
#> Running at Laplace(88.4748) None  Grid-point #5 at 0.0002555 	Fold #1 Rep #1 pred log like = -206.645
#> Running at Laplace(88.4748) None  Grid-point #5 at 0.0002555 	Fold #2 Rep #1 pred log like = -223.525
#> Running at Laplace(88.4748) None  Grid-point #5 at 0.0002555 	Fold #3 Rep #1 pred log like = -196.843
#> Running at Laplace(88.4748) None  Grid-point #5 at 0.0002555 	Fold #4 Rep #1 pred log like = -190.618
#> Running at Laplace(88.4748) None  Grid-point #5 at 0.0002555 	Fold #5 Rep #1 pred log like = -201.676
#> Running at Laplace(88.4748) None  Grid-point #5 at 0.0002555 	Fold #6 Rep #1 pred log like = -198.089
#> Running at Laplace(88.4748) None  Grid-point #5 at 0.0002555 	Fold #7 Rep #1 pred log like = -179.376
#> Running at Laplace(88.4748) None  Grid-point #5 at 0.0002555 	Fold #8 Rep #1 pred log like = -185.157
#> Running at Laplace(88.4748) None  Grid-point #5 at 0.0002555 	Fold #9 Rep #1 pred log like = -198.56
#> Running at Laplace(88.4748) None  Grid-point #5 at 0.0002555 	Fold #10 Rep #1 pred log like = -193.643
#> AvgPred = -197.413 with stdev = 11.5032
#> Completed at 0.0002555
#> Next point at 2.555e-005 with value 0 and continue = 1
#> search[ 0.0002555 ] = -197.413(11.5032)
#> search[ 0.002555 ] = -197.413(11.5032)
#> search[ 0.02555 ] = -197.444(11.5256)
#> search[ 0.2555 ] = -197.497(11.5396)
#> search[ 2.555 ] = -197.519(11.541)
#> 
#> Running at Laplace(279.782) None  Grid-point #6 at 2.555e-005 	Fold #1 Rep #1 pred log like = -206.645
#> Running at Laplace(279.782) None  Grid-point #6 at 2.555e-005 	Fold #2 Rep #1 pred log like = -223.525
#> Running at Laplace(279.782) None  Grid-point #6 at 2.555e-005 	Fold #3 Rep #1 pred log like = -196.843
#> Running at Laplace(279.782) None  Grid-point #6 at 2.555e-005 	Fold #4 Rep #1 pred log like = -190.618
#> Running at Laplace(279.782) None  Grid-point #6 at 2.555e-005 	Fold #5 Rep #1 pred log like = -201.676
#> Running at Laplace(279.782) None  Grid-point #6 at 2.555e-005 	Fold #6 Rep #1 pred log like = -198.089
#> Running at Laplace(279.782) None  Grid-point #6 at 2.555e-005 	Fold #7 Rep #1 pred log like = -179.376
#> Running at Laplace(279.782) None  Grid-point #6 at 2.555e-005 	Fold #8 Rep #1 pred log like = -185.157
#> Running at Laplace(279.782) None  Grid-point #6 at 2.555e-005 	Fold #9 Rep #1 pred log like = -198.56
#> Running at Laplace(279.782) None  Grid-point #6 at 2.555e-005 	Fold #10 Rep #1 pred log like = -193.643
#> AvgPred = -197.413 with stdev = 11.5032
#> Completed at 2.555e-005
#> Next point at 7.92081e-005 with value -197.41 and continue = 0
#> search[ 2.555e-005 ] = -197.413(11.5032)
#> search[ 0.0002555 ] = -197.413(11.5032)
#> search[ 0.002555 ] = -197.413(11.5032)
#> search[ 0.02555 ] = -197.444(11.5256)
#> search[ 0.2555 ] = -197.497(11.5396)
#> search[ 2.555 ] = -197.519(11.541)
#> 
#> 
#> Maximum predicted log likelihood (-197.41) estimated at:
#> 	7.92081e-005 (variance)
#> 	158.902 (lambda)
#> 
#> Fitting model at optimal hyperparameter
#> Using prior: Laplace(158.902) None 

#Find out what the optimal hyperparameter was:
getHyperParameter(fit)

#> [1] 7.920808e-05

#Extract the current log-likelihood, and coefficients
logLik(fit)

#> 'log Lik.' -1973.343 (df=3)

coef(fit)

#> (Intercept)           1           2 
#>   -3.937029    0.000000    0.000000 

#We can only retrieve the confidence interval for unregularized coefficients:
confint(fit, c(0))

#> Using 1 thread(s)

#>      covariate     2.5 %   97.5 % evaluations
#> [1,]         0 -3.965462 -3.90885          24