createAutoGridCrossValidationControl {Cyclops}	R Documentation

Create a Cyclops control object that supports multiple hyperparameters

Description

createCrossValidationControl creates a Cyclops control object for use with fitCyclopsModel that supports multiple hyperparameters through an auto-search in one dimension and a grid-search over the remaining dimensions

Usage

createAutoGridCrossValidationControl(
  outerGrid,
  autoPosition = 1,
  refitAtMaximum = TRUE,
  cvType = "auto",
  initialValue = 1,
  ...
)

Arguments

outerGrid
List or data.frame of grid parameters to explore

autoPosition
Vector position for auto-search parameter (concatenated into outerGrid)

refitAtMaximum
Logical: re-fit Cyclops object at maximal cross-validation parameters

cvType
Must equal "auto"

initialValue
Initial value for auto-search parameter

...
Additional parameters passed through to createControl

Value

A Cyclops prior object of class inheriting from "cyclopsPrior" and "cyclopsFunctionalPrior" for use with fitCyclopsModel.

[Package Cyclops version 3.2.0 Index]