createParameterizedPrior {Cyclops}	R Documentation

Create a Cyclops parameterized prior object

Description

createParameterizedPrior creates a Cyclops prior object for use with fitCyclopsModel in which arbitrary R functions parameterize the prior location and variance.

Usage

createParameterizedPrior(
  priorType,
  parameterize,
  values,
  useCrossValidation = FALSE,
  forceIntercept = FALSE
)

Arguments

priorType
Character vector: specifies prior distribution. See below for options

parameterize
Function list: parameterizes location and variance

values
Numeric vector: initial parameter values

useCrossValidation
Logical: Perform cross-validation to determine parameters.

forceIntercept
Logical: Force intercept coefficient into prior

Value

A Cyclops prior object of class inheriting from "cyclopsPrior" and "cyclopsFunctionalPrior" for use with fitCyclopsModel.

[Package Cyclops version 3.2.1 Index]