Create a Cyclops prior object

createPrior creates a Cyclops prior object for use with fitCyclopsModel.

createPrior(
  priorType,
  variance = 1,
  exclude = c(),
  graph = NULL,
  neighborhood = NULL,
  useCrossValidation = FALSE,
  forceIntercept = FALSE
)

Arguments

priorType
Character: specifies prior distribution. See below for options

variance
Numeric: prior distribution variance

exclude
A vector of numbers or covariateId names to exclude from prior

graph
Child-to-parent mapping for a hierarchical prior

neighborhood
A list of first-order neighborhoods for a partially fused prior

useCrossValidation
Logical: Perform cross-validation to determine prior variance.

forceIntercept
Logical: Force intercept coefficient into prior

Value

A Cyclops prior object of class inheriting from "cyclopsPrior" for use with fitCyclopsModel.

Prior types

We specify all priors in terms of their variance parameters. Similar fitting tools for regularized regression often parameterize the Laplace distribution in terms of a rate "lambda" per observation. See "glmnet", for example.

variance = 2 * / (nobs * lambda)^2 or lambda = sqrt(2 / variance) / nobs

Examples

#Generate some simulated data:
sim <- simulateCyclopsData(nstrata = 1, nrows = 1000, ncovars = 2, eCovarsPerRow = 0.5,
                           model = "poisson")

#> Sparseness = 75.8 %

cyclopsData <- convertToCyclopsData(sim$outcomes, sim$covariates, modelType = "pr",
                                    addIntercept = TRUE)

#> Sorting covariates by covariateId and rowId

#Define the prior and control objects to use cross-validation for finding the 
#optimal hyperparameter:
prior <- createPrior("laplace", exclude = 0, useCrossValidation = TRUE)
control <- createControl(cvType = "auto", noiseLevel = "quiet")

#Fit the model
fit <- fitCyclopsModel(cyclopsData,prior = prior, control = control)

#> Using cross-validation selector type byRow
#> Performing 10-fold cross-validation [seed = 1591596883] with data partitions of sizes 100 100 100 100 100 100 100 100 100 100
#> Using 1 thread(s)
#> Starting var = 0.242 (default)
#> Running at Laplace(2.8748) None  Grid-point #1 at 0.242 	Fold #1 Rep #1 pred log like = 830.493
#> Running at Laplace(2.8748) None  Grid-point #1 at 0.242 	Fold #2 Rep #1 pred log like = 865.958
#> Running at Laplace(2.8748) None  Grid-point #1 at 0.242 	Fold #3 Rep #1 pred log like = 744.579
#> Running at Laplace(2.8748) None  Grid-point #1 at 0.242 	Fold #4 Rep #1 pred log like = 747.384
#> Running at Laplace(2.8748) None  Grid-point #1 at 0.242 	Fold #5 Rep #1 pred log like = 700.417
#> Running at Laplace(2.8748) None  Grid-point #1 at 0.242 	Fold #6 Rep #1 pred log like = 882.724
#> Running at Laplace(2.8748) None  Grid-point #1 at 0.242 	Fold #7 Rep #1 pred log like = 717.863
#> Running at Laplace(2.8748) None  Grid-point #1 at 0.242 	Fold #8 Rep #1 pred log like = 786.315
#> Running at Laplace(2.8748) None  Grid-point #1 at 0.242 	Fold #9 Rep #1 pred log like = 684.439
#> Running at Laplace(2.8748) None  Grid-point #1 at 0.242 	Fold #10 Rep #1 pred log like = 872.989
#> AvgPred = 783.316 with stdev = 71.2127
#> Completed at 0.242
#> Next point at 2.42 with value 0 and continue = 1
#> search[ 0.242 ] = 783.316(71.2127)
#> 
#> Running at Laplace(0.909091) None  Grid-point #2 at 2.42 	Fold #1 Rep #1 pred log like = 830.565
#> Running at Laplace(0.909091) None  Grid-point #2 at 2.42 	Fold #2 Rep #1 pred log like = 865.988
#> Running at Laplace(0.909091) None  Grid-point #2 at 2.42 	Fold #3 Rep #1 pred log like = 744.572
#> Running at Laplace(0.909091) None  Grid-point #2 at 2.42 	Fold #4 Rep #1 pred log like = 747.324
#> Running at Laplace(0.909091) None  Grid-point #2 at 2.42 	Fold #5 Rep #1 pred log like = 700.383
#> Running at Laplace(0.909091) None  Grid-point #2 at 2.42 	Fold #6 Rep #1 pred log like = 882.721
#> Running at Laplace(0.909091) None  Grid-point #2 at 2.42 	Fold #7 Rep #1 pred log like = 717.857
#> Running at Laplace(0.909091) None  Grid-point #2 at 2.42 	Fold #8 Rep #1 pred log like = 786.321
#> Running at Laplace(0.909091) None  Grid-point #2 at 2.42 	Fold #9 Rep #1 pred log like = 684.432
#> Running at Laplace(0.909091) None  Grid-point #2 at 2.42 	Fold #10 Rep #1 pred log like = 873.008
#> AvgPred = 783.317 with stdev = 71.2317
#> Completed at 2.42
#> Next point at 24.2 with value 0 and continue = 1
#> search[ 0.242 ] = 783.316(71.2127)
#> search[ 2.42 ] = 783.317(71.2317)
#> 
#> Running at Laplace(0.28748) None  Grid-point #3 at 24.2 	Fold #1 Rep #1 pred log like = 830.587
#> Running at Laplace(0.28748) None  Grid-point #3 at 24.2 	Fold #2 Rep #1 pred log like = 865.997
#> Running at Laplace(0.28748) None  Grid-point #3 at 24.2 	Fold #3 Rep #1 pred log like = 744.57
#> Running at Laplace(0.28748) None  Grid-point #3 at 24.2 	Fold #4 Rep #1 pred log like = 747.305
#> Running at Laplace(0.28748) None  Grid-point #3 at 24.2 	Fold #5 Rep #1 pred log like = 700.372
#> Running at Laplace(0.28748) None  Grid-point #3 at 24.2 	Fold #6 Rep #1 pred log like = 882.72
#> Running at Laplace(0.28748) None  Grid-point #3 at 24.2 	Fold #7 Rep #1 pred log like = 717.855
#> Running at Laplace(0.28748) None  Grid-point #3 at 24.2 	Fold #8 Rep #1 pred log like = 786.323
#> Running at Laplace(0.28748) None  Grid-point #3 at 24.2 	Fold #9 Rep #1 pred log like = 684.429
#> Running at Laplace(0.28748) None  Grid-point #3 at 24.2 	Fold #10 Rep #1 pred log like = 873.014
#> AvgPred = 783.317 with stdev = 71.2377
#> Completed at 24.2
#> Next point at 242 with value 0 and continue = 1
#> search[ 0.242 ] = 783.316(71.2127)
#> search[ 2.42 ] = 783.317(71.2317)
#> search[ 24.2 ] = 783.317(71.2377)
#> 
#> Running at Laplace(0.0909091) None  Grid-point #4 at 242 	Fold #1 Rep #1 pred log like = 830.594
#> Running at Laplace(0.0909091) None  Grid-point #4 at 242 	Fold #2 Rep #1 pred log like = 866
#> Running at Laplace(0.0909091) None  Grid-point #4 at 242 	Fold #3 Rep #1 pred log like = 744.569
#> Running at Laplace(0.0909091) None  Grid-point #4 at 242 	Fold #4 Rep #1 pred log like = 747.299
#> Running at Laplace(0.0909091) None  Grid-point #4 at 242 	Fold #5 Rep #1 pred log like = 700.369
#> Running at Laplace(0.0909091) None  Grid-point #4 at 242 	Fold #6 Rep #1 pred log like = 882.72
#> Running at Laplace(0.0909091) None  Grid-point #4 at 242 	Fold #7 Rep #1 pred log like = 717.855
#> Running at Laplace(0.0909091) None  Grid-point #4 at 242 	Fold #8 Rep #1 pred log like = 786.323
#> Running at Laplace(0.0909091) None  Grid-point #4 at 242 	Fold #9 Rep #1 pred log like = 684.428
#> Running at Laplace(0.0909091) None  Grid-point #4 at 242 	Fold #10 Rep #1 pred log like = 873.016
#> AvgPred = 783.317 with stdev = 71.2396
#> Completed at 242
#> Next point at 2420 with value 0 and continue = 1
#> search[ 0.242 ] = 783.316(71.2127)
#> search[ 2.42 ] = 783.317(71.2317)
#> search[ 24.2 ] = 783.317(71.2377)
#> search[ 242 ] = 783.317(71.2396)
#> 
#> Running at Laplace(0.028748) None  Grid-point #5 at 2420 	Fold #1 Rep #1 pred log like = 830.596
#> Running at Laplace(0.028748) None  Grid-point #5 at 2420 	Fold #2 Rep #1 pred log like = 866.001
#> Running at Laplace(0.028748) None  Grid-point #5 at 2420 	Fold #3 Rep #1 pred log like = 744.569
#> Running at Laplace(0.028748) None  Grid-point #5 at 2420 	Fold #4 Rep #1 pred log like = 747.297
#> Running at Laplace(0.028748) None  Grid-point #5 at 2420 	Fold #5 Rep #1 pred log like = 700.368
#> Running at Laplace(0.028748) None  Grid-point #5 at 2420 	Fold #6 Rep #1 pred log like = 882.72
#> Running at Laplace(0.028748) None  Grid-point #5 at 2420 	Fold #7 Rep #1 pred log like = 717.854
#> Running at Laplace(0.028748) None  Grid-point #5 at 2420 	Fold #8 Rep #1 pred log like = 786.323
#> Running at Laplace(0.028748) None  Grid-point #5 at 2420 	Fold #9 Rep #1 pred log like = 684.428
#> Running at Laplace(0.028748) None  Grid-point #5 at 2420 	Fold #10 Rep #1 pred log like = 873.016
#> AvgPred = 783.317 with stdev = 71.2402
#> Completed at 2420
#> Next point at 24200 with value 0 and continue = 1
#> search[ 0.242 ] = 783.316(71.2127)
#> search[ 2.42 ] = 783.317(71.2317)
#> search[ 24.2 ] = 783.317(71.2377)
#> search[ 242 ] = 783.317(71.2396)
#> search[ 2420 ] = 783.317(71.2402)
#> 
#> Running at Laplace(0.00909091) None  Grid-point #6 at 24200 	Fold #1 Rep #1 pred log like = 830.597
#> Running at Laplace(0.00909091) None  Grid-point #6 at 24200 	Fold #2 Rep #1 pred log like = 866.001
#> Running at Laplace(0.00909091) None  Grid-point #6 at 24200 	Fold #3 Rep #1 pred log like = 744.569
#> Running at Laplace(0.00909091) None  Grid-point #6 at 24200 	Fold #4 Rep #1 pred log like = 747.297
#> Running at Laplace(0.00909091) None  Grid-point #6 at 24200 	Fold #5 Rep #1 pred log like = 700.368
#> Running at Laplace(0.00909091) None  Grid-point #6 at 24200 	Fold #6 Rep #1 pred log like = 882.719
#> Running at Laplace(0.00909091) None  Grid-point #6 at 24200 	Fold #7 Rep #1 pred log like = 717.854
#> Running at Laplace(0.00909091) None  Grid-point #6 at 24200 	Fold #8 Rep #1 pred log like = 786.323
#> Running at Laplace(0.00909091) None  Grid-point #6 at 24200 	Fold #9 Rep #1 pred log like = 684.428
#> Running at Laplace(0.00909091) None  Grid-point #6 at 24200 	Fold #10 Rep #1 pred log like = 873.017
#> AvgPred = 783.317 with stdev = 71.2403
#> Completed at 24200
#> Next point at 242000 with value 0 and continue = 1
#> search[ 0.242 ] = 783.316(71.2127)
#> search[ 2.42 ] = 783.317(71.2317)
#> search[ 24.2 ] = 783.317(71.2377)
#> search[ 242 ] = 783.317(71.2396)
#> search[ 2420 ] = 783.317(71.2402)
#> search[ 24200 ] = 783.317(71.2403)
#> 
#> Running at Laplace(0.0028748) None  Grid-point #7 at 242000 	Fold #1 Rep #1 pred log like = 830.597
#> Running at Laplace(0.0028748) None  Grid-point #7 at 242000 	Fold #2 Rep #1 pred log like = 866.001
#> Running at Laplace(0.0028748) None  Grid-point #7 at 242000 	Fold #3 Rep #1 pred log like = 744.569
#> Running at Laplace(0.0028748) None  Grid-point #7 at 242000 	Fold #4 Rep #1 pred log like = 747.296
#> Running at Laplace(0.0028748) None  Grid-point #7 at 242000 	Fold #5 Rep #1 pred log like = 700.368
#> Running at Laplace(0.0028748) None  Grid-point #7 at 242000 	Fold #6 Rep #1 pred log like = 882.719
#> Running at Laplace(0.0028748) None  Grid-point #7 at 242000 	Fold #7 Rep #1 pred log like = 717.854
#> Running at Laplace(0.0028748) None  Grid-point #7 at 242000 	Fold #8 Rep #1 pred log like = 786.323
#> Running at Laplace(0.0028748) None  Grid-point #7 at 242000 	Fold #9 Rep #1 pred log like = 684.428
#> Running at Laplace(0.0028748) None  Grid-point #7 at 242000 	Fold #10 Rep #1 pred log like = 873.017
#> AvgPred = 783.317 with stdev = 71.2404
#> Completed at 242000
#> Next point at 2.42e+06 with value 0 and continue = 1
#> search[ 0.242 ] = 783.316(71.2127)
#> search[ 2.42 ] = 783.317(71.2317)
#> search[ 24.2 ] = 783.317(71.2377)
#> search[ 242 ] = 783.317(71.2396)
#> search[ 2420 ] = 783.317(71.2402)
#> search[ 24200 ] = 783.317(71.2403)
#> search[ 242000 ] = 783.317(71.2404)
#> 
#> Running at Laplace(0.000909091) None  Grid-point #8 at 2.42e+06 	Fold #1 Rep #1 pred log like = 830.597
#> Running at Laplace(0.000909091) None  Grid-point #8 at 2.42e+06 	Fold #2 Rep #1 pred log like = 866.001
#> Running at Laplace(0.000909091) None  Grid-point #8 at 2.42e+06 	Fold #3 Rep #1 pred log like = 744.569
#> Running at Laplace(0.000909091) None  Grid-point #8 at 2.42e+06 	Fold #4 Rep #1 pred log like = 747.296
#> Running at Laplace(0.000909091) None  Grid-point #8 at 2.42e+06 	Fold #5 Rep #1 pred log like = 700.367
#> Running at Laplace(0.000909091) None  Grid-point #8 at 2.42e+06 	Fold #6 Rep #1 pred log like = 882.719
#> Running at Laplace(0.000909091) None  Grid-point #8 at 2.42e+06 	Fold #7 Rep #1 pred log like = 717.854
#> Running at Laplace(0.000909091) None  Grid-point #8 at 2.42e+06 	Fold #8 Rep #1 pred log like = 786.323
#> Running at Laplace(0.000909091) None  Grid-point #8 at 2.42e+06 	Fold #9 Rep #1 pred log like = 684.428
#> Running at Laplace(0.000909091) None  Grid-point #8 at 2.42e+06 	Fold #10 Rep #1 pred log like = 873.017
#> AvgPred = 783.317 with stdev = 71.2404
#> Completed at 2.42e+06
#> Next point at 2.42e+07 with value 0 and continue = 1
#> search[ 0.242 ] = 783.316(71.2127)
#> search[ 2.42 ] = 783.317(71.2317)
#> search[ 24.2 ] = 783.317(71.2377)
#> search[ 242 ] = 783.317(71.2396)
#> search[ 2420 ] = 783.317(71.2402)
#> search[ 24200 ] = 783.317(71.2403)
#> search[ 242000 ] = 783.317(71.2404)
#> search[ 2.42e+06 ] = 783.317(71.2404)
#> 
#> Running at Laplace(0.00028748) None  Grid-point #9 at 2.42e+07 	Fold #1 Rep #1 pred log like = 830.598
#> Running at Laplace(0.00028748) None  Grid-point #9 at 2.42e+07 	Fold #2 Rep #1 pred log like = 866.001
#> Running at Laplace(0.00028748) None  Grid-point #9 at 2.42e+07 	Fold #3 Rep #1 pred log like = 744.569
#> Running at Laplace(0.00028748) None  Grid-point #9 at 2.42e+07 	Fold #4 Rep #1 pred log like = 747.296
#> Running at Laplace(0.00028748) None  Grid-point #9 at 2.42e+07 	Fold #5 Rep #1 pred log like = 700.367
#> Running at Laplace(0.00028748) None  Grid-point #9 at 2.42e+07 	Fold #6 Rep #1 pred log like = 882.719
#> Running at Laplace(0.00028748) None  Grid-point #9 at 2.42e+07 	Fold #7 Rep #1 pred log like = 717.854
#> Running at Laplace(0.00028748) None  Grid-point #9 at 2.42e+07 	Fold #8 Rep #1 pred log like = 786.323
#> Running at Laplace(0.00028748) None  Grid-point #9 at 2.42e+07 	Fold #9 Rep #1 pred log like = 684.428
#> Running at Laplace(0.00028748) None  Grid-point #9 at 2.42e+07 	Fold #10 Rep #1 pred log like = 873.017
#> AvgPred = 783.317 with stdev = 71.2404
#> Completed at 2.42e+07
#> Next point at 2.42e+08 with value 0 and continue = 1
#> search[ 0.242 ] = 783.316(71.2127)
#> search[ 2.42 ] = 783.317(71.2317)
#> search[ 24.2 ] = 783.317(71.2377)
#> search[ 242 ] = 783.317(71.2396)
#> search[ 2420 ] = 783.317(71.2402)
#> search[ 24200 ] = 783.317(71.2403)
#> search[ 242000 ] = 783.317(71.2404)
#> search[ 2.42e+06 ] = 783.317(71.2404)
#> search[ 2.42e+07 ] = 783.317(71.2404)
#> 
#> Running at Laplace(9.09091e-05) None  Grid-point #10 at 2.42e+08 	Fold #1 Rep #1 pred log like = 830.598
#> Running at Laplace(9.09091e-05) None  Grid-point #10 at 2.42e+08 	Fold #2 Rep #1 pred log like = 866.001
#> Running at Laplace(9.09091e-05) None  Grid-point #10 at 2.42e+08 	Fold #3 Rep #1 pred log like = 744.569
#> Running at Laplace(9.09091e-05) None  Grid-point #10 at 2.42e+08 	Fold #4 Rep #1 pred log like = 747.296
#> Running at Laplace(9.09091e-05) None  Grid-point #10 at 2.42e+08 	Fold #5 Rep #1 pred log like = 700.367
#> Running at Laplace(9.09091e-05) None  Grid-point #10 at 2.42e+08 	Fold #6 Rep #1 pred log like = 882.719
#> Running at Laplace(9.09091e-05) None  Grid-point #10 at 2.42e+08 	Fold #7 Rep #1 pred log like = 717.854
#> Running at Laplace(9.09091e-05) None  Grid-point #10 at 2.42e+08 	Fold #8 Rep #1 pred log like = 786.323
#> Running at Laplace(9.09091e-05) None  Grid-point #10 at 2.42e+08 	Fold #9 Rep #1 pred log like = 684.428
#> Running at Laplace(9.09091e-05) None  Grid-point #10 at 2.42e+08 	Fold #10 Rep #1 pred log like = 873.017
#> AvgPred = 783.317 with stdev = 71.2404
#> Completed at 2.42e+08
#> Next point at 2.42e+09 with value 0 and continue = 1
#> search[ 0.242 ] = 783.316(71.2127)
#> search[ 2.42 ] = 783.317(71.2317)
#> search[ 24.2 ] = 783.317(71.2377)
#> search[ 242 ] = 783.317(71.2396)
#> search[ 2420 ] = 783.317(71.2402)
#> search[ 24200 ] = 783.317(71.2403)
#> search[ 242000 ] = 783.317(71.2404)
#> search[ 2.42e+06 ] = 783.317(71.2404)
#> search[ 2.42e+07 ] = 783.317(71.2404)
#> search[ 2.42e+08 ] = 783.317(71.2404)
#> 
#> Running at Laplace(2.8748e-05) None  Grid-point #11 at 2.42e+09 	Fold #1 Rep #1 pred log like = 830.598
#> Running at Laplace(2.8748e-05) None  Grid-point #11 at 2.42e+09 	Fold #2 Rep #1 pred log like = 866.001
#> Running at Laplace(2.8748e-05) None  Grid-point #11 at 2.42e+09 	Fold #3 Rep #1 pred log like = 744.569
#> Running at Laplace(2.8748e-05) None  Grid-point #11 at 2.42e+09 	Fold #4 Rep #1 pred log like = 747.296
#> Running at Laplace(2.8748e-05) None  Grid-point #11 at 2.42e+09 	Fold #5 Rep #1 pred log like = 700.367
#> Running at Laplace(2.8748e-05) None  Grid-point #11 at 2.42e+09 	Fold #6 Rep #1 pred log like = 882.719
#> Running at Laplace(2.8748e-05) None  Grid-point #11 at 2.42e+09 	Fold #7 Rep #1 pred log like = 717.854
#> Running at Laplace(2.8748e-05) None  Grid-point #11 at 2.42e+09 	Fold #8 Rep #1 pred log like = 786.323
#> Running at Laplace(2.8748e-05) None  Grid-point #11 at 2.42e+09 	Fold #9 Rep #1 pred log like = 684.428
#> Running at Laplace(2.8748e-05) None  Grid-point #11 at 2.42e+09 	Fold #10 Rep #1 pred log like = 873.017
#> AvgPred = 783.317 with stdev = 71.2404
#> Completed at 2.42e+09
#> Next point at 2.42e+10 with value 0 and continue = 1
#> search[ 0.242 ] = 783.316(71.2127)
#> search[ 2.42 ] = 783.317(71.2317)
#> search[ 24.2 ] = 783.317(71.2377)
#> search[ 242 ] = 783.317(71.2396)
#> search[ 2420 ] = 783.317(71.2402)
#> search[ 24200 ] = 783.317(71.2403)
#> search[ 242000 ] = 783.317(71.2404)
#> search[ 2.42e+06 ] = 783.317(71.2404)
#> search[ 2.42e+07 ] = 783.317(71.2404)
#> search[ 2.42e+08 ] = 783.317(71.2404)
#> search[ 2.42e+09 ] = 783.317(71.2404)
#> 
#> Running at Laplace(9.09091e-06) None  Grid-point #12 at 2.42e+10 	Fold #1 Rep #1 pred log like = 830.598
#> Running at Laplace(9.09091e-06) None  Grid-point #12 at 2.42e+10 	Fold #2 Rep #1 pred log like = 866.001
#> Running at Laplace(9.09091e-06) None  Grid-point #12 at 2.42e+10 	Fold #3 Rep #1 pred log like = 744.569
#> Running at Laplace(9.09091e-06) None  Grid-point #12 at 2.42e+10 	Fold #4 Rep #1 pred log like = 747.296
#> Running at Laplace(9.09091e-06) None  Grid-point #12 at 2.42e+10 	Fold #5 Rep #1 pred log like = 700.367
#> Running at Laplace(9.09091e-06) None  Grid-point #12 at 2.42e+10 	Fold #6 Rep #1 pred log like = 882.719
#> Running at Laplace(9.09091e-06) None  Grid-point #12 at 2.42e+10 	Fold #7 Rep #1 pred log like = 717.854
#> Running at Laplace(9.09091e-06) None  Grid-point #12 at 2.42e+10 	Fold #8 Rep #1 pred log like = 786.323
#> Running at Laplace(9.09091e-06) None  Grid-point #12 at 2.42e+10 	Fold #9 Rep #1 pred log like = 684.428
#> Running at Laplace(9.09091e-06) None  Grid-point #12 at 2.42e+10 	Fold #10 Rep #1 pred log like = 873.017
#> AvgPred = 783.317 with stdev = 71.2404
#> Completed at 2.42e+10
#> Next point at 2.42e+11 with value 0 and continue = 1
#> search[ 0.242 ] = 783.316(71.2127)
#> search[ 2.42 ] = 783.317(71.2317)
#> search[ 24.2 ] = 783.317(71.2377)
#> search[ 242 ] = 783.317(71.2396)
#> search[ 2420 ] = 783.317(71.2402)
#> search[ 24200 ] = 783.317(71.2403)
#> search[ 242000 ] = 783.317(71.2404)
#> search[ 2.42e+06 ] = 783.317(71.2404)
#> search[ 2.42e+07 ] = 783.317(71.2404)
#> search[ 2.42e+08 ] = 783.317(71.2404)
#> search[ 2.42e+09 ] = 783.317(71.2404)
#> search[ 2.42e+10 ] = 783.317(71.2404)
#> 
#> Running at Laplace(2.8748e-06) None  Grid-point #13 at 2.42e+11 	Fold #1 Rep #1 pred log like = 830.598
#> Running at Laplace(2.8748e-06) None  Grid-point #13 at 2.42e+11 	Fold #2 Rep #1 pred log like = 866.001
#> Running at Laplace(2.8748e-06) None  Grid-point #13 at 2.42e+11 	Fold #3 Rep #1 pred log like = 744.569
#> Running at Laplace(2.8748e-06) None  Grid-point #13 at 2.42e+11 	Fold #4 Rep #1 pred log like = 747.296
#> Running at Laplace(2.8748e-06) None  Grid-point #13 at 2.42e+11 	Fold #5 Rep #1 pred log like = 700.367
#> Running at Laplace(2.8748e-06) None  Grid-point #13 at 2.42e+11 	Fold #6 Rep #1 pred log like = 882.719
#> Running at Laplace(2.8748e-06) None  Grid-point #13 at 2.42e+11 	Fold #7 Rep #1 pred log like = 717.854
#> Running at Laplace(2.8748e-06) None  Grid-point #13 at 2.42e+11 	Fold #8 Rep #1 pred log like = 786.323
#> Running at Laplace(2.8748e-06) None  Grid-point #13 at 2.42e+11 	Fold #9 Rep #1 pred log like = 684.428
#> Running at Laplace(2.8748e-06) None  Grid-point #13 at 2.42e+11 	Fold #10 Rep #1 pred log like = 873.017
#> AvgPred = 783.317 with stdev = 71.2404
#> Completed at 2.42e+11
#> Next point at 2.42e+12 with value 0 and continue = 1
#> search[ 0.242 ] = 783.316(71.2127)
#> search[ 2.42 ] = 783.317(71.2317)
#> search[ 24.2 ] = 783.317(71.2377)
#> search[ 242 ] = 783.317(71.2396)
#> search[ 2420 ] = 783.317(71.2402)
#> search[ 24200 ] = 783.317(71.2403)
#> search[ 242000 ] = 783.317(71.2404)
#> search[ 2.42e+06 ] = 783.317(71.2404)
#> search[ 2.42e+07 ] = 783.317(71.2404)
#> search[ 2.42e+08 ] = 783.317(71.2404)
#> search[ 2.42e+09 ] = 783.317(71.2404)
#> search[ 2.42e+10 ] = 783.317(71.2404)
#> search[ 2.42e+11 ] = 783.317(71.2404)
#> 
#> Running at Laplace(9.09091e-07) None  Grid-point #14 at 2.42e+12 	Fold #1 Rep #1 pred log like = 830.598
#> Running at Laplace(9.09091e-07) None  Grid-point #14 at 2.42e+12 	Fold #2 Rep #1 pred log like = 866.001
#> Running at Laplace(9.09091e-07) None  Grid-point #14 at 2.42e+12 	Fold #3 Rep #1 pred log like = 744.569
#> Running at Laplace(9.09091e-07) None  Grid-point #14 at 2.42e+12 	Fold #4 Rep #1 pred log like = 747.296
#> Running at Laplace(9.09091e-07) None  Grid-point #14 at 2.42e+12 	Fold #5 Rep #1 pred log like = 700.367
#> Running at Laplace(9.09091e-07) None  Grid-point #14 at 2.42e+12 	Fold #6 Rep #1 pred log like = 882.719
#> Running at Laplace(9.09091e-07) None  Grid-point #14 at 2.42e+12 	Fold #7 Rep #1 pred log like = 717.854
#> Running at Laplace(9.09091e-07) None  Grid-point #14 at 2.42e+12 	Fold #8 Rep #1 pred log like = 786.323
#> Running at Laplace(9.09091e-07) None  Grid-point #14 at 2.42e+12 	Fold #9 Rep #1 pred log like = 684.428
#> Running at Laplace(9.09091e-07) None  Grid-point #14 at 2.42e+12 	Fold #10 Rep #1 pred log like = 873.017
#> AvgPred = 783.317 with stdev = 71.2404
#> Completed at 2.42e+12
#> Next point at 2.42e+13 with value 0 and continue = 1
#> search[ 0.242 ] = 783.316(71.2127)
#> search[ 2.42 ] = 783.317(71.2317)
#> search[ 24.2 ] = 783.317(71.2377)
#> search[ 242 ] = 783.317(71.2396)
#> search[ 2420 ] = 783.317(71.2402)
#> search[ 24200 ] = 783.317(71.2403)
#> search[ 242000 ] = 783.317(71.2404)
#> search[ 2.42e+06 ] = 783.317(71.2404)
#> search[ 2.42e+07 ] = 783.317(71.2404)
#> search[ 2.42e+08 ] = 783.317(71.2404)
#> search[ 2.42e+09 ] = 783.317(71.2404)
#> search[ 2.42e+10 ] = 783.317(71.2404)
#> search[ 2.42e+11 ] = 783.317(71.2404)
#> search[ 2.42e+12 ] = 783.317(71.2404)
#> 
#> Running at Laplace(2.8748e-07) None  Grid-point #15 at 2.42e+13 	Fold #1 Rep #1 pred log like = 830.598
#> Running at Laplace(2.8748e-07) None  Grid-point #15 at 2.42e+13 	Fold #2 Rep #1 pred log like = 866.001
#> Running at Laplace(2.8748e-07) None  Grid-point #15 at 2.42e+13 	Fold #3 Rep #1 pred log like = 744.569
#> Running at Laplace(2.8748e-07) None  Grid-point #15 at 2.42e+13 	Fold #4 Rep #1 pred log like = 747.296
#> Running at Laplace(2.8748e-07) None  Grid-point #15 at 2.42e+13 	Fold #5 Rep #1 pred log like = 700.367
#> Running at Laplace(2.8748e-07) None  Grid-point #15 at 2.42e+13 	Fold #6 Rep #1 pred log like = 882.719
#> Running at Laplace(2.8748e-07) None  Grid-point #15 at 2.42e+13 	Fold #7 Rep #1 pred log like = 717.854
#> Running at Laplace(2.8748e-07) None  Grid-point #15 at 2.42e+13 	Fold #8 Rep #1 pred log like = 786.323
#> Running at Laplace(2.8748e-07) None  Grid-point #15 at 2.42e+13 	Fold #9 Rep #1 pred log like = 684.428
#> Running at Laplace(2.8748e-07) None  Grid-point #15 at 2.42e+13 	Fold #10 Rep #1 pred log like = 873.017
#> AvgPred = 783.317 with stdev = 71.2404
#> Completed at 2.42e+13
#> Next point at 2.42e+14 with value 0 and continue = 1
#> search[ 0.242 ] = 783.316(71.2127)
#> search[ 2.42 ] = 783.317(71.2317)
#> search[ 24.2 ] = 783.317(71.2377)
#> search[ 242 ] = 783.317(71.2396)
#> search[ 2420 ] = 783.317(71.2402)
#> search[ 24200 ] = 783.317(71.2403)
#> search[ 242000 ] = 783.317(71.2404)
#> search[ 2.42e+06 ] = 783.317(71.2404)
#> search[ 2.42e+07 ] = 783.317(71.2404)
#> search[ 2.42e+08 ] = 783.317(71.2404)
#> search[ 2.42e+09 ] = 783.317(71.2404)
#> search[ 2.42e+10 ] = 783.317(71.2404)
#> search[ 2.42e+11 ] = 783.317(71.2404)
#> search[ 2.42e+12 ] = 783.317(71.2404)
#> search[ 2.42e+13 ] = 783.317(71.2404)
#> 
#> Running at Laplace(9.09091e-08) None  Grid-point #16 at 2.42e+14 	Fold #1 Rep #1 pred log like = 830.598
#> Running at Laplace(9.09091e-08) None  Grid-point #16 at 2.42e+14 	Fold #2 Rep #1 pred log like = 866.001
#> Running at Laplace(9.09091e-08) None  Grid-point #16 at 2.42e+14 	Fold #3 Rep #1 pred log like = 744.569
#> Running at Laplace(9.09091e-08) None  Grid-point #16 at 2.42e+14 	Fold #4 Rep #1 pred log like = 747.296
#> Running at Laplace(9.09091e-08) None  Grid-point #16 at 2.42e+14 	Fold #5 Rep #1 pred log like = 700.367
#> Running at Laplace(9.09091e-08) None  Grid-point #16 at 2.42e+14 	Fold #6 Rep #1 pred log like = 882.719
#> Running at Laplace(9.09091e-08) None  Grid-point #16 at 2.42e+14 	Fold #7 Rep #1 pred log like = 717.854
#> Running at Laplace(9.09091e-08) None  Grid-point #16 at 2.42e+14 	Fold #8 Rep #1 pred log like = 786.323
#> Running at Laplace(9.09091e-08) None  Grid-point #16 at 2.42e+14 	Fold #9 Rep #1 pred log like = 684.428
#> Running at Laplace(9.09091e-08) None  Grid-point #16 at 2.42e+14 	Fold #10 Rep #1 pred log like = 873.017
#> AvgPred = 783.317 with stdev = 71.2404
#> Completed at 2.42e+14
#> Next point at 2.42e+15 with value 0 and continue = 1
#> search[ 0.242 ] = 783.316(71.2127)
#> search[ 2.42 ] = 783.317(71.2317)
#> search[ 24.2 ] = 783.317(71.2377)
#> search[ 242 ] = 783.317(71.2396)
#> search[ 2420 ] = 783.317(71.2402)
#> search[ 24200 ] = 783.317(71.2403)
#> search[ 242000 ] = 783.317(71.2404)
#> search[ 2.42e+06 ] = 783.317(71.2404)
#> search[ 2.42e+07 ] = 783.317(71.2404)
#> search[ 2.42e+08 ] = 783.317(71.2404)
#> search[ 2.42e+09 ] = 783.317(71.2404)
#> search[ 2.42e+10 ] = 783.317(71.2404)
#> search[ 2.42e+11 ] = 783.317(71.2404)
#> search[ 2.42e+12 ] = 783.317(71.2404)
#> search[ 2.42e+13 ] = 783.317(71.2404)
#> search[ 2.42e+14 ] = 783.317(71.2404)
#> 
#> Running at Laplace(2.8748e-08) None  Grid-point #17 at 2.42e+15 	Fold #1 Rep #1 pred log like = 830.598
#> Running at Laplace(2.8748e-08) None  Grid-point #17 at 2.42e+15 	Fold #2 Rep #1 pred log like = 866.001
#> Running at Laplace(2.8748e-08) None  Grid-point #17 at 2.42e+15 	Fold #3 Rep #1 pred log like = 744.569
#> Running at Laplace(2.8748e-08) None  Grid-point #17 at 2.42e+15 	Fold #4 Rep #1 pred log like = 747.296
#> Running at Laplace(2.8748e-08) None  Grid-point #17 at 2.42e+15 	Fold #5 Rep #1 pred log like = 700.367
#> Running at Laplace(2.8748e-08) None  Grid-point #17 at 2.42e+15 	Fold #6 Rep #1 pred log like = 882.719
#> Running at Laplace(2.8748e-08) None  Grid-point #17 at 2.42e+15 	Fold #7 Rep #1 pred log like = 717.854
#> Running at Laplace(2.8748e-08) None  Grid-point #17 at 2.42e+15 	Fold #8 Rep #1 pred log like = 786.323
#> Running at Laplace(2.8748e-08) None  Grid-point #17 at 2.42e+15 	Fold #9 Rep #1 pred log like = 684.428
#> Running at Laplace(2.8748e-08) None  Grid-point #17 at 2.42e+15 	Fold #10 Rep #1 pred log like = 873.017
#> AvgPred = 783.317 with stdev = 71.2404
#> Completed at 2.42e+15
#> Next point at 2.42e+16 with value 0 and continue = 1
#> search[ 0.242 ] = 783.316(71.2127)
#> search[ 2.42 ] = 783.317(71.2317)
#> search[ 24.2 ] = 783.317(71.2377)
#> search[ 242 ] = 783.317(71.2396)
#> search[ 2420 ] = 783.317(71.2402)
#> search[ 24200 ] = 783.317(71.2403)
#> search[ 242000 ] = 783.317(71.2404)
#> search[ 2.42e+06 ] = 783.317(71.2404)
#> search[ 2.42e+07 ] = 783.317(71.2404)
#> search[ 2.42e+08 ] = 783.317(71.2404)
#> search[ 2.42e+09 ] = 783.317(71.2404)
#> search[ 2.42e+10 ] = 783.317(71.2404)
#> search[ 2.42e+11 ] = 783.317(71.2404)
#> search[ 2.42e+12 ] = 783.317(71.2404)
#> search[ 2.42e+13 ] = 783.317(71.2404)
#> search[ 2.42e+14 ] = 783.317(71.2404)
#> search[ 2.42e+15 ] = 783.317(71.2404)
#> 
#> Running at Laplace(9.09091e-09) None  Grid-point #18 at 2.42e+16 	Fold #1 Rep #1 pred log like = 830.598
#> Running at Laplace(9.09091e-09) None  Grid-point #18 at 2.42e+16 	Fold #2 Rep #1 pred log like = 866.001
#> Running at Laplace(9.09091e-09) None  Grid-point #18 at 2.42e+16 	Fold #3 Rep #1 pred log like = 744.569
#> Running at Laplace(9.09091e-09) None  Grid-point #18 at 2.42e+16 	Fold #4 Rep #1 pred log like = 747.296
#> Running at Laplace(9.09091e-09) None  Grid-point #18 at 2.42e+16 	Fold #5 Rep #1 pred log like = 700.367
#> Running at Laplace(9.09091e-09) None  Grid-point #18 at 2.42e+16 	Fold #6 Rep #1 pred log like = 882.719
#> Running at Laplace(9.09091e-09) None  Grid-point #18 at 2.42e+16 	Fold #7 Rep #1 pred log like = 717.854
#> Running at Laplace(9.09091e-09) None  Grid-point #18 at 2.42e+16 	Fold #8 Rep #1 pred log like = 786.323
#> Running at Laplace(9.09091e-09) None  Grid-point #18 at 2.42e+16 	Fold #9 Rep #1 pred log like = 684.428
#> Running at Laplace(9.09091e-09) None  Grid-point #18 at 2.42e+16 	Fold #10 Rep #1 pred log like = 873.017
#> AvgPred = 783.317 with stdev = 71.2404
#> Completed at 2.42e+16
#> Next point at 2.42e+17 with value 0 and continue = 1
#> search[ 0.242 ] = 783.316(71.2127)
#> search[ 2.42 ] = 783.317(71.2317)
#> search[ 24.2 ] = 783.317(71.2377)
#> search[ 242 ] = 783.317(71.2396)
#> search[ 2420 ] = 783.317(71.2402)
#> search[ 24200 ] = 783.317(71.2403)
#> search[ 242000 ] = 783.317(71.2404)
#> search[ 2.42e+06 ] = 783.317(71.2404)
#> search[ 2.42e+07 ] = 783.317(71.2404)
#> search[ 2.42e+08 ] = 783.317(71.2404)
#> search[ 2.42e+09 ] = 783.317(71.2404)
#> search[ 2.42e+10 ] = 783.317(71.2404)
#> search[ 2.42e+11 ] = 783.317(71.2404)
#> search[ 2.42e+12 ] = 783.317(71.2404)
#> search[ 2.42e+13 ] = 783.317(71.2404)
#> search[ 2.42e+14 ] = 783.317(71.2404)
#> search[ 2.42e+15 ] = 783.317(71.2404)
#> search[ 2.42e+16 ] = 783.317(71.2404)
#> 
#> Running at Laplace(2.8748e-09) None  Grid-point #19 at 2.42e+17 	Fold #1 Rep #1 pred log like = 830.598
#> Running at Laplace(2.8748e-09) None  Grid-point #19 at 2.42e+17 	Fold #2 Rep #1 pred log like = 866.001
#> Running at Laplace(2.8748e-09) None  Grid-point #19 at 2.42e+17 	Fold #3 Rep #1 pred log like = 744.569
#> Running at Laplace(2.8748e-09) None  Grid-point #19 at 2.42e+17 	Fold #4 Rep #1 pred log like = 747.296
#> Running at Laplace(2.8748e-09) None  Grid-point #19 at 2.42e+17 	Fold #5 Rep #1 pred log like = 700.367
#> Running at Laplace(2.8748e-09) None  Grid-point #19 at 2.42e+17 	Fold #6 Rep #1 pred log like = 882.719
#> Running at Laplace(2.8748e-09) None  Grid-point #19 at 2.42e+17 	Fold #7 Rep #1 pred log like = 717.854
#> Running at Laplace(2.8748e-09) None  Grid-point #19 at 2.42e+17 	Fold #8 Rep #1 pred log like = 786.323
#> Running at Laplace(2.8748e-09) None  Grid-point #19 at 2.42e+17 	Fold #9 Rep #1 pred log like = 684.428
#> Running at Laplace(2.8748e-09) None  Grid-point #19 at 2.42e+17 	Fold #10 Rep #1 pred log like = 873.017
#> AvgPred = 783.317 with stdev = 71.2404
#> Completed at 2.42e+17
#> Next point at 2.42e+18 with value 0 and continue = 1
#> search[ 0.242 ] = 783.316(71.2127)
#> search[ 2.42 ] = 783.317(71.2317)
#> search[ 24.2 ] = 783.317(71.2377)
#> search[ 242 ] = 783.317(71.2396)
#> search[ 2420 ] = 783.317(71.2402)
#> search[ 24200 ] = 783.317(71.2403)
#> search[ 242000 ] = 783.317(71.2404)
#> search[ 2.42e+06 ] = 783.317(71.2404)
#> search[ 2.42e+07 ] = 783.317(71.2404)
#> search[ 2.42e+08 ] = 783.317(71.2404)
#> search[ 2.42e+09 ] = 783.317(71.2404)
#> search[ 2.42e+10 ] = 783.317(71.2404)
#> search[ 2.42e+11 ] = 783.317(71.2404)
#> search[ 2.42e+12 ] = 783.317(71.2404)
#> search[ 2.42e+13 ] = 783.317(71.2404)
#> search[ 2.42e+14 ] = 783.317(71.2404)
#> search[ 2.42e+15 ] = 783.317(71.2404)
#> search[ 2.42e+16 ] = 783.317(71.2404)
#> search[ 2.42e+17 ] = 783.317(71.2404)
#> 
#> Running at Laplace(9.09091e-10) None  Grid-point #20 at 2.42e+18 	Fold #1 Rep #1 pred log like = 830.598
#> Running at Laplace(9.09091e-10) None  Grid-point #20 at 2.42e+18 	Fold #2 Rep #1 pred log like = 866.001
#> Running at Laplace(9.09091e-10) None  Grid-point #20 at 2.42e+18 	Fold #3 Rep #1 pred log like = 744.569
#> Running at Laplace(9.09091e-10) None  Grid-point #20 at 2.42e+18 	Fold #4 Rep #1 pred log like = 747.296
#> Running at Laplace(9.09091e-10) None  Grid-point #20 at 2.42e+18 	Fold #5 Rep #1 pred log like = 700.367
#> Running at Laplace(9.09091e-10) None  Grid-point #20 at 2.42e+18 	Fold #6 Rep #1 pred log like = 882.719
#> Running at Laplace(9.09091e-10) None  Grid-point #20 at 2.42e+18 	Fold #7 Rep #1 pred log like = 717.854
#> Running at Laplace(9.09091e-10) None  Grid-point #20 at 2.42e+18 	Fold #8 Rep #1 pred log like = 786.323
#> Running at Laplace(9.09091e-10) None  Grid-point #20 at 2.42e+18 	Fold #9 Rep #1 pred log like = 684.428
#> Running at Laplace(9.09091e-10) None  Grid-point #20 at 2.42e+18 	Fold #10 Rep #1 pred log like = 873.017
#> AvgPred = 783.317 with stdev = 71.2404
#> Completed at 2.42e+18
#> Next point at 2.42e+19 with value 0 and continue = 1
#> search[ 0.242 ] = 783.316(71.2127)
#> search[ 2.42 ] = 783.317(71.2317)
#> search[ 24.2 ] = 783.317(71.2377)
#> search[ 242 ] = 783.317(71.2396)
#> search[ 2420 ] = 783.317(71.2402)
#> search[ 24200 ] = 783.317(71.2403)
#> search[ 242000 ] = 783.317(71.2404)
#> search[ 2.42e+06 ] = 783.317(71.2404)
#> search[ 2.42e+07 ] = 783.317(71.2404)
#> search[ 2.42e+08 ] = 783.317(71.2404)
#> search[ 2.42e+09 ] = 783.317(71.2404)
#> search[ 2.42e+10 ] = 783.317(71.2404)
#> search[ 2.42e+11 ] = 783.317(71.2404)
#> search[ 2.42e+12 ] = 783.317(71.2404)
#> search[ 2.42e+13 ] = 783.317(71.2404)
#> search[ 2.42e+14 ] = 783.317(71.2404)
#> search[ 2.42e+15 ] = 783.317(71.2404)
#> search[ 2.42e+16 ] = 783.317(71.2404)
#> search[ 2.42e+17 ] = 783.317(71.2404)
#> search[ 2.42e+18 ] = 783.317(71.2404)
#> 
#> Running at Laplace(2.8748e-10) None  Grid-point #21 at 2.42e+19 	Fold #1 Rep #1 pred log like = 830.598
#> Running at Laplace(2.8748e-10) None  Grid-point #21 at 2.42e+19 	Fold #2 Rep #1 pred log like = 866.001
#> Running at Laplace(2.8748e-10) None  Grid-point #21 at 2.42e+19 	Fold #3 Rep #1 pred log like = 744.569
#> Running at Laplace(2.8748e-10) None  Grid-point #21 at 2.42e+19 	Fold #4 Rep #1 pred log like = 747.296
#> Running at Laplace(2.8748e-10) None  Grid-point #21 at 2.42e+19 	Fold #5 Rep #1 pred log like = 700.367
#> Running at Laplace(2.8748e-10) None  Grid-point #21 at 2.42e+19 	Fold #6 Rep #1 pred log like = 882.719
#> Running at Laplace(2.8748e-10) None  Grid-point #21 at 2.42e+19 	Fold #7 Rep #1 pred log like = 717.854
#> Running at Laplace(2.8748e-10) None  Grid-point #21 at 2.42e+19 	Fold #8 Rep #1 pred log like = 786.323
#> Running at Laplace(2.8748e-10) None  Grid-point #21 at 2.42e+19 	Fold #9 Rep #1 pred log like = 684.428
#> Running at Laplace(2.8748e-10) None  Grid-point #21 at 2.42e+19 	Fold #10 Rep #1 pred log like = 873.017
#> AvgPred = 783.317 with stdev = 71.2404
#> Completed at 2.42e+19
#> Next point at 2.42e+20 with value 0 and continue = 1
#> search[ 0.242 ] = 783.316(71.2127)
#> search[ 2.42 ] = 783.317(71.2317)
#> search[ 24.2 ] = 783.317(71.2377)
#> search[ 242 ] = 783.317(71.2396)
#> search[ 2420 ] = 783.317(71.2402)
#> search[ 24200 ] = 783.317(71.2403)
#> search[ 242000 ] = 783.317(71.2404)
#> search[ 2.42e+06 ] = 783.317(71.2404)
#> search[ 2.42e+07 ] = 783.317(71.2404)
#> search[ 2.42e+08 ] = 783.317(71.2404)
#> search[ 2.42e+09 ] = 783.317(71.2404)
#> search[ 2.42e+10 ] = 783.317(71.2404)
#> search[ 2.42e+11 ] = 783.317(71.2404)
#> search[ 2.42e+12 ] = 783.317(71.2404)
#> search[ 2.42e+13 ] = 783.317(71.2404)
#> search[ 2.42e+14 ] = 783.317(71.2404)
#> search[ 2.42e+15 ] = 783.317(71.2404)
#> search[ 2.42e+16 ] = 783.317(71.2404)
#> search[ 2.42e+17 ] = 783.317(71.2404)
#> search[ 2.42e+18 ] = 783.317(71.2404)
#> search[ 2.42e+19 ] = 783.317(71.2404)
#> 
#> Running at Laplace(9.09091e-11) None  Grid-point #22 at 2.42e+20 	Fold #1 Rep #1 pred log like = 830.598
#> Running at Laplace(9.09091e-11) None  Grid-point #22 at 2.42e+20 	Fold #2 Rep #1 pred log like = 866.001
#> Running at Laplace(9.09091e-11) None  Grid-point #22 at 2.42e+20 	Fold #3 Rep #1 pred log like = 744.569
#> Running at Laplace(9.09091e-11) None  Grid-point #22 at 2.42e+20 	Fold #4 Rep #1 pred log like = 747.296
#> Running at Laplace(9.09091e-11) None  Grid-point #22 at 2.42e+20 	Fold #5 Rep #1 pred log like = 700.367
#> Running at Laplace(9.09091e-11) None  Grid-point #22 at 2.42e+20 	Fold #6 Rep #1 pred log like = 882.719
#> Running at Laplace(9.09091e-11) None  Grid-point #22 at 2.42e+20 	Fold #7 Rep #1 pred log like = 717.854
#> Running at Laplace(9.09091e-11) None  Grid-point #22 at 2.42e+20 	Fold #8 Rep #1 pred log like = 786.323
#> Running at Laplace(9.09091e-11) None  Grid-point #22 at 2.42e+20 	Fold #9 Rep #1 pred log like = 684.428
#> Running at Laplace(9.09091e-11) None  Grid-point #22 at 2.42e+20 	Fold #10 Rep #1 pred log like = 873.017
#> AvgPred = 783.317 with stdev = 71.2404
#> Completed at 2.42e+20
#> Next point at 2.46184e+12 with value 783.317 and continue = 0
#> search[ 0.242 ] = 783.316(71.2127)
#> search[ 2.42 ] = 783.317(71.2317)
#> search[ 24.2 ] = 783.317(71.2377)
#> search[ 242 ] = 783.317(71.2396)
#> search[ 2420 ] = 783.317(71.2402)
#> search[ 24200 ] = 783.317(71.2403)
#> search[ 242000 ] = 783.317(71.2404)
#> search[ 2.42e+06 ] = 783.317(71.2404)
#> search[ 2.42e+07 ] = 783.317(71.2404)
#> search[ 2.42e+08 ] = 783.317(71.2404)
#> search[ 2.42e+09 ] = 783.317(71.2404)
#> search[ 2.42e+10 ] = 783.317(71.2404)
#> search[ 2.42e+11 ] = 783.317(71.2404)
#> search[ 2.42e+12 ] = 783.317(71.2404)
#> search[ 2.42e+13 ] = 783.317(71.2404)
#> search[ 2.42e+14 ] = 783.317(71.2404)
#> search[ 2.42e+15 ] = 783.317(71.2404)
#> search[ 2.42e+16 ] = 783.317(71.2404)
#> search[ 2.42e+17 ] = 783.317(71.2404)
#> search[ 2.42e+18 ] = 783.317(71.2404)
#> search[ 2.42e+19 ] = 783.317(71.2404)
#> search[ 2.42e+20 ] = 783.317(71.2404)
#> 
#> 
#> Maximum predicted log likelihood (783.317) estimated at:
#> 	2.46184e+12 (variance)
#> 	9.01332e-07 (lambda)
#> 
#> Fitting model at optimal hyperparameter
#> Using prior: Laplace(9.01332e-07) None 

#Find out what the optimal hyperparameter was:
getHyperParameter(fit)

#> [1] 2.461844e+12

#Extract the current log-likelihood, and coefficients
logLik(fit)

#> 'log Lik.' -2207.566 (df=3)

coef(fit)

#> (Intercept)           1           2 
#> -3.58685740  0.05610044 -0.02769318 

#We can only retrieve the confidence interval for unregularized coefficients:
confint(fit, c(0))

#> Using 1 thread(s)

#>      covariate     2.5 %    97.5 % evaluations
#> [1,]         0 -3.616605 -3.557373          22