Create a Cyclops prior object

createPrior creates a Cyclops prior object for use with fitCyclopsModel.

createPrior(
  priorType,
  variance = 1,
  exclude = c(),
  graph = NULL,
  neighborhood = NULL,
  useCrossValidation = FALSE,
  forceIntercept = FALSE
)

Arguments

priorType

Character: specifies prior distribution. See below for options

variance

Numeric: prior distribution variance

exclude

A vector of numbers or covariateId names to exclude from prior

graph

Child-to-parent mapping for a hierarchical prior

neighborhood

A list of first-order neighborhoods for a partially fused prior

useCrossValidation

Logical: Perform cross-validation to determine prior variance.

forceIntercept

Logical: Force intercept coefficient into prior

Value

A Cyclops prior object of class inheriting from "cyclopsPrior" for use with fitCyclopsModel.

Prior types

We specify all priors in terms of their variance parameters. Similar fitting tools for regularized regression often parameterize the Laplace distribution in terms of a rate "lambda" per observation. See "glmnet", for example.

variance = 2 * / (nobs * lambda)^2 or lambda = sqrt(2 / variance) / nobs

Examples

#Generate some simulated data:
sim <- simulateCyclopsData(nstrata = 1, nrows = 1000, ncovars = 2, eCovarsPerRow = 0.5, 
                           model = "poisson")
#> Sparseness = 76.2 %
cyclopsData <- convertToCyclopsData(sim$outcomes, sim$covariates, modelType = "pr", 
                                    addIntercept = TRUE)
#> Sorting covariates by covariateId and rowId

#Define the prior and control objects to use cross-validation for finding the 
#optimal hyperparameter:
prior <- createPrior("laplace", exclude = 0, useCrossValidation = TRUE)
control <- createControl(cvType = "auto", noiseLevel = "quiet")

#Fit the model
fit <- fitCyclopsModel(cyclopsData,prior = prior, control = control)  
#> Using cross-validation selector type byRow
#> Performing 10-fold cross-validation [seed = 1698757934] with data partitions of sizes 100 100 100 100 100 100 100 100 100 100
#> Using 1 thread(s)
#> Starting var = 0.238 (default)
#> Running at Laplace(2.89886) None  Grid-point #1 at 0.238 	Fold #1 Rep #1 pred log like = 898.631
#> Running at Laplace(2.89886) None  Grid-point #1 at 0.238 	Fold #2 Rep #1 pred log like = 1378.61
#> Running at Laplace(2.89886) None  Grid-point #1 at 0.238 	Fold #3 Rep #1 pred log like = 1439.51
#> Running at Laplace(2.89886) None  Grid-point #1 at 0.238 	Fold #4 Rep #1 pred log like = 1227.83
#> Running at Laplace(2.89886) None  Grid-point #1 at 0.238 	Fold #5 Rep #1 pred log like = 1717.57
#> Running at Laplace(2.89886) None  Grid-point #1 at 0.238 	Fold #6 Rep #1 pred log like = 1272.34
#> Running at Laplace(2.89886) None  Grid-point #1 at 0.238 	Fold #7 Rep #1 pred log like = 1047.47
#> Running at Laplace(2.89886) None  Grid-point #1 at 0.238 	Fold #8 Rep #1 pred log like = 1266.38
#> Running at Laplace(2.89886) None  Grid-point #1 at 0.238 	Fold #9 Rep #1 pred log like = 1253.53
#> Running at Laplace(2.89886) None  Grid-point #1 at 0.238 	Fold #10 Rep #1 pred log like = 1346.96
#> AvgPred = 1284.88 with stdev = 208.341
#> Completed at 0.238
#> Next point at 2.38 with value 0 and continue = 1
#> search[ 0.238 ] = 1284.88(208.341)
#> 
#> Running at Laplace(0.916698) None  Grid-point #2 at 2.38 	Fold #1 Rep #1 pred log like = 898.623
#> Running at Laplace(0.916698) None  Grid-point #2 at 2.38 	Fold #2 Rep #1 pred log like = 1378.57
#> Running at Laplace(0.916698) None  Grid-point #2 at 2.38 	Fold #3 Rep #1 pred log like = 1439.49
#> Running at Laplace(0.916698) None  Grid-point #2 at 2.38 	Fold #4 Rep #1 pred log like = 1227.81
#> Running at Laplace(0.916698) None  Grid-point #2 at 2.38 	Fold #5 Rep #1 pred log like = 1717.55
#> Running at Laplace(0.916698) None  Grid-point #2 at 2.38 	Fold #6 Rep #1 pred log like = 1272.36
#> Running at Laplace(0.916698) None  Grid-point #2 at 2.38 	Fold #7 Rep #1 pred log like = 1047.55
#> Running at Laplace(0.916698) None  Grid-point #2 at 2.38 	Fold #8 Rep #1 pred log like = 1266.43
#> Running at Laplace(0.916698) None  Grid-point #2 at 2.38 	Fold #9 Rep #1 pred log like = 1253.54
#> Running at Laplace(0.916698) None  Grid-point #2 at 2.38 	Fold #10 Rep #1 pred log like = 1346.91
#> AvgPred = 1284.88 with stdev = 208.324
#> Completed at 2.38
#> Next point at 23.8 with value 0 and continue = 1
#> search[ 0.238 ] = 1284.88(208.341)
#> search[ 2.38 ] = 1284.88(208.324)
#> 
#> Running at Laplace(0.289886) None  Grid-point #3 at 23.8 	Fold #1 Rep #1 pred log like = 898.62
#> Running at Laplace(0.289886) None  Grid-point #3 at 23.8 	Fold #2 Rep #1 pred log like = 1378.56
#> Running at Laplace(0.289886) None  Grid-point #3 at 23.8 	Fold #3 Rep #1 pred log like = 1439.48
#> Running at Laplace(0.289886) None  Grid-point #3 at 23.8 	Fold #4 Rep #1 pred log like = 1227.81
#> Running at Laplace(0.289886) None  Grid-point #3 at 23.8 	Fold #5 Rep #1 pred log like = 1717.54
#> Running at Laplace(0.289886) None  Grid-point #3 at 23.8 	Fold #6 Rep #1 pred log like = 1272.37
#> Running at Laplace(0.289886) None  Grid-point #3 at 23.8 	Fold #7 Rep #1 pred log like = 1047.58
#> Running at Laplace(0.289886) None  Grid-point #3 at 23.8 	Fold #8 Rep #1 pred log like = 1266.45
#> Running at Laplace(0.289886) None  Grid-point #3 at 23.8 	Fold #9 Rep #1 pred log like = 1253.54
#> Running at Laplace(0.289886) None  Grid-point #3 at 23.8 	Fold #10 Rep #1 pred log like = 1346.9
#> AvgPred = 1284.88 with stdev = 208.318
#> Completed at 23.8
#> Next point at 5.37835 with value 1284.88 and continue = 0
#> search[ 0.238 ] = 1284.88(208.341)
#> search[ 2.38 ] = 1284.88(208.324)
#> search[ 23.8 ] = 1284.88(208.318)
#> 
#> 
#> Maximum predicted log likelihood (1284.88) estimated at:
#> 	5.37835 (variance)
#> 	0.609804 (lambda)
#> 
#> Fitting model at optimal hyperparameter
#> Using prior: Laplace(0.609804) None 

#Find out what the optimal hyperparameter was:
getHyperParameter(fit)
#> [1] 5.37835

#Extract the current log-likelihood, and coefficients
logLik(fit)
#> 'log Lik.' -2158.207 (df=3)
coef(fit)
#> (Intercept)           1           2 
#> -3.92974607 -0.02127371  1.41692709 

#We can only retrieve the confidence interval for unregularized coefficients:
confint(fit, c(0))
#> Using 1 thread(s)
#>             covariate     2.5 %  97.5 % evaluations
#> (Intercept)         0 -3.962925 -3.8969          22