Create a Cyclops prior object

createPrior creates a Cyclops prior object for use with fitCyclopsModel.

createPrior(priorType, variance = 1, exclude = c(), graph = NULL,
  neighborhood = NULL, useCrossValidation = FALSE,
  forceIntercept = FALSE)

Arguments

priorType
Character: specifies prior distribution. See below for options

variance
Numeric: prior distribution variance

exclude
A vector of numbers or covariateId names to exclude from prior

graph
Child-to-parent mapping for a hierarchical prior

neighborhood
A list of first-order neighborhoods for a partially fused prior

useCrossValidation
Logical: Perform cross-validation to determine prior variance.

forceIntercept
Logical: Force intercept coefficient into prior

Value

A Cyclops prior object of class inheriting from "cyclopsPrior" for use with fitCyclopsModel.

Prior types

We specify all priors in terms of their variance parameters. Similar fitting tools for regularized regression often parameterize the Laplace distribution in terms of a rate "lambda" per observation. See "glmnet", for example.

variance = 2 * / (nobs * lambda)^2 or lambda = sqrt(2 / variance) / nobs

Examples

#Generate some simulated data:
sim <- simulateCyclopsData(nstrata = 1, nrows = 1000, ncovars = 2, eCovarsPerRow = 0.5,
                           model = "poisson")

#> Sparseness = 75.8 %

cyclopsData <- convertToCyclopsData(sim$outcomes, sim$covariates, modelType = "pr",
                                    addIntercept = TRUE)

#> Sorting covariates by covariateId and rowId

#Define the prior and control objects to use cross-validation for finding the 
#optimal hyperparameter:
prior <- createPrior("laplace", exclude = 0, useCrossValidation = TRUE)
control <- createControl(cvType = "auto", noiseLevel = "quiet")

#Fit the model
fit <- fitCyclopsModel(cyclopsData,prior = prior, control = control)

#> Using cross-validation selector type byRow
#> Performing 10-fold cross-validation [seed = 1543260552] with data partitions of sizes 100 100 100 100 100 100 100 100 100 100
#> Using 1 thread(s)
#> Starting var = 0.242 (default)
#> Running at Laplace(2.8748) None  Grid-point #1 at 0.242 	Fold #1 Rep #1 pred log like = -217.992
#> Running at Laplace(2.8748) None  Grid-point #1 at 0.242 	Fold #2 Rep #1 pred log like = -227.865
#> Running at Laplace(2.8748) None  Grid-point #1 at 0.242 	Fold #3 Rep #1 pred log like = -220.588
#> Running at Laplace(2.8748) None  Grid-point #1 at 0.242 	Fold #4 Rep #1 pred log like = -218.034
#> Running at Laplace(2.8748) None  Grid-point #1 at 0.242 	Fold #5 Rep #1 pred log like = -224.958
#> Running at Laplace(2.8748) None  Grid-point #1 at 0.242 	Fold #6 Rep #1 pred log like = -227.795
#> Running at Laplace(2.8748) None  Grid-point #1 at 0.242 	Fold #7 Rep #1 pred log like = -233.7
#> Running at Laplace(2.8748) None  Grid-point #1 at 0.242 	Fold #8 Rep #1 pred log like = -211.695
#> Running at Laplace(2.8748) None  Grid-point #1 at 0.242 	Fold #9 Rep #1 pred log like = -218.535
#> Running at Laplace(2.8748) None  Grid-point #1 at 0.242 	Fold #10 Rep #1 pred log like = -210.513
#> AvgPred = -221.167 with stdev = 7.00456
#> Completed at 0.242
#> Next point at 2.42 with value 0 and continue = 1
#> search[ 0.242 ] = -221.167(7.00456)
#> 
#> Running at Laplace(0.909091) None  Grid-point #2 at 2.42 	Fold #1 Rep #1 pred log like = -217.983
#> Running at Laplace(0.909091) None  Grid-point #2 at 2.42 	Fold #2 Rep #1 pred log like = -227.862
#> Running at Laplace(0.909091) None  Grid-point #2 at 2.42 	Fold #3 Rep #1 pred log like = -220.627
#> Running at Laplace(0.909091) None  Grid-point #2 at 2.42 	Fold #4 Rep #1 pred log like = -218.017
#> Running at Laplace(0.909091) None  Grid-point #2 at 2.42 	Fold #5 Rep #1 pred log like = -224.96
#> Running at Laplace(0.909091) None  Grid-point #2 at 2.42 	Fold #6 Rep #1 pred log like = -227.778
#> Running at Laplace(0.909091) None  Grid-point #2 at 2.42 	Fold #7 Rep #1 pred log like = -233.7
#> Running at Laplace(0.909091) None  Grid-point #2 at 2.42 	Fold #8 Rep #1 pred log like = -211.694
#> Running at Laplace(0.909091) None  Grid-point #2 at 2.42 	Fold #9 Rep #1 pred log like = -218.541
#> Running at Laplace(0.909091) None  Grid-point #2 at 2.42 	Fold #10 Rep #1 pred log like = -210.584
#> AvgPred = -221.175 with stdev = 6.99249
#> Completed at 2.42
#> Next point at 0.0242 with value 0 and continue = 1
#> search[ 0.242 ] = -221.167(7.00456)
#> search[ 2.42 ] = -221.175(6.99249)
#> 
#> Running at Laplace(9.09091) None  Grid-point #3 at 0.0242 	Fold #1 Rep #1 pred log like = -218.023
#> Running at Laplace(9.09091) None  Grid-point #3 at 0.0242 	Fold #2 Rep #1 pred log like = -227.877
#> Running at Laplace(9.09091) None  Grid-point #3 at 0.0242 	Fold #3 Rep #1 pred log like = -220.47
#> Running at Laplace(9.09091) None  Grid-point #3 at 0.0242 	Fold #4 Rep #1 pred log like = -218.091
#> Running at Laplace(9.09091) None  Grid-point #3 at 0.0242 	Fold #5 Rep #1 pred log like = -224.954
#> Running at Laplace(9.09091) None  Grid-point #3 at 0.0242 	Fold #6 Rep #1 pred log like = -227.855
#> Running at Laplace(9.09091) None  Grid-point #3 at 0.0242 	Fold #7 Rep #1 pred log like = -233.706
#> Running at Laplace(9.09091) None  Grid-point #3 at 0.0242 	Fold #8 Rep #1 pred log like = -211.699
#> Running at Laplace(9.09091) None  Grid-point #3 at 0.0242 	Fold #9 Rep #1 pred log like = -218.512
#> Running at Laplace(9.09091) None  Grid-point #3 at 0.0242 	Fold #10 Rep #1 pred log like = -210.321
#> AvgPred = -221.151 with stdev = 7.03881
#> Completed at 0.0242
#> Next point at 0.00242 with value 0 and continue = 1
#> search[ 0.0242 ] = -221.151(7.03881)
#> search[ 0.242 ] = -221.167(7.00456)
#> search[ 2.42 ] = -221.175(6.99249)
#> 
#> Running at Laplace(28.748) None  Grid-point #4 at 0.00242 	Fold #1 Rep #1 pred log like = -218.101
#> Running at Laplace(28.748) None  Grid-point #4 at 0.00242 	Fold #2 Rep #1 pred log like = -227.923
#> Running at Laplace(28.748) None  Grid-point #4 at 0.00242 	Fold #3 Rep #1 pred log like = -220.147
#> Running at Laplace(28.748) None  Grid-point #4 at 0.00242 	Fold #4 Rep #1 pred log like = -218.179
#> Running at Laplace(28.748) None  Grid-point #4 at 0.00242 	Fold #5 Rep #1 pred log like = -224.982
#> Running at Laplace(28.748) None  Grid-point #4 at 0.00242 	Fold #6 Rep #1 pred log like = -227.937
#> Running at Laplace(28.748) None  Grid-point #4 at 0.00242 	Fold #7 Rep #1 pred log like = -233.76
#> Running at Laplace(28.748) None  Grid-point #4 at 0.00242 	Fold #8 Rep #1 pred log like = -211.708
#> Running at Laplace(28.748) None  Grid-point #4 at 0.00242 	Fold #9 Rep #1 pred log like = -218.43
#> Running at Laplace(28.748) None  Grid-point #4 at 0.00242 	Fold #10 Rep #1 pred log like = -210.022
#> AvgPred = -221.119 with stdev = 7.10723
#> Completed at 0.00242
#> Next point at 0.000242 with value 0 and continue = 1
#> search[ 0.00242 ] = -221.119(7.10723)
#> search[ 0.0242 ] = -221.151(7.03881)
#> search[ 0.242 ] = -221.167(7.00456)
#> search[ 2.42 ] = -221.175(6.99249)
#> 
#> Running at Laplace(90.9091) None  Grid-point #5 at 0.000242 	Fold #1 Rep #1 pred log like = -218.101
#> Running at Laplace(90.9091) None  Grid-point #5 at 0.000242 	Fold #2 Rep #1 pred log like = -227.923
#> Running at Laplace(90.9091) None  Grid-point #5 at 0.000242 	Fold #3 Rep #1 pred log like = -220.112
#> Running at Laplace(90.9091) None  Grid-point #5 at 0.000242 	Fold #4 Rep #1 pred log like = -218.179
#> Running at Laplace(90.9091) None  Grid-point #5 at 0.000242 	Fold #5 Rep #1 pred log like = -224.982
#> Running at Laplace(90.9091) None  Grid-point #5 at 0.000242 	Fold #6 Rep #1 pred log like = -227.937
#> Running at Laplace(90.9091) None  Grid-point #5 at 0.000242 	Fold #7 Rep #1 pred log like = -233.76
#> Running at Laplace(90.9091) None  Grid-point #5 at 0.000242 	Fold #8 Rep #1 pred log like = -211.708
#> Running at Laplace(90.9091) None  Grid-point #5 at 0.000242 	Fold #9 Rep #1 pred log like = -218.415
#> Running at Laplace(90.9091) None  Grid-point #5 at 0.000242 	Fold #10 Rep #1 pred log like = -209.796
#> AvgPred = -221.091 with stdev = 7.14382
#> Completed at 0.000242
#> Next point at 2.42e-005 with value 0 and continue = 1
#> search[ 0.000242 ] = -221.091(7.14382)
#> search[ 0.00242 ] = -221.119(7.10723)
#> search[ 0.0242 ] = -221.151(7.03881)
#> search[ 0.242 ] = -221.167(7.00456)
#> search[ 2.42 ] = -221.175(6.99249)
#> 
#> Running at Laplace(287.48) None  Grid-point #6 at 2.42e-005 	Fold #1 Rep #1 pred log like = -218.101
#> Running at Laplace(287.48) None  Grid-point #6 at 2.42e-005 	Fold #2 Rep #1 pred log like = -227.923
#> Running at Laplace(287.48) None  Grid-point #6 at 2.42e-005 	Fold #3 Rep #1 pred log like = -220.112
#> Running at Laplace(287.48) None  Grid-point #6 at 2.42e-005 	Fold #4 Rep #1 pred log like = -218.179
#> Running at Laplace(287.48) None  Grid-point #6 at 2.42e-005 	Fold #5 Rep #1 pred log like = -224.982
#> Running at Laplace(287.48) None  Grid-point #6 at 2.42e-005 	Fold #6 Rep #1 pred log like = -227.937
#> Running at Laplace(287.48) None  Grid-point #6 at 2.42e-005 	Fold #7 Rep #1 pred log like = -233.76
#> Running at Laplace(287.48) None  Grid-point #6 at 2.42e-005 	Fold #8 Rep #1 pred log like = -211.708
#> Running at Laplace(287.48) None  Grid-point #6 at 2.42e-005 	Fold #9 Rep #1 pred log like = -218.415
#> Running at Laplace(287.48) None  Grid-point #6 at 2.42e-005 	Fold #10 Rep #1 pred log like = -209.796
#> AvgPred = -221.091 with stdev = 7.14382
#> Completed at 2.42e-005
#> Next point at 1.04441e+056 with value -221.696 and continue = 0
#> search[ 2.42e-005 ] = -221.091(7.14382)
#> search[ 0.000242 ] = -221.091(7.14382)
#> search[ 0.00242 ] = -221.119(7.10723)
#> search[ 0.0242 ] = -221.151(7.03881)
#> search[ 0.242 ] = -221.167(7.00456)
#> search[ 2.42 ] = -221.175(6.99249)
#> 
#> 
#> Maximum predicted log likelihood (-221.696) estimated at:
#> 	1.04441e+056 (variance)
#> 	1.38382e-028 (lambda)
#> 
#> Fitting model at optimal hyperparameter
#> Using prior: Laplace(1.38382e-028) None 

#Find out what the optimal hyperparameter was:
getHyperParameter(fit)

#> [1] 1.044412e+56

#Extract the current log-likelihood, and coefficients
logLik(fit)

#> 'log Lik.' -2209.46 (df=3)

coef(fit)

#>  (Intercept)            1            2 
#> -3.586737592  0.021710306  0.007347772 

#We can only retrieve the confidence interval for unregularized coefficients:
confint(fit, c(0))

#> Using 1 thread(s)

#>      covariate     2.5 %    97.5 % evaluations
#> [1,]         0 -3.616484 -3.557257          22