getCyclopsProfileLogLikelihood {Cyclops}	R Documentation

Profile likelihood for Cyclops model parameters

Description

getCyclopsProfileLogLikelihood evaluates the profile likelihood at a grid of parameter values.

Usage

getCyclopsProfileLogLikelihood(
  object,
  parm,
  x = NULL,
  bounds = NULL,
  tolerance = 0.001,
  initialGridSize = 10,
  includePenalty = TRUE
)

Arguments

object
Fitted Cyclops model object

parm
Specification of which parameter requires profiling, either a vector of numbers of covariateId names

x
Vector of values of the parameter

bounds
Pair of values to bound adaptive profiling

tolerance
Absolute tolerance allowed for adaptive profiling

initialGridSize
Initial grid size for adaptive profiling

includePenalty
Logical: Include regularized covariate penalty in profile

Value

A data frame containing the profile log likelihood. Returns NULL when the adaptive profiling fails to converge.

[Package Cyclops version 3.2.0 Index]