Get hyperparameter

getHyperParameter returns the current hyper parameter in a Cyclops model fit object

getHyperParameter(object)

Arguments

object
A Cyclops model fit object

Examples

#Generate some simulated data:
sim <- simulateCyclopsData(nstrata = 1, nrows = 1000, ncovars = 2, eCovarsPerRow = 0.5,
                           model = "poisson")

#> Sparseness = 75.45 %

cyclopsData <- convertToCyclopsData(sim$outcomes, sim$covariates, modelType = "pr",
                                    addIntercept = TRUE)

#> Sorting covariates by covariateId and rowId

#Define the prior and control objects to use cross-validation for finding the 
#optimal hyperparameter:
prior <- createPrior("laplace", exclude = 0, useCrossValidation = TRUE)
control <- createControl(cvType = "auto", noiseLevel = "quiet")

#Fit the model
fit <- fitCyclopsModel(cyclopsData,prior = prior, control = control)

#> Using cross-validation selector type byRow
#> Performing 10-fold cross-validation [seed = 1591596885] with data partitions of sizes 100 100 100 100 100 100 100 100 100 100
#> Using 1 thread(s)
#> Starting var = 0.2455 (default)
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #1 Rep #1 pred log like = 351.045
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #2 Rep #1 pred log like = 243.403
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #3 Rep #1 pred log like = 208.28
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #4 Rep #1 pred log like = 243.849
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #5 Rep #1 pred log like = 254.646
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #6 Rep #1 pred log like = 206.725
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #7 Rep #1 pred log like = 269.921
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #8 Rep #1 pred log like = 212.76
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #9 Rep #1 pred log like = 258.347
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #10 Rep #1 pred log like = 269.523
#> AvgPred = 251.85 with stdev = 40.158
#> Completed at 0.2455
#> Next point at 2.455 with value 0 and continue = 1
#> search[ 0.2455 ] = 251.85(40.158)
#> 
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #1 Rep #1 pred log like = 350.967
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #2 Rep #1 pred log like = 243.354
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #3 Rep #1 pred log like = 208.268
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #4 Rep #1 pred log like = 243.868
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #5 Rep #1 pred log like = 254.63
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #6 Rep #1 pred log like = 206.716
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #7 Rep #1 pred log like = 269.92
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #8 Rep #1 pred log like = 212.703
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #9 Rep #1 pred log like = 258.376
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #10 Rep #1 pred log like = 269.522
#> AvgPred = 251.832 with stdev = 40.1474
#> Completed at 2.455
#> Next point at 0.02455 with value 0 and continue = 1
#> search[ 0.2455 ] = 251.85(40.158)
#> search[ 2.455 ] = 251.832(40.1474)
#> 
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #1 Rep #1 pred log like = 351.282
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #2 Rep #1 pred log like = 243.545
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #3 Rep #1 pred log like = 208.23
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #4 Rep #1 pred log like = 243.793
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #5 Rep #1 pred log like = 254.689
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #6 Rep #1 pred log like = 206.71
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #7 Rep #1 pred log like = 269.911
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #8 Rep #1 pred log like = 212.928
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #9 Rep #1 pred log like = 258.267
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #10 Rep #1 pred log like = 269.524
#> AvgPred = 251.888 with stdev = 40.2041
#> Completed at 0.02455
#> Next point at 0.002455 with value 0 and continue = 1
#> search[ 0.02455 ] = 251.888(40.2041)
#> search[ 0.2455 ] = 251.85(40.158)
#> search[ 2.455 ] = 251.832(40.1474)
#> 
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #1 Rep #1 pred log like = 351.7
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #2 Rep #1 pred log like = 243.66
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #3 Rep #1 pred log like = 208.199
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #4 Rep #1 pred log like = 243.759
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #5 Rep #1 pred log like = 254.746
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #6 Rep #1 pred log like = 206.678
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #7 Rep #1 pred log like = 269.875
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #8 Rep #1 pred log like = 213.302
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #9 Rep #1 pred log like = 258.253
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #10 Rep #1 pred log like = 269.516
#> AvgPred = 251.969 with stdev = 40.2748
#> Completed at 0.002455
#> Next point at 0.0002455 with value 0 and continue = 1
#> search[ 0.002455 ] = 251.969(40.2748)
#> search[ 0.02455 ] = 251.888(40.2041)
#> search[ 0.2455 ] = 251.85(40.158)
#> search[ 2.455 ] = 251.832(40.1474)
#> 
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #1 Rep #1 pred log like = 351.7
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #2 Rep #1 pred log like = 243.66
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #3 Rep #1 pred log like = 208.199
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #4 Rep #1 pred log like = 243.759
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #5 Rep #1 pred log like = 254.746
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #6 Rep #1 pred log like = 206.678
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #7 Rep #1 pred log like = 269.875
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #8 Rep #1 pred log like = 213.341
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #9 Rep #1 pred log like = 258.253
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #10 Rep #1 pred log like = 269.516
#> AvgPred = 251.973 with stdev = 40.2711
#> Completed at 0.0002455
#> Next point at 2.455e-05 with value 0 and continue = 1
#> search[ 0.0002455 ] = 251.973(40.2711)
#> search[ 0.002455 ] = 251.969(40.2748)
#> search[ 0.02455 ] = 251.888(40.2041)
#> search[ 0.2455 ] = 251.85(40.158)
#> search[ 2.455 ] = 251.832(40.1474)
#> 
#> Running at Laplace(285.423) None  Grid-point #6 at 2.455e-05 	Fold #1 Rep #1 pred log like = 351.7
#> Running at Laplace(285.423) None  Grid-point #6 at 2.455e-05 	Fold #2 Rep #1 pred log like = 243.66
#> Running at Laplace(285.423) None  Grid-point #6 at 2.455e-05 	Fold #3 Rep #1 pred log like = 208.199
#> Running at Laplace(285.423) None  Grid-point #6 at 2.455e-05 	Fold #4 Rep #1 pred log like = 243.759
#> Running at Laplace(285.423) None  Grid-point #6 at 2.455e-05 	Fold #5 Rep #1 pred log like = 254.746
#> Running at Laplace(285.423) None  Grid-point #6 at 2.455e-05 	Fold #6 Rep #1 pred log like = 206.678
#> Running at Laplace(285.423) None  Grid-point #6 at 2.455e-05 	Fold #7 Rep #1 pred log like = 269.875
#> Running at Laplace(285.423) None  Grid-point #6 at 2.455e-05 	Fold #8 Rep #1 pred log like = 213.341
#> Running at Laplace(285.423) None  Grid-point #6 at 2.455e-05 	Fold #9 Rep #1 pred log like = 258.253
#> Running at Laplace(285.423) None  Grid-point #6 at 2.455e-05 	Fold #10 Rep #1 pred log like = 269.516
#> AvgPred = 251.973 with stdev = 40.2711
#> Completed at 2.455e-05
#> Next point at 5.92362e-07 with value 251.993 and continue = 0
#> search[ 2.455e-05 ] = 251.973(40.2711)
#> search[ 0.0002455 ] = 251.973(40.2711)
#> search[ 0.002455 ] = 251.969(40.2748)
#> search[ 0.02455 ] = 251.888(40.2041)
#> search[ 0.2455 ] = 251.85(40.158)
#> search[ 2.455 ] = 251.832(40.1474)
#> 
#> 
#> Maximum predicted log likelihood (251.993) estimated at:
#> 	5.92362e-07 (variance)
#> 	1837.47 (lambda)
#> 
#> Fitting model at optimal hyperparameter
#> Using prior: Laplace(1837.47) None 

#Find out what the optimal hyperparameter was:
getHyperParameter(fit)

#> [1] 5.92362e-07

#Extract the current log-likelihood, and coefficients
logLik(fit)

#> 'log Lik.' -1927.858 (df=3)

coef(fit)

#> (Intercept)           1           2 
#>   -4.118361    0.000000    0.000000 

#We can only retrieve the confidence interval for unregularized coefficients:
confint(fit, c(0))

#> Using 1 thread(s)

#>      covariate     2.5 %    97.5 % evaluations
#> [1,]         0 -4.149079 -4.087942          24