Get hyperparameter

getHyperParameter returns the current hyper parameter in a Cyclops model fit object

getHyperParameter(object)

Arguments

object
A Cyclops model fit object

Examples

#Generate some simulated data:
sim <- simulateCyclopsData(nstrata = 1, nrows = 1000, ncovars = 2, eCovarsPerRow = 0.5, 
                           model = "poisson")

#> Sparseness = 75.45 %

cyclopsData <- convertToCyclopsData(sim$outcomes, sim$covariates, modelType = "pr", 
                                    addIntercept = TRUE)

#> Sorting covariates by covariateId and rowId

#Define the prior and control objects to use cross-validation for finding the 
#optimal hyperparameter:
prior <- createPrior("laplace", exclude = 0, useCrossValidation = TRUE)
control <- createControl(cvType = "auto", noiseLevel = "quiet")

#Fit the model
fit <- fitCyclopsModel(cyclopsData,prior = prior, control = control)  

#> Using cross-validation selector type byRow
#> Performing 10-fold cross-validation [seed = 1623936646] with data partitions of sizes 100 100 100 100 100 100 100 100 100 100
#> Using 1 thread(s)
#> Starting var = 0.2455 (default)
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #1 Rep #1 pred log like = 200.993
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #2 Rep #1 pred log like = 206.093
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #3 Rep #1 pred log like = 245.023
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #4 Rep #1 pred log like = 264.254
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #5 Rep #1 pred log like = 332.052
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #6 Rep #1 pred log like = 241.658
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #7 Rep #1 pred log like = 267.085
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #8 Rep #1 pred log like = 309.134
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #9 Rep #1 pred log like = 215.561
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #10 Rep #1 pred log like = 236.844
#> AvgPred = 251.87 with stdev = 40.5863
#> Completed at 0.2455
#> Next point at 2.455 with value 0 and continue = 1
#> search[ 0.2455 ] = 251.87(40.5863)
#> 
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #1 Rep #1 pred log like = 201.025
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #2 Rep #1 pred log like = 206.105
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #3 Rep #1 pred log like = 244.985
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #4 Rep #1 pred log like = 264.222
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #5 Rep #1 pred log like = 332.034
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #6 Rep #1 pred log like = 241.635
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #7 Rep #1 pred log like = 267.08
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #8 Rep #1 pred log like = 309.102
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #9 Rep #1 pred log like = 215.544
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #10 Rep #1 pred log like = 236.821
#> AvgPred = 251.855 with stdev = 40.5755
#> Completed at 2.455
#> Next point at 0.02455 with value 0 and continue = 1
#> search[ 0.2455 ] = 251.87(40.5863)
#> search[ 2.455 ] = 251.855(40.5755)
#> 
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #1 Rep #1 pred log like = 200.993
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #2 Rep #1 pred log like = 206.063
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #3 Rep #1 pred log like = 245.129
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #4 Rep #1 pred log like = 264.304
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #5 Rep #1 pred log like = 332.096
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #6 Rep #1 pred log like = 241.667
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #7 Rep #1 pred log like = 267.089
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #8 Rep #1 pred log like = 309.223
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #9 Rep #1 pred log like = 215.609
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #10 Rep #1 pred log like = 236.912
#> AvgPred = 251.908 with stdev = 40.6039
#> Completed at 0.02455
#> Next point at 0.002455 with value 0 and continue = 1
#> search[ 0.02455 ] = 251.908(40.6039)
#> search[ 0.2455 ] = 251.87(40.5863)
#> search[ 2.455 ] = 251.855(40.5755)
#> 
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #1 Rep #1 pred log like = 200.993
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #2 Rep #1 pred log like = 206.03
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #3 Rep #1 pred log like = 245.18
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #4 Rep #1 pred log like = 264.306
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #5 Rep #1 pred log like = 332.076
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #6 Rep #1 pred log like = 241.668
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #7 Rep #1 pred log like = 267.058
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #8 Rep #1 pred log like = 309.405
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #9 Rep #1 pred log like = 215.596
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #10 Rep #1 pred log like = 237.054
#> AvgPred = 251.936 with stdev = 40.6232
#> Completed at 0.002455
#> Next point at 0.0002455 with value 0 and continue = 1
#> search[ 0.002455 ] = 251.936(40.6232)
#> search[ 0.02455 ] = 251.908(40.6039)
#> search[ 0.2455 ] = 251.87(40.5863)
#> search[ 2.455 ] = 251.855(40.5755)
#> 
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #1 Rep #1 pred log like = 200.993
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #2 Rep #1 pred log like = 206.03
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #3 Rep #1 pred log like = 245.18
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #4 Rep #1 pred log like = 264.306
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #5 Rep #1 pred log like = 332.076
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #6 Rep #1 pred log like = 241.668
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #7 Rep #1 pred log like = 267.058
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #8 Rep #1 pred log like = 309.405
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #9 Rep #1 pred log like = 215.596
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #10 Rep #1 pred log like = 237.054
#> AvgPred = 251.936 with stdev = 40.6232
#> Completed at 0.0002455
#> Next point at 1.99853e-06 with value 251.954 and continue = 0
#> search[ 0.0002455 ] = 251.936(40.6232)
#> search[ 0.002455 ] = 251.936(40.6232)
#> search[ 0.02455 ] = 251.908(40.6039)
#> search[ 0.2455 ] = 251.87(40.5863)
#> search[ 2.455 ] = 251.855(40.5755)
#> 
#> 
#> Maximum predicted log likelihood (251.954) estimated at:
#> 	1.99853e-06 (variance)
#> 	1000.37 (lambda)
#> 
#> Fitting model at optimal hyperparameter
#> Using prior: Laplace(1000.37) None 

#Find out what the optimal hyperparameter was:
getHyperParameter(fit)

#> [1] 1.998532e-06

#Extract the current log-likelihood, and coefficients
logLik(fit)

#> 'log Lik.' -1927.858 (df=3)

coef(fit)

#> (Intercept)           1           2 
#>   -4.118361    0.000000    0.000000 

#We can only retrieve the confidence interval for unregularized coefficients:
confint(fit, c(0))

#> Using 1 thread(s)

#>      covariate     2.5 %    97.5 % evaluations
#> [1,]         0 -4.149079 -4.087942          24