Get hyperparameter

getHyperParameter returns the current hyper parameter in a Cyclops model fit object

getHyperParameter(object)

Arguments

object
A Cyclops model fit object

Examples

#Generate some simulated data:
sim <- simulateCyclopsData(nstrata = 1, nrows = 1000, ncovars = 2, eCovarsPerRow = 0.5, 
                           model = "poisson")

#> Sparseness = 75.45 %

cyclopsData <- convertToCyclopsData(sim$outcomes, sim$covariates, modelType = "pr", 
                                    addIntercept = TRUE)

#> Sorting covariates by covariateId and rowId

#Define the prior and control objects to use cross-validation for finding the 
#optimal hyperparameter:
prior <- createPrior("laplace", exclude = 0, useCrossValidation = TRUE)
control <- createControl(cvType = "auto", noiseLevel = "quiet")

#Fit the model
fit <- fitCyclopsModel(cyclopsData,prior = prior, control = control)  

#> Using cross-validation selector type byRow
#> Performing 10-fold cross-validation [seed = 1604638982] with data partitions of sizes 100 100 100 100 100 100 100 100 100 100
#> Using 1 thread(s)
#> Starting var = 0.2455 (default)
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #1 Rep #1 pred log like = 297.87
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #2 Rep #1 pred log like = 331.665
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #3 Rep #1 pred log like = 195.623
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #4 Rep #1 pred log like = 273.631
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #5 Rep #1 pred log like = 228.769
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #6 Rep #1 pred log like = 199.189
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #7 Rep #1 pred log like = 250.077
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #8 Rep #1 pred log like = 233.389
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #9 Rep #1 pred log like = 253.775
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #10 Rep #1 pred log like = 252.617
#> AvgPred = 251.661 with stdev = 39.7213
#> Completed at 0.2455
#> Next point at 2.455 with value 0 and continue = 1
#> search[ 0.2455 ] = 251.661(39.7213)
#> 
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #1 Rep #1 pred log like = 297.881
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #2 Rep #1 pred log like = 331.578
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #3 Rep #1 pred log like = 195.633
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #4 Rep #1 pred log like = 273.59
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #5 Rep #1 pred log like = 228.771
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #6 Rep #1 pred log like = 199.147
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #7 Rep #1 pred log like = 250.019
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #8 Rep #1 pred log like = 233.349
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #9 Rep #1 pred log like = 253.769
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #10 Rep #1 pred log like = 252.658
#> AvgPred = 251.639 with stdev = 39.7092
#> Completed at 2.455
#> Next point at 0.02455 with value 0 and continue = 1
#> search[ 0.2455 ] = 251.661(39.7213)
#> search[ 2.455 ] = 251.639(39.7092)
#> 
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #1 Rep #1 pred log like = 297.833
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #2 Rep #1 pred log like = 331.753
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #3 Rep #1 pred log like = 195.601
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #4 Rep #1 pred log like = 273.752
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #5 Rep #1 pred log like = 228.746
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #6 Rep #1 pred log like = 199.299
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #7 Rep #1 pred log like = 250.25
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #8 Rep #1 pred log like = 233.507
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #9 Rep #1 pred log like = 253.793
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #10 Rep #1 pred log like = 252.534
#> AvgPred = 251.707 with stdev = 39.7253
#> Completed at 0.02455
#> Next point at 0.002455 with value 0 and continue = 1
#> search[ 0.02455 ] = 251.707(39.7253)
#> search[ 0.2455 ] = 251.661(39.7213)
#> search[ 2.455 ] = 251.639(39.7092)
#> 
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #1 Rep #1 pred log like = 297.794
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #2 Rep #1 pred log like = 331.774
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #3 Rep #1 pred log like = 195.572
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #4 Rep #1 pred log like = 273.879
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #5 Rep #1 pred log like = 228.716
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #6 Rep #1 pred log like = 199.596
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #7 Rep #1 pred log like = 250.75
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #8 Rep #1 pred log like = 233.63
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #9 Rep #1 pred log like = 253.816
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #10 Rep #1 pred log like = 252.534
#> AvgPred = 251.806 with stdev = 39.6917
#> Completed at 0.002455
#> Next point at 0.0002455 with value 0 and continue = 1
#> search[ 0.002455 ] = 251.806(39.6917)
#> search[ 0.02455 ] = 251.707(39.7253)
#> search[ 0.2455 ] = 251.661(39.7213)
#> search[ 2.455 ] = 251.639(39.7092)
#> 
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #1 Rep #1 pred log like = 297.794
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #2 Rep #1 pred log like = 331.774
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #3 Rep #1 pred log like = 195.572
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #4 Rep #1 pred log like = 273.879
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #5 Rep #1 pred log like = 228.716
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #6 Rep #1 pred log like = 199.596
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #7 Rep #1 pred log like = 250.894
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #8 Rep #1 pred log like = 233.63
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #9 Rep #1 pred log like = 253.816
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #10 Rep #1 pred log like = 252.534
#> AvgPred = 251.82 with stdev = 39.6913
#> Completed at 0.0002455
#> Next point at 2.455e-05 with value 0 and continue = 1
#> search[ 0.0002455 ] = 251.82(39.6913)
#> search[ 0.002455 ] = 251.806(39.6917)
#> search[ 0.02455 ] = 251.707(39.7253)
#> search[ 0.2455 ] = 251.661(39.7213)
#> search[ 2.455 ] = 251.639(39.7092)
#> 
#> Running at Laplace(285.423) None  Grid-point #6 at 2.455e-05 	Fold #1 Rep #1 pred log like = 297.794
#> Running at Laplace(285.423) None  Grid-point #6 at 2.455e-05 	Fold #2 Rep #1 pred log like = 331.774
#> Running at Laplace(285.423) None  Grid-point #6 at 2.455e-05 	Fold #3 Rep #1 pred log like = 195.572
#> Running at Laplace(285.423) None  Grid-point #6 at 2.455e-05 	Fold #4 Rep #1 pred log like = 273.879
#> Running at Laplace(285.423) None  Grid-point #6 at 2.455e-05 	Fold #5 Rep #1 pred log like = 228.716
#> Running at Laplace(285.423) None  Grid-point #6 at 2.455e-05 	Fold #6 Rep #1 pred log like = 199.596
#> Running at Laplace(285.423) None  Grid-point #6 at 2.455e-05 	Fold #7 Rep #1 pred log like = 250.894
#> Running at Laplace(285.423) None  Grid-point #6 at 2.455e-05 	Fold #8 Rep #1 pred log like = 233.63
#> Running at Laplace(285.423) None  Grid-point #6 at 2.455e-05 	Fold #9 Rep #1 pred log like = 253.816
#> Running at Laplace(285.423) None  Grid-point #6 at 2.455e-05 	Fold #10 Rep #1 pred log like = 252.534
#> AvgPred = 251.82 with stdev = 39.6913
#> Completed at 2.455e-05
#> Next point at 6.34355e-08 with value 251.862 and continue = 0
#> search[ 2.455e-05 ] = 251.82(39.6913)
#> search[ 0.0002455 ] = 251.82(39.6913)
#> search[ 0.002455 ] = 251.806(39.6917)
#> search[ 0.02455 ] = 251.707(39.7253)
#> search[ 0.2455 ] = 251.661(39.7213)
#> search[ 2.455 ] = 251.639(39.7092)
#> 
#> 
#> Maximum predicted log likelihood (251.862) estimated at:
#> 	6.34355e-08 (variance)
#> 	5614.99 (lambda)
#> 
#> Fitting model at optimal hyperparameter
#> Using prior: Laplace(5614.99) None 

#Find out what the optimal hyperparameter was:
getHyperParameter(fit)

#> [1] 6.343549e-08

#Extract the current log-likelihood, and coefficients
logLik(fit)

#> 'log Lik.' -1927.858 (df=3)

coef(fit)

#> (Intercept)           1           2 
#>   -4.118361    0.000000    0.000000 

#We can only retrieve the confidence interval for unregularized coefficients:
confint(fit, c(0))

#> Using 1 thread(s)

#>      covariate     2.5 %    97.5 % evaluations
#> [1,]         0 -4.149079 -4.087942          24