Get hyperparameter

getHyperParameter returns the current hyper parameter in a Cyclops model fit object

getHyperParameter(object)

Arguments

object

A Cyclops model fit object

Examples

#Generate some simulated data:
sim <- simulateCyclopsData(nstrata = 1, nrows = 1000, ncovars = 2, eCovarsPerRow = 0.5, 
                           model = "poisson")
#> Sparseness = 73.85 %
cyclopsData <- convertToCyclopsData(sim$outcomes, sim$covariates, modelType = "pr", 
                                    addIntercept = TRUE)
#> Sorting covariates by covariateId and rowId

#Define the prior and control objects to use cross-validation for finding the 
#optimal hyperparameter:
prior <- createPrior("laplace", exclude = 0, useCrossValidation = TRUE)
control <- createControl(cvType = "auto", noiseLevel = "quiet")

#Fit the model
fit <- fitCyclopsModel(cyclopsData,prior = prior, control = control)  
#> Using cross-validation selector type byRow
#> Performing 10-fold cross-validation [seed = 1698757937] with data partitions of sizes 100 100 100 100 100 100 100 100 100 100
#> Using 1 thread(s)
#> Starting var = 0.2615 (default)
#> Running at Laplace(2.76553) None  Grid-point #1 at 0.2615 	Fold #1 Rep #1 pred log like = 372.484
#> Running at Laplace(2.76553) None  Grid-point #1 at 0.2615 	Fold #2 Rep #1 pred log like = 427.249
#> Running at Laplace(2.76553) None  Grid-point #1 at 0.2615 	Fold #3 Rep #1 pred log like = 468.957
#> Running at Laplace(2.76553) None  Grid-point #1 at 0.2615 	Fold #4 Rep #1 pred log like = 505.492
#> Running at Laplace(2.76553) None  Grid-point #1 at 0.2615 	Fold #5 Rep #1 pred log like = 364.273
#> Running at Laplace(2.76553) None  Grid-point #1 at 0.2615 	Fold #6 Rep #1 pred log like = 433.21
#> Running at Laplace(2.76553) None  Grid-point #1 at 0.2615 	Fold #7 Rep #1 pred log like = 443.256
#> Running at Laplace(2.76553) None  Grid-point #1 at 0.2615 	Fold #8 Rep #1 pred log like = 438.388
#> Running at Laplace(2.76553) None  Grid-point #1 at 0.2615 	Fold #9 Rep #1 pred log like = 475.904
#> Running at Laplace(2.76553) None  Grid-point #1 at 0.2615 	Fold #10 Rep #1 pred log like = 470.303
#> AvgPred = 439.952 with stdev = 42.2492
#> Completed at 0.2615
#> Next point at 2.615 with value 0 and continue = 1
#> search[ 0.2615 ] = 439.952(42.2492)
#> 
#> Running at Laplace(0.874539) None  Grid-point #2 at 2.615 	Fold #1 Rep #1 pred log like = 372.478
#> Running at Laplace(0.874539) None  Grid-point #2 at 2.615 	Fold #2 Rep #1 pred log like = 427.302
#> Running at Laplace(0.874539) None  Grid-point #2 at 2.615 	Fold #3 Rep #1 pred log like = 468.968
#> Running at Laplace(0.874539) None  Grid-point #2 at 2.615 	Fold #4 Rep #1 pred log like = 505.462
#> Running at Laplace(0.874539) None  Grid-point #2 at 2.615 	Fold #5 Rep #1 pred log like = 364.28
#> Running at Laplace(0.874539) None  Grid-point #2 at 2.615 	Fold #6 Rep #1 pred log like = 433.206
#> Running at Laplace(0.874539) None  Grid-point #2 at 2.615 	Fold #7 Rep #1 pred log like = 443.204
#> Running at Laplace(0.874539) None  Grid-point #2 at 2.615 	Fold #8 Rep #1 pred log like = 438.398
#> Running at Laplace(0.874539) None  Grid-point #2 at 2.615 	Fold #9 Rep #1 pred log like = 475.914
#> Running at Laplace(0.874539) None  Grid-point #2 at 2.615 	Fold #10 Rep #1 pred log like = 470.313
#> AvgPred = 439.952 with stdev = 42.2447
#> Completed at 2.615
#> Next point at 26.15 with value 0 and continue = 1
#> search[ 0.2615 ] = 439.952(42.2492)
#> search[ 2.615 ] = 439.952(42.2447)
#> 
#> Running at Laplace(0.276553) None  Grid-point #3 at 26.15 	Fold #1 Rep #1 pred log like = 372.477
#> Running at Laplace(0.276553) None  Grid-point #3 at 26.15 	Fold #2 Rep #1 pred log like = 427.318
#> Running at Laplace(0.276553) None  Grid-point #3 at 26.15 	Fold #3 Rep #1 pred log like = 468.971
#> Running at Laplace(0.276553) None  Grid-point #3 at 26.15 	Fold #4 Rep #1 pred log like = 505.453
#> Running at Laplace(0.276553) None  Grid-point #3 at 26.15 	Fold #5 Rep #1 pred log like = 364.282
#> Running at Laplace(0.276553) None  Grid-point #3 at 26.15 	Fold #6 Rep #1 pred log like = 433.205
#> Running at Laplace(0.276553) None  Grid-point #3 at 26.15 	Fold #7 Rep #1 pred log like = 443.188
#> Running at Laplace(0.276553) None  Grid-point #3 at 26.15 	Fold #8 Rep #1 pred log like = 438.4
#> Running at Laplace(0.276553) None  Grid-point #3 at 26.15 	Fold #9 Rep #1 pred log like = 475.917
#> Running at Laplace(0.276553) None  Grid-point #3 at 26.15 	Fold #10 Rep #1 pred log like = 470.316
#> AvgPred = 439.953 with stdev = 42.243
#> Completed at 26.15
#> Next point at 261.5 with value 0 and continue = 1
#> search[ 0.2615 ] = 439.952(42.2492)
#> search[ 2.615 ] = 439.952(42.2447)
#> search[ 26.15 ] = 439.953(42.243)
#> 
#> Running at Laplace(0.0874539) None  Grid-point #4 at 261.5 	Fold #1 Rep #1 pred log like = 372.477
#> Running at Laplace(0.0874539) None  Grid-point #4 at 261.5 	Fold #2 Rep #1 pred log like = 427.323
#> Running at Laplace(0.0874539) None  Grid-point #4 at 261.5 	Fold #3 Rep #1 pred log like = 468.972
#> Running at Laplace(0.0874539) None  Grid-point #4 at 261.5 	Fold #4 Rep #1 pred log like = 505.45
#> Running at Laplace(0.0874539) None  Grid-point #4 at 261.5 	Fold #5 Rep #1 pred log like = 364.282
#> Running at Laplace(0.0874539) None  Grid-point #4 at 261.5 	Fold #6 Rep #1 pred log like = 433.204
#> Running at Laplace(0.0874539) None  Grid-point #4 at 261.5 	Fold #7 Rep #1 pred log like = 443.182
#> Running at Laplace(0.0874539) None  Grid-point #4 at 261.5 	Fold #8 Rep #1 pred log like = 438.401
#> Running at Laplace(0.0874539) None  Grid-point #4 at 261.5 	Fold #9 Rep #1 pred log like = 475.917
#> Running at Laplace(0.0874539) None  Grid-point #4 at 261.5 	Fold #10 Rep #1 pred log like = 470.317
#> AvgPred = 439.953 with stdev = 42.2425
#> Completed at 261.5
#> Next point at 2615 with value 0 and continue = 1
#> search[ 0.2615 ] = 439.952(42.2492)
#> search[ 2.615 ] = 439.952(42.2447)
#> search[ 26.15 ] = 439.953(42.243)
#> search[ 261.5 ] = 439.953(42.2425)
#> 
#> Running at Laplace(0.0276553) None  Grid-point #5 at 2615 	Fold #1 Rep #1 pred log like = 372.477
#> Running at Laplace(0.0276553) None  Grid-point #5 at 2615 	Fold #2 Rep #1 pred log like = 427.325
#> Running at Laplace(0.0276553) None  Grid-point #5 at 2615 	Fold #3 Rep #1 pred log like = 468.972
#> Running at Laplace(0.0276553) None  Grid-point #5 at 2615 	Fold #4 Rep #1 pred log like = 505.449
#> Running at Laplace(0.0276553) None  Grid-point #5 at 2615 	Fold #5 Rep #1 pred log like = 364.282
#> Running at Laplace(0.0276553) None  Grid-point #5 at 2615 	Fold #6 Rep #1 pred log like = 433.204
#> Running at Laplace(0.0276553) None  Grid-point #5 at 2615 	Fold #7 Rep #1 pred log like = 443.181
#> Running at Laplace(0.0276553) None  Grid-point #5 at 2615 	Fold #8 Rep #1 pred log like = 438.402
#> Running at Laplace(0.0276553) None  Grid-point #5 at 2615 	Fold #9 Rep #1 pred log like = 475.918
#> Running at Laplace(0.0276553) None  Grid-point #5 at 2615 	Fold #10 Rep #1 pred log like = 470.317
#> AvgPred = 439.953 with stdev = 42.2423
#> Completed at 2615
#> Next point at 26150 with value 0 and continue = 1
#> search[ 0.2615 ] = 439.952(42.2492)
#> search[ 2.615 ] = 439.952(42.2447)
#> search[ 26.15 ] = 439.953(42.243)
#> search[ 261.5 ] = 439.953(42.2425)
#> search[ 2615 ] = 439.953(42.2423)
#> 
#> Running at Laplace(0.00874539) None  Grid-point #6 at 26150 	Fold #1 Rep #1 pred log like = 372.477
#> Running at Laplace(0.00874539) None  Grid-point #6 at 26150 	Fold #2 Rep #1 pred log like = 427.325
#> Running at Laplace(0.00874539) None  Grid-point #6 at 26150 	Fold #3 Rep #1 pred log like = 468.972
#> Running at Laplace(0.00874539) None  Grid-point #6 at 26150 	Fold #4 Rep #1 pred log like = 505.449
#> Running at Laplace(0.00874539) None  Grid-point #6 at 26150 	Fold #5 Rep #1 pred log like = 364.283
#> Running at Laplace(0.00874539) None  Grid-point #6 at 26150 	Fold #6 Rep #1 pred log like = 433.204
#> Running at Laplace(0.00874539) None  Grid-point #6 at 26150 	Fold #7 Rep #1 pred log like = 443.18
#> Running at Laplace(0.00874539) None  Grid-point #6 at 26150 	Fold #8 Rep #1 pred log like = 438.402
#> Running at Laplace(0.00874539) None  Grid-point #6 at 26150 	Fold #9 Rep #1 pred log like = 475.918
#> Running at Laplace(0.00874539) None  Grid-point #6 at 26150 	Fold #10 Rep #1 pred log like = 470.318
#> AvgPred = 439.953 with stdev = 42.2423
#> Completed at 26150
#> Next point at 261500 with value 0 and continue = 1
#> search[ 0.2615 ] = 439.952(42.2492)
#> search[ 2.615 ] = 439.952(42.2447)
#> search[ 26.15 ] = 439.953(42.243)
#> search[ 261.5 ] = 439.953(42.2425)
#> search[ 2615 ] = 439.953(42.2423)
#> search[ 26150 ] = 439.953(42.2423)
#> 
#> Running at Laplace(0.00276553) None  Grid-point #7 at 261500 	Fold #1 Rep #1 pred log like = 372.477
#> Running at Laplace(0.00276553) None  Grid-point #7 at 261500 	Fold #2 Rep #1 pred log like = 427.325
#> Running at Laplace(0.00276553) None  Grid-point #7 at 261500 	Fold #3 Rep #1 pred log like = 468.972
#> Running at Laplace(0.00276553) None  Grid-point #7 at 261500 	Fold #4 Rep #1 pred log like = 505.448
#> Running at Laplace(0.00276553) None  Grid-point #7 at 261500 	Fold #5 Rep #1 pred log like = 364.283
#> Running at Laplace(0.00276553) None  Grid-point #7 at 261500 	Fold #6 Rep #1 pred log like = 433.204
#> Running at Laplace(0.00276553) None  Grid-point #7 at 261500 	Fold #7 Rep #1 pred log like = 443.18
#> Running at Laplace(0.00276553) None  Grid-point #7 at 261500 	Fold #8 Rep #1 pred log like = 438.402
#> Running at Laplace(0.00276553) None  Grid-point #7 at 261500 	Fold #9 Rep #1 pred log like = 475.918
#> Running at Laplace(0.00276553) None  Grid-point #7 at 261500 	Fold #10 Rep #1 pred log like = 470.318
#> AvgPred = 439.953 with stdev = 42.2422
#> Completed at 261500
#> Next point at 2.615e+06 with value 0 and continue = 1
#> search[ 0.2615 ] = 439.952(42.2492)
#> search[ 2.615 ] = 439.952(42.2447)
#> search[ 26.15 ] = 439.953(42.243)
#> search[ 261.5 ] = 439.953(42.2425)
#> search[ 2615 ] = 439.953(42.2423)
#> search[ 26150 ] = 439.953(42.2423)
#> search[ 261500 ] = 439.953(42.2422)
#> 
#> Running at Laplace(0.000874539) None  Grid-point #8 at 2.615e+06 	Fold #1 Rep #1 pred log like = 372.477
#> Running at Laplace(0.000874539) None  Grid-point #8 at 2.615e+06 	Fold #2 Rep #1 pred log like = 427.325
#> Running at Laplace(0.000874539) None  Grid-point #8 at 2.615e+06 	Fold #3 Rep #1 pred log like = 468.972
#> Running at Laplace(0.000874539) None  Grid-point #8 at 2.615e+06 	Fold #4 Rep #1 pred log like = 505.448
#> Running at Laplace(0.000874539) None  Grid-point #8 at 2.615e+06 	Fold #5 Rep #1 pred log like = 364.283
#> Running at Laplace(0.000874539) None  Grid-point #8 at 2.615e+06 	Fold #6 Rep #1 pred log like = 433.204
#> Running at Laplace(0.000874539) None  Grid-point #8 at 2.615e+06 	Fold #7 Rep #1 pred log like = 443.18
#> Running at Laplace(0.000874539) None  Grid-point #8 at 2.615e+06 	Fold #8 Rep #1 pred log like = 438.402
#> Running at Laplace(0.000874539) None  Grid-point #8 at 2.615e+06 	Fold #9 Rep #1 pred log like = 475.918
#> Running at Laplace(0.000874539) None  Grid-point #8 at 2.615e+06 	Fold #10 Rep #1 pred log like = 470.318
#> AvgPred = 439.953 with stdev = 42.2422
#> Completed at 2.615e+06
#> Next point at 2.615e+07 with value 0 and continue = 1
#> search[ 0.2615 ] = 439.952(42.2492)
#> search[ 2.615 ] = 439.952(42.2447)
#> search[ 26.15 ] = 439.953(42.243)
#> search[ 261.5 ] = 439.953(42.2425)
#> search[ 2615 ] = 439.953(42.2423)
#> search[ 26150 ] = 439.953(42.2423)
#> search[ 261500 ] = 439.953(42.2422)
#> search[ 2.615e+06 ] = 439.953(42.2422)
#> 
#> Running at Laplace(0.000276553) None  Grid-point #9 at 2.615e+07 	Fold #1 Rep #1 pred log like = 372.477
#> Running at Laplace(0.000276553) None  Grid-point #9 at 2.615e+07 	Fold #2 Rep #1 pred log like = 427.325
#> Running at Laplace(0.000276553) None  Grid-point #9 at 2.615e+07 	Fold #3 Rep #1 pred log like = 468.972
#> Running at Laplace(0.000276553) None  Grid-point #9 at 2.615e+07 	Fold #4 Rep #1 pred log like = 505.448
#> Running at Laplace(0.000276553) None  Grid-point #9 at 2.615e+07 	Fold #5 Rep #1 pred log like = 364.283
#> Running at Laplace(0.000276553) None  Grid-point #9 at 2.615e+07 	Fold #6 Rep #1 pred log like = 433.204
#> Running at Laplace(0.000276553) None  Grid-point #9 at 2.615e+07 	Fold #7 Rep #1 pred log like = 443.18
#> Running at Laplace(0.000276553) None  Grid-point #9 at 2.615e+07 	Fold #8 Rep #1 pred log like = 438.402
#> Running at Laplace(0.000276553) None  Grid-point #9 at 2.615e+07 	Fold #9 Rep #1 pred log like = 475.918
#> Running at Laplace(0.000276553) None  Grid-point #9 at 2.615e+07 	Fold #10 Rep #1 pred log like = 470.318
#> AvgPred = 439.953 with stdev = 42.2422
#> Completed at 2.615e+07
#> Next point at 2.615e+08 with value 0 and continue = 1
#> search[ 0.2615 ] = 439.952(42.2492)
#> search[ 2.615 ] = 439.952(42.2447)
#> search[ 26.15 ] = 439.953(42.243)
#> search[ 261.5 ] = 439.953(42.2425)
#> search[ 2615 ] = 439.953(42.2423)
#> search[ 26150 ] = 439.953(42.2423)
#> search[ 261500 ] = 439.953(42.2422)
#> search[ 2.615e+06 ] = 439.953(42.2422)
#> search[ 2.615e+07 ] = 439.953(42.2422)
#> 
#> Running at Laplace(8.74539e-05) None  Grid-point #10 at 2.615e+08 	Fold #1 Rep #1 pred log like = 372.477
#> Running at Laplace(8.74539e-05) None  Grid-point #10 at 2.615e+08 	Fold #2 Rep #1 pred log like = 427.325
#> Running at Laplace(8.74539e-05) None  Grid-point #10 at 2.615e+08 	Fold #3 Rep #1 pred log like = 468.972
#> Running at Laplace(8.74539e-05) None  Grid-point #10 at 2.615e+08 	Fold #4 Rep #1 pred log like = 505.448
#> Running at Laplace(8.74539e-05) None  Grid-point #10 at 2.615e+08 	Fold #5 Rep #1 pred log like = 364.283
#> Running at Laplace(8.74539e-05) None  Grid-point #10 at 2.615e+08 	Fold #6 Rep #1 pred log like = 433.204
#> Running at Laplace(8.74539e-05) None  Grid-point #10 at 2.615e+08 	Fold #7 Rep #1 pred log like = 443.18
#> Running at Laplace(8.74539e-05) None  Grid-point #10 at 2.615e+08 	Fold #8 Rep #1 pred log like = 438.402
#> Running at Laplace(8.74539e-05) None  Grid-point #10 at 2.615e+08 	Fold #9 Rep #1 pred log like = 475.918
#> Running at Laplace(8.74539e-05) None  Grid-point #10 at 2.615e+08 	Fold #10 Rep #1 pred log like = 470.318
#> AvgPred = 439.953 with stdev = 42.2422
#> Completed at 2.615e+08
#> Next point at 2.615e+09 with value 0 and continue = 1
#> search[ 0.2615 ] = 439.952(42.2492)
#> search[ 2.615 ] = 439.952(42.2447)
#> search[ 26.15 ] = 439.953(42.243)
#> search[ 261.5 ] = 439.953(42.2425)
#> search[ 2615 ] = 439.953(42.2423)
#> search[ 26150 ] = 439.953(42.2423)
#> search[ 261500 ] = 439.953(42.2422)
#> search[ 2.615e+06 ] = 439.953(42.2422)
#> search[ 2.615e+07 ] = 439.953(42.2422)
#> search[ 2.615e+08 ] = 439.953(42.2422)
#> 
#> Running at Laplace(2.76553e-05) None  Grid-point #11 at 2.615e+09 	Fold #1 Rep #1 pred log like = 372.477
#> Running at Laplace(2.76553e-05) None  Grid-point #11 at 2.615e+09 	Fold #2 Rep #1 pred log like = 427.325
#> Running at Laplace(2.76553e-05) None  Grid-point #11 at 2.615e+09 	Fold #3 Rep #1 pred log like = 468.972
#> Running at Laplace(2.76553e-05) None  Grid-point #11 at 2.615e+09 	Fold #4 Rep #1 pred log like = 505.448
#> Running at Laplace(2.76553e-05) None  Grid-point #11 at 2.615e+09 	Fold #5 Rep #1 pred log like = 364.283
#> Running at Laplace(2.76553e-05) None  Grid-point #11 at 2.615e+09 	Fold #6 Rep #1 pred log like = 433.204
#> Running at Laplace(2.76553e-05) None  Grid-point #11 at 2.615e+09 	Fold #7 Rep #1 pred log like = 443.18
#> Running at Laplace(2.76553e-05) None  Grid-point #11 at 2.615e+09 	Fold #8 Rep #1 pred log like = 438.402
#> Running at Laplace(2.76553e-05) None  Grid-point #11 at 2.615e+09 	Fold #9 Rep #1 pred log like = 475.918
#> Running at Laplace(2.76553e-05) None  Grid-point #11 at 2.615e+09 	Fold #10 Rep #1 pred log like = 470.318
#> AvgPred = 439.953 with stdev = 42.2422
#> Completed at 2.615e+09
#> Next point at 2.615e+10 with value 0 and continue = 1
#> search[ 0.2615 ] = 439.952(42.2492)
#> search[ 2.615 ] = 439.952(42.2447)
#> search[ 26.15 ] = 439.953(42.243)
#> search[ 261.5 ] = 439.953(42.2425)
#> search[ 2615 ] = 439.953(42.2423)
#> search[ 26150 ] = 439.953(42.2423)
#> search[ 261500 ] = 439.953(42.2422)
#> search[ 2.615e+06 ] = 439.953(42.2422)
#> search[ 2.615e+07 ] = 439.953(42.2422)
#> search[ 2.615e+08 ] = 439.953(42.2422)
#> search[ 2.615e+09 ] = 439.953(42.2422)
#> 
#> Running at Laplace(8.74539e-06) None  Grid-point #12 at 2.615e+10 	Fold #1 Rep #1 pred log like = 372.477
#> Running at Laplace(8.74539e-06) None  Grid-point #12 at 2.615e+10 	Fold #2 Rep #1 pred log like = 427.325
#> Running at Laplace(8.74539e-06) None  Grid-point #12 at 2.615e+10 	Fold #3 Rep #1 pred log like = 468.972
#> Running at Laplace(8.74539e-06) None  Grid-point #12 at 2.615e+10 	Fold #4 Rep #1 pred log like = 505.448
#> Running at Laplace(8.74539e-06) None  Grid-point #12 at 2.615e+10 	Fold #5 Rep #1 pred log like = 364.283
#> Running at Laplace(8.74539e-06) None  Grid-point #12 at 2.615e+10 	Fold #6 Rep #1 pred log like = 433.204
#> Running at Laplace(8.74539e-06) None  Grid-point #12 at 2.615e+10 	Fold #7 Rep #1 pred log like = 443.18
#> Running at Laplace(8.74539e-06) None  Grid-point #12 at 2.615e+10 	Fold #8 Rep #1 pred log like = 438.402
#> Running at Laplace(8.74539e-06) None  Grid-point #12 at 2.615e+10 	Fold #9 Rep #1 pred log like = 475.918
#> Running at Laplace(8.74539e-06) None  Grid-point #12 at 2.615e+10 	Fold #10 Rep #1 pred log like = 470.318
#> AvgPred = 439.953 with stdev = 42.2422
#> Completed at 2.615e+10
#> Next point at 2.615e+11 with value 0 and continue = 1
#> search[ 0.2615 ] = 439.952(42.2492)
#> search[ 2.615 ] = 439.952(42.2447)
#> search[ 26.15 ] = 439.953(42.243)
#> search[ 261.5 ] = 439.953(42.2425)
#> search[ 2615 ] = 439.953(42.2423)
#> search[ 26150 ] = 439.953(42.2423)
#> search[ 261500 ] = 439.953(42.2422)
#> search[ 2.615e+06 ] = 439.953(42.2422)
#> search[ 2.615e+07 ] = 439.953(42.2422)
#> search[ 2.615e+08 ] = 439.953(42.2422)
#> search[ 2.615e+09 ] = 439.953(42.2422)
#> search[ 2.615e+10 ] = 439.953(42.2422)
#> 
#> Running at Laplace(2.76553e-06) None  Grid-point #13 at 2.615e+11 	Fold #1 Rep #1 pred log like = 372.477
#> Running at Laplace(2.76553e-06) None  Grid-point #13 at 2.615e+11 	Fold #2 Rep #1 pred log like = 427.325
#> Running at Laplace(2.76553e-06) None  Grid-point #13 at 2.615e+11 	Fold #3 Rep #1 pred log like = 468.972
#> Running at Laplace(2.76553e-06) None  Grid-point #13 at 2.615e+11 	Fold #4 Rep #1 pred log like = 505.448
#> Running at Laplace(2.76553e-06) None  Grid-point #13 at 2.615e+11 	Fold #5 Rep #1 pred log like = 364.283
#> Running at Laplace(2.76553e-06) None  Grid-point #13 at 2.615e+11 	Fold #6 Rep #1 pred log like = 433.204
#> Running at Laplace(2.76553e-06) None  Grid-point #13 at 2.615e+11 	Fold #7 Rep #1 pred log like = 443.18
#> Running at Laplace(2.76553e-06) None  Grid-point #13 at 2.615e+11 	Fold #8 Rep #1 pred log like = 438.402
#> Running at Laplace(2.76553e-06) None  Grid-point #13 at 2.615e+11 	Fold #9 Rep #1 pred log like = 475.918
#> Running at Laplace(2.76553e-06) None  Grid-point #13 at 2.615e+11 	Fold #10 Rep #1 pred log like = 470.318
#> AvgPred = 439.953 with stdev = 42.2422
#> Completed at 2.615e+11
#> Next point at 2.615e+12 with value 0 and continue = 1
#> search[ 0.2615 ] = 439.952(42.2492)
#> search[ 2.615 ] = 439.952(42.2447)
#> search[ 26.15 ] = 439.953(42.243)
#> search[ 261.5 ] = 439.953(42.2425)
#> search[ 2615 ] = 439.953(42.2423)
#> search[ 26150 ] = 439.953(42.2423)
#> search[ 261500 ] = 439.953(42.2422)
#> search[ 2.615e+06 ] = 439.953(42.2422)
#> search[ 2.615e+07 ] = 439.953(42.2422)
#> search[ 2.615e+08 ] = 439.953(42.2422)
#> search[ 2.615e+09 ] = 439.953(42.2422)
#> search[ 2.615e+10 ] = 439.953(42.2422)
#> search[ 2.615e+11 ] = 439.953(42.2422)
#> 
#> Running at Laplace(8.74539e-07) None  Grid-point #14 at 2.615e+12 	Fold #1 Rep #1 pred log like = 372.477
#> Running at Laplace(8.74539e-07) None  Grid-point #14 at 2.615e+12 	Fold #2 Rep #1 pred log like = 427.325
#> Running at Laplace(8.74539e-07) None  Grid-point #14 at 2.615e+12 	Fold #3 Rep #1 pred log like = 468.972
#> Running at Laplace(8.74539e-07) None  Grid-point #14 at 2.615e+12 	Fold #4 Rep #1 pred log like = 505.448
#> Running at Laplace(8.74539e-07) None  Grid-point #14 at 2.615e+12 	Fold #5 Rep #1 pred log like = 364.283
#> Running at Laplace(8.74539e-07) None  Grid-point #14 at 2.615e+12 	Fold #6 Rep #1 pred log like = 433.204
#> Running at Laplace(8.74539e-07) None  Grid-point #14 at 2.615e+12 	Fold #7 Rep #1 pred log like = 443.18
#> Running at Laplace(8.74539e-07) None  Grid-point #14 at 2.615e+12 	Fold #8 Rep #1 pred log like = 438.402
#> Running at Laplace(8.74539e-07) None  Grid-point #14 at 2.615e+12 	Fold #9 Rep #1 pred log like = 475.918
#> Running at Laplace(8.74539e-07) None  Grid-point #14 at 2.615e+12 	Fold #10 Rep #1 pred log like = 470.318
#> AvgPred = 439.953 with stdev = 42.2422
#> Completed at 2.615e+12
#> Next point at 2.615e+13 with value 0 and continue = 1
#> search[ 0.2615 ] = 439.952(42.2492)
#> search[ 2.615 ] = 439.952(42.2447)
#> search[ 26.15 ] = 439.953(42.243)
#> search[ 261.5 ] = 439.953(42.2425)
#> search[ 2615 ] = 439.953(42.2423)
#> search[ 26150 ] = 439.953(42.2423)
#> search[ 261500 ] = 439.953(42.2422)
#> search[ 2.615e+06 ] = 439.953(42.2422)
#> search[ 2.615e+07 ] = 439.953(42.2422)
#> search[ 2.615e+08 ] = 439.953(42.2422)
#> search[ 2.615e+09 ] = 439.953(42.2422)
#> search[ 2.615e+10 ] = 439.953(42.2422)
#> search[ 2.615e+11 ] = 439.953(42.2422)
#> search[ 2.615e+12 ] = 439.953(42.2422)
#> 
#> Running at Laplace(2.76553e-07) None  Grid-point #15 at 2.615e+13 	Fold #1 Rep #1 pred log like = 372.477
#> Running at Laplace(2.76553e-07) None  Grid-point #15 at 2.615e+13 	Fold #2 Rep #1 pred log like = 427.325
#> Running at Laplace(2.76553e-07) None  Grid-point #15 at 2.615e+13 	Fold #3 Rep #1 pred log like = 468.972
#> Running at Laplace(2.76553e-07) None  Grid-point #15 at 2.615e+13 	Fold #4 Rep #1 pred log like = 505.448
#> Running at Laplace(2.76553e-07) None  Grid-point #15 at 2.615e+13 	Fold #5 Rep #1 pred log like = 364.283
#> Running at Laplace(2.76553e-07) None  Grid-point #15 at 2.615e+13 	Fold #6 Rep #1 pred log like = 433.204
#> Running at Laplace(2.76553e-07) None  Grid-point #15 at 2.615e+13 	Fold #7 Rep #1 pred log like = 443.18
#> Running at Laplace(2.76553e-07) None  Grid-point #15 at 2.615e+13 	Fold #8 Rep #1 pred log like = 438.402
#> Running at Laplace(2.76553e-07) None  Grid-point #15 at 2.615e+13 	Fold #9 Rep #1 pred log like = 475.918
#> Running at Laplace(2.76553e-07) None  Grid-point #15 at 2.615e+13 	Fold #10 Rep #1 pred log like = 470.318
#> AvgPred = 439.953 with stdev = 42.2422
#> Completed at 2.615e+13
#> Next point at 2.615e+14 with value 0 and continue = 1
#> search[ 0.2615 ] = 439.952(42.2492)
#> search[ 2.615 ] = 439.952(42.2447)
#> search[ 26.15 ] = 439.953(42.243)
#> search[ 261.5 ] = 439.953(42.2425)
#> search[ 2615 ] = 439.953(42.2423)
#> search[ 26150 ] = 439.953(42.2423)
#> search[ 261500 ] = 439.953(42.2422)
#> search[ 2.615e+06 ] = 439.953(42.2422)
#> search[ 2.615e+07 ] = 439.953(42.2422)
#> search[ 2.615e+08 ] = 439.953(42.2422)
#> search[ 2.615e+09 ] = 439.953(42.2422)
#> search[ 2.615e+10 ] = 439.953(42.2422)
#> search[ 2.615e+11 ] = 439.953(42.2422)
#> search[ 2.615e+12 ] = 439.953(42.2422)
#> search[ 2.615e+13 ] = 439.953(42.2422)
#> 
#> Running at Laplace(8.74539e-08) None  Grid-point #16 at 2.615e+14 	Fold #1 Rep #1 pred log like = 372.477
#> Running at Laplace(8.74539e-08) None  Grid-point #16 at 2.615e+14 	Fold #2 Rep #1 pred log like = 427.325
#> Running at Laplace(8.74539e-08) None  Grid-point #16 at 2.615e+14 	Fold #3 Rep #1 pred log like = 468.972
#> Running at Laplace(8.74539e-08) None  Grid-point #16 at 2.615e+14 	Fold #4 Rep #1 pred log like = 505.448
#> Running at Laplace(8.74539e-08) None  Grid-point #16 at 2.615e+14 	Fold #5 Rep #1 pred log like = 364.283
#> Running at Laplace(8.74539e-08) None  Grid-point #16 at 2.615e+14 	Fold #6 Rep #1 pred log like = 433.204
#> Running at Laplace(8.74539e-08) None  Grid-point #16 at 2.615e+14 	Fold #7 Rep #1 pred log like = 443.18
#> Running at Laplace(8.74539e-08) None  Grid-point #16 at 2.615e+14 	Fold #8 Rep #1 pred log like = 438.402
#> Running at Laplace(8.74539e-08) None  Grid-point #16 at 2.615e+14 	Fold #9 Rep #1 pred log like = 475.918
#> Running at Laplace(8.74539e-08) None  Grid-point #16 at 2.615e+14 	Fold #10 Rep #1 pred log like = 470.318
#> AvgPred = 439.953 with stdev = 42.2422
#> Completed at 2.615e+14
#> Next point at 2.615e+15 with value 0 and continue = 1
#> search[ 0.2615 ] = 439.952(42.2492)
#> search[ 2.615 ] = 439.952(42.2447)
#> search[ 26.15 ] = 439.953(42.243)
#> search[ 261.5 ] = 439.953(42.2425)
#> search[ 2615 ] = 439.953(42.2423)
#> search[ 26150 ] = 439.953(42.2423)
#> search[ 261500 ] = 439.953(42.2422)
#> search[ 2.615e+06 ] = 439.953(42.2422)
#> search[ 2.615e+07 ] = 439.953(42.2422)
#> search[ 2.615e+08 ] = 439.953(42.2422)
#> search[ 2.615e+09 ] = 439.953(42.2422)
#> search[ 2.615e+10 ] = 439.953(42.2422)
#> search[ 2.615e+11 ] = 439.953(42.2422)
#> search[ 2.615e+12 ] = 439.953(42.2422)
#> search[ 2.615e+13 ] = 439.953(42.2422)
#> search[ 2.615e+14 ] = 439.953(42.2422)
#> 
#> Running at Laplace(2.76553e-08) None  Grid-point #17 at 2.615e+15 	Fold #1 Rep #1 pred log like = 372.477
#> Running at Laplace(2.76553e-08) None  Grid-point #17 at 2.615e+15 	Fold #2 Rep #1 pred log like = 427.325
#> Running at Laplace(2.76553e-08) None  Grid-point #17 at 2.615e+15 	Fold #3 Rep #1 pred log like = 468.972
#> Running at Laplace(2.76553e-08) None  Grid-point #17 at 2.615e+15 	Fold #4 Rep #1 pred log like = 505.448
#> Running at Laplace(2.76553e-08) None  Grid-point #17 at 2.615e+15 	Fold #5 Rep #1 pred log like = 364.283
#> Running at Laplace(2.76553e-08) None  Grid-point #17 at 2.615e+15 	Fold #6 Rep #1 pred log like = 433.204
#> Running at Laplace(2.76553e-08) None  Grid-point #17 at 2.615e+15 	Fold #7 Rep #1 pred log like = 443.18
#> Running at Laplace(2.76553e-08) None  Grid-point #17 at 2.615e+15 	Fold #8 Rep #1 pred log like = 438.402
#> Running at Laplace(2.76553e-08) None  Grid-point #17 at 2.615e+15 	Fold #9 Rep #1 pred log like = 475.918
#> Running at Laplace(2.76553e-08) None  Grid-point #17 at 2.615e+15 	Fold #10 Rep #1 pred log like = 470.318
#> AvgPred = 439.953 with stdev = 42.2422
#> Completed at 2.615e+15
#> Next point at 2.615e+16 with value 0 and continue = 1
#> search[ 0.2615 ] = 439.952(42.2492)
#> search[ 2.615 ] = 439.952(42.2447)
#> search[ 26.15 ] = 439.953(42.243)
#> search[ 261.5 ] = 439.953(42.2425)
#> search[ 2615 ] = 439.953(42.2423)
#> search[ 26150 ] = 439.953(42.2423)
#> search[ 261500 ] = 439.953(42.2422)
#> search[ 2.615e+06 ] = 439.953(42.2422)
#> search[ 2.615e+07 ] = 439.953(42.2422)
#> search[ 2.615e+08 ] = 439.953(42.2422)
#> search[ 2.615e+09 ] = 439.953(42.2422)
#> search[ 2.615e+10 ] = 439.953(42.2422)
#> search[ 2.615e+11 ] = 439.953(42.2422)
#> search[ 2.615e+12 ] = 439.953(42.2422)
#> search[ 2.615e+13 ] = 439.953(42.2422)
#> search[ 2.615e+14 ] = 439.953(42.2422)
#> search[ 2.615e+15 ] = 439.953(42.2422)
#> 
#> Running at Laplace(8.74539e-09) None  Grid-point #18 at 2.615e+16 	Fold #1 Rep #1 pred log like = 372.477
#> Running at Laplace(8.74539e-09) None  Grid-point #18 at 2.615e+16 	Fold #2 Rep #1 pred log like = 427.325
#> Running at Laplace(8.74539e-09) None  Grid-point #18 at 2.615e+16 	Fold #3 Rep #1 pred log like = 468.972
#> Running at Laplace(8.74539e-09) None  Grid-point #18 at 2.615e+16 	Fold #4 Rep #1 pred log like = 505.448
#> Running at Laplace(8.74539e-09) None  Grid-point #18 at 2.615e+16 	Fold #5 Rep #1 pred log like = 364.283
#> Running at Laplace(8.74539e-09) None  Grid-point #18 at 2.615e+16 	Fold #6 Rep #1 pred log like = 433.204
#> Running at Laplace(8.74539e-09) None  Grid-point #18 at 2.615e+16 	Fold #7 Rep #1 pred log like = 443.18
#> Running at Laplace(8.74539e-09) None  Grid-point #18 at 2.615e+16 	Fold #8 Rep #1 pred log like = 438.402
#> Running at Laplace(8.74539e-09) None  Grid-point #18 at 2.615e+16 	Fold #9 Rep #1 pred log like = 475.918
#> Running at Laplace(8.74539e-09) None  Grid-point #18 at 2.615e+16 	Fold #10 Rep #1 pred log like = 470.318
#> AvgPred = 439.953 with stdev = 42.2422
#> Completed at 2.615e+16
#> Next point at 2.615e+17 with value 0 and continue = 1
#> search[ 0.2615 ] = 439.952(42.2492)
#> search[ 2.615 ] = 439.952(42.2447)
#> search[ 26.15 ] = 439.953(42.243)
#> search[ 261.5 ] = 439.953(42.2425)
#> search[ 2615 ] = 439.953(42.2423)
#> search[ 26150 ] = 439.953(42.2423)
#> search[ 261500 ] = 439.953(42.2422)
#> search[ 2.615e+06 ] = 439.953(42.2422)
#> search[ 2.615e+07 ] = 439.953(42.2422)
#> search[ 2.615e+08 ] = 439.953(42.2422)
#> search[ 2.615e+09 ] = 439.953(42.2422)
#> search[ 2.615e+10 ] = 439.953(42.2422)
#> search[ 2.615e+11 ] = 439.953(42.2422)
#> search[ 2.615e+12 ] = 439.953(42.2422)
#> search[ 2.615e+13 ] = 439.953(42.2422)
#> search[ 2.615e+14 ] = 439.953(42.2422)
#> search[ 2.615e+15 ] = 439.953(42.2422)
#> search[ 2.615e+16 ] = 439.953(42.2422)
#> 
#> Running at Laplace(2.76553e-09) None  Grid-point #19 at 2.615e+17 	Fold #1 Rep #1 pred log like = 372.477
#> Running at Laplace(2.76553e-09) None  Grid-point #19 at 2.615e+17 	Fold #2 Rep #1 pred log like = 427.325
#> Running at Laplace(2.76553e-09) None  Grid-point #19 at 2.615e+17 	Fold #3 Rep #1 pred log like = 468.972
#> Running at Laplace(2.76553e-09) None  Grid-point #19 at 2.615e+17 	Fold #4 Rep #1 pred log like = 505.448
#> Running at Laplace(2.76553e-09) None  Grid-point #19 at 2.615e+17 	Fold #5 Rep #1 pred log like = 364.283
#> Running at Laplace(2.76553e-09) None  Grid-point #19 at 2.615e+17 	Fold #6 Rep #1 pred log like = 433.204
#> Running at Laplace(2.76553e-09) None  Grid-point #19 at 2.615e+17 	Fold #7 Rep #1 pred log like = 443.18
#> Running at Laplace(2.76553e-09) None  Grid-point #19 at 2.615e+17 	Fold #8 Rep #1 pred log like = 438.402
#> Running at Laplace(2.76553e-09) None  Grid-point #19 at 2.615e+17 	Fold #9 Rep #1 pred log like = 475.918
#> Running at Laplace(2.76553e-09) None  Grid-point #19 at 2.615e+17 	Fold #10 Rep #1 pred log like = 470.318
#> AvgPred = 439.953 with stdev = 42.2422
#> Completed at 2.615e+17
#> Next point at 2.615e+18 with value 0 and continue = 1
#> search[ 0.2615 ] = 439.952(42.2492)
#> search[ 2.615 ] = 439.952(42.2447)
#> search[ 26.15 ] = 439.953(42.243)
#> search[ 261.5 ] = 439.953(42.2425)
#> search[ 2615 ] = 439.953(42.2423)
#> search[ 26150 ] = 439.953(42.2423)
#> search[ 261500 ] = 439.953(42.2422)
#> search[ 2.615e+06 ] = 439.953(42.2422)
#> search[ 2.615e+07 ] = 439.953(42.2422)
#> search[ 2.615e+08 ] = 439.953(42.2422)
#> search[ 2.615e+09 ] = 439.953(42.2422)
#> search[ 2.615e+10 ] = 439.953(42.2422)
#> search[ 2.615e+11 ] = 439.953(42.2422)
#> search[ 2.615e+12 ] = 439.953(42.2422)
#> search[ 2.615e+13 ] = 439.953(42.2422)
#> search[ 2.615e+14 ] = 439.953(42.2422)
#> search[ 2.615e+15 ] = 439.953(42.2422)
#> search[ 2.615e+16 ] = 439.953(42.2422)
#> search[ 2.615e+17 ] = 439.953(42.2422)
#> 
#> Running at Laplace(8.74539e-10) None  Grid-point #20 at 2.615e+18 	Fold #1 Rep #1 pred log like = 372.477
#> Running at Laplace(8.74539e-10) None  Grid-point #20 at 2.615e+18 	Fold #2 Rep #1 pred log like = 427.325
#> Running at Laplace(8.74539e-10) None  Grid-point #20 at 2.615e+18 	Fold #3 Rep #1 pred log like = 468.972
#> Running at Laplace(8.74539e-10) None  Grid-point #20 at 2.615e+18 	Fold #4 Rep #1 pred log like = 505.448
#> Running at Laplace(8.74539e-10) None  Grid-point #20 at 2.615e+18 	Fold #5 Rep #1 pred log like = 364.283
#> Running at Laplace(8.74539e-10) None  Grid-point #20 at 2.615e+18 	Fold #6 Rep #1 pred log like = 433.204
#> Running at Laplace(8.74539e-10) None  Grid-point #20 at 2.615e+18 	Fold #7 Rep #1 pred log like = 443.18
#> Running at Laplace(8.74539e-10) None  Grid-point #20 at 2.615e+18 	Fold #8 Rep #1 pred log like = 438.402
#> Running at Laplace(8.74539e-10) None  Grid-point #20 at 2.615e+18 	Fold #9 Rep #1 pred log like = 475.918
#> Running at Laplace(8.74539e-10) None  Grid-point #20 at 2.615e+18 	Fold #10 Rep #1 pred log like = 470.318
#> AvgPred = 439.953 with stdev = 42.2422
#> Completed at 2.615e+18
#> Next point at 2.615e+19 with value 0 and continue = 1
#> search[ 0.2615 ] = 439.952(42.2492)
#> search[ 2.615 ] = 439.952(42.2447)
#> search[ 26.15 ] = 439.953(42.243)
#> search[ 261.5 ] = 439.953(42.2425)
#> search[ 2615 ] = 439.953(42.2423)
#> search[ 26150 ] = 439.953(42.2423)
#> search[ 261500 ] = 439.953(42.2422)
#> search[ 2.615e+06 ] = 439.953(42.2422)
#> search[ 2.615e+07 ] = 439.953(42.2422)
#> search[ 2.615e+08 ] = 439.953(42.2422)
#> search[ 2.615e+09 ] = 439.953(42.2422)
#> search[ 2.615e+10 ] = 439.953(42.2422)
#> search[ 2.615e+11 ] = 439.953(42.2422)
#> search[ 2.615e+12 ] = 439.953(42.2422)
#> search[ 2.615e+13 ] = 439.953(42.2422)
#> search[ 2.615e+14 ] = 439.953(42.2422)
#> search[ 2.615e+15 ] = 439.953(42.2422)
#> search[ 2.615e+16 ] = 439.953(42.2422)
#> search[ 2.615e+17 ] = 439.953(42.2422)
#> search[ 2.615e+18 ] = 439.953(42.2422)
#> 
#> Running at Laplace(2.76553e-10) None  Grid-point #21 at 2.615e+19 	Fold #1 Rep #1 pred log like = 372.477
#> Running at Laplace(2.76553e-10) None  Grid-point #21 at 2.615e+19 	Fold #2 Rep #1 pred log like = 427.325
#> Running at Laplace(2.76553e-10) None  Grid-point #21 at 2.615e+19 	Fold #3 Rep #1 pred log like = 468.972
#> Running at Laplace(2.76553e-10) None  Grid-point #21 at 2.615e+19 	Fold #4 Rep #1 pred log like = 505.448
#> Running at Laplace(2.76553e-10) None  Grid-point #21 at 2.615e+19 	Fold #5 Rep #1 pred log like = 364.283
#> Running at Laplace(2.76553e-10) None  Grid-point #21 at 2.615e+19 	Fold #6 Rep #1 pred log like = 433.204
#> Running at Laplace(2.76553e-10) None  Grid-point #21 at 2.615e+19 	Fold #7 Rep #1 pred log like = 443.18
#> Running at Laplace(2.76553e-10) None  Grid-point #21 at 2.615e+19 	Fold #8 Rep #1 pred log like = 438.402
#> Running at Laplace(2.76553e-10) None  Grid-point #21 at 2.615e+19 	Fold #9 Rep #1 pred log like = 475.918
#> Running at Laplace(2.76553e-10) None  Grid-point #21 at 2.615e+19 	Fold #10 Rep #1 pred log like = 470.318
#> AvgPred = 439.953 with stdev = 42.2422
#> Completed at 2.615e+19
#> Next point at 2.615e+20 with value 0 and continue = 1
#> search[ 0.2615 ] = 439.952(42.2492)
#> search[ 2.615 ] = 439.952(42.2447)
#> search[ 26.15 ] = 439.953(42.243)
#> search[ 261.5 ] = 439.953(42.2425)
#> search[ 2615 ] = 439.953(42.2423)
#> search[ 26150 ] = 439.953(42.2423)
#> search[ 261500 ] = 439.953(42.2422)
#> search[ 2.615e+06 ] = 439.953(42.2422)
#> search[ 2.615e+07 ] = 439.953(42.2422)
#> search[ 2.615e+08 ] = 439.953(42.2422)
#> search[ 2.615e+09 ] = 439.953(42.2422)
#> search[ 2.615e+10 ] = 439.953(42.2422)
#> search[ 2.615e+11 ] = 439.953(42.2422)
#> search[ 2.615e+12 ] = 439.953(42.2422)
#> search[ 2.615e+13 ] = 439.953(42.2422)
#> search[ 2.615e+14 ] = 439.953(42.2422)
#> search[ 2.615e+15 ] = 439.953(42.2422)
#> search[ 2.615e+16 ] = 439.953(42.2422)
#> search[ 2.615e+17 ] = 439.953(42.2422)
#> search[ 2.615e+18 ] = 439.953(42.2422)
#> search[ 2.615e+19 ] = 439.953(42.2422)
#> 
#> Running at Laplace(8.74539e-11) None  Grid-point #22 at 2.615e+20 	Fold #1 Rep #1 pred log like = 372.477
#> Running at Laplace(8.74539e-11) None  Grid-point #22 at 2.615e+20 	Fold #2 Rep #1 pred log like = 427.325
#> Running at Laplace(8.74539e-11) None  Grid-point #22 at 2.615e+20 	Fold #3 Rep #1 pred log like = 468.972
#> Running at Laplace(8.74539e-11) None  Grid-point #22 at 2.615e+20 	Fold #4 Rep #1 pred log like = 505.448
#> Running at Laplace(8.74539e-11) None  Grid-point #22 at 2.615e+20 	Fold #5 Rep #1 pred log like = 364.283
#> Running at Laplace(8.74539e-11) None  Grid-point #22 at 2.615e+20 	Fold #6 Rep #1 pred log like = 433.204
#> Running at Laplace(8.74539e-11) None  Grid-point #22 at 2.615e+20 	Fold #7 Rep #1 pred log like = 443.18
#> Running at Laplace(8.74539e-11) None  Grid-point #22 at 2.615e+20 	Fold #8 Rep #1 pred log like = 438.402
#> Running at Laplace(8.74539e-11) None  Grid-point #22 at 2.615e+20 	Fold #9 Rep #1 pred log like = 475.918
#> Running at Laplace(8.74539e-11) None  Grid-point #22 at 2.615e+20 	Fold #10 Rep #1 pred log like = 470.318
#> AvgPred = 439.953 with stdev = 42.2422
#> Completed at 2.615e+20
#> Next point at 3.27402e+12 with value 439.953 and continue = 0
#> search[ 0.2615 ] = 439.952(42.2492)
#> search[ 2.615 ] = 439.952(42.2447)
#> search[ 26.15 ] = 439.953(42.243)
#> search[ 261.5 ] = 439.953(42.2425)
#> search[ 2615 ] = 439.953(42.2423)
#> search[ 26150 ] = 439.953(42.2423)
#> search[ 261500 ] = 439.953(42.2422)
#> search[ 2.615e+06 ] = 439.953(42.2422)
#> search[ 2.615e+07 ] = 439.953(42.2422)
#> search[ 2.615e+08 ] = 439.953(42.2422)
#> search[ 2.615e+09 ] = 439.953(42.2422)
#> search[ 2.615e+10 ] = 439.953(42.2422)
#> search[ 2.615e+11 ] = 439.953(42.2422)
#> search[ 2.615e+12 ] = 439.953(42.2422)
#> search[ 2.615e+13 ] = 439.953(42.2422)
#> search[ 2.615e+14 ] = 439.953(42.2422)
#> search[ 2.615e+15 ] = 439.953(42.2422)
#> search[ 2.615e+16 ] = 439.953(42.2422)
#> search[ 2.615e+17 ] = 439.953(42.2422)
#> search[ 2.615e+18 ] = 439.953(42.2422)
#> search[ 2.615e+19 ] = 439.953(42.2422)
#> search[ 2.615e+20 ] = 439.953(42.2422)
#> 
#> 
#> Maximum predicted log likelihood (439.953) estimated at:
#> 	3.27402e+12 (variance)
#> 	7.81582e-07 (lambda)
#> 
#> Fitting model at optimal hyperparameter
#> Using prior: Laplace(7.81582e-07) None 

#Find out what the optimal hyperparameter was:
getHyperParameter(fit)
#> [1] 3.274021e+12

#Extract the current log-likelihood, and coefficients
logLik(fit)
#> 'log Lik.' -2097.62 (df=3)
coef(fit)
#> (Intercept)           1           2 
#> -3.87566911  0.05844875  0.01875022 

#We can only retrieve the confidence interval for unregularized coefficients:
confint(fit, c(0))
#> Using 1 thread(s)
#>             covariate     2.5 %    97.5 % evaluations
#> (Intercept)         0 -3.910302 -3.841399          22