Get hyperparameter

getHyperParameter returns the current hyper parameter in a Cyclops model fit object

getHyperParameter(object)

Arguments

object
A Cyclops model fit object

Examples

#Generate some simulated data:
sim <- simulateCyclopsData(nstrata = 1, nrows = 1000, ncovars = 2, eCovarsPerRow = 0.5,
                           model = "poisson")

#> Sparseness = 75.45 %

cyclopsData <- convertToCyclopsData(sim$outcomes, sim$covariates, modelType = "pr",
                                    addIntercept = TRUE)

#> Sorting covariates by covariateId and rowId

#Define the prior and control objects to use cross-validation for finding the 
#optimal hyperparameter:
prior <- createPrior("laplace", exclude = 0, useCrossValidation = TRUE)
control <- createControl(cvType = "auto", noiseLevel = "quiet")

#Fit the model
fit <- fitCyclopsModel(cyclopsData,prior = prior, control = control)

#> Using cross-validation selector type byRow
#> Performing 10-fold cross-validation [seed = 1543260557] with data partitions of sizes 100 100 100 100 100 100 100 100 100 100
#> Using 1 thread(s)
#> Starting var = 0.2455 (default)
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #1 Rep #1 pred log like = -192.832
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #2 Rep #1 pred log like = -183.902
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #3 Rep #1 pred log like = -202.428
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #4 Rep #1 pred log like = -185.294
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #5 Rep #1 pred log like = -198.921
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #6 Rep #1 pred log like = -196.471
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #7 Rep #1 pred log like = -194.329
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #8 Rep #1 pred log like = -190.492
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #9 Rep #1 pred log like = -204.559
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #10 Rep #1 pred log like = -180.695
#> AvgPred = -192.992 with stdev = 7.55817
#> Completed at 0.2455
#> Next point at 2.455 with value 0 and continue = 1
#> search[ 0.2455 ] = -192.992(7.55817)
#> 
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #1 Rep #1 pred log like = -192.762
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #2 Rep #1 pred log like = -183.917
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #3 Rep #1 pred log like = -202.374
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #4 Rep #1 pred log like = -185.325
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #5 Rep #1 pred log like = -198.964
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #6 Rep #1 pred log like = -196.494
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #7 Rep #1 pred log like = -194.326
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #8 Rep #1 pred log like = -190.446
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #9 Rep #1 pred log like = -204.581
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #10 Rep #1 pred log like = -180.719
#> AvgPred = -192.991 with stdev = 7.55194
#> Completed at 2.455
#> Next point at 24.55 with value 0 and continue = 1
#> search[ 0.2455 ] = -192.992(7.55817)
#> search[ 2.455 ] = -192.991(7.55194)
#> 
#> Running at Laplace(0.285423) None  Grid-point #3 at 24.55 	Fold #1 Rep #1 pred log like = -192.74
#> Running at Laplace(0.285423) None  Grid-point #3 at 24.55 	Fold #2 Rep #1 pred log like = -183.922
#> Running at Laplace(0.285423) None  Grid-point #3 at 24.55 	Fold #3 Rep #1 pred log like = -202.357
#> Running at Laplace(0.285423) None  Grid-point #3 at 24.55 	Fold #4 Rep #1 pred log like = -185.336
#> Running at Laplace(0.285423) None  Grid-point #3 at 24.55 	Fold #5 Rep #1 pred log like = -198.978
#> Running at Laplace(0.285423) None  Grid-point #3 at 24.55 	Fold #6 Rep #1 pred log like = -196.502
#> Running at Laplace(0.285423) None  Grid-point #3 at 24.55 	Fold #7 Rep #1 pred log like = -194.325
#> Running at Laplace(0.285423) None  Grid-point #3 at 24.55 	Fold #8 Rep #1 pred log like = -190.431
#> Running at Laplace(0.285423) None  Grid-point #3 at 24.55 	Fold #9 Rep #1 pred log like = -204.588
#> Running at Laplace(0.285423) None  Grid-point #3 at 24.55 	Fold #10 Rep #1 pred log like = -180.727
#> AvgPred = -192.991 with stdev = 7.55001
#> Completed at 24.55
#> Next point at 245.5 with value 0 and continue = 1
#> search[ 0.2455 ] = -192.992(7.55817)
#> search[ 2.455 ] = -192.991(7.55194)
#> search[ 24.55 ] = -192.991(7.55001)
#> 
#> Running at Laplace(0.0902587) None  Grid-point #4 at 245.5 	Fold #1 Rep #1 pred log like = -192.734
#> Running at Laplace(0.0902587) None  Grid-point #4 at 245.5 	Fold #2 Rep #1 pred log like = -183.924
#> Running at Laplace(0.0902587) None  Grid-point #4 at 245.5 	Fold #3 Rep #1 pred log like = -202.352
#> Running at Laplace(0.0902587) None  Grid-point #4 at 245.5 	Fold #4 Rep #1 pred log like = -185.339
#> Running at Laplace(0.0902587) None  Grid-point #4 at 245.5 	Fold #5 Rep #1 pred log like = -198.982
#> Running at Laplace(0.0902587) None  Grid-point #4 at 245.5 	Fold #6 Rep #1 pred log like = -196.504
#> Running at Laplace(0.0902587) None  Grid-point #4 at 245.5 	Fold #7 Rep #1 pred log like = -194.325
#> Running at Laplace(0.0902587) None  Grid-point #4 at 245.5 	Fold #8 Rep #1 pred log like = -190.427
#> Running at Laplace(0.0902587) None  Grid-point #4 at 245.5 	Fold #9 Rep #1 pred log like = -204.59
#> Running at Laplace(0.0902587) None  Grid-point #4 at 245.5 	Fold #10 Rep #1 pred log like = -180.73
#> AvgPred = -192.991 with stdev = 7.5494
#> Completed at 245.5
#> Next point at 2455 with value 0 and continue = 1
#> search[ 0.2455 ] = -192.992(7.55817)
#> search[ 2.455 ] = -192.991(7.55194)
#> search[ 24.55 ] = -192.991(7.55001)
#> search[ 245.5 ] = -192.991(7.5494)
#> 
#> Running at Laplace(0.0285423) None  Grid-point #5 at 2455 	Fold #1 Rep #1 pred log like = -192.731
#> Running at Laplace(0.0285423) None  Grid-point #5 at 2455 	Fold #2 Rep #1 pred log like = -183.924
#> Running at Laplace(0.0285423) None  Grid-point #5 at 2455 	Fold #3 Rep #1 pred log like = -202.35
#> Running at Laplace(0.0285423) None  Grid-point #5 at 2455 	Fold #4 Rep #1 pred log like = -185.34
#> Running at Laplace(0.0285423) None  Grid-point #5 at 2455 	Fold #5 Rep #1 pred log like = -198.984
#> Running at Laplace(0.0285423) None  Grid-point #5 at 2455 	Fold #6 Rep #1 pred log like = -196.505
#> Running at Laplace(0.0285423) None  Grid-point #5 at 2455 	Fold #7 Rep #1 pred log like = -194.325
#> Running at Laplace(0.0285423) None  Grid-point #5 at 2455 	Fold #8 Rep #1 pred log like = -190.425
#> Running at Laplace(0.0285423) None  Grid-point #5 at 2455 	Fold #9 Rep #1 pred log like = -204.591
#> Running at Laplace(0.0285423) None  Grid-point #5 at 2455 	Fold #10 Rep #1 pred log like = -180.731
#> AvgPred = -192.991 with stdev = 7.5492
#> Completed at 2455
#> Next point at 24550 with value 0 and continue = 1
#> search[ 0.2455 ] = -192.992(7.55817)
#> search[ 2.455 ] = -192.991(7.55194)
#> search[ 24.55 ] = -192.991(7.55001)
#> search[ 245.5 ] = -192.991(7.5494)
#> search[ 2455 ] = -192.991(7.5492)
#> 
#> Running at Laplace(0.00902587) None  Grid-point #6 at 24550 	Fold #1 Rep #1 pred log like = -192.731
#> Running at Laplace(0.00902587) None  Grid-point #6 at 24550 	Fold #2 Rep #1 pred log like = -183.925
#> Running at Laplace(0.00902587) None  Grid-point #6 at 24550 	Fold #3 Rep #1 pred log like = -202.35
#> Running at Laplace(0.00902587) None  Grid-point #6 at 24550 	Fold #4 Rep #1 pred log like = -185.34
#> Running at Laplace(0.00902587) None  Grid-point #6 at 24550 	Fold #5 Rep #1 pred log like = -198.984
#> Running at Laplace(0.00902587) None  Grid-point #6 at 24550 	Fold #6 Rep #1 pred log like = -196.505
#> Running at Laplace(0.00902587) None  Grid-point #6 at 24550 	Fold #7 Rep #1 pred log like = -194.325
#> Running at Laplace(0.00902587) None  Grid-point #6 at 24550 	Fold #8 Rep #1 pred log like = -190.425
#> Running at Laplace(0.00902587) None  Grid-point #6 at 24550 	Fold #9 Rep #1 pred log like = -204.591
#> Running at Laplace(0.00902587) None  Grid-point #6 at 24550 	Fold #10 Rep #1 pred log like = -180.731
#> AvgPred = -192.991 with stdev = 7.54914
#> Completed at 24550
#> Next point at 245500 with value 0 and continue = 1
#> search[ 0.2455 ] = -192.992(7.55817)
#> search[ 2.455 ] = -192.991(7.55194)
#> search[ 24.55 ] = -192.991(7.55001)
#> search[ 245.5 ] = -192.991(7.5494)
#> search[ 2455 ] = -192.991(7.5492)
#> search[ 24550 ] = -192.991(7.54914)
#> 
#> Running at Laplace(0.00285423) None  Grid-point #7 at 245500 	Fold #1 Rep #1 pred log like = -192.731
#> Running at Laplace(0.00285423) None  Grid-point #7 at 245500 	Fold #2 Rep #1 pred log like = -183.925
#> Running at Laplace(0.00285423) None  Grid-point #7 at 245500 	Fold #3 Rep #1 pred log like = -202.35
#> Running at Laplace(0.00285423) None  Grid-point #7 at 245500 	Fold #4 Rep #1 pred log like = -185.34
#> Running at Laplace(0.00285423) None  Grid-point #7 at 245500 	Fold #5 Rep #1 pred log like = -198.984
#> Running at Laplace(0.00285423) None  Grid-point #7 at 245500 	Fold #6 Rep #1 pred log like = -196.506
#> Running at Laplace(0.00285423) None  Grid-point #7 at 245500 	Fold #7 Rep #1 pred log like = -194.325
#> Running at Laplace(0.00285423) None  Grid-point #7 at 245500 	Fold #8 Rep #1 pred log like = -190.425
#> Running at Laplace(0.00285423) None  Grid-point #7 at 245500 	Fold #9 Rep #1 pred log like = -204.591
#> Running at Laplace(0.00285423) None  Grid-point #7 at 245500 	Fold #10 Rep #1 pred log like = -180.731
#> AvgPred = -192.991 with stdev = 7.54912
#> Completed at 245500
#> Next point at 2.455e+006 with value 0 and continue = 1
#> search[ 0.2455 ] = -192.992(7.55817)
#> search[ 2.455 ] = -192.991(7.55194)
#> search[ 24.55 ] = -192.991(7.55001)
#> search[ 245.5 ] = -192.991(7.5494)
#> search[ 2455 ] = -192.991(7.5492)
#> search[ 24550 ] = -192.991(7.54914)
#> search[ 245500 ] = -192.991(7.54912)
#> 
#> Running at Laplace(0.000902587) None  Grid-point #8 at 2.455e+006 	Fold #1 Rep #1 pred log like = -192.73
#> Running at Laplace(0.000902587) None  Grid-point #8 at 2.455e+006 	Fold #2 Rep #1 pred log like = -183.925
#> Running at Laplace(0.000902587) None  Grid-point #8 at 2.455e+006 	Fold #3 Rep #1 pred log like = -202.35
#> Running at Laplace(0.000902587) None  Grid-point #8 at 2.455e+006 	Fold #4 Rep #1 pred log like = -185.34
#> Running at Laplace(0.000902587) None  Grid-point #8 at 2.455e+006 	Fold #5 Rep #1 pred log like = -198.984
#> Running at Laplace(0.000902587) None  Grid-point #8 at 2.455e+006 	Fold #6 Rep #1 pred log like = -196.506
#> Running at Laplace(0.000902587) None  Grid-point #8 at 2.455e+006 	Fold #7 Rep #1 pred log like = -194.325
#> Running at Laplace(0.000902587) None  Grid-point #8 at 2.455e+006 	Fold #8 Rep #1 pred log like = -190.424
#> Running at Laplace(0.000902587) None  Grid-point #8 at 2.455e+006 	Fold #9 Rep #1 pred log like = -204.591
#> Running at Laplace(0.000902587) None  Grid-point #8 at 2.455e+006 	Fold #10 Rep #1 pred log like = -180.731
#> AvgPred = -192.991 with stdev = 7.54912
#> Completed at 2.455e+006
#> Next point at 2.455e+007 with value 0 and continue = 1
#> search[ 0.2455 ] = -192.992(7.55817)
#> search[ 2.455 ] = -192.991(7.55194)
#> search[ 24.55 ] = -192.991(7.55001)
#> search[ 245.5 ] = -192.991(7.5494)
#> search[ 2455 ] = -192.991(7.5492)
#> search[ 24550 ] = -192.991(7.54914)
#> search[ 245500 ] = -192.991(7.54912)
#> search[ 2.455e+006 ] = -192.991(7.54912)
#> 
#> Running at Laplace(0.000285423) None  Grid-point #9 at 2.455e+007 	Fold #1 Rep #1 pred log like = -192.73
#> Running at Laplace(0.000285423) None  Grid-point #9 at 2.455e+007 	Fold #2 Rep #1 pred log like = -183.925
#> Running at Laplace(0.000285423) None  Grid-point #9 at 2.455e+007 	Fold #3 Rep #1 pred log like = -202.35
#> Running at Laplace(0.000285423) None  Grid-point #9 at 2.455e+007 	Fold #4 Rep #1 pred log like = -185.34
#> Running at Laplace(0.000285423) None  Grid-point #9 at 2.455e+007 	Fold #5 Rep #1 pred log like = -198.984
#> Running at Laplace(0.000285423) None  Grid-point #9 at 2.455e+007 	Fold #6 Rep #1 pred log like = -196.506
#> Running at Laplace(0.000285423) None  Grid-point #9 at 2.455e+007 	Fold #7 Rep #1 pred log like = -194.325
#> Running at Laplace(0.000285423) None  Grid-point #9 at 2.455e+007 	Fold #8 Rep #1 pred log like = -190.424
#> Running at Laplace(0.000285423) None  Grid-point #9 at 2.455e+007 	Fold #9 Rep #1 pred log like = -204.591
#> Running at Laplace(0.000285423) None  Grid-point #9 at 2.455e+007 	Fold #10 Rep #1 pred log like = -180.731
#> AvgPred = -192.991 with stdev = 7.54912
#> Completed at 2.455e+007
#> Next point at 2.455e+008 with value 0 and continue = 1
#> search[ 0.2455 ] = -192.992(7.55817)
#> search[ 2.455 ] = -192.991(7.55194)
#> search[ 24.55 ] = -192.991(7.55001)
#> search[ 245.5 ] = -192.991(7.5494)
#> search[ 2455 ] = -192.991(7.5492)
#> search[ 24550 ] = -192.991(7.54914)
#> search[ 245500 ] = -192.991(7.54912)
#> search[ 2.455e+006 ] = -192.991(7.54912)
#> search[ 2.455e+007 ] = -192.991(7.54912)
#> 
#> Running at Laplace(9.02587e-005) None  Grid-point #10 at 2.455e+008 	Fold #1 Rep #1 pred log like = -192.73
#> Running at Laplace(9.02587e-005) None  Grid-point #10 at 2.455e+008 	Fold #2 Rep #1 pred log like = -183.925
#> Running at Laplace(9.02587e-005) None  Grid-point #10 at 2.455e+008 	Fold #3 Rep #1 pred log like = -202.35
#> Running at Laplace(9.02587e-005) None  Grid-point #10 at 2.455e+008 	Fold #4 Rep #1 pred log like = -185.34
#> Running at Laplace(9.02587e-005) None  Grid-point #10 at 2.455e+008 	Fold #5 Rep #1 pred log like = -198.984
#> Running at Laplace(9.02587e-005) None  Grid-point #10 at 2.455e+008 	Fold #6 Rep #1 pred log like = -196.506
#> Running at Laplace(9.02587e-005) None  Grid-point #10 at 2.455e+008 	Fold #7 Rep #1 pred log like = -194.325
#> Running at Laplace(9.02587e-005) None  Grid-point #10 at 2.455e+008 	Fold #8 Rep #1 pred log like = -190.424
#> Running at Laplace(9.02587e-005) None  Grid-point #10 at 2.455e+008 	Fold #9 Rep #1 pred log like = -204.591
#> Running at Laplace(9.02587e-005) None  Grid-point #10 at 2.455e+008 	Fold #10 Rep #1 pred log like = -180.731
#> AvgPred = -192.991 with stdev = 7.54912
#> Completed at 2.455e+008
#> Next point at 2.455e+009 with value 0 and continue = 1
#> search[ 0.2455 ] = -192.992(7.55817)
#> search[ 2.455 ] = -192.991(7.55194)
#> search[ 24.55 ] = -192.991(7.55001)
#> search[ 245.5 ] = -192.991(7.5494)
#> search[ 2455 ] = -192.991(7.5492)
#> search[ 24550 ] = -192.991(7.54914)
#> search[ 245500 ] = -192.991(7.54912)
#> search[ 2.455e+006 ] = -192.991(7.54912)
#> search[ 2.455e+007 ] = -192.991(7.54912)
#> search[ 2.455e+008 ] = -192.991(7.54912)
#> 
#> Running at Laplace(2.85423e-005) None  Grid-point #11 at 2.455e+009 	Fold #1 Rep #1 pred log like = -192.73
#> Running at Laplace(2.85423e-005) None  Grid-point #11 at 2.455e+009 	Fold #2 Rep #1 pred log like = -183.925
#> Running at Laplace(2.85423e-005) None  Grid-point #11 at 2.455e+009 	Fold #3 Rep #1 pred log like = -202.35
#> Running at Laplace(2.85423e-005) None  Grid-point #11 at 2.455e+009 	Fold #4 Rep #1 pred log like = -185.34
#> Running at Laplace(2.85423e-005) None  Grid-point #11 at 2.455e+009 	Fold #5 Rep #1 pred log like = -198.984
#> Running at Laplace(2.85423e-005) None  Grid-point #11 at 2.455e+009 	Fold #6 Rep #1 pred log like = -196.506
#> Running at Laplace(2.85423e-005) None  Grid-point #11 at 2.455e+009 	Fold #7 Rep #1 pred log like = -194.325
#> Running at Laplace(2.85423e-005) None  Grid-point #11 at 2.455e+009 	Fold #8 Rep #1 pred log like = -190.424
#> Running at Laplace(2.85423e-005) None  Grid-point #11 at 2.455e+009 	Fold #9 Rep #1 pred log like = -204.591
#> Running at Laplace(2.85423e-005) None  Grid-point #11 at 2.455e+009 	Fold #10 Rep #1 pred log like = -180.731
#> AvgPred = -192.991 with stdev = 7.54912
#> Completed at 2.455e+009
#> Next point at 2.455e+010 with value 0 and continue = 1
#> search[ 0.2455 ] = -192.992(7.55817)
#> search[ 2.455 ] = -192.991(7.55194)
#> search[ 24.55 ] = -192.991(7.55001)
#> search[ 245.5 ] = -192.991(7.5494)
#> search[ 2455 ] = -192.991(7.5492)
#> search[ 24550 ] = -192.991(7.54914)
#> search[ 245500 ] = -192.991(7.54912)
#> search[ 2.455e+006 ] = -192.991(7.54912)
#> search[ 2.455e+007 ] = -192.991(7.54912)
#> search[ 2.455e+008 ] = -192.991(7.54912)
#> search[ 2.455e+009 ] = -192.991(7.54912)
#> 
#> Running at Laplace(9.02587e-006) None  Grid-point #12 at 2.455e+010 	Fold #1 Rep #1 pred log like = -192.73
#> Running at Laplace(9.02587e-006) None  Grid-point #12 at 2.455e+010 	Fold #2 Rep #1 pred log like = -183.925
#> Running at Laplace(9.02587e-006) None  Grid-point #12 at 2.455e+010 	Fold #3 Rep #1 pred log like = -202.35
#> Running at Laplace(9.02587e-006) None  Grid-point #12 at 2.455e+010 	Fold #4 Rep #1 pred log like = -185.34
#> Running at Laplace(9.02587e-006) None  Grid-point #12 at 2.455e+010 	Fold #5 Rep #1 pred log like = -198.984
#> Running at Laplace(9.02587e-006) None  Grid-point #12 at 2.455e+010 	Fold #6 Rep #1 pred log like = -196.506
#> Running at Laplace(9.02587e-006) None  Grid-point #12 at 2.455e+010 	Fold #7 Rep #1 pred log like = -194.325
#> Running at Laplace(9.02587e-006) None  Grid-point #12 at 2.455e+010 	Fold #8 Rep #1 pred log like = -190.424
#> Running at Laplace(9.02587e-006) None  Grid-point #12 at 2.455e+010 	Fold #9 Rep #1 pred log like = -204.591
#> Running at Laplace(9.02587e-006) None  Grid-point #12 at 2.455e+010 	Fold #10 Rep #1 pred log like = -180.731
#> AvgPred = -192.991 with stdev = 7.54912
#> Completed at 2.455e+010
#> Next point at 2.455e+011 with value 0 and continue = 1
#> search[ 0.2455 ] = -192.992(7.55817)
#> search[ 2.455 ] = -192.991(7.55194)
#> search[ 24.55 ] = -192.991(7.55001)
#> search[ 245.5 ] = -192.991(7.5494)
#> search[ 2455 ] = -192.991(7.5492)
#> search[ 24550 ] = -192.991(7.54914)
#> search[ 245500 ] = -192.991(7.54912)
#> search[ 2.455e+006 ] = -192.991(7.54912)
#> search[ 2.455e+007 ] = -192.991(7.54912)
#> search[ 2.455e+008 ] = -192.991(7.54912)
#> search[ 2.455e+009 ] = -192.991(7.54912)
#> search[ 2.455e+010 ] = -192.991(7.54912)
#> 
#> Running at Laplace(2.85423e-006) None  Grid-point #13 at 2.455e+011 	Fold #1 Rep #1 pred log like = -192.73
#> Running at Laplace(2.85423e-006) None  Grid-point #13 at 2.455e+011 	Fold #2 Rep #1 pred log like = -183.925
#> Running at Laplace(2.85423e-006) None  Grid-point #13 at 2.455e+011 	Fold #3 Rep #1 pred log like = -202.35
#> Running at Laplace(2.85423e-006) None  Grid-point #13 at 2.455e+011 	Fold #4 Rep #1 pred log like = -185.34
#> Running at Laplace(2.85423e-006) None  Grid-point #13 at 2.455e+011 	Fold #5 Rep #1 pred log like = -198.984
#> Running at Laplace(2.85423e-006) None  Grid-point #13 at 2.455e+011 	Fold #6 Rep #1 pred log like = -196.506
#> Running at Laplace(2.85423e-006) None  Grid-point #13 at 2.455e+011 	Fold #7 Rep #1 pred log like = -194.325
#> Running at Laplace(2.85423e-006) None  Grid-point #13 at 2.455e+011 	Fold #8 Rep #1 pred log like = -190.424
#> Running at Laplace(2.85423e-006) None  Grid-point #13 at 2.455e+011 	Fold #9 Rep #1 pred log like = -204.591
#> Running at Laplace(2.85423e-006) None  Grid-point #13 at 2.455e+011 	Fold #10 Rep #1 pred log like = -180.731
#> AvgPred = -192.991 with stdev = 7.54912
#> Completed at 2.455e+011
#> Next point at 2.455e+012 with value 0 and continue = 1
#> search[ 0.2455 ] = -192.992(7.55817)
#> search[ 2.455 ] = -192.991(7.55194)
#> search[ 24.55 ] = -192.991(7.55001)
#> search[ 245.5 ] = -192.991(7.5494)
#> search[ 2455 ] = -192.991(7.5492)
#> search[ 24550 ] = -192.991(7.54914)
#> search[ 245500 ] = -192.991(7.54912)
#> search[ 2.455e+006 ] = -192.991(7.54912)
#> search[ 2.455e+007 ] = -192.991(7.54912)
#> search[ 2.455e+008 ] = -192.991(7.54912)
#> search[ 2.455e+009 ] = -192.991(7.54912)
#> search[ 2.455e+010 ] = -192.991(7.54912)
#> search[ 2.455e+011 ] = -192.991(7.54912)
#> 
#> Running at Laplace(9.02587e-007) None  Grid-point #14 at 2.455e+012 	Fold #1 Rep #1 pred log like = -192.73
#> Running at Laplace(9.02587e-007) None  Grid-point #14 at 2.455e+012 	Fold #2 Rep #1 pred log like = -183.925
#> Running at Laplace(9.02587e-007) None  Grid-point #14 at 2.455e+012 	Fold #3 Rep #1 pred log like = -202.35
#> Running at Laplace(9.02587e-007) None  Grid-point #14 at 2.455e+012 	Fold #4 Rep #1 pred log like = -185.34
#> Running at Laplace(9.02587e-007) None  Grid-point #14 at 2.455e+012 	Fold #5 Rep #1 pred log like = -198.984
#> Running at Laplace(9.02587e-007) None  Grid-point #14 at 2.455e+012 	Fold #6 Rep #1 pred log like = -196.506
#> Running at Laplace(9.02587e-007) None  Grid-point #14 at 2.455e+012 	Fold #7 Rep #1 pred log like = -194.325
#> Running at Laplace(9.02587e-007) None  Grid-point #14 at 2.455e+012 	Fold #8 Rep #1 pred log like = -190.424
#> Running at Laplace(9.02587e-007) None  Grid-point #14 at 2.455e+012 	Fold #9 Rep #1 pred log like = -204.591
#> Running at Laplace(9.02587e-007) None  Grid-point #14 at 2.455e+012 	Fold #10 Rep #1 pred log like = -180.731
#> AvgPred = -192.991 with stdev = 7.54912
#> Completed at 2.455e+012
#> Next point at 2.455e+013 with value 0 and continue = 1
#> search[ 0.2455 ] = -192.992(7.55817)
#> search[ 2.455 ] = -192.991(7.55194)
#> search[ 24.55 ] = -192.991(7.55001)
#> search[ 245.5 ] = -192.991(7.5494)
#> search[ 2455 ] = -192.991(7.5492)
#> search[ 24550 ] = -192.991(7.54914)
#> search[ 245500 ] = -192.991(7.54912)
#> search[ 2.455e+006 ] = -192.991(7.54912)
#> search[ 2.455e+007 ] = -192.991(7.54912)
#> search[ 2.455e+008 ] = -192.991(7.54912)
#> search[ 2.455e+009 ] = -192.991(7.54912)
#> search[ 2.455e+010 ] = -192.991(7.54912)
#> search[ 2.455e+011 ] = -192.991(7.54912)
#> search[ 2.455e+012 ] = -192.991(7.54912)
#> 
#> Running at Laplace(2.85423e-007) None  Grid-point #15 at 2.455e+013 	Fold #1 Rep #1 pred log like = -192.73
#> Running at Laplace(2.85423e-007) None  Grid-point #15 at 2.455e+013 	Fold #2 Rep #1 pred log like = -183.925
#> Running at Laplace(2.85423e-007) None  Grid-point #15 at 2.455e+013 	Fold #3 Rep #1 pred log like = -202.35
#> Running at Laplace(2.85423e-007) None  Grid-point #15 at 2.455e+013 	Fold #4 Rep #1 pred log like = -185.34
#> Running at Laplace(2.85423e-007) None  Grid-point #15 at 2.455e+013 	Fold #5 Rep #1 pred log like = -198.984
#> Running at Laplace(2.85423e-007) None  Grid-point #15 at 2.455e+013 	Fold #6 Rep #1 pred log like = -196.506
#> Running at Laplace(2.85423e-007) None  Grid-point #15 at 2.455e+013 	Fold #7 Rep #1 pred log like = -194.325
#> Running at Laplace(2.85423e-007) None  Grid-point #15 at 2.455e+013 	Fold #8 Rep #1 pred log like = -190.424
#> Running at Laplace(2.85423e-007) None  Grid-point #15 at 2.455e+013 	Fold #9 Rep #1 pred log like = -204.591
#> Running at Laplace(2.85423e-007) None  Grid-point #15 at 2.455e+013 	Fold #10 Rep #1 pred log like = -180.731
#> AvgPred = -192.991 with stdev = 7.54912
#> Completed at 2.455e+013
#> Next point at 2.455e+014 with value 0 and continue = 1
#> search[ 0.2455 ] = -192.992(7.55817)
#> search[ 2.455 ] = -192.991(7.55194)
#> search[ 24.55 ] = -192.991(7.55001)
#> search[ 245.5 ] = -192.991(7.5494)
#> search[ 2455 ] = -192.991(7.5492)
#> search[ 24550 ] = -192.991(7.54914)
#> search[ 245500 ] = -192.991(7.54912)
#> search[ 2.455e+006 ] = -192.991(7.54912)
#> search[ 2.455e+007 ] = -192.991(7.54912)
#> search[ 2.455e+008 ] = -192.991(7.54912)
#> search[ 2.455e+009 ] = -192.991(7.54912)
#> search[ 2.455e+010 ] = -192.991(7.54912)
#> search[ 2.455e+011 ] = -192.991(7.54912)
#> search[ 2.455e+012 ] = -192.991(7.54912)
#> search[ 2.455e+013 ] = -192.991(7.54912)
#> 
#> Running at Laplace(9.02587e-008) None  Grid-point #16 at 2.455e+014 	Fold #1 Rep #1 pred log like = -192.73
#> Running at Laplace(9.02587e-008) None  Grid-point #16 at 2.455e+014 	Fold #2 Rep #1 pred log like = -183.925
#> Running at Laplace(9.02587e-008) None  Grid-point #16 at 2.455e+014 	Fold #3 Rep #1 pred log like = -202.35
#> Running at Laplace(9.02587e-008) None  Grid-point #16 at 2.455e+014 	Fold #4 Rep #1 pred log like = -185.34
#> Running at Laplace(9.02587e-008) None  Grid-point #16 at 2.455e+014 	Fold #5 Rep #1 pred log like = -198.984
#> Running at Laplace(9.02587e-008) None  Grid-point #16 at 2.455e+014 	Fold #6 Rep #1 pred log like = -196.506
#> Running at Laplace(9.02587e-008) None  Grid-point #16 at 2.455e+014 	Fold #7 Rep #1 pred log like = -194.325
#> Running at Laplace(9.02587e-008) None  Grid-point #16 at 2.455e+014 	Fold #8 Rep #1 pred log like = -190.424
#> Running at Laplace(9.02587e-008) None  Grid-point #16 at 2.455e+014 	Fold #9 Rep #1 pred log like = -204.591
#> Running at Laplace(9.02587e-008) None  Grid-point #16 at 2.455e+014 	Fold #10 Rep #1 pred log like = -180.731
#> AvgPred = -192.991 with stdev = 7.54912
#> Completed at 2.455e+014
#> Next point at 2.455e+015 with value 0 and continue = 1
#> search[ 0.2455 ] = -192.992(7.55817)
#> search[ 2.455 ] = -192.991(7.55194)
#> search[ 24.55 ] = -192.991(7.55001)
#> search[ 245.5 ] = -192.991(7.5494)
#> search[ 2455 ] = -192.991(7.5492)
#> search[ 24550 ] = -192.991(7.54914)
#> search[ 245500 ] = -192.991(7.54912)
#> search[ 2.455e+006 ] = -192.991(7.54912)
#> search[ 2.455e+007 ] = -192.991(7.54912)
#> search[ 2.455e+008 ] = -192.991(7.54912)
#> search[ 2.455e+009 ] = -192.991(7.54912)
#> search[ 2.455e+010 ] = -192.991(7.54912)
#> search[ 2.455e+011 ] = -192.991(7.54912)
#> search[ 2.455e+012 ] = -192.991(7.54912)
#> search[ 2.455e+013 ] = -192.991(7.54912)
#> search[ 2.455e+014 ] = -192.991(7.54912)
#> 
#> Running at Laplace(2.85423e-008) None  Grid-point #17 at 2.455e+015 	Fold #1 Rep #1 pred log like = -192.73
#> Running at Laplace(2.85423e-008) None  Grid-point #17 at 2.455e+015 	Fold #2 Rep #1 pred log like = -183.925
#> Running at Laplace(2.85423e-008) None  Grid-point #17 at 2.455e+015 	Fold #3 Rep #1 pred log like = -202.35
#> Running at Laplace(2.85423e-008) None  Grid-point #17 at 2.455e+015 	Fold #4 Rep #1 pred log like = -185.34
#> Running at Laplace(2.85423e-008) None  Grid-point #17 at 2.455e+015 	Fold #5 Rep #1 pred log like = -198.984
#> Running at Laplace(2.85423e-008) None  Grid-point #17 at 2.455e+015 	Fold #6 Rep #1 pred log like = -196.506
#> Running at Laplace(2.85423e-008) None  Grid-point #17 at 2.455e+015 	Fold #7 Rep #1 pred log like = -194.325
#> Running at Laplace(2.85423e-008) None  Grid-point #17 at 2.455e+015 	Fold #8 Rep #1 pred log like = -190.424
#> Running at Laplace(2.85423e-008) None  Grid-point #17 at 2.455e+015 	Fold #9 Rep #1 pred log like = -204.591
#> Running at Laplace(2.85423e-008) None  Grid-point #17 at 2.455e+015 	Fold #10 Rep #1 pred log like = -180.731
#> AvgPred = -192.991 with stdev = 7.54912
#> Completed at 2.455e+015
#> Next point at 2.455e+016 with value 0 and continue = 1
#> search[ 0.2455 ] = -192.992(7.55817)
#> search[ 2.455 ] = -192.991(7.55194)
#> search[ 24.55 ] = -192.991(7.55001)
#> search[ 245.5 ] = -192.991(7.5494)
#> search[ 2455 ] = -192.991(7.5492)
#> search[ 24550 ] = -192.991(7.54914)
#> search[ 245500 ] = -192.991(7.54912)
#> search[ 2.455e+006 ] = -192.991(7.54912)
#> search[ 2.455e+007 ] = -192.991(7.54912)
#> search[ 2.455e+008 ] = -192.991(7.54912)
#> search[ 2.455e+009 ] = -192.991(7.54912)
#> search[ 2.455e+010 ] = -192.991(7.54912)
#> search[ 2.455e+011 ] = -192.991(7.54912)
#> search[ 2.455e+012 ] = -192.991(7.54912)
#> search[ 2.455e+013 ] = -192.991(7.54912)
#> search[ 2.455e+014 ] = -192.991(7.54912)
#> search[ 2.455e+015 ] = -192.991(7.54912)
#> 
#> Running at Laplace(9.02587e-009) None  Grid-point #18 at 2.455e+016 	Fold #1 Rep #1 pred log like = -192.73
#> Running at Laplace(9.02587e-009) None  Grid-point #18 at 2.455e+016 	Fold #2 Rep #1 pred log like = -183.925
#> Running at Laplace(9.02587e-009) None  Grid-point #18 at 2.455e+016 	Fold #3 Rep #1 pred log like = -202.35
#> Running at Laplace(9.02587e-009) None  Grid-point #18 at 2.455e+016 	Fold #4 Rep #1 pred log like = -185.34
#> Running at Laplace(9.02587e-009) None  Grid-point #18 at 2.455e+016 	Fold #5 Rep #1 pred log like = -198.984
#> Running at Laplace(9.02587e-009) None  Grid-point #18 at 2.455e+016 	Fold #6 Rep #1 pred log like = -196.506
#> Running at Laplace(9.02587e-009) None  Grid-point #18 at 2.455e+016 	Fold #7 Rep #1 pred log like = -194.325
#> Running at Laplace(9.02587e-009) None  Grid-point #18 at 2.455e+016 	Fold #8 Rep #1 pred log like = -190.424
#> Running at Laplace(9.02587e-009) None  Grid-point #18 at 2.455e+016 	Fold #9 Rep #1 pred log like = -204.591
#> Running at Laplace(9.02587e-009) None  Grid-point #18 at 2.455e+016 	Fold #10 Rep #1 pred log like = -180.731
#> AvgPred = -192.991 with stdev = 7.54912
#> Completed at 2.455e+016
#> Next point at 2.455e+017 with value 0 and continue = 1
#> search[ 0.2455 ] = -192.992(7.55817)
#> search[ 2.455 ] = -192.991(7.55194)
#> search[ 24.55 ] = -192.991(7.55001)
#> search[ 245.5 ] = -192.991(7.5494)
#> search[ 2455 ] = -192.991(7.5492)
#> search[ 24550 ] = -192.991(7.54914)
#> search[ 245500 ] = -192.991(7.54912)
#> search[ 2.455e+006 ] = -192.991(7.54912)
#> search[ 2.455e+007 ] = -192.991(7.54912)
#> search[ 2.455e+008 ] = -192.991(7.54912)
#> search[ 2.455e+009 ] = -192.991(7.54912)
#> search[ 2.455e+010 ] = -192.991(7.54912)
#> search[ 2.455e+011 ] = -192.991(7.54912)
#> search[ 2.455e+012 ] = -192.991(7.54912)
#> search[ 2.455e+013 ] = -192.991(7.54912)
#> search[ 2.455e+014 ] = -192.991(7.54912)
#> search[ 2.455e+015 ] = -192.991(7.54912)
#> search[ 2.455e+016 ] = -192.991(7.54912)
#> 
#> Running at Laplace(2.85423e-009) None  Grid-point #19 at 2.455e+017 	Fold #1 Rep #1 pred log like = -192.73
#> Running at Laplace(2.85423e-009) None  Grid-point #19 at 2.455e+017 	Fold #2 Rep #1 pred log like = -183.925
#> Running at Laplace(2.85423e-009) None  Grid-point #19 at 2.455e+017 	Fold #3 Rep #1 pred log like = -202.35
#> Running at Laplace(2.85423e-009) None  Grid-point #19 at 2.455e+017 	Fold #4 Rep #1 pred log like = -185.34
#> Running at Laplace(2.85423e-009) None  Grid-point #19 at 2.455e+017 	Fold #5 Rep #1 pred log like = -198.984
#> Running at Laplace(2.85423e-009) None  Grid-point #19 at 2.455e+017 	Fold #6 Rep #1 pred log like = -196.506
#> Running at Laplace(2.85423e-009) None  Grid-point #19 at 2.455e+017 	Fold #7 Rep #1 pred log like = -194.325
#> Running at Laplace(2.85423e-009) None  Grid-point #19 at 2.455e+017 	Fold #8 Rep #1 pred log like = -190.424
#> Running at Laplace(2.85423e-009) None  Grid-point #19 at 2.455e+017 	Fold #9 Rep #1 pred log like = -204.591
#> Running at Laplace(2.85423e-009) None  Grid-point #19 at 2.455e+017 	Fold #10 Rep #1 pred log like = -180.731
#> AvgPred = -192.991 with stdev = 7.54912
#> Completed at 2.455e+017
#> Next point at 2.455e+018 with value 0 and continue = 1
#> search[ 0.2455 ] = -192.992(7.55817)
#> search[ 2.455 ] = -192.991(7.55194)
#> search[ 24.55 ] = -192.991(7.55001)
#> search[ 245.5 ] = -192.991(7.5494)
#> search[ 2455 ] = -192.991(7.5492)
#> search[ 24550 ] = -192.991(7.54914)
#> search[ 245500 ] = -192.991(7.54912)
#> search[ 2.455e+006 ] = -192.991(7.54912)
#> search[ 2.455e+007 ] = -192.991(7.54912)
#> search[ 2.455e+008 ] = -192.991(7.54912)
#> search[ 2.455e+009 ] = -192.991(7.54912)
#> search[ 2.455e+010 ] = -192.991(7.54912)
#> search[ 2.455e+011 ] = -192.991(7.54912)
#> search[ 2.455e+012 ] = -192.991(7.54912)
#> search[ 2.455e+013 ] = -192.991(7.54912)
#> search[ 2.455e+014 ] = -192.991(7.54912)
#> search[ 2.455e+015 ] = -192.991(7.54912)
#> search[ 2.455e+016 ] = -192.991(7.54912)
#> search[ 2.455e+017 ] = -192.991(7.54912)
#> 
#> Running at Laplace(9.02587e-010) None  Grid-point #20 at 2.455e+018 	Fold #1 Rep #1 pred log like = -192.73
#> Running at Laplace(9.02587e-010) None  Grid-point #20 at 2.455e+018 	Fold #2 Rep #1 pred log like = -183.925
#> Running at Laplace(9.02587e-010) None  Grid-point #20 at 2.455e+018 	Fold #3 Rep #1 pred log like = -202.35
#> Running at Laplace(9.02587e-010) None  Grid-point #20 at 2.455e+018 	Fold #4 Rep #1 pred log like = -185.34
#> Running at Laplace(9.02587e-010) None  Grid-point #20 at 2.455e+018 	Fold #5 Rep #1 pred log like = -198.984
#> Running at Laplace(9.02587e-010) None  Grid-point #20 at 2.455e+018 	Fold #6 Rep #1 pred log like = -196.506
#> Running at Laplace(9.02587e-010) None  Grid-point #20 at 2.455e+018 	Fold #7 Rep #1 pred log like = -194.325
#> Running at Laplace(9.02587e-010) None  Grid-point #20 at 2.455e+018 	Fold #8 Rep #1 pred log like = -190.424
#> Running at Laplace(9.02587e-010) None  Grid-point #20 at 2.455e+018 	Fold #9 Rep #1 pred log like = -204.591
#> Running at Laplace(9.02587e-010) None  Grid-point #20 at 2.455e+018 	Fold #10 Rep #1 pred log like = -180.731
#> AvgPred = -192.991 with stdev = 7.54912
#> Completed at 2.455e+018
#> Next point at 2.455e+019 with value 0 and continue = 1
#> search[ 0.2455 ] = -192.992(7.55817)
#> search[ 2.455 ] = -192.991(7.55194)
#> search[ 24.55 ] = -192.991(7.55001)
#> search[ 245.5 ] = -192.991(7.5494)
#> search[ 2455 ] = -192.991(7.5492)
#> search[ 24550 ] = -192.991(7.54914)
#> search[ 245500 ] = -192.991(7.54912)
#> search[ 2.455e+006 ] = -192.991(7.54912)
#> search[ 2.455e+007 ] = -192.991(7.54912)
#> search[ 2.455e+008 ] = -192.991(7.54912)
#> search[ 2.455e+009 ] = -192.991(7.54912)
#> search[ 2.455e+010 ] = -192.991(7.54912)
#> search[ 2.455e+011 ] = -192.991(7.54912)
#> search[ 2.455e+012 ] = -192.991(7.54912)
#> search[ 2.455e+013 ] = -192.991(7.54912)
#> search[ 2.455e+014 ] = -192.991(7.54912)
#> search[ 2.455e+015 ] = -192.991(7.54912)
#> search[ 2.455e+016 ] = -192.991(7.54912)
#> search[ 2.455e+017 ] = -192.991(7.54912)
#> search[ 2.455e+018 ] = -192.991(7.54912)
#> 
#> Running at Laplace(2.85423e-010) None  Grid-point #21 at 2.455e+019 	Fold #1 Rep #1 pred log like = -192.73
#> Running at Laplace(2.85423e-010) None  Grid-point #21 at 2.455e+019 	Fold #2 Rep #1 pred log like = -183.925
#> Running at Laplace(2.85423e-010) None  Grid-point #21 at 2.455e+019 	Fold #3 Rep #1 pred log like = -202.35
#> Running at Laplace(2.85423e-010) None  Grid-point #21 at 2.455e+019 	Fold #4 Rep #1 pred log like = -185.34
#> Running at Laplace(2.85423e-010) None  Grid-point #21 at 2.455e+019 	Fold #5 Rep #1 pred log like = -198.984
#> Running at Laplace(2.85423e-010) None  Grid-point #21 at 2.455e+019 	Fold #6 Rep #1 pred log like = -196.506
#> Running at Laplace(2.85423e-010) None  Grid-point #21 at 2.455e+019 	Fold #7 Rep #1 pred log like = -194.325
#> Running at Laplace(2.85423e-010) None  Grid-point #21 at 2.455e+019 	Fold #8 Rep #1 pred log like = -190.424
#> Running at Laplace(2.85423e-010) None  Grid-point #21 at 2.455e+019 	Fold #9 Rep #1 pred log like = -204.591
#> Running at Laplace(2.85423e-010) None  Grid-point #21 at 2.455e+019 	Fold #10 Rep #1 pred log like = -180.731
#> AvgPred = -192.991 with stdev = 7.54912
#> Completed at 2.455e+019
#> Next point at 2.455e+020 with value 0 and continue = 1
#> search[ 0.2455 ] = -192.992(7.55817)
#> search[ 2.455 ] = -192.991(7.55194)
#> search[ 24.55 ] = -192.991(7.55001)
#> search[ 245.5 ] = -192.991(7.5494)
#> search[ 2455 ] = -192.991(7.5492)
#> search[ 24550 ] = -192.991(7.54914)
#> search[ 245500 ] = -192.991(7.54912)
#> search[ 2.455e+006 ] = -192.991(7.54912)
#> search[ 2.455e+007 ] = -192.991(7.54912)
#> search[ 2.455e+008 ] = -192.991(7.54912)
#> search[ 2.455e+009 ] = -192.991(7.54912)
#> search[ 2.455e+010 ] = -192.991(7.54912)
#> search[ 2.455e+011 ] = -192.991(7.54912)
#> search[ 2.455e+012 ] = -192.991(7.54912)
#> search[ 2.455e+013 ] = -192.991(7.54912)
#> search[ 2.455e+014 ] = -192.991(7.54912)
#> search[ 2.455e+015 ] = -192.991(7.54912)
#> search[ 2.455e+016 ] = -192.991(7.54912)
#> search[ 2.455e+017 ] = -192.991(7.54912)
#> search[ 2.455e+018 ] = -192.991(7.54912)
#> search[ 2.455e+019 ] = -192.991(7.54912)
#> 
#> Running at Laplace(9.02587e-011) None  Grid-point #22 at 2.455e+020 	Fold #1 Rep #1 pred log like = -192.73
#> Running at Laplace(9.02587e-011) None  Grid-point #22 at 2.455e+020 	Fold #2 Rep #1 pred log like = -183.925
#> Running at Laplace(9.02587e-011) None  Grid-point #22 at 2.455e+020 	Fold #3 Rep #1 pred log like = -202.35
#> Running at Laplace(9.02587e-011) None  Grid-point #22 at 2.455e+020 	Fold #4 Rep #1 pred log like = -185.34
#> Running at Laplace(9.02587e-011) None  Grid-point #22 at 2.455e+020 	Fold #5 Rep #1 pred log like = -198.984
#> Running at Laplace(9.02587e-011) None  Grid-point #22 at 2.455e+020 	Fold #6 Rep #1 pred log like = -196.506
#> Running at Laplace(9.02587e-011) None  Grid-point #22 at 2.455e+020 	Fold #7 Rep #1 pred log like = -194.325
#> Running at Laplace(9.02587e-011) None  Grid-point #22 at 2.455e+020 	Fold #8 Rep #1 pred log like = -190.424
#> Running at Laplace(9.02587e-011) None  Grid-point #22 at 2.455e+020 	Fold #9 Rep #1 pred log like = -204.591
#> Running at Laplace(9.02587e-011) None  Grid-point #22 at 2.455e+020 	Fold #10 Rep #1 pred log like = -180.731
#> AvgPred = -192.991 with stdev = 7.54912
#> Completed at 2.455e+020
#> Next point at 2.455e+021 with value 0 and continue = 1
#> search[ 0.2455 ] = -192.992(7.55817)
#> search[ 2.455 ] = -192.991(7.55194)
#> search[ 24.55 ] = -192.991(7.55001)
#> search[ 245.5 ] = -192.991(7.5494)
#> search[ 2455 ] = -192.991(7.5492)
#> search[ 24550 ] = -192.991(7.54914)
#> search[ 245500 ] = -192.991(7.54912)
#> search[ 2.455e+006 ] = -192.991(7.54912)
#> search[ 2.455e+007 ] = -192.991(7.54912)
#> search[ 2.455e+008 ] = -192.991(7.54912)
#> search[ 2.455e+009 ] = -192.991(7.54912)
#> search[ 2.455e+010 ] = -192.991(7.54912)
#> search[ 2.455e+011 ] = -192.991(7.54912)
#> search[ 2.455e+012 ] = -192.991(7.54912)
#> search[ 2.455e+013 ] = -192.991(7.54912)
#> search[ 2.455e+014 ] = -192.991(7.54912)
#> search[ 2.455e+015 ] = -192.991(7.54912)
#> search[ 2.455e+016 ] = -192.991(7.54912)
#> search[ 2.455e+017 ] = -192.991(7.54912)
#> search[ 2.455e+018 ] = -192.991(7.54912)
#> search[ 2.455e+019 ] = -192.991(7.54912)
#> search[ 2.455e+020 ] = -192.991(7.54912)
#> 
#> Running at Laplace(2.85423e-011) None  Grid-point #23 at 2.455e+021 	Fold #1 Rep #1 pred log like = -192.73
#> Running at Laplace(2.85423e-011) None  Grid-point #23 at 2.455e+021 	Fold #2 Rep #1 pred log like = -183.925
#> Running at Laplace(2.85423e-011) None  Grid-point #23 at 2.455e+021 	Fold #3 Rep #1 pred log like = -202.35
#> Running at Laplace(2.85423e-011) None  Grid-point #23 at 2.455e+021 	Fold #4 Rep #1 pred log like = -185.34
#> Running at Laplace(2.85423e-011) None  Grid-point #23 at 2.455e+021 	Fold #5 Rep #1 pred log like = -198.984
#> Running at Laplace(2.85423e-011) None  Grid-point #23 at 2.455e+021 	Fold #6 Rep #1 pred log like = -196.506
#> Running at Laplace(2.85423e-011) None  Grid-point #23 at 2.455e+021 	Fold #7 Rep #1 pred log like = -194.325
#> Running at Laplace(2.85423e-011) None  Grid-point #23 at 2.455e+021 	Fold #8 Rep #1 pred log like = -190.424
#> Running at Laplace(2.85423e-011) None  Grid-point #23 at 2.455e+021 	Fold #9 Rep #1 pred log like = -204.591
#> Running at Laplace(2.85423e-011) None  Grid-point #23 at 2.455e+021 	Fold #10 Rep #1 pred log like = -180.731
#> AvgPred = -192.991 with stdev = 7.54912
#> Completed at 2.455e+021
#> Next point at 8.83295e+012 with value -192.991 and continue = 0
#> search[ 0.2455 ] = -192.992(7.55817)
#> search[ 2.455 ] = -192.991(7.55194)
#> search[ 24.55 ] = -192.991(7.55001)
#> search[ 245.5 ] = -192.991(7.5494)
#> search[ 2455 ] = -192.991(7.5492)
#> search[ 24550 ] = -192.991(7.54914)
#> search[ 245500 ] = -192.991(7.54912)
#> search[ 2.455e+006 ] = -192.991(7.54912)
#> search[ 2.455e+007 ] = -192.991(7.54912)
#> search[ 2.455e+008 ] = -192.991(7.54912)
#> search[ 2.455e+009 ] = -192.991(7.54912)
#> search[ 2.455e+010 ] = -192.991(7.54912)
#> search[ 2.455e+011 ] = -192.991(7.54912)
#> search[ 2.455e+012 ] = -192.991(7.54912)
#> search[ 2.455e+013 ] = -192.991(7.54912)
#> search[ 2.455e+014 ] = -192.991(7.54912)
#> search[ 2.455e+015 ] = -192.991(7.54912)
#> search[ 2.455e+016 ] = -192.991(7.54912)
#> search[ 2.455e+017 ] = -192.991(7.54912)
#> search[ 2.455e+018 ] = -192.991(7.54912)
#> search[ 2.455e+019 ] = -192.991(7.54912)
#> search[ 2.455e+020 ] = -192.991(7.54912)
#> search[ 2.455e+021 ] = -192.991(7.54912)
#> 
#> 
#> Maximum predicted log likelihood (-192.991) estimated at:
#> 	8.83295e+012 (variance)
#> 	4.75841e-007 (lambda)
#> 
#> Fitting model at optimal hyperparameter
#> Using prior: Laplace(4.75841e-007) None 

#Find out what the optimal hyperparameter was:
getHyperParameter(fit)

#> [1] 8.832946e+12

#Extract the current log-likelihood, and coefficients
logLik(fit)

#> 'log Lik.' -1926.897 (df=3)

coef(fit)

#> (Intercept)           1           2 
#> -4.12256724  0.04683372 -0.02827391 

#We can only retrieve the confidence interval for unregularized coefficients:
confint(fit, c(0))

#> Using 1 thread(s)

#>      covariate     2.5 %    97.5 % evaluations
#> [1,]         0 -4.160864 -4.084717          22