Extract log-likelihood

logLik returns the current log-likelihood of the fit in a Cyclops model fit object

# S3 method for cyclopsFit
logLik(object, ...)

Arguments

object

A Cyclops model fit object

...

Additional arguments

Examples

#Generate some simulated data:
sim <- simulateCyclopsData(nstrata = 1, nrows = 1000, ncovars = 2, eCovarsPerRow = 0.5, 
                           model = "poisson")
#> Sparseness = 77.4 %
cyclopsData <- convertToCyclopsData(sim$outcomes, sim$covariates, modelType = "pr", 
                                    addIntercept = TRUE)
#> Sorting covariates by covariateId and rowId

#Define the prior and control objects to use cross-validation for finding the 
#optimal hyperparameter:
prior <- createPrior("laplace", exclude = 0, useCrossValidation = TRUE)
control <- createControl(cvType = "auto", noiseLevel = "quiet")

#Fit the model
fit <- fitCyclopsModel(cyclopsData,prior = prior, control = control)  
#> Using cross-validation selector type byRow
#> Performing 10-fold cross-validation [seed = 1698757939] with data partitions of sizes 100 100 100 100 100 100 100 100 100 100
#> Using 1 thread(s)
#> Starting var = 0.226 (default)
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #1 Rep #1 pred log like = 424.312
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #2 Rep #1 pred log like = 286.112
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #3 Rep #1 pred log like = 542.244
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #4 Rep #1 pred log like = 462.595
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #5 Rep #1 pred log like = 523.632
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #6 Rep #1 pred log like = 390.675
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #7 Rep #1 pred log like = 360.237
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #8 Rep #1 pred log like = 399.312
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #9 Rep #1 pred log like = 395.077
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #10 Rep #1 pred log like = 331.197
#> AvgPred = 411.539 with stdev = 76.2186
#> Completed at 0.226
#> Next point at 2.26 with value 0 and continue = 1
#> search[ 0.226 ] = 411.539(76.2186)
#> 
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #1 Rep #1 pred log like = 424.271
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #2 Rep #1 pred log like = 286.187
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #3 Rep #1 pred log like = 542.304
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #4 Rep #1 pred log like = 462.57
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #5 Rep #1 pred log like = 523.608
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #6 Rep #1 pred log like = 390.669
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #7 Rep #1 pred log like = 360.212
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #8 Rep #1 pred log like = 399.333
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #9 Rep #1 pred log like = 395.126
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #10 Rep #1 pred log like = 331.12
#> AvgPred = 411.54 with stdev = 76.2193
#> Completed at 2.26
#> Next point at 22.6 with value 0 and continue = 1
#> search[ 0.226 ] = 411.539(76.2186)
#> search[ 2.26 ] = 411.54(76.2193)
#> 
#> Running at Laplace(0.297482) None  Grid-point #3 at 22.6 	Fold #1 Rep #1 pred log like = 424.258
#> Running at Laplace(0.297482) None  Grid-point #3 at 22.6 	Fold #2 Rep #1 pred log like = 286.211
#> Running at Laplace(0.297482) None  Grid-point #3 at 22.6 	Fold #3 Rep #1 pred log like = 542.322
#> Running at Laplace(0.297482) None  Grid-point #3 at 22.6 	Fold #4 Rep #1 pred log like = 462.562
#> Running at Laplace(0.297482) None  Grid-point #3 at 22.6 	Fold #5 Rep #1 pred log like = 523.601
#> Running at Laplace(0.297482) None  Grid-point #3 at 22.6 	Fold #6 Rep #1 pred log like = 390.667
#> Running at Laplace(0.297482) None  Grid-point #3 at 22.6 	Fold #7 Rep #1 pred log like = 360.204
#> Running at Laplace(0.297482) None  Grid-point #3 at 22.6 	Fold #8 Rep #1 pred log like = 399.34
#> Running at Laplace(0.297482) None  Grid-point #3 at 22.6 	Fold #9 Rep #1 pred log like = 395.142
#> Running at Laplace(0.297482) None  Grid-point #3 at 22.6 	Fold #10 Rep #1 pred log like = 331.095
#> AvgPred = 411.54 with stdev = 76.2195
#> Completed at 22.6
#> Next point at 7.08693 with value 411.54 and continue = 0
#> search[ 0.226 ] = 411.539(76.2186)
#> search[ 2.26 ] = 411.54(76.2193)
#> search[ 22.6 ] = 411.54(76.2195)
#> 
#> 
#> Maximum predicted log likelihood (411.54) estimated at:
#> 	7.08693 (variance)
#> 	0.531234 (lambda)
#> 
#> Fitting model at optimal hyperparameter
#> Using prior: Laplace(0.531234) None 

#Find out what the optimal hyperparameter was:
getHyperParameter(fit)
#> [1] 7.086931

#Extract the current log-likelihood, and coefficients
logLik(fit)
#> 'log Lik.' -2045.115 (df=3)
coef(fit)
#> (Intercept)           1           2 
#> -3.89896635 -0.03932575  0.03150517 

#We can only retrieve the confidence interval for unregularized coefficients:
confint(fit, c(0))
#> Using 1 thread(s)
#>             covariate     2.5 %    97.5 % evaluations
#> (Intercept)         0 -3.932993 -3.865291          22