Extract log-likelihood

logLik returns the current log-likelihood of the fit in a Cyclops model fit object

# S3 method for cyclopsFit
logLik(object, ...)

Arguments

object

A Cyclops model fit object

...

Additional arguments

Examples

#Generate some simulated data:
sim <- simulateCyclopsData(nstrata = 1, nrows = 1000, ncovars = 2, eCovarsPerRow = 0.5, 
                           model = "poisson")
#> Sparseness = 75.25 %
cyclopsData <- convertToCyclopsData(sim$outcomes, sim$covariates, modelType = "pr", 
                                    addIntercept = TRUE)
#> Sorting covariates by covariateId and rowId

#Define the prior and control objects to use cross-validation for finding the 
#optimal hyperparameter:
prior <- createPrior("laplace", exclude = 0, useCrossValidation = TRUE)
control <- createControl(cvType = "auto", noiseLevel = "quiet")

#Fit the model
fit <- fitCyclopsModel(cyclopsData,prior = prior, control = control)  
#> Using cross-validation selector type byRow
#> Performing 10-fold cross-validation [seed = 1698789252] with data partitions of sizes 100 100 100 100 100 100 100 100 100 100
#> Using 1 thread(s)
#> Starting var = 0.2475 (default)
#> Running at Laplace(2.84268) None  Grid-point #1 at 0.2475 	Fold #1 Rep #1 pred log like = 764.157
#> Running at Laplace(2.84268) None  Grid-point #1 at 0.2475 	Fold #2 Rep #1 pred log like = 802.169
#> Running at Laplace(2.84268) None  Grid-point #1 at 0.2475 	Fold #3 Rep #1 pred log like = 809.884
#> Running at Laplace(2.84268) None  Grid-point #1 at 0.2475 	Fold #4 Rep #1 pred log like = 829.664
#> Running at Laplace(2.84268) None  Grid-point #1 at 0.2475 	Fold #5 Rep #1 pred log like = 803.191
#> Running at Laplace(2.84268) None  Grid-point #1 at 0.2475 	Fold #6 Rep #1 pred log like = 774.997
#> Running at Laplace(2.84268) None  Grid-point #1 at 0.2475 	Fold #7 Rep #1 pred log like = 771.955
#> Running at Laplace(2.84268) None  Grid-point #1 at 0.2475 	Fold #8 Rep #1 pred log like = 809.568
#> Running at Laplace(2.84268) None  Grid-point #1 at 0.2475 	Fold #9 Rep #1 pred log like = 730.503
#> Running at Laplace(2.84268) None  Grid-point #1 at 0.2475 	Fold #10 Rep #1 pred log like = 661.936
#> AvgPred = 775.802 with stdev = 46.7085
#> Completed at 0.2475
#> Next point at 2.475 with value 0 and continue = 1
#> search[ 0.2475 ] = 775.802(46.7085)
#> 
#> Running at Laplace(0.898933) None  Grid-point #2 at 2.475 	Fold #1 Rep #1 pred log like = 764.135
#> Running at Laplace(0.898933) None  Grid-point #2 at 2.475 	Fold #2 Rep #1 pred log like = 802.181
#> Running at Laplace(0.898933) None  Grid-point #2 at 2.475 	Fold #3 Rep #1 pred log like = 809.903
#> Running at Laplace(0.898933) None  Grid-point #2 at 2.475 	Fold #4 Rep #1 pred log like = 829.66
#> Running at Laplace(0.898933) None  Grid-point #2 at 2.475 	Fold #5 Rep #1 pred log like = 803.233
#> Running at Laplace(0.898933) None  Grid-point #2 at 2.475 	Fold #6 Rep #1 pred log like = 774.984
#> Running at Laplace(0.898933) None  Grid-point #2 at 2.475 	Fold #7 Rep #1 pred log like = 771.951
#> Running at Laplace(0.898933) None  Grid-point #2 at 2.475 	Fold #8 Rep #1 pred log like = 809.56
#> Running at Laplace(0.898933) None  Grid-point #2 at 2.475 	Fold #9 Rep #1 pred log like = 730.492
#> Running at Laplace(0.898933) None  Grid-point #2 at 2.475 	Fold #10 Rep #1 pred log like = 661.932
#> AvgPred = 775.803 with stdev = 46.7144
#> Completed at 2.475
#> Next point at 24.75 with value 0 and continue = 1
#> search[ 0.2475 ] = 775.802(46.7085)
#> search[ 2.475 ] = 775.803(46.7144)
#> 
#> Running at Laplace(0.284268) None  Grid-point #3 at 24.75 	Fold #1 Rep #1 pred log like = 764.128
#> Running at Laplace(0.284268) None  Grid-point #3 at 24.75 	Fold #2 Rep #1 pred log like = 802.185
#> Running at Laplace(0.284268) None  Grid-point #3 at 24.75 	Fold #3 Rep #1 pred log like = 809.909
#> Running at Laplace(0.284268) None  Grid-point #3 at 24.75 	Fold #4 Rep #1 pred log like = 829.658
#> Running at Laplace(0.284268) None  Grid-point #3 at 24.75 	Fold #5 Rep #1 pred log like = 803.246
#> Running at Laplace(0.284268) None  Grid-point #3 at 24.75 	Fold #6 Rep #1 pred log like = 774.98
#> Running at Laplace(0.284268) None  Grid-point #3 at 24.75 	Fold #7 Rep #1 pred log like = 771.949
#> Running at Laplace(0.284268) None  Grid-point #3 at 24.75 	Fold #8 Rep #1 pred log like = 809.558
#> Running at Laplace(0.284268) None  Grid-point #3 at 24.75 	Fold #9 Rep #1 pred log like = 730.488
#> Running at Laplace(0.284268) None  Grid-point #3 at 24.75 	Fold #10 Rep #1 pred log like = 661.931
#> AvgPred = 775.803 with stdev = 46.7163
#> Completed at 24.75
#> Next point at 247.5 with value 0 and continue = 1
#> search[ 0.2475 ] = 775.802(46.7085)
#> search[ 2.475 ] = 775.803(46.7144)
#> search[ 24.75 ] = 775.803(46.7163)
#> 
#> Running at Laplace(0.0898933) None  Grid-point #4 at 247.5 	Fold #1 Rep #1 pred log like = 764.126
#> Running at Laplace(0.0898933) None  Grid-point #4 at 247.5 	Fold #2 Rep #1 pred log like = 802.186
#> Running at Laplace(0.0898933) None  Grid-point #4 at 247.5 	Fold #3 Rep #1 pred log like = 809.91
#> Running at Laplace(0.0898933) None  Grid-point #4 at 247.5 	Fold #4 Rep #1 pred log like = 829.658
#> Running at Laplace(0.0898933) None  Grid-point #4 at 247.5 	Fold #5 Rep #1 pred log like = 803.25
#> Running at Laplace(0.0898933) None  Grid-point #4 at 247.5 	Fold #6 Rep #1 pred log like = 774.979
#> Running at Laplace(0.0898933) None  Grid-point #4 at 247.5 	Fold #7 Rep #1 pred log like = 771.949
#> Running at Laplace(0.0898933) None  Grid-point #4 at 247.5 	Fold #8 Rep #1 pred log like = 809.557
#> Running at Laplace(0.0898933) None  Grid-point #4 at 247.5 	Fold #9 Rep #1 pred log like = 730.487
#> Running at Laplace(0.0898933) None  Grid-point #4 at 247.5 	Fold #10 Rep #1 pred log like = 661.93
#> AvgPred = 775.803 with stdev = 46.7169
#> Completed at 247.5
#> Next point at 2475 with value 0 and continue = 1
#> search[ 0.2475 ] = 775.802(46.7085)
#> search[ 2.475 ] = 775.803(46.7144)
#> search[ 24.75 ] = 775.803(46.7163)
#> search[ 247.5 ] = 775.803(46.7169)
#> 
#> Running at Laplace(0.0284268) None  Grid-point #5 at 2475 	Fold #1 Rep #1 pred log like = 764.125
#> Running at Laplace(0.0284268) None  Grid-point #5 at 2475 	Fold #2 Rep #1 pred log like = 802.186
#> Running at Laplace(0.0284268) None  Grid-point #5 at 2475 	Fold #3 Rep #1 pred log like = 809.911
#> Running at Laplace(0.0284268) None  Grid-point #5 at 2475 	Fold #4 Rep #1 pred log like = 829.657
#> Running at Laplace(0.0284268) None  Grid-point #5 at 2475 	Fold #5 Rep #1 pred log like = 803.251
#> Running at Laplace(0.0284268) None  Grid-point #5 at 2475 	Fold #6 Rep #1 pred log like = 774.978
#> Running at Laplace(0.0284268) None  Grid-point #5 at 2475 	Fold #7 Rep #1 pred log like = 771.948
#> Running at Laplace(0.0284268) None  Grid-point #5 at 2475 	Fold #8 Rep #1 pred log like = 809.556
#> Running at Laplace(0.0284268) None  Grid-point #5 at 2475 	Fold #9 Rep #1 pred log like = 730.487
#> Running at Laplace(0.0284268) None  Grid-point #5 at 2475 	Fold #10 Rep #1 pred log like = 661.93
#> AvgPred = 775.803 with stdev = 46.7171
#> Completed at 2475
#> Next point at 24750 with value 0 and continue = 1
#> search[ 0.2475 ] = 775.802(46.7085)
#> search[ 2.475 ] = 775.803(46.7144)
#> search[ 24.75 ] = 775.803(46.7163)
#> search[ 247.5 ] = 775.803(46.7169)
#> search[ 2475 ] = 775.803(46.7171)
#> 
#> Running at Laplace(0.00898933) None  Grid-point #6 at 24750 	Fold #1 Rep #1 pred log like = 764.125
#> Running at Laplace(0.00898933) None  Grid-point #6 at 24750 	Fold #2 Rep #1 pred log like = 802.186
#> Running at Laplace(0.00898933) None  Grid-point #6 at 24750 	Fold #3 Rep #1 pred log like = 809.911
#> Running at Laplace(0.00898933) None  Grid-point #6 at 24750 	Fold #4 Rep #1 pred log like = 829.657
#> Running at Laplace(0.00898933) None  Grid-point #6 at 24750 	Fold #5 Rep #1 pred log like = 803.251
#> Running at Laplace(0.00898933) None  Grid-point #6 at 24750 	Fold #6 Rep #1 pred log like = 774.978
#> Running at Laplace(0.00898933) None  Grid-point #6 at 24750 	Fold #7 Rep #1 pred log like = 771.948
#> Running at Laplace(0.00898933) None  Grid-point #6 at 24750 	Fold #8 Rep #1 pred log like = 809.556
#> Running at Laplace(0.00898933) None  Grid-point #6 at 24750 	Fold #9 Rep #1 pred log like = 730.487
#> Running at Laplace(0.00898933) None  Grid-point #6 at 24750 	Fold #10 Rep #1 pred log like = 661.93
#> AvgPred = 775.803 with stdev = 46.7171
#> Completed at 24750
#> Next point at 247500 with value 0 and continue = 1
#> search[ 0.2475 ] = 775.802(46.7085)
#> search[ 2.475 ] = 775.803(46.7144)
#> search[ 24.75 ] = 775.803(46.7163)
#> search[ 247.5 ] = 775.803(46.7169)
#> search[ 2475 ] = 775.803(46.7171)
#> search[ 24750 ] = 775.803(46.7171)
#> 
#> Running at Laplace(0.00284268) None  Grid-point #7 at 247500 	Fold #1 Rep #1 pred log like = 764.125
#> Running at Laplace(0.00284268) None  Grid-point #7 at 247500 	Fold #2 Rep #1 pred log like = 802.186
#> Running at Laplace(0.00284268) None  Grid-point #7 at 247500 	Fold #3 Rep #1 pred log like = 809.911
#> Running at Laplace(0.00284268) None  Grid-point #7 at 247500 	Fold #4 Rep #1 pred log like = 829.657
#> Running at Laplace(0.00284268) None  Grid-point #7 at 247500 	Fold #5 Rep #1 pred log like = 803.251
#> Running at Laplace(0.00284268) None  Grid-point #7 at 247500 	Fold #6 Rep #1 pred log like = 774.978
#> Running at Laplace(0.00284268) None  Grid-point #7 at 247500 	Fold #7 Rep #1 pred log like = 771.948
#> Running at Laplace(0.00284268) None  Grid-point #7 at 247500 	Fold #8 Rep #1 pred log like = 809.556
#> Running at Laplace(0.00284268) None  Grid-point #7 at 247500 	Fold #9 Rep #1 pred log like = 730.487
#> Running at Laplace(0.00284268) None  Grid-point #7 at 247500 	Fold #10 Rep #1 pred log like = 661.93
#> AvgPred = 775.803 with stdev = 46.7171
#> Completed at 247500
#> Next point at 2.475e+06 with value 0 and continue = 1
#> search[ 0.2475 ] = 775.802(46.7085)
#> search[ 2.475 ] = 775.803(46.7144)
#> search[ 24.75 ] = 775.803(46.7163)
#> search[ 247.5 ] = 775.803(46.7169)
#> search[ 2475 ] = 775.803(46.7171)
#> search[ 24750 ] = 775.803(46.7171)
#> search[ 247500 ] = 775.803(46.7171)
#> 
#> Running at Laplace(0.000898933) None  Grid-point #8 at 2.475e+06 	Fold #1 Rep #1 pred log like = 764.125
#> Running at Laplace(0.000898933) None  Grid-point #8 at 2.475e+06 	Fold #2 Rep #1 pred log like = 802.186
#> Running at Laplace(0.000898933) None  Grid-point #8 at 2.475e+06 	Fold #3 Rep #1 pred log like = 809.911
#> Running at Laplace(0.000898933) None  Grid-point #8 at 2.475e+06 	Fold #4 Rep #1 pred log like = 829.657
#> Running at Laplace(0.000898933) None  Grid-point #8 at 2.475e+06 	Fold #5 Rep #1 pred log like = 803.252
#> Running at Laplace(0.000898933) None  Grid-point #8 at 2.475e+06 	Fold #6 Rep #1 pred log like = 774.978
#> Running at Laplace(0.000898933) None  Grid-point #8 at 2.475e+06 	Fold #7 Rep #1 pred log like = 771.948
#> Running at Laplace(0.000898933) None  Grid-point #8 at 2.475e+06 	Fold #8 Rep #1 pred log like = 809.556
#> Running at Laplace(0.000898933) None  Grid-point #8 at 2.475e+06 	Fold #9 Rep #1 pred log like = 730.487
#> Running at Laplace(0.000898933) None  Grid-point #8 at 2.475e+06 	Fold #10 Rep #1 pred log like = 661.93
#> AvgPred = 775.803 with stdev = 46.7172
#> Completed at 2.475e+06
#> Next point at 2.475e+07 with value 0 and continue = 1
#> search[ 0.2475 ] = 775.802(46.7085)
#> search[ 2.475 ] = 775.803(46.7144)
#> search[ 24.75 ] = 775.803(46.7163)
#> search[ 247.5 ] = 775.803(46.7169)
#> search[ 2475 ] = 775.803(46.7171)
#> search[ 24750 ] = 775.803(46.7171)
#> search[ 247500 ] = 775.803(46.7171)
#> search[ 2.475e+06 ] = 775.803(46.7172)
#> 
#> Running at Laplace(0.000284268) None  Grid-point #9 at 2.475e+07 	Fold #1 Rep #1 pred log like = 764.125
#> Running at Laplace(0.000284268) None  Grid-point #9 at 2.475e+07 	Fold #2 Rep #1 pred log like = 802.186
#> Running at Laplace(0.000284268) None  Grid-point #9 at 2.475e+07 	Fold #3 Rep #1 pred log like = 809.911
#> Running at Laplace(0.000284268) None  Grid-point #9 at 2.475e+07 	Fold #4 Rep #1 pred log like = 829.657
#> Running at Laplace(0.000284268) None  Grid-point #9 at 2.475e+07 	Fold #5 Rep #1 pred log like = 803.252
#> Running at Laplace(0.000284268) None  Grid-point #9 at 2.475e+07 	Fold #6 Rep #1 pred log like = 774.978
#> Running at Laplace(0.000284268) None  Grid-point #9 at 2.475e+07 	Fold #7 Rep #1 pred log like = 771.948
#> Running at Laplace(0.000284268) None  Grid-point #9 at 2.475e+07 	Fold #8 Rep #1 pred log like = 809.556
#> Running at Laplace(0.000284268) None  Grid-point #9 at 2.475e+07 	Fold #9 Rep #1 pred log like = 730.487
#> Running at Laplace(0.000284268) None  Grid-point #9 at 2.475e+07 	Fold #10 Rep #1 pred log like = 661.93
#> AvgPred = 775.803 with stdev = 46.7172
#> Completed at 2.475e+07
#> Next point at 2.475e+08 with value 0 and continue = 1
#> search[ 0.2475 ] = 775.802(46.7085)
#> search[ 2.475 ] = 775.803(46.7144)
#> search[ 24.75 ] = 775.803(46.7163)
#> search[ 247.5 ] = 775.803(46.7169)
#> search[ 2475 ] = 775.803(46.7171)
#> search[ 24750 ] = 775.803(46.7171)
#> search[ 247500 ] = 775.803(46.7171)
#> search[ 2.475e+06 ] = 775.803(46.7172)
#> search[ 2.475e+07 ] = 775.803(46.7172)
#> 
#> Running at Laplace(8.98933e-05) None  Grid-point #10 at 2.475e+08 	Fold #1 Rep #1 pred log like = 764.125
#> Running at Laplace(8.98933e-05) None  Grid-point #10 at 2.475e+08 	Fold #2 Rep #1 pred log like = 802.186
#> Running at Laplace(8.98933e-05) None  Grid-point #10 at 2.475e+08 	Fold #3 Rep #1 pred log like = 809.911
#> Running at Laplace(8.98933e-05) None  Grid-point #10 at 2.475e+08 	Fold #4 Rep #1 pred log like = 829.657
#> Running at Laplace(8.98933e-05) None  Grid-point #10 at 2.475e+08 	Fold #5 Rep #1 pred log like = 803.252
#> Running at Laplace(8.98933e-05) None  Grid-point #10 at 2.475e+08 	Fold #6 Rep #1 pred log like = 774.978
#> Running at Laplace(8.98933e-05) None  Grid-point #10 at 2.475e+08 	Fold #7 Rep #1 pred log like = 771.948
#> Running at Laplace(8.98933e-05) None  Grid-point #10 at 2.475e+08 	Fold #8 Rep #1 pred log like = 809.556
#> Running at Laplace(8.98933e-05) None  Grid-point #10 at 2.475e+08 	Fold #9 Rep #1 pred log like = 730.487
#> Running at Laplace(8.98933e-05) None  Grid-point #10 at 2.475e+08 	Fold #10 Rep #1 pred log like = 661.93
#> AvgPred = 775.803 with stdev = 46.7172
#> Completed at 2.475e+08
#> Next point at 2.475e+09 with value 0 and continue = 1
#> search[ 0.2475 ] = 775.802(46.7085)
#> search[ 2.475 ] = 775.803(46.7144)
#> search[ 24.75 ] = 775.803(46.7163)
#> search[ 247.5 ] = 775.803(46.7169)
#> search[ 2475 ] = 775.803(46.7171)
#> search[ 24750 ] = 775.803(46.7171)
#> search[ 247500 ] = 775.803(46.7171)
#> search[ 2.475e+06 ] = 775.803(46.7172)
#> search[ 2.475e+07 ] = 775.803(46.7172)
#> search[ 2.475e+08 ] = 775.803(46.7172)
#> 
#> Running at Laplace(2.84268e-05) None  Grid-point #11 at 2.475e+09 	Fold #1 Rep #1 pred log like = 764.125
#> Running at Laplace(2.84268e-05) None  Grid-point #11 at 2.475e+09 	Fold #2 Rep #1 pred log like = 802.186
#> Running at Laplace(2.84268e-05) None  Grid-point #11 at 2.475e+09 	Fold #3 Rep #1 pred log like = 809.911
#> Running at Laplace(2.84268e-05) None  Grid-point #11 at 2.475e+09 	Fold #4 Rep #1 pred log like = 829.657
#> Running at Laplace(2.84268e-05) None  Grid-point #11 at 2.475e+09 	Fold #5 Rep #1 pred log like = 803.252
#> Running at Laplace(2.84268e-05) None  Grid-point #11 at 2.475e+09 	Fold #6 Rep #1 pred log like = 774.978
#> Running at Laplace(2.84268e-05) None  Grid-point #11 at 2.475e+09 	Fold #7 Rep #1 pred log like = 771.948
#> Running at Laplace(2.84268e-05) None  Grid-point #11 at 2.475e+09 	Fold #8 Rep #1 pred log like = 809.556
#> Running at Laplace(2.84268e-05) None  Grid-point #11 at 2.475e+09 	Fold #9 Rep #1 pred log like = 730.487
#> Running at Laplace(2.84268e-05) None  Grid-point #11 at 2.475e+09 	Fold #10 Rep #1 pred log like = 661.93
#> AvgPred = 775.803 with stdev = 46.7172
#> Completed at 2.475e+09
#> Next point at 2.475e+10 with value 0 and continue = 1
#> search[ 0.2475 ] = 775.802(46.7085)
#> search[ 2.475 ] = 775.803(46.7144)
#> search[ 24.75 ] = 775.803(46.7163)
#> search[ 247.5 ] = 775.803(46.7169)
#> search[ 2475 ] = 775.803(46.7171)
#> search[ 24750 ] = 775.803(46.7171)
#> search[ 247500 ] = 775.803(46.7171)
#> search[ 2.475e+06 ] = 775.803(46.7172)
#> search[ 2.475e+07 ] = 775.803(46.7172)
#> search[ 2.475e+08 ] = 775.803(46.7172)
#> search[ 2.475e+09 ] = 775.803(46.7172)
#> 
#> Running at Laplace(8.98933e-06) None  Grid-point #12 at 2.475e+10 	Fold #1 Rep #1 pred log like = 764.125
#> Running at Laplace(8.98933e-06) None  Grid-point #12 at 2.475e+10 	Fold #2 Rep #1 pred log like = 802.186
#> Running at Laplace(8.98933e-06) None  Grid-point #12 at 2.475e+10 	Fold #3 Rep #1 pred log like = 809.911
#> Running at Laplace(8.98933e-06) None  Grid-point #12 at 2.475e+10 	Fold #4 Rep #1 pred log like = 829.657
#> Running at Laplace(8.98933e-06) None  Grid-point #12 at 2.475e+10 	Fold #5 Rep #1 pred log like = 803.252
#> Running at Laplace(8.98933e-06) None  Grid-point #12 at 2.475e+10 	Fold #6 Rep #1 pred log like = 774.978
#> Running at Laplace(8.98933e-06) None  Grid-point #12 at 2.475e+10 	Fold #7 Rep #1 pred log like = 771.948
#> Running at Laplace(8.98933e-06) None  Grid-point #12 at 2.475e+10 	Fold #8 Rep #1 pred log like = 809.556
#> Running at Laplace(8.98933e-06) None  Grid-point #12 at 2.475e+10 	Fold #9 Rep #1 pred log like = 730.487
#> Running at Laplace(8.98933e-06) None  Grid-point #12 at 2.475e+10 	Fold #10 Rep #1 pred log like = 661.93
#> AvgPred = 775.803 with stdev = 46.7172
#> Completed at 2.475e+10
#> Next point at 2.475e+11 with value 0 and continue = 1
#> search[ 0.2475 ] = 775.802(46.7085)
#> search[ 2.475 ] = 775.803(46.7144)
#> search[ 24.75 ] = 775.803(46.7163)
#> search[ 247.5 ] = 775.803(46.7169)
#> search[ 2475 ] = 775.803(46.7171)
#> search[ 24750 ] = 775.803(46.7171)
#> search[ 247500 ] = 775.803(46.7171)
#> search[ 2.475e+06 ] = 775.803(46.7172)
#> search[ 2.475e+07 ] = 775.803(46.7172)
#> search[ 2.475e+08 ] = 775.803(46.7172)
#> search[ 2.475e+09 ] = 775.803(46.7172)
#> search[ 2.475e+10 ] = 775.803(46.7172)
#> 
#> Running at Laplace(2.84268e-06) None  Grid-point #13 at 2.475e+11 	Fold #1 Rep #1 pred log like = 764.125
#> Running at Laplace(2.84268e-06) None  Grid-point #13 at 2.475e+11 	Fold #2 Rep #1 pred log like = 802.186
#> Running at Laplace(2.84268e-06) None  Grid-point #13 at 2.475e+11 	Fold #3 Rep #1 pred log like = 809.911
#> Running at Laplace(2.84268e-06) None  Grid-point #13 at 2.475e+11 	Fold #4 Rep #1 pred log like = 829.657
#> Running at Laplace(2.84268e-06) None  Grid-point #13 at 2.475e+11 	Fold #5 Rep #1 pred log like = 803.252
#> Running at Laplace(2.84268e-06) None  Grid-point #13 at 2.475e+11 	Fold #6 Rep #1 pred log like = 774.978
#> Running at Laplace(2.84268e-06) None  Grid-point #13 at 2.475e+11 	Fold #7 Rep #1 pred log like = 771.948
#> Running at Laplace(2.84268e-06) None  Grid-point #13 at 2.475e+11 	Fold #8 Rep #1 pred log like = 809.556
#> Running at Laplace(2.84268e-06) None  Grid-point #13 at 2.475e+11 	Fold #9 Rep #1 pred log like = 730.487
#> Running at Laplace(2.84268e-06) None  Grid-point #13 at 2.475e+11 	Fold #10 Rep #1 pred log like = 661.93
#> AvgPred = 775.803 with stdev = 46.7172
#> Completed at 2.475e+11
#> Next point at 2.475e+12 with value 0 and continue = 1
#> search[ 0.2475 ] = 775.802(46.7085)
#> search[ 2.475 ] = 775.803(46.7144)
#> search[ 24.75 ] = 775.803(46.7163)
#> search[ 247.5 ] = 775.803(46.7169)
#> search[ 2475 ] = 775.803(46.7171)
#> search[ 24750 ] = 775.803(46.7171)
#> search[ 247500 ] = 775.803(46.7171)
#> search[ 2.475e+06 ] = 775.803(46.7172)
#> search[ 2.475e+07 ] = 775.803(46.7172)
#> search[ 2.475e+08 ] = 775.803(46.7172)
#> search[ 2.475e+09 ] = 775.803(46.7172)
#> search[ 2.475e+10 ] = 775.803(46.7172)
#> search[ 2.475e+11 ] = 775.803(46.7172)
#> 
#> Running at Laplace(8.98933e-07) None  Grid-point #14 at 2.475e+12 	Fold #1 Rep #1 pred log like = 764.125
#> Running at Laplace(8.98933e-07) None  Grid-point #14 at 2.475e+12 	Fold #2 Rep #1 pred log like = 802.186
#> Running at Laplace(8.98933e-07) None  Grid-point #14 at 2.475e+12 	Fold #3 Rep #1 pred log like = 809.911
#> Running at Laplace(8.98933e-07) None  Grid-point #14 at 2.475e+12 	Fold #4 Rep #1 pred log like = 829.657
#> Running at Laplace(8.98933e-07) None  Grid-point #14 at 2.475e+12 	Fold #5 Rep #1 pred log like = 803.252
#> Running at Laplace(8.98933e-07) None  Grid-point #14 at 2.475e+12 	Fold #6 Rep #1 pred log like = 774.978
#> Running at Laplace(8.98933e-07) None  Grid-point #14 at 2.475e+12 	Fold #7 Rep #1 pred log like = 771.948
#> Running at Laplace(8.98933e-07) None  Grid-point #14 at 2.475e+12 	Fold #8 Rep #1 pred log like = 809.556
#> Running at Laplace(8.98933e-07) None  Grid-point #14 at 2.475e+12 	Fold #9 Rep #1 pred log like = 730.487
#> Running at Laplace(8.98933e-07) None  Grid-point #14 at 2.475e+12 	Fold #10 Rep #1 pred log like = 661.93
#> AvgPred = 775.803 with stdev = 46.7172
#> Completed at 2.475e+12
#> Next point at 2.475e+13 with value 0 and continue = 1
#> search[ 0.2475 ] = 775.802(46.7085)
#> search[ 2.475 ] = 775.803(46.7144)
#> search[ 24.75 ] = 775.803(46.7163)
#> search[ 247.5 ] = 775.803(46.7169)
#> search[ 2475 ] = 775.803(46.7171)
#> search[ 24750 ] = 775.803(46.7171)
#> search[ 247500 ] = 775.803(46.7171)
#> search[ 2.475e+06 ] = 775.803(46.7172)
#> search[ 2.475e+07 ] = 775.803(46.7172)
#> search[ 2.475e+08 ] = 775.803(46.7172)
#> search[ 2.475e+09 ] = 775.803(46.7172)
#> search[ 2.475e+10 ] = 775.803(46.7172)
#> search[ 2.475e+11 ] = 775.803(46.7172)
#> search[ 2.475e+12 ] = 775.803(46.7172)
#> 
#> Running at Laplace(2.84268e-07) None  Grid-point #15 at 2.475e+13 	Fold #1 Rep #1 pred log like = 764.125
#> Running at Laplace(2.84268e-07) None  Grid-point #15 at 2.475e+13 	Fold #2 Rep #1 pred log like = 802.186
#> Running at Laplace(2.84268e-07) None  Grid-point #15 at 2.475e+13 	Fold #3 Rep #1 pred log like = 809.911
#> Running at Laplace(2.84268e-07) None  Grid-point #15 at 2.475e+13 	Fold #4 Rep #1 pred log like = 829.657
#> Running at Laplace(2.84268e-07) None  Grid-point #15 at 2.475e+13 	Fold #5 Rep #1 pred log like = 803.252
#> Running at Laplace(2.84268e-07) None  Grid-point #15 at 2.475e+13 	Fold #6 Rep #1 pred log like = 774.978
#> Running at Laplace(2.84268e-07) None  Grid-point #15 at 2.475e+13 	Fold #7 Rep #1 pred log like = 771.948
#> Running at Laplace(2.84268e-07) None  Grid-point #15 at 2.475e+13 	Fold #8 Rep #1 pred log like = 809.556
#> Running at Laplace(2.84268e-07) None  Grid-point #15 at 2.475e+13 	Fold #9 Rep #1 pred log like = 730.487
#> Running at Laplace(2.84268e-07) None  Grid-point #15 at 2.475e+13 	Fold #10 Rep #1 pred log like = 661.93
#> AvgPred = 775.803 with stdev = 46.7172
#> Completed at 2.475e+13
#> Next point at 2.475e+14 with value 0 and continue = 1
#> search[ 0.2475 ] = 775.802(46.7085)
#> search[ 2.475 ] = 775.803(46.7144)
#> search[ 24.75 ] = 775.803(46.7163)
#> search[ 247.5 ] = 775.803(46.7169)
#> search[ 2475 ] = 775.803(46.7171)
#> search[ 24750 ] = 775.803(46.7171)
#> search[ 247500 ] = 775.803(46.7171)
#> search[ 2.475e+06 ] = 775.803(46.7172)
#> search[ 2.475e+07 ] = 775.803(46.7172)
#> search[ 2.475e+08 ] = 775.803(46.7172)
#> search[ 2.475e+09 ] = 775.803(46.7172)
#> search[ 2.475e+10 ] = 775.803(46.7172)
#> search[ 2.475e+11 ] = 775.803(46.7172)
#> search[ 2.475e+12 ] = 775.803(46.7172)
#> search[ 2.475e+13 ] = 775.803(46.7172)
#> 
#> Running at Laplace(8.98933e-08) None  Grid-point #16 at 2.475e+14 	Fold #1 Rep #1 pred log like = 764.125
#> Running at Laplace(8.98933e-08) None  Grid-point #16 at 2.475e+14 	Fold #2 Rep #1 pred log like = 802.186
#> Running at Laplace(8.98933e-08) None  Grid-point #16 at 2.475e+14 	Fold #3 Rep #1 pred log like = 809.911
#> Running at Laplace(8.98933e-08) None  Grid-point #16 at 2.475e+14 	Fold #4 Rep #1 pred log like = 829.657
#> Running at Laplace(8.98933e-08) None  Grid-point #16 at 2.475e+14 	Fold #5 Rep #1 pred log like = 803.252
#> Running at Laplace(8.98933e-08) None  Grid-point #16 at 2.475e+14 	Fold #6 Rep #1 pred log like = 774.978
#> Running at Laplace(8.98933e-08) None  Grid-point #16 at 2.475e+14 	Fold #7 Rep #1 pred log like = 771.948
#> Running at Laplace(8.98933e-08) None  Grid-point #16 at 2.475e+14 	Fold #8 Rep #1 pred log like = 809.556
#> Running at Laplace(8.98933e-08) None  Grid-point #16 at 2.475e+14 	Fold #9 Rep #1 pred log like = 730.487
#> Running at Laplace(8.98933e-08) None  Grid-point #16 at 2.475e+14 	Fold #10 Rep #1 pred log like = 661.93
#> AvgPred = 775.803 with stdev = 46.7172
#> Completed at 2.475e+14
#> Next point at 2.475e+15 with value 0 and continue = 1
#> search[ 0.2475 ] = 775.802(46.7085)
#> search[ 2.475 ] = 775.803(46.7144)
#> search[ 24.75 ] = 775.803(46.7163)
#> search[ 247.5 ] = 775.803(46.7169)
#> search[ 2475 ] = 775.803(46.7171)
#> search[ 24750 ] = 775.803(46.7171)
#> search[ 247500 ] = 775.803(46.7171)
#> search[ 2.475e+06 ] = 775.803(46.7172)
#> search[ 2.475e+07 ] = 775.803(46.7172)
#> search[ 2.475e+08 ] = 775.803(46.7172)
#> search[ 2.475e+09 ] = 775.803(46.7172)
#> search[ 2.475e+10 ] = 775.803(46.7172)
#> search[ 2.475e+11 ] = 775.803(46.7172)
#> search[ 2.475e+12 ] = 775.803(46.7172)
#> search[ 2.475e+13 ] = 775.803(46.7172)
#> search[ 2.475e+14 ] = 775.803(46.7172)
#> 
#> Running at Laplace(2.84268e-08) None  Grid-point #17 at 2.475e+15 	Fold #1 Rep #1 pred log like = 764.125
#> Running at Laplace(2.84268e-08) None  Grid-point #17 at 2.475e+15 	Fold #2 Rep #1 pred log like = 802.186
#> Running at Laplace(2.84268e-08) None  Grid-point #17 at 2.475e+15 	Fold #3 Rep #1 pred log like = 809.911
#> Running at Laplace(2.84268e-08) None  Grid-point #17 at 2.475e+15 	Fold #4 Rep #1 pred log like = 829.657
#> Running at Laplace(2.84268e-08) None  Grid-point #17 at 2.475e+15 	Fold #5 Rep #1 pred log like = 803.252
#> Running at Laplace(2.84268e-08) None  Grid-point #17 at 2.475e+15 	Fold #6 Rep #1 pred log like = 774.978
#> Running at Laplace(2.84268e-08) None  Grid-point #17 at 2.475e+15 	Fold #7 Rep #1 pred log like = 771.948
#> Running at Laplace(2.84268e-08) None  Grid-point #17 at 2.475e+15 	Fold #8 Rep #1 pred log like = 809.556
#> Running at Laplace(2.84268e-08) None  Grid-point #17 at 2.475e+15 	Fold #9 Rep #1 pred log like = 730.487
#> Running at Laplace(2.84268e-08) None  Grid-point #17 at 2.475e+15 	Fold #10 Rep #1 pred log like = 661.93
#> AvgPred = 775.803 with stdev = 46.7172
#> Completed at 2.475e+15
#> Next point at 2.475e+16 with value 0 and continue = 1
#> search[ 0.2475 ] = 775.802(46.7085)
#> search[ 2.475 ] = 775.803(46.7144)
#> search[ 24.75 ] = 775.803(46.7163)
#> search[ 247.5 ] = 775.803(46.7169)
#> search[ 2475 ] = 775.803(46.7171)
#> search[ 24750 ] = 775.803(46.7171)
#> search[ 247500 ] = 775.803(46.7171)
#> search[ 2.475e+06 ] = 775.803(46.7172)
#> search[ 2.475e+07 ] = 775.803(46.7172)
#> search[ 2.475e+08 ] = 775.803(46.7172)
#> search[ 2.475e+09 ] = 775.803(46.7172)
#> search[ 2.475e+10 ] = 775.803(46.7172)
#> search[ 2.475e+11 ] = 775.803(46.7172)
#> search[ 2.475e+12 ] = 775.803(46.7172)
#> search[ 2.475e+13 ] = 775.803(46.7172)
#> search[ 2.475e+14 ] = 775.803(46.7172)
#> search[ 2.475e+15 ] = 775.803(46.7172)
#> 
#> Running at Laplace(8.98933e-09) None  Grid-point #18 at 2.475e+16 	Fold #1 Rep #1 pred log like = 764.125
#> Running at Laplace(8.98933e-09) None  Grid-point #18 at 2.475e+16 	Fold #2 Rep #1 pred log like = 802.186
#> Running at Laplace(8.98933e-09) None  Grid-point #18 at 2.475e+16 	Fold #3 Rep #1 pred log like = 809.911
#> Running at Laplace(8.98933e-09) None  Grid-point #18 at 2.475e+16 	Fold #4 Rep #1 pred log like = 829.657
#> Running at Laplace(8.98933e-09) None  Grid-point #18 at 2.475e+16 	Fold #5 Rep #1 pred log like = 803.252
#> Running at Laplace(8.98933e-09) None  Grid-point #18 at 2.475e+16 	Fold #6 Rep #1 pred log like = 774.978
#> Running at Laplace(8.98933e-09) None  Grid-point #18 at 2.475e+16 	Fold #7 Rep #1 pred log like = 771.948
#> Running at Laplace(8.98933e-09) None  Grid-point #18 at 2.475e+16 	Fold #8 Rep #1 pred log like = 809.556
#> Running at Laplace(8.98933e-09) None  Grid-point #18 at 2.475e+16 	Fold #9 Rep #1 pred log like = 730.487
#> Running at Laplace(8.98933e-09) None  Grid-point #18 at 2.475e+16 	Fold #10 Rep #1 pred log like = 661.93
#> AvgPred = 775.803 with stdev = 46.7172
#> Completed at 2.475e+16
#> Next point at 2.475e+17 with value 0 and continue = 1
#> search[ 0.2475 ] = 775.802(46.7085)
#> search[ 2.475 ] = 775.803(46.7144)
#> search[ 24.75 ] = 775.803(46.7163)
#> search[ 247.5 ] = 775.803(46.7169)
#> search[ 2475 ] = 775.803(46.7171)
#> search[ 24750 ] = 775.803(46.7171)
#> search[ 247500 ] = 775.803(46.7171)
#> search[ 2.475e+06 ] = 775.803(46.7172)
#> search[ 2.475e+07 ] = 775.803(46.7172)
#> search[ 2.475e+08 ] = 775.803(46.7172)
#> search[ 2.475e+09 ] = 775.803(46.7172)
#> search[ 2.475e+10 ] = 775.803(46.7172)
#> search[ 2.475e+11 ] = 775.803(46.7172)
#> search[ 2.475e+12 ] = 775.803(46.7172)
#> search[ 2.475e+13 ] = 775.803(46.7172)
#> search[ 2.475e+14 ] = 775.803(46.7172)
#> search[ 2.475e+15 ] = 775.803(46.7172)
#> search[ 2.475e+16 ] = 775.803(46.7172)
#> 
#> Running at Laplace(2.84268e-09) None  Grid-point #19 at 2.475e+17 	Fold #1 Rep #1 pred log like = 764.125
#> Running at Laplace(2.84268e-09) None  Grid-point #19 at 2.475e+17 	Fold #2 Rep #1 pred log like = 802.186
#> Running at Laplace(2.84268e-09) None  Grid-point #19 at 2.475e+17 	Fold #3 Rep #1 pred log like = 809.911
#> Running at Laplace(2.84268e-09) None  Grid-point #19 at 2.475e+17 	Fold #4 Rep #1 pred log like = 829.657
#> Running at Laplace(2.84268e-09) None  Grid-point #19 at 2.475e+17 	Fold #5 Rep #1 pred log like = 803.252
#> Running at Laplace(2.84268e-09) None  Grid-point #19 at 2.475e+17 	Fold #6 Rep #1 pred log like = 774.978
#> Running at Laplace(2.84268e-09) None  Grid-point #19 at 2.475e+17 	Fold #7 Rep #1 pred log like = 771.948
#> Running at Laplace(2.84268e-09) None  Grid-point #19 at 2.475e+17 	Fold #8 Rep #1 pred log like = 809.556
#> Running at Laplace(2.84268e-09) None  Grid-point #19 at 2.475e+17 	Fold #9 Rep #1 pred log like = 730.487
#> Running at Laplace(2.84268e-09) None  Grid-point #19 at 2.475e+17 	Fold #10 Rep #1 pred log like = 661.93
#> AvgPred = 775.803 with stdev = 46.7172
#> Completed at 2.475e+17
#> Next point at 2.475e+18 with value 0 and continue = 1
#> search[ 0.2475 ] = 775.802(46.7085)
#> search[ 2.475 ] = 775.803(46.7144)
#> search[ 24.75 ] = 775.803(46.7163)
#> search[ 247.5 ] = 775.803(46.7169)
#> search[ 2475 ] = 775.803(46.7171)
#> search[ 24750 ] = 775.803(46.7171)
#> search[ 247500 ] = 775.803(46.7171)
#> search[ 2.475e+06 ] = 775.803(46.7172)
#> search[ 2.475e+07 ] = 775.803(46.7172)
#> search[ 2.475e+08 ] = 775.803(46.7172)
#> search[ 2.475e+09 ] = 775.803(46.7172)
#> search[ 2.475e+10 ] = 775.803(46.7172)
#> search[ 2.475e+11 ] = 775.803(46.7172)
#> search[ 2.475e+12 ] = 775.803(46.7172)
#> search[ 2.475e+13 ] = 775.803(46.7172)
#> search[ 2.475e+14 ] = 775.803(46.7172)
#> search[ 2.475e+15 ] = 775.803(46.7172)
#> search[ 2.475e+16 ] = 775.803(46.7172)
#> search[ 2.475e+17 ] = 775.803(46.7172)
#> 
#> Running at Laplace(8.98933e-10) None  Grid-point #20 at 2.475e+18 	Fold #1 Rep #1 pred log like = 764.125
#> Running at Laplace(8.98933e-10) None  Grid-point #20 at 2.475e+18 	Fold #2 Rep #1 pred log like = 802.186
#> Running at Laplace(8.98933e-10) None  Grid-point #20 at 2.475e+18 	Fold #3 Rep #1 pred log like = 809.911
#> Running at Laplace(8.98933e-10) None  Grid-point #20 at 2.475e+18 	Fold #4 Rep #1 pred log like = 829.657
#> Running at Laplace(8.98933e-10) None  Grid-point #20 at 2.475e+18 	Fold #5 Rep #1 pred log like = 803.252
#> Running at Laplace(8.98933e-10) None  Grid-point #20 at 2.475e+18 	Fold #6 Rep #1 pred log like = 774.978
#> Running at Laplace(8.98933e-10) None  Grid-point #20 at 2.475e+18 	Fold #7 Rep #1 pred log like = 771.948
#> Running at Laplace(8.98933e-10) None  Grid-point #20 at 2.475e+18 	Fold #8 Rep #1 pred log like = 809.556
#> Running at Laplace(8.98933e-10) None  Grid-point #20 at 2.475e+18 	Fold #9 Rep #1 pred log like = 730.487
#> Running at Laplace(8.98933e-10) None  Grid-point #20 at 2.475e+18 	Fold #10 Rep #1 pred log like = 661.93
#> AvgPred = 775.803 with stdev = 46.7172
#> Completed at 2.475e+18
#> Next point at 2.475e+19 with value 0 and continue = 1
#> search[ 0.2475 ] = 775.802(46.7085)
#> search[ 2.475 ] = 775.803(46.7144)
#> search[ 24.75 ] = 775.803(46.7163)
#> search[ 247.5 ] = 775.803(46.7169)
#> search[ 2475 ] = 775.803(46.7171)
#> search[ 24750 ] = 775.803(46.7171)
#> search[ 247500 ] = 775.803(46.7171)
#> search[ 2.475e+06 ] = 775.803(46.7172)
#> search[ 2.475e+07 ] = 775.803(46.7172)
#> search[ 2.475e+08 ] = 775.803(46.7172)
#> search[ 2.475e+09 ] = 775.803(46.7172)
#> search[ 2.475e+10 ] = 775.803(46.7172)
#> search[ 2.475e+11 ] = 775.803(46.7172)
#> search[ 2.475e+12 ] = 775.803(46.7172)
#> search[ 2.475e+13 ] = 775.803(46.7172)
#> search[ 2.475e+14 ] = 775.803(46.7172)
#> search[ 2.475e+15 ] = 775.803(46.7172)
#> search[ 2.475e+16 ] = 775.803(46.7172)
#> search[ 2.475e+17 ] = 775.803(46.7172)
#> search[ 2.475e+18 ] = 775.803(46.7172)
#> 
#> Running at Laplace(2.84268e-10) None  Grid-point #21 at 2.475e+19 	Fold #1 Rep #1 pred log like = 764.125
#> Running at Laplace(2.84268e-10) None  Grid-point #21 at 2.475e+19 	Fold #2 Rep #1 pred log like = 802.186
#> Running at Laplace(2.84268e-10) None  Grid-point #21 at 2.475e+19 	Fold #3 Rep #1 pred log like = 809.911
#> Running at Laplace(2.84268e-10) None  Grid-point #21 at 2.475e+19 	Fold #4 Rep #1 pred log like = 829.657
#> Running at Laplace(2.84268e-10) None  Grid-point #21 at 2.475e+19 	Fold #5 Rep #1 pred log like = 803.252
#> Running at Laplace(2.84268e-10) None  Grid-point #21 at 2.475e+19 	Fold #6 Rep #1 pred log like = 774.978
#> Running at Laplace(2.84268e-10) None  Grid-point #21 at 2.475e+19 	Fold #7 Rep #1 pred log like = 771.948
#> Running at Laplace(2.84268e-10) None  Grid-point #21 at 2.475e+19 	Fold #8 Rep #1 pred log like = 809.556
#> Running at Laplace(2.84268e-10) None  Grid-point #21 at 2.475e+19 	Fold #9 Rep #1 pred log like = 730.487
#> Running at Laplace(2.84268e-10) None  Grid-point #21 at 2.475e+19 	Fold #10 Rep #1 pred log like = 661.93
#> AvgPred = 775.803 with stdev = 46.7172
#> Completed at 2.475e+19
#> Next point at 6.1351e+11 with value 775.803 and continue = 0
#> search[ 0.2475 ] = 775.802(46.7085)
#> search[ 2.475 ] = 775.803(46.7144)
#> search[ 24.75 ] = 775.803(46.7163)
#> search[ 247.5 ] = 775.803(46.7169)
#> search[ 2475 ] = 775.803(46.7171)
#> search[ 24750 ] = 775.803(46.7171)
#> search[ 247500 ] = 775.803(46.7171)
#> search[ 2.475e+06 ] = 775.803(46.7172)
#> search[ 2.475e+07 ] = 775.803(46.7172)
#> search[ 2.475e+08 ] = 775.803(46.7172)
#> search[ 2.475e+09 ] = 775.803(46.7172)
#> search[ 2.475e+10 ] = 775.803(46.7172)
#> search[ 2.475e+11 ] = 775.803(46.7172)
#> search[ 2.475e+12 ] = 775.803(46.7172)
#> search[ 2.475e+13 ] = 775.803(46.7172)
#> search[ 2.475e+14 ] = 775.803(46.7172)
#> search[ 2.475e+15 ] = 775.803(46.7172)
#> search[ 2.475e+16 ] = 775.803(46.7172)
#> search[ 2.475e+17 ] = 775.803(46.7172)
#> search[ 2.475e+18 ] = 775.803(46.7172)
#> search[ 2.475e+19 ] = 775.803(46.7172)
#> 
#> 
#> Maximum predicted log likelihood (775.803) estimated at:
#> 	6.1351e+11 (variance)
#> 	1.80553e-06 (lambda)
#> 
#> Fitting model at optimal hyperparameter
#> Using prior: Laplace(1.80553e-06) None 

#Find out what the optimal hyperparameter was:
getHyperParameter(fit)
#> [1] 6.1351e+11

#Extract the current log-likelihood, and coefficients
logLik(fit)
#> 'log Lik.' -2193.651 (df=3)
coef(fit)
#> (Intercept)           1           2 
#> -3.61883104  0.06239337  0.04106083 

#We can only retrieve the confidence interval for unregularized coefficients:
confint(fit, c(0))
#> Using 1 thread(s)
#>             covariate     2.5 %    97.5 % evaluations
#> (Intercept)         0 -3.649159 -3.588779          22