Extract log-likelihood

logLik returns the current log-likelihood of the fit in a Cyclops model fit object

# S3 method for cyclopsFit
logLik(object, ...)

Arguments

object
A Cyclops model fit object

...
Additional arguments

Examples

#Generate some simulated data:
sim <- simulateCyclopsData(nstrata = 1, nrows = 1000, ncovars = 2, eCovarsPerRow = 0.5, 
                           model = "poisson")

#> Sparseness = 76.3 %

cyclopsData <- convertToCyclopsData(sim$outcomes, sim$covariates, modelType = "pr", 
                                    addIntercept = TRUE)

#> Sorting covariates by covariateId and rowId

#Define the prior and control objects to use cross-validation for finding the 
#optimal hyperparameter:
prior <- createPrior("laplace", exclude = 0, useCrossValidation = TRUE)
control <- createControl(cvType = "auto", noiseLevel = "quiet")

#Fit the model
fit <- fitCyclopsModel(cyclopsData,prior = prior, control = control)  

#> Using cross-validation selector type byRow
#> Performing 10-fold cross-validation [seed = 1604638984] with data partitions of sizes 100 100 100 100 100 100 100 100 100 100
#> Using 1 thread(s)
#> Starting var = 0.237 (default)
#> Running at Laplace(2.90496) None  Grid-point #1 at 0.237 	Fold #1 Rep #1 pred log like = 49.1689
#> Running at Laplace(2.90496) None  Grid-point #1 at 0.237 	Fold #2 Rep #1 pred log like = 46.2885
#> Running at Laplace(2.90496) None  Grid-point #1 at 0.237 	Fold #3 Rep #1 pred log like = 46.0454
#> Running at Laplace(2.90496) None  Grid-point #1 at 0.237 	Fold #4 Rep #1 pred log like = 82.6947
#> Running at Laplace(2.90496) None  Grid-point #1 at 0.237 	Fold #5 Rep #1 pred log like = 65.2779
#> Running at Laplace(2.90496) None  Grid-point #1 at 0.237 	Fold #6 Rep #1 pred log like = 57.3497
#> Running at Laplace(2.90496) None  Grid-point #1 at 0.237 	Fold #7 Rep #1 pred log like = 51.0739
#> Running at Laplace(2.90496) None  Grid-point #1 at 0.237 	Fold #8 Rep #1 pred log like = 104.447
#> Running at Laplace(2.90496) None  Grid-point #1 at 0.237 	Fold #9 Rep #1 pred log like = 65.074
#> Running at Laplace(2.90496) None  Grid-point #1 at 0.237 	Fold #10 Rep #1 pred log like = 56.0256
#> AvgPred = 62.3445 with stdev = 17.5833
#> Completed at 0.237
#> Next point at 2.37 with value 0 and continue = 1
#> search[ 0.237 ] = 62.3445(17.5833)
#> 
#> Running at Laplace(0.91863) None  Grid-point #2 at 2.37 	Fold #1 Rep #1 pred log like = 49.1923
#> Running at Laplace(0.91863) None  Grid-point #2 at 2.37 	Fold #2 Rep #1 pred log like = 46.2714
#> Running at Laplace(0.91863) None  Grid-point #2 at 2.37 	Fold #3 Rep #1 pred log like = 45.9551
#> Running at Laplace(0.91863) None  Grid-point #2 at 2.37 	Fold #4 Rep #1 pred log like = 82.6351
#> Running at Laplace(0.91863) None  Grid-point #2 at 2.37 	Fold #5 Rep #1 pred log like = 65.2641
#> Running at Laplace(0.91863) None  Grid-point #2 at 2.37 	Fold #6 Rep #1 pred log like = 57.3012
#> Running at Laplace(0.91863) None  Grid-point #2 at 2.37 	Fold #7 Rep #1 pred log like = 51.1562
#> Running at Laplace(0.91863) None  Grid-point #2 at 2.37 	Fold #8 Rep #1 pred log like = 104.537
#> Running at Laplace(0.91863) None  Grid-point #2 at 2.37 	Fold #9 Rep #1 pred log like = 65.1193
#> Running at Laplace(0.91863) None  Grid-point #2 at 2.37 	Fold #10 Rep #1 pred log like = 56.0385
#> AvgPred = 62.347 with stdev = 17.6025
#> Completed at 2.37
#> Next point at 23.7 with value 0 and continue = 1
#> search[ 0.237 ] = 62.3445(17.5833)
#> search[ 2.37 ] = 62.347(17.6025)
#> 
#> Running at Laplace(0.290496) None  Grid-point #3 at 23.7 	Fold #1 Rep #1 pred log like = 49.1992
#> Running at Laplace(0.290496) None  Grid-point #3 at 23.7 	Fold #2 Rep #1 pred log like = 46.2656
#> Running at Laplace(0.290496) None  Grid-point #3 at 23.7 	Fold #3 Rep #1 pred log like = 45.9259
#> Running at Laplace(0.290496) None  Grid-point #3 at 23.7 	Fold #4 Rep #1 pred log like = 82.6157
#> Running at Laplace(0.290496) None  Grid-point #3 at 23.7 	Fold #5 Rep #1 pred log like = 65.2593
#> Running at Laplace(0.290496) None  Grid-point #3 at 23.7 	Fold #6 Rep #1 pred log like = 57.2854
#> Running at Laplace(0.290496) None  Grid-point #3 at 23.7 	Fold #7 Rep #1 pred log like = 51.1817
#> Running at Laplace(0.290496) None  Grid-point #3 at 23.7 	Fold #8 Rep #1 pred log like = 104.565
#> Running at Laplace(0.290496) None  Grid-point #3 at 23.7 	Fold #9 Rep #1 pred log like = 65.1333
#> Running at Laplace(0.290496) None  Grid-point #3 at 23.7 	Fold #10 Rep #1 pred log like = 56.0423
#> AvgPred = 62.3474 with stdev = 17.6085
#> Completed at 23.7
#> Next point at 237 with value 0 and continue = 1
#> search[ 0.237 ] = 62.3445(17.5833)
#> search[ 2.37 ] = 62.347(17.6025)
#> search[ 23.7 ] = 62.3474(17.6085)
#> 
#> Running at Laplace(0.091863) None  Grid-point #4 at 237 	Fold #1 Rep #1 pred log like = 49.2014
#> Running at Laplace(0.091863) None  Grid-point #4 at 237 	Fold #2 Rep #1 pred log like = 46.2637
#> Running at Laplace(0.091863) None  Grid-point #4 at 237 	Fold #3 Rep #1 pred log like = 45.9167
#> Running at Laplace(0.091863) None  Grid-point #4 at 237 	Fold #4 Rep #1 pred log like = 82.6095
#> Running at Laplace(0.091863) None  Grid-point #4 at 237 	Fold #5 Rep #1 pred log like = 65.2578
#> Running at Laplace(0.091863) None  Grid-point #4 at 237 	Fold #6 Rep #1 pred log like = 57.2804
#> Running at Laplace(0.091863) None  Grid-point #4 at 237 	Fold #7 Rep #1 pred log like = 51.1897
#> Running at Laplace(0.091863) None  Grid-point #4 at 237 	Fold #8 Rep #1 pred log like = 104.574
#> Running at Laplace(0.091863) None  Grid-point #4 at 237 	Fold #9 Rep #1 pred log like = 65.1377
#> Running at Laplace(0.091863) None  Grid-point #4 at 237 	Fold #10 Rep #1 pred log like = 56.0434
#> AvgPred = 62.3474 with stdev = 17.6104
#> Completed at 237
#> Next point at 2370 with value 0 and continue = 1
#> search[ 0.237 ] = 62.3445(17.5833)
#> search[ 2.37 ] = 62.347(17.6025)
#> search[ 23.7 ] = 62.3474(17.6085)
#> search[ 237 ] = 62.3474(17.6104)
#> 
#> Running at Laplace(0.0290496) None  Grid-point #5 at 2370 	Fold #1 Rep #1 pred log like = 49.202
#> Running at Laplace(0.0290496) None  Grid-point #5 at 2370 	Fold #2 Rep #1 pred log like = 46.2631
#> Running at Laplace(0.0290496) None  Grid-point #5 at 2370 	Fold #3 Rep #1 pred log like = 45.9137
#> Running at Laplace(0.0290496) None  Grid-point #5 at 2370 	Fold #4 Rep #1 pred log like = 82.6075
#> Running at Laplace(0.0290496) None  Grid-point #5 at 2370 	Fold #5 Rep #1 pred log like = 65.2573
#> Running at Laplace(0.0290496) None  Grid-point #5 at 2370 	Fold #6 Rep #1 pred log like = 57.2788
#> Running at Laplace(0.0290496) None  Grid-point #5 at 2370 	Fold #7 Rep #1 pred log like = 51.1923
#> Running at Laplace(0.0290496) None  Grid-point #5 at 2370 	Fold #8 Rep #1 pred log like = 104.577
#> Running at Laplace(0.0290496) None  Grid-point #5 at 2370 	Fold #9 Rep #1 pred log like = 65.1391
#> Running at Laplace(0.0290496) None  Grid-point #5 at 2370 	Fold #10 Rep #1 pred log like = 56.0438
#> AvgPred = 62.3474 with stdev = 17.611
#> Completed at 2370
#> Next point at 23700 with value 0 and continue = 1
#> search[ 0.237 ] = 62.3445(17.5833)
#> search[ 2.37 ] = 62.347(17.6025)
#> search[ 23.7 ] = 62.3474(17.6085)
#> search[ 237 ] = 62.3474(17.6104)
#> search[ 2370 ] = 62.3474(17.611)
#> 
#> Running at Laplace(0.0091863) None  Grid-point #6 at 23700 	Fold #1 Rep #1 pred log like = 49.2022
#> Running at Laplace(0.0091863) None  Grid-point #6 at 23700 	Fold #2 Rep #1 pred log like = 46.263
#> Running at Laplace(0.0091863) None  Grid-point #6 at 23700 	Fold #3 Rep #1 pred log like = 45.9128
#> Running at Laplace(0.0091863) None  Grid-point #6 at 23700 	Fold #4 Rep #1 pred log like = 82.6069
#> Running at Laplace(0.0091863) None  Grid-point #6 at 23700 	Fold #5 Rep #1 pred log like = 65.2571
#> Running at Laplace(0.0091863) None  Grid-point #6 at 23700 	Fold #6 Rep #1 pred log like = 57.2783
#> Running at Laplace(0.0091863) None  Grid-point #6 at 23700 	Fold #7 Rep #1 pred log like = 51.1931
#> Running at Laplace(0.0091863) None  Grid-point #6 at 23700 	Fold #8 Rep #1 pred log like = 104.578
#> Running at Laplace(0.0091863) None  Grid-point #6 at 23700 	Fold #9 Rep #1 pred log like = 65.1396
#> Running at Laplace(0.0091863) None  Grid-point #6 at 23700 	Fold #10 Rep #1 pred log like = 56.0439
#> AvgPred = 62.3474 with stdev = 17.6112
#> Completed at 23700
#> Next point at 237000 with value 0 and continue = 1
#> search[ 0.237 ] = 62.3445(17.5833)
#> search[ 2.37 ] = 62.347(17.6025)
#> search[ 23.7 ] = 62.3474(17.6085)
#> search[ 237 ] = 62.3474(17.6104)
#> search[ 2370 ] = 62.3474(17.611)
#> search[ 23700 ] = 62.3474(17.6112)
#> 
#> Running at Laplace(0.00290496) None  Grid-point #7 at 237000 	Fold #1 Rep #1 pred log like = 49.2023
#> Running at Laplace(0.00290496) None  Grid-point #7 at 237000 	Fold #2 Rep #1 pred log like = 46.2629
#> Running at Laplace(0.00290496) None  Grid-point #7 at 237000 	Fold #3 Rep #1 pred log like = 45.9125
#> Running at Laplace(0.00290496) None  Grid-point #7 at 237000 	Fold #4 Rep #1 pred log like = 82.6067
#> Running at Laplace(0.00290496) None  Grid-point #7 at 237000 	Fold #5 Rep #1 pred log like = 65.2571
#> Running at Laplace(0.00290496) None  Grid-point #7 at 237000 	Fold #6 Rep #1 pred log like = 57.2781
#> Running at Laplace(0.00290496) None  Grid-point #7 at 237000 	Fold #7 Rep #1 pred log like = 51.1933
#> Running at Laplace(0.00290496) None  Grid-point #7 at 237000 	Fold #8 Rep #1 pred log like = 104.578
#> Running at Laplace(0.00290496) None  Grid-point #7 at 237000 	Fold #9 Rep #1 pred log like = 65.1397
#> Running at Laplace(0.00290496) None  Grid-point #7 at 237000 	Fold #10 Rep #1 pred log like = 56.0439
#> AvgPred = 62.3474 with stdev = 17.6113
#> Completed at 237000
#> Next point at 2.37e+06 with value 0 and continue = 1
#> search[ 0.237 ] = 62.3445(17.5833)
#> search[ 2.37 ] = 62.347(17.6025)
#> search[ 23.7 ] = 62.3474(17.6085)
#> search[ 237 ] = 62.3474(17.6104)
#> search[ 2370 ] = 62.3474(17.611)
#> search[ 23700 ] = 62.3474(17.6112)
#> search[ 237000 ] = 62.3474(17.6113)
#> 
#> Running at Laplace(0.00091863) None  Grid-point #8 at 2.37e+06 	Fold #1 Rep #1 pred log like = 49.2023
#> Running at Laplace(0.00091863) None  Grid-point #8 at 2.37e+06 	Fold #2 Rep #1 pred log like = 46.2629
#> Running at Laplace(0.00091863) None  Grid-point #8 at 2.37e+06 	Fold #3 Rep #1 pred log like = 45.9124
#> Running at Laplace(0.00091863) None  Grid-point #8 at 2.37e+06 	Fold #4 Rep #1 pred log like = 82.6066
#> Running at Laplace(0.00091863) None  Grid-point #8 at 2.37e+06 	Fold #5 Rep #1 pred log like = 65.2571
#> Running at Laplace(0.00091863) None  Grid-point #8 at 2.37e+06 	Fold #6 Rep #1 pred log like = 57.2781
#> Running at Laplace(0.00091863) None  Grid-point #8 at 2.37e+06 	Fold #7 Rep #1 pred log like = 51.1934
#> Running at Laplace(0.00091863) None  Grid-point #8 at 2.37e+06 	Fold #8 Rep #1 pred log like = 104.578
#> Running at Laplace(0.00091863) None  Grid-point #8 at 2.37e+06 	Fold #9 Rep #1 pred log like = 65.1397
#> Running at Laplace(0.00091863) None  Grid-point #8 at 2.37e+06 	Fold #10 Rep #1 pred log like = 56.0439
#> AvgPred = 62.3474 with stdev = 17.6113
#> Completed at 2.37e+06
#> Next point at 2.37e+07 with value 0 and continue = 1
#> search[ 0.237 ] = 62.3445(17.5833)
#> search[ 2.37 ] = 62.347(17.6025)
#> search[ 23.7 ] = 62.3474(17.6085)
#> search[ 237 ] = 62.3474(17.6104)
#> search[ 2370 ] = 62.3474(17.611)
#> search[ 23700 ] = 62.3474(17.6112)
#> search[ 237000 ] = 62.3474(17.6113)
#> search[ 2.37e+06 ] = 62.3474(17.6113)
#> 
#> Running at Laplace(0.000290496) None  Grid-point #9 at 2.37e+07 	Fold #1 Rep #1 pred log like = 49.2023
#> Running at Laplace(0.000290496) None  Grid-point #9 at 2.37e+07 	Fold #2 Rep #1 pred log like = 46.2629
#> Running at Laplace(0.000290496) None  Grid-point #9 at 2.37e+07 	Fold #3 Rep #1 pred log like = 45.9124
#> Running at Laplace(0.000290496) None  Grid-point #9 at 2.37e+07 	Fold #4 Rep #1 pred log like = 82.6066
#> Running at Laplace(0.000290496) None  Grid-point #9 at 2.37e+07 	Fold #5 Rep #1 pred log like = 65.2571
#> Running at Laplace(0.000290496) None  Grid-point #9 at 2.37e+07 	Fold #6 Rep #1 pred log like = 57.278
#> Running at Laplace(0.000290496) None  Grid-point #9 at 2.37e+07 	Fold #7 Rep #1 pred log like = 51.1934
#> Running at Laplace(0.000290496) None  Grid-point #9 at 2.37e+07 	Fold #8 Rep #1 pred log like = 104.578
#> Running at Laplace(0.000290496) None  Grid-point #9 at 2.37e+07 	Fold #9 Rep #1 pred log like = 65.1398
#> Running at Laplace(0.000290496) None  Grid-point #9 at 2.37e+07 	Fold #10 Rep #1 pred log like = 56.044
#> AvgPred = 62.3474 with stdev = 17.6113
#> Completed at 2.37e+07
#> Next point at 2.37e+08 with value 0 and continue = 1
#> search[ 0.237 ] = 62.3445(17.5833)
#> search[ 2.37 ] = 62.347(17.6025)
#> search[ 23.7 ] = 62.3474(17.6085)
#> search[ 237 ] = 62.3474(17.6104)
#> search[ 2370 ] = 62.3474(17.611)
#> search[ 23700 ] = 62.3474(17.6112)
#> search[ 237000 ] = 62.3474(17.6113)
#> search[ 2.37e+06 ] = 62.3474(17.6113)
#> search[ 2.37e+07 ] = 62.3474(17.6113)
#> 
#> Running at Laplace(9.1863e-05) None  Grid-point #10 at 2.37e+08 	Fold #1 Rep #1 pred log like = 49.2023
#> Running at Laplace(9.1863e-05) None  Grid-point #10 at 2.37e+08 	Fold #2 Rep #1 pred log like = 46.2629
#> Running at Laplace(9.1863e-05) None  Grid-point #10 at 2.37e+08 	Fold #3 Rep #1 pred log like = 45.9124
#> Running at Laplace(9.1863e-05) None  Grid-point #10 at 2.37e+08 	Fold #4 Rep #1 pred log like = 82.6066
#> Running at Laplace(9.1863e-05) None  Grid-point #10 at 2.37e+08 	Fold #5 Rep #1 pred log like = 65.2571
#> Running at Laplace(9.1863e-05) None  Grid-point #10 at 2.37e+08 	Fold #6 Rep #1 pred log like = 57.278
#> Running at Laplace(9.1863e-05) None  Grid-point #10 at 2.37e+08 	Fold #7 Rep #1 pred log like = 51.1934
#> Running at Laplace(9.1863e-05) None  Grid-point #10 at 2.37e+08 	Fold #8 Rep #1 pred log like = 104.578
#> Running at Laplace(9.1863e-05) None  Grid-point #10 at 2.37e+08 	Fold #9 Rep #1 pred log like = 65.1398
#> Running at Laplace(9.1863e-05) None  Grid-point #10 at 2.37e+08 	Fold #10 Rep #1 pred log like = 56.044
#> AvgPred = 62.3474 with stdev = 17.6113
#> Completed at 2.37e+08
#> Next point at 2.37e+09 with value 0 and continue = 1
#> search[ 0.237 ] = 62.3445(17.5833)
#> search[ 2.37 ] = 62.347(17.6025)
#> search[ 23.7 ] = 62.3474(17.6085)
#> search[ 237 ] = 62.3474(17.6104)
#> search[ 2370 ] = 62.3474(17.611)
#> search[ 23700 ] = 62.3474(17.6112)
#> search[ 237000 ] = 62.3474(17.6113)
#> search[ 2.37e+06 ] = 62.3474(17.6113)
#> search[ 2.37e+07 ] = 62.3474(17.6113)
#> search[ 2.37e+08 ] = 62.3474(17.6113)
#> 
#> Running at Laplace(2.90496e-05) None  Grid-point #11 at 2.37e+09 	Fold #1 Rep #1 pred log like = 49.2023
#> Running at Laplace(2.90496e-05) None  Grid-point #11 at 2.37e+09 	Fold #2 Rep #1 pred log like = 46.2629
#> Running at Laplace(2.90496e-05) None  Grid-point #11 at 2.37e+09 	Fold #3 Rep #1 pred log like = 45.9124
#> Running at Laplace(2.90496e-05) None  Grid-point #11 at 2.37e+09 	Fold #4 Rep #1 pred log like = 82.6066
#> Running at Laplace(2.90496e-05) None  Grid-point #11 at 2.37e+09 	Fold #5 Rep #1 pred log like = 65.2571
#> Running at Laplace(2.90496e-05) None  Grid-point #11 at 2.37e+09 	Fold #6 Rep #1 pred log like = 57.278
#> Running at Laplace(2.90496e-05) None  Grid-point #11 at 2.37e+09 	Fold #7 Rep #1 pred log like = 51.1934
#> Running at Laplace(2.90496e-05) None  Grid-point #11 at 2.37e+09 	Fold #8 Rep #1 pred log like = 104.578
#> Running at Laplace(2.90496e-05) None  Grid-point #11 at 2.37e+09 	Fold #9 Rep #1 pred log like = 65.1398
#> Running at Laplace(2.90496e-05) None  Grid-point #11 at 2.37e+09 	Fold #10 Rep #1 pred log like = 56.044
#> AvgPred = 62.3474 with stdev = 17.6113
#> Completed at 2.37e+09
#> Next point at 2.37e+10 with value 0 and continue = 1
#> search[ 0.237 ] = 62.3445(17.5833)
#> search[ 2.37 ] = 62.347(17.6025)
#> search[ 23.7 ] = 62.3474(17.6085)
#> search[ 237 ] = 62.3474(17.6104)
#> search[ 2370 ] = 62.3474(17.611)
#> search[ 23700 ] = 62.3474(17.6112)
#> search[ 237000 ] = 62.3474(17.6113)
#> search[ 2.37e+06 ] = 62.3474(17.6113)
#> search[ 2.37e+07 ] = 62.3474(17.6113)
#> search[ 2.37e+08 ] = 62.3474(17.6113)
#> search[ 2.37e+09 ] = 62.3474(17.6113)
#> 
#> Running at Laplace(9.1863e-06) None  Grid-point #12 at 2.37e+10 	Fold #1 Rep #1 pred log like = 49.2023
#> Running at Laplace(9.1863e-06) None  Grid-point #12 at 2.37e+10 	Fold #2 Rep #1 pred log like = 46.2629
#> Running at Laplace(9.1863e-06) None  Grid-point #12 at 2.37e+10 	Fold #3 Rep #1 pred log like = 45.9124
#> Running at Laplace(9.1863e-06) None  Grid-point #12 at 2.37e+10 	Fold #4 Rep #1 pred log like = 82.6066
#> Running at Laplace(9.1863e-06) None  Grid-point #12 at 2.37e+10 	Fold #5 Rep #1 pred log like = 65.2571
#> Running at Laplace(9.1863e-06) None  Grid-point #12 at 2.37e+10 	Fold #6 Rep #1 pred log like = 57.278
#> Running at Laplace(9.1863e-06) None  Grid-point #12 at 2.37e+10 	Fold #7 Rep #1 pred log like = 51.1934
#> Running at Laplace(9.1863e-06) None  Grid-point #12 at 2.37e+10 	Fold #8 Rep #1 pred log like = 104.578
#> Running at Laplace(9.1863e-06) None  Grid-point #12 at 2.37e+10 	Fold #9 Rep #1 pred log like = 65.1398
#> Running at Laplace(9.1863e-06) None  Grid-point #12 at 2.37e+10 	Fold #10 Rep #1 pred log like = 56.044
#> AvgPred = 62.3474 with stdev = 17.6113
#> Completed at 2.37e+10
#> Next point at 2.37e+11 with value 0 and continue = 1
#> search[ 0.237 ] = 62.3445(17.5833)
#> search[ 2.37 ] = 62.347(17.6025)
#> search[ 23.7 ] = 62.3474(17.6085)
#> search[ 237 ] = 62.3474(17.6104)
#> search[ 2370 ] = 62.3474(17.611)
#> search[ 23700 ] = 62.3474(17.6112)
#> search[ 237000 ] = 62.3474(17.6113)
#> search[ 2.37e+06 ] = 62.3474(17.6113)
#> search[ 2.37e+07 ] = 62.3474(17.6113)
#> search[ 2.37e+08 ] = 62.3474(17.6113)
#> search[ 2.37e+09 ] = 62.3474(17.6113)
#> search[ 2.37e+10 ] = 62.3474(17.6113)
#> 
#> Running at Laplace(2.90496e-06) None  Grid-point #13 at 2.37e+11 	Fold #1 Rep #1 pred log like = 49.2023
#> Running at Laplace(2.90496e-06) None  Grid-point #13 at 2.37e+11 	Fold #2 Rep #1 pred log like = 46.2629
#> Running at Laplace(2.90496e-06) None  Grid-point #13 at 2.37e+11 	Fold #3 Rep #1 pred log like = 45.9124
#> Running at Laplace(2.90496e-06) None  Grid-point #13 at 2.37e+11 	Fold #4 Rep #1 pred log like = 82.6066
#> Running at Laplace(2.90496e-06) None  Grid-point #13 at 2.37e+11 	Fold #5 Rep #1 pred log like = 65.2571
#> Running at Laplace(2.90496e-06) None  Grid-point #13 at 2.37e+11 	Fold #6 Rep #1 pred log like = 57.278
#> Running at Laplace(2.90496e-06) None  Grid-point #13 at 2.37e+11 	Fold #7 Rep #1 pred log like = 51.1934
#> Running at Laplace(2.90496e-06) None  Grid-point #13 at 2.37e+11 	Fold #8 Rep #1 pred log like = 104.578
#> Running at Laplace(2.90496e-06) None  Grid-point #13 at 2.37e+11 	Fold #9 Rep #1 pred log like = 65.1398
#> Running at Laplace(2.90496e-06) None  Grid-point #13 at 2.37e+11 	Fold #10 Rep #1 pred log like = 56.044
#> AvgPred = 62.3474 with stdev = 17.6113
#> Completed at 2.37e+11
#> Next point at 2.37e+12 with value 0 and continue = 1
#> search[ 0.237 ] = 62.3445(17.5833)
#> search[ 2.37 ] = 62.347(17.6025)
#> search[ 23.7 ] = 62.3474(17.6085)
#> search[ 237 ] = 62.3474(17.6104)
#> search[ 2370 ] = 62.3474(17.611)
#> search[ 23700 ] = 62.3474(17.6112)
#> search[ 237000 ] = 62.3474(17.6113)
#> search[ 2.37e+06 ] = 62.3474(17.6113)
#> search[ 2.37e+07 ] = 62.3474(17.6113)
#> search[ 2.37e+08 ] = 62.3474(17.6113)
#> search[ 2.37e+09 ] = 62.3474(17.6113)
#> search[ 2.37e+10 ] = 62.3474(17.6113)
#> search[ 2.37e+11 ] = 62.3474(17.6113)
#> 
#> Running at Laplace(9.1863e-07) None  Grid-point #14 at 2.37e+12 	Fold #1 Rep #1 pred log like = 49.2023
#> Running at Laplace(9.1863e-07) None  Grid-point #14 at 2.37e+12 	Fold #2 Rep #1 pred log like = 46.2629
#> Running at Laplace(9.1863e-07) None  Grid-point #14 at 2.37e+12 	Fold #3 Rep #1 pred log like = 45.9124
#> Running at Laplace(9.1863e-07) None  Grid-point #14 at 2.37e+12 	Fold #4 Rep #1 pred log like = 82.6066
#> Running at Laplace(9.1863e-07) None  Grid-point #14 at 2.37e+12 	Fold #5 Rep #1 pred log like = 65.2571
#> Running at Laplace(9.1863e-07) None  Grid-point #14 at 2.37e+12 	Fold #6 Rep #1 pred log like = 57.278
#> Running at Laplace(9.1863e-07) None  Grid-point #14 at 2.37e+12 	Fold #7 Rep #1 pred log like = 51.1934
#> Running at Laplace(9.1863e-07) None  Grid-point #14 at 2.37e+12 	Fold #8 Rep #1 pred log like = 104.578
#> Running at Laplace(9.1863e-07) None  Grid-point #14 at 2.37e+12 	Fold #9 Rep #1 pred log like = 65.1398
#> Running at Laplace(9.1863e-07) None  Grid-point #14 at 2.37e+12 	Fold #10 Rep #1 pred log like = 56.044
#> AvgPred = 62.3474 with stdev = 17.6113
#> Completed at 2.37e+12
#> Next point at 2.37e+13 with value 0 and continue = 1
#> search[ 0.237 ] = 62.3445(17.5833)
#> search[ 2.37 ] = 62.347(17.6025)
#> search[ 23.7 ] = 62.3474(17.6085)
#> search[ 237 ] = 62.3474(17.6104)
#> search[ 2370 ] = 62.3474(17.611)
#> search[ 23700 ] = 62.3474(17.6112)
#> search[ 237000 ] = 62.3474(17.6113)
#> search[ 2.37e+06 ] = 62.3474(17.6113)
#> search[ 2.37e+07 ] = 62.3474(17.6113)
#> search[ 2.37e+08 ] = 62.3474(17.6113)
#> search[ 2.37e+09 ] = 62.3474(17.6113)
#> search[ 2.37e+10 ] = 62.3474(17.6113)
#> search[ 2.37e+11 ] = 62.3474(17.6113)
#> search[ 2.37e+12 ] = 62.3474(17.6113)
#> 
#> Running at Laplace(2.90496e-07) None  Grid-point #15 at 2.37e+13 	Fold #1 Rep #1 pred log like = 49.2023
#> Running at Laplace(2.90496e-07) None  Grid-point #15 at 2.37e+13 	Fold #2 Rep #1 pred log like = 46.2629
#> Running at Laplace(2.90496e-07) None  Grid-point #15 at 2.37e+13 	Fold #3 Rep #1 pred log like = 45.9124
#> Running at Laplace(2.90496e-07) None  Grid-point #15 at 2.37e+13 	Fold #4 Rep #1 pred log like = 82.6066
#> Running at Laplace(2.90496e-07) None  Grid-point #15 at 2.37e+13 	Fold #5 Rep #1 pred log like = 65.2571
#> Running at Laplace(2.90496e-07) None  Grid-point #15 at 2.37e+13 	Fold #6 Rep #1 pred log like = 57.278
#> Running at Laplace(2.90496e-07) None  Grid-point #15 at 2.37e+13 	Fold #7 Rep #1 pred log like = 51.1934
#> Running at Laplace(2.90496e-07) None  Grid-point #15 at 2.37e+13 	Fold #8 Rep #1 pred log like = 104.578
#> Running at Laplace(2.90496e-07) None  Grid-point #15 at 2.37e+13 	Fold #9 Rep #1 pred log like = 65.1398
#> Running at Laplace(2.90496e-07) None  Grid-point #15 at 2.37e+13 	Fold #10 Rep #1 pred log like = 56.044
#> AvgPred = 62.3474 with stdev = 17.6113
#> Completed at 2.37e+13
#> Next point at 2.37e+14 with value 0 and continue = 1
#> search[ 0.237 ] = 62.3445(17.5833)
#> search[ 2.37 ] = 62.347(17.6025)
#> search[ 23.7 ] = 62.3474(17.6085)
#> search[ 237 ] = 62.3474(17.6104)
#> search[ 2370 ] = 62.3474(17.611)
#> search[ 23700 ] = 62.3474(17.6112)
#> search[ 237000 ] = 62.3474(17.6113)
#> search[ 2.37e+06 ] = 62.3474(17.6113)
#> search[ 2.37e+07 ] = 62.3474(17.6113)
#> search[ 2.37e+08 ] = 62.3474(17.6113)
#> search[ 2.37e+09 ] = 62.3474(17.6113)
#> search[ 2.37e+10 ] = 62.3474(17.6113)
#> search[ 2.37e+11 ] = 62.3474(17.6113)
#> search[ 2.37e+12 ] = 62.3474(17.6113)
#> search[ 2.37e+13 ] = 62.3474(17.6113)
#> 
#> Running at Laplace(9.1863e-08) None  Grid-point #16 at 2.37e+14 	Fold #1 Rep #1 pred log like = 49.2023
#> Running at Laplace(9.1863e-08) None  Grid-point #16 at 2.37e+14 	Fold #2 Rep #1 pred log like = 46.2629
#> Running at Laplace(9.1863e-08) None  Grid-point #16 at 2.37e+14 	Fold #3 Rep #1 pred log like = 45.9124
#> Running at Laplace(9.1863e-08) None  Grid-point #16 at 2.37e+14 	Fold #4 Rep #1 pred log like = 82.6066
#> Running at Laplace(9.1863e-08) None  Grid-point #16 at 2.37e+14 	Fold #5 Rep #1 pred log like = 65.2571
#> Running at Laplace(9.1863e-08) None  Grid-point #16 at 2.37e+14 	Fold #6 Rep #1 pred log like = 57.278
#> Running at Laplace(9.1863e-08) None  Grid-point #16 at 2.37e+14 	Fold #7 Rep #1 pred log like = 51.1934
#> Running at Laplace(9.1863e-08) None  Grid-point #16 at 2.37e+14 	Fold #8 Rep #1 pred log like = 104.578
#> Running at Laplace(9.1863e-08) None  Grid-point #16 at 2.37e+14 	Fold #9 Rep #1 pred log like = 65.1398
#> Running at Laplace(9.1863e-08) None  Grid-point #16 at 2.37e+14 	Fold #10 Rep #1 pred log like = 56.044
#> AvgPred = 62.3474 with stdev = 17.6113
#> Completed at 2.37e+14
#> Next point at 2.37e+15 with value 0 and continue = 1
#> search[ 0.237 ] = 62.3445(17.5833)
#> search[ 2.37 ] = 62.347(17.6025)
#> search[ 23.7 ] = 62.3474(17.6085)
#> search[ 237 ] = 62.3474(17.6104)
#> search[ 2370 ] = 62.3474(17.611)
#> search[ 23700 ] = 62.3474(17.6112)
#> search[ 237000 ] = 62.3474(17.6113)
#> search[ 2.37e+06 ] = 62.3474(17.6113)
#> search[ 2.37e+07 ] = 62.3474(17.6113)
#> search[ 2.37e+08 ] = 62.3474(17.6113)
#> search[ 2.37e+09 ] = 62.3474(17.6113)
#> search[ 2.37e+10 ] = 62.3474(17.6113)
#> search[ 2.37e+11 ] = 62.3474(17.6113)
#> search[ 2.37e+12 ] = 62.3474(17.6113)
#> search[ 2.37e+13 ] = 62.3474(17.6113)
#> search[ 2.37e+14 ] = 62.3474(17.6113)
#> 
#> Running at Laplace(2.90496e-08) None  Grid-point #17 at 2.37e+15 	Fold #1 Rep #1 pred log like = 49.2023
#> Running at Laplace(2.90496e-08) None  Grid-point #17 at 2.37e+15 	Fold #2 Rep #1 pred log like = 46.2629
#> Running at Laplace(2.90496e-08) None  Grid-point #17 at 2.37e+15 	Fold #3 Rep #1 pred log like = 45.9124
#> Running at Laplace(2.90496e-08) None  Grid-point #17 at 2.37e+15 	Fold #4 Rep #1 pred log like = 82.6066
#> Running at Laplace(2.90496e-08) None  Grid-point #17 at 2.37e+15 	Fold #5 Rep #1 pred log like = 65.2571
#> Running at Laplace(2.90496e-08) None  Grid-point #17 at 2.37e+15 	Fold #6 Rep #1 pred log like = 57.278
#> Running at Laplace(2.90496e-08) None  Grid-point #17 at 2.37e+15 	Fold #7 Rep #1 pred log like = 51.1934
#> Running at Laplace(2.90496e-08) None  Grid-point #17 at 2.37e+15 	Fold #8 Rep #1 pred log like = 104.578
#> Running at Laplace(2.90496e-08) None  Grid-point #17 at 2.37e+15 	Fold #9 Rep #1 pred log like = 65.1398
#> Running at Laplace(2.90496e-08) None  Grid-point #17 at 2.37e+15 	Fold #10 Rep #1 pred log like = 56.044
#> AvgPred = 62.3474 with stdev = 17.6113
#> Completed at 2.37e+15
#> Next point at 2.37e+16 with value 0 and continue = 1
#> search[ 0.237 ] = 62.3445(17.5833)
#> search[ 2.37 ] = 62.347(17.6025)
#> search[ 23.7 ] = 62.3474(17.6085)
#> search[ 237 ] = 62.3474(17.6104)
#> search[ 2370 ] = 62.3474(17.611)
#> search[ 23700 ] = 62.3474(17.6112)
#> search[ 237000 ] = 62.3474(17.6113)
#> search[ 2.37e+06 ] = 62.3474(17.6113)
#> search[ 2.37e+07 ] = 62.3474(17.6113)
#> search[ 2.37e+08 ] = 62.3474(17.6113)
#> search[ 2.37e+09 ] = 62.3474(17.6113)
#> search[ 2.37e+10 ] = 62.3474(17.6113)
#> search[ 2.37e+11 ] = 62.3474(17.6113)
#> search[ 2.37e+12 ] = 62.3474(17.6113)
#> search[ 2.37e+13 ] = 62.3474(17.6113)
#> search[ 2.37e+14 ] = 62.3474(17.6113)
#> search[ 2.37e+15 ] = 62.3474(17.6113)
#> 
#> Running at Laplace(9.1863e-09) None  Grid-point #18 at 2.37e+16 	Fold #1 Rep #1 pred log like = 49.2023
#> Running at Laplace(9.1863e-09) None  Grid-point #18 at 2.37e+16 	Fold #2 Rep #1 pred log like = 46.2629
#> Running at Laplace(9.1863e-09) None  Grid-point #18 at 2.37e+16 	Fold #3 Rep #1 pred log like = 45.9124
#> Running at Laplace(9.1863e-09) None  Grid-point #18 at 2.37e+16 	Fold #4 Rep #1 pred log like = 82.6066
#> Running at Laplace(9.1863e-09) None  Grid-point #18 at 2.37e+16 	Fold #5 Rep #1 pred log like = 65.2571
#> Running at Laplace(9.1863e-09) None  Grid-point #18 at 2.37e+16 	Fold #6 Rep #1 pred log like = 57.278
#> Running at Laplace(9.1863e-09) None  Grid-point #18 at 2.37e+16 	Fold #7 Rep #1 pred log like = 51.1934
#> Running at Laplace(9.1863e-09) None  Grid-point #18 at 2.37e+16 	Fold #8 Rep #1 pred log like = 104.578
#> Running at Laplace(9.1863e-09) None  Grid-point #18 at 2.37e+16 	Fold #9 Rep #1 pred log like = 65.1398
#> Running at Laplace(9.1863e-09) None  Grid-point #18 at 2.37e+16 	Fold #10 Rep #1 pred log like = 56.044
#> AvgPred = 62.3474 with stdev = 17.6113
#> Completed at 2.37e+16
#> Next point at 2.37e+17 with value 0 and continue = 1
#> search[ 0.237 ] = 62.3445(17.5833)
#> search[ 2.37 ] = 62.347(17.6025)
#> search[ 23.7 ] = 62.3474(17.6085)
#> search[ 237 ] = 62.3474(17.6104)
#> search[ 2370 ] = 62.3474(17.611)
#> search[ 23700 ] = 62.3474(17.6112)
#> search[ 237000 ] = 62.3474(17.6113)
#> search[ 2.37e+06 ] = 62.3474(17.6113)
#> search[ 2.37e+07 ] = 62.3474(17.6113)
#> search[ 2.37e+08 ] = 62.3474(17.6113)
#> search[ 2.37e+09 ] = 62.3474(17.6113)
#> search[ 2.37e+10 ] = 62.3474(17.6113)
#> search[ 2.37e+11 ] = 62.3474(17.6113)
#> search[ 2.37e+12 ] = 62.3474(17.6113)
#> search[ 2.37e+13 ] = 62.3474(17.6113)
#> search[ 2.37e+14 ] = 62.3474(17.6113)
#> search[ 2.37e+15 ] = 62.3474(17.6113)
#> search[ 2.37e+16 ] = 62.3474(17.6113)
#> 
#> Running at Laplace(2.90496e-09) None  Grid-point #19 at 2.37e+17 	Fold #1 Rep #1 pred log like = 49.2023
#> Running at Laplace(2.90496e-09) None  Grid-point #19 at 2.37e+17 	Fold #2 Rep #1 pred log like = 46.2629
#> Running at Laplace(2.90496e-09) None  Grid-point #19 at 2.37e+17 	Fold #3 Rep #1 pred log like = 45.9124
#> Running at Laplace(2.90496e-09) None  Grid-point #19 at 2.37e+17 	Fold #4 Rep #1 pred log like = 82.6066
#> Running at Laplace(2.90496e-09) None  Grid-point #19 at 2.37e+17 	Fold #5 Rep #1 pred log like = 65.2571
#> Running at Laplace(2.90496e-09) None  Grid-point #19 at 2.37e+17 	Fold #6 Rep #1 pred log like = 57.278
#> Running at Laplace(2.90496e-09) None  Grid-point #19 at 2.37e+17 	Fold #7 Rep #1 pred log like = 51.1934
#> Running at Laplace(2.90496e-09) None  Grid-point #19 at 2.37e+17 	Fold #8 Rep #1 pred log like = 104.578
#> Running at Laplace(2.90496e-09) None  Grid-point #19 at 2.37e+17 	Fold #9 Rep #1 pred log like = 65.1398
#> Running at Laplace(2.90496e-09) None  Grid-point #19 at 2.37e+17 	Fold #10 Rep #1 pred log like = 56.044
#> AvgPred = 62.3474 with stdev = 17.6113
#> Completed at 2.37e+17
#> Next point at 2.37e+18 with value 0 and continue = 1
#> search[ 0.237 ] = 62.3445(17.5833)
#> search[ 2.37 ] = 62.347(17.6025)
#> search[ 23.7 ] = 62.3474(17.6085)
#> search[ 237 ] = 62.3474(17.6104)
#> search[ 2370 ] = 62.3474(17.611)
#> search[ 23700 ] = 62.3474(17.6112)
#> search[ 237000 ] = 62.3474(17.6113)
#> search[ 2.37e+06 ] = 62.3474(17.6113)
#> search[ 2.37e+07 ] = 62.3474(17.6113)
#> search[ 2.37e+08 ] = 62.3474(17.6113)
#> search[ 2.37e+09 ] = 62.3474(17.6113)
#> search[ 2.37e+10 ] = 62.3474(17.6113)
#> search[ 2.37e+11 ] = 62.3474(17.6113)
#> search[ 2.37e+12 ] = 62.3474(17.6113)
#> search[ 2.37e+13 ] = 62.3474(17.6113)
#> search[ 2.37e+14 ] = 62.3474(17.6113)
#> search[ 2.37e+15 ] = 62.3474(17.6113)
#> search[ 2.37e+16 ] = 62.3474(17.6113)
#> search[ 2.37e+17 ] = 62.3474(17.6113)
#> 
#> Running at Laplace(9.1863e-10) None  Grid-point #20 at 2.37e+18 	Fold #1 Rep #1 pred log like = 49.2023
#> Running at Laplace(9.1863e-10) None  Grid-point #20 at 2.37e+18 	Fold #2 Rep #1 pred log like = 46.2629
#> Running at Laplace(9.1863e-10) None  Grid-point #20 at 2.37e+18 	Fold #3 Rep #1 pred log like = 45.9124
#> Running at Laplace(9.1863e-10) None  Grid-point #20 at 2.37e+18 	Fold #4 Rep #1 pred log like = 82.6066
#> Running at Laplace(9.1863e-10) None  Grid-point #20 at 2.37e+18 	Fold #5 Rep #1 pred log like = 65.2571
#> Running at Laplace(9.1863e-10) None  Grid-point #20 at 2.37e+18 	Fold #6 Rep #1 pred log like = 57.278
#> Running at Laplace(9.1863e-10) None  Grid-point #20 at 2.37e+18 	Fold #7 Rep #1 pred log like = 51.1934
#> Running at Laplace(9.1863e-10) None  Grid-point #20 at 2.37e+18 	Fold #8 Rep #1 pred log like = 104.578
#> Running at Laplace(9.1863e-10) None  Grid-point #20 at 2.37e+18 	Fold #9 Rep #1 pred log like = 65.1398
#> Running at Laplace(9.1863e-10) None  Grid-point #20 at 2.37e+18 	Fold #10 Rep #1 pred log like = 56.044
#> AvgPred = 62.3474 with stdev = 17.6113
#> Completed at 2.37e+18
#> Next point at 2.37e+19 with value 0 and continue = 1
#> search[ 0.237 ] = 62.3445(17.5833)
#> search[ 2.37 ] = 62.347(17.6025)
#> search[ 23.7 ] = 62.3474(17.6085)
#> search[ 237 ] = 62.3474(17.6104)
#> search[ 2370 ] = 62.3474(17.611)
#> search[ 23700 ] = 62.3474(17.6112)
#> search[ 237000 ] = 62.3474(17.6113)
#> search[ 2.37e+06 ] = 62.3474(17.6113)
#> search[ 2.37e+07 ] = 62.3474(17.6113)
#> search[ 2.37e+08 ] = 62.3474(17.6113)
#> search[ 2.37e+09 ] = 62.3474(17.6113)
#> search[ 2.37e+10 ] = 62.3474(17.6113)
#> search[ 2.37e+11 ] = 62.3474(17.6113)
#> search[ 2.37e+12 ] = 62.3474(17.6113)
#> search[ 2.37e+13 ] = 62.3474(17.6113)
#> search[ 2.37e+14 ] = 62.3474(17.6113)
#> search[ 2.37e+15 ] = 62.3474(17.6113)
#> search[ 2.37e+16 ] = 62.3474(17.6113)
#> search[ 2.37e+17 ] = 62.3474(17.6113)
#> search[ 2.37e+18 ] = 62.3474(17.6113)
#> 
#> Running at Laplace(2.90496e-10) None  Grid-point #21 at 2.37e+19 	Fold #1 Rep #1 pred log like = 49.2023
#> Running at Laplace(2.90496e-10) None  Grid-point #21 at 2.37e+19 	Fold #2 Rep #1 pred log like = 46.2629
#> Running at Laplace(2.90496e-10) None  Grid-point #21 at 2.37e+19 	Fold #3 Rep #1 pred log like = 45.9124
#> Running at Laplace(2.90496e-10) None  Grid-point #21 at 2.37e+19 	Fold #4 Rep #1 pred log like = 82.6066
#> Running at Laplace(2.90496e-10) None  Grid-point #21 at 2.37e+19 	Fold #5 Rep #1 pred log like = 65.2571
#> Running at Laplace(2.90496e-10) None  Grid-point #21 at 2.37e+19 	Fold #6 Rep #1 pred log like = 57.278
#> Running at Laplace(2.90496e-10) None  Grid-point #21 at 2.37e+19 	Fold #7 Rep #1 pred log like = 51.1934
#> Running at Laplace(2.90496e-10) None  Grid-point #21 at 2.37e+19 	Fold #8 Rep #1 pred log like = 104.578
#> Running at Laplace(2.90496e-10) None  Grid-point #21 at 2.37e+19 	Fold #9 Rep #1 pred log like = 65.1398
#> Running at Laplace(2.90496e-10) None  Grid-point #21 at 2.37e+19 	Fold #10 Rep #1 pred log like = 56.044
#> AvgPred = 62.3474 with stdev = 17.6113
#> Completed at 2.37e+19
#> Next point at 2.37e+20 with value 0 and continue = 1
#> search[ 0.237 ] = 62.3445(17.5833)
#> search[ 2.37 ] = 62.347(17.6025)
#> search[ 23.7 ] = 62.3474(17.6085)
#> search[ 237 ] = 62.3474(17.6104)
#> search[ 2370 ] = 62.3474(17.611)
#> search[ 23700 ] = 62.3474(17.6112)
#> search[ 237000 ] = 62.3474(17.6113)
#> search[ 2.37e+06 ] = 62.3474(17.6113)
#> search[ 2.37e+07 ] = 62.3474(17.6113)
#> search[ 2.37e+08 ] = 62.3474(17.6113)
#> search[ 2.37e+09 ] = 62.3474(17.6113)
#> search[ 2.37e+10 ] = 62.3474(17.6113)
#> search[ 2.37e+11 ] = 62.3474(17.6113)
#> search[ 2.37e+12 ] = 62.3474(17.6113)
#> search[ 2.37e+13 ] = 62.3474(17.6113)
#> search[ 2.37e+14 ] = 62.3474(17.6113)
#> search[ 2.37e+15 ] = 62.3474(17.6113)
#> search[ 2.37e+16 ] = 62.3474(17.6113)
#> search[ 2.37e+17 ] = 62.3474(17.6113)
#> search[ 2.37e+18 ] = 62.3474(17.6113)
#> search[ 2.37e+19 ] = 62.3474(17.6113)
#> 
#> Running at Laplace(9.1863e-11) None  Grid-point #22 at 2.37e+20 	Fold #1 Rep #1 pred log like = 49.2023
#> Running at Laplace(9.1863e-11) None  Grid-point #22 at 2.37e+20 	Fold #2 Rep #1 pred log like = 46.2629
#> Running at Laplace(9.1863e-11) None  Grid-point #22 at 2.37e+20 	Fold #3 Rep #1 pred log like = 45.9124
#> Running at Laplace(9.1863e-11) None  Grid-point #22 at 2.37e+20 	Fold #4 Rep #1 pred log like = 82.6066
#> Running at Laplace(9.1863e-11) None  Grid-point #22 at 2.37e+20 	Fold #5 Rep #1 pred log like = 65.2571
#> Running at Laplace(9.1863e-11) None  Grid-point #22 at 2.37e+20 	Fold #6 Rep #1 pred log like = 57.278
#> Running at Laplace(9.1863e-11) None  Grid-point #22 at 2.37e+20 	Fold #7 Rep #1 pred log like = 51.1934
#> Running at Laplace(9.1863e-11) None  Grid-point #22 at 2.37e+20 	Fold #8 Rep #1 pred log like = 104.578
#> Running at Laplace(9.1863e-11) None  Grid-point #22 at 2.37e+20 	Fold #9 Rep #1 pred log like = 65.1398
#> Running at Laplace(9.1863e-11) None  Grid-point #22 at 2.37e+20 	Fold #10 Rep #1 pred log like = 56.044
#> AvgPred = 62.3474 with stdev = 17.6113
#> Completed at 2.37e+20
#> Next point at 2.37e+21 with value 0 and continue = 1
#> search[ 0.237 ] = 62.3445(17.5833)
#> search[ 2.37 ] = 62.347(17.6025)
#> search[ 23.7 ] = 62.3474(17.6085)
#> search[ 237 ] = 62.3474(17.6104)
#> search[ 2370 ] = 62.3474(17.611)
#> search[ 23700 ] = 62.3474(17.6112)
#> search[ 237000 ] = 62.3474(17.6113)
#> search[ 2.37e+06 ] = 62.3474(17.6113)
#> search[ 2.37e+07 ] = 62.3474(17.6113)
#> search[ 2.37e+08 ] = 62.3474(17.6113)
#> search[ 2.37e+09 ] = 62.3474(17.6113)
#> search[ 2.37e+10 ] = 62.3474(17.6113)
#> search[ 2.37e+11 ] = 62.3474(17.6113)
#> search[ 2.37e+12 ] = 62.3474(17.6113)
#> search[ 2.37e+13 ] = 62.3474(17.6113)
#> search[ 2.37e+14 ] = 62.3474(17.6113)
#> search[ 2.37e+15 ] = 62.3474(17.6113)
#> search[ 2.37e+16 ] = 62.3474(17.6113)
#> search[ 2.37e+17 ] = 62.3474(17.6113)
#> search[ 2.37e+18 ] = 62.3474(17.6113)
#> search[ 2.37e+19 ] = 62.3474(17.6113)
#> search[ 2.37e+20 ] = 62.3474(17.6113)
#> 
#> Running at Laplace(2.90496e-11) None  Grid-point #23 at 2.37e+21 	Fold #1 Rep #1 pred log like = 49.2023
#> Running at Laplace(2.90496e-11) None  Grid-point #23 at 2.37e+21 	Fold #2 Rep #1 pred log like = 46.2629
#> Running at Laplace(2.90496e-11) None  Grid-point #23 at 2.37e+21 	Fold #3 Rep #1 pred log like = 45.9124
#> Running at Laplace(2.90496e-11) None  Grid-point #23 at 2.37e+21 	Fold #4 Rep #1 pred log like = 82.6066
#> Running at Laplace(2.90496e-11) None  Grid-point #23 at 2.37e+21 	Fold #5 Rep #1 pred log like = 65.2571
#> Running at Laplace(2.90496e-11) None  Grid-point #23 at 2.37e+21 	Fold #6 Rep #1 pred log like = 57.278
#> Running at Laplace(2.90496e-11) None  Grid-point #23 at 2.37e+21 	Fold #7 Rep #1 pred log like = 51.1934
#> Running at Laplace(2.90496e-11) None  Grid-point #23 at 2.37e+21 	Fold #8 Rep #1 pred log like = 104.578
#> Running at Laplace(2.90496e-11) None  Grid-point #23 at 2.37e+21 	Fold #9 Rep #1 pred log like = 65.1398
#> Running at Laplace(2.90496e-11) None  Grid-point #23 at 2.37e+21 	Fold #10 Rep #1 pred log like = 56.044
#> AvgPred = 62.3474 with stdev = 17.6113
#> Completed at 2.37e+21
#> Next point at 2.37e+22 with value 0 and continue = 1
#> search[ 0.237 ] = 62.3445(17.5833)
#> search[ 2.37 ] = 62.347(17.6025)
#> search[ 23.7 ] = 62.3474(17.6085)
#> search[ 237 ] = 62.3474(17.6104)
#> search[ 2370 ] = 62.3474(17.611)
#> search[ 23700 ] = 62.3474(17.6112)
#> search[ 237000 ] = 62.3474(17.6113)
#> search[ 2.37e+06 ] = 62.3474(17.6113)
#> search[ 2.37e+07 ] = 62.3474(17.6113)
#> search[ 2.37e+08 ] = 62.3474(17.6113)
#> search[ 2.37e+09 ] = 62.3474(17.6113)
#> search[ 2.37e+10 ] = 62.3474(17.6113)
#> search[ 2.37e+11 ] = 62.3474(17.6113)
#> search[ 2.37e+12 ] = 62.3474(17.6113)
#> search[ 2.37e+13 ] = 62.3474(17.6113)
#> search[ 2.37e+14 ] = 62.3474(17.6113)
#> search[ 2.37e+15 ] = 62.3474(17.6113)
#> search[ 2.37e+16 ] = 62.3474(17.6113)
#> search[ 2.37e+17 ] = 62.3474(17.6113)
#> search[ 2.37e+18 ] = 62.3474(17.6113)
#> search[ 2.37e+19 ] = 62.3474(17.6113)
#> search[ 2.37e+20 ] = 62.3474(17.6113)
#> search[ 2.37e+21 ] = 62.3474(17.6113)
#> 
#> Running at Laplace(9.1863e-12) None  Grid-point #24 at 2.37e+22 	Fold #1 Rep #1 pred log like = 49.2023
#> Running at Laplace(9.1863e-12) None  Grid-point #24 at 2.37e+22 	Fold #2 Rep #1 pred log like = 46.2629
#> Running at Laplace(9.1863e-12) None  Grid-point #24 at 2.37e+22 	Fold #3 Rep #1 pred log like = 45.9124
#> Running at Laplace(9.1863e-12) None  Grid-point #24 at 2.37e+22 	Fold #4 Rep #1 pred log like = 82.6066
#> Running at Laplace(9.1863e-12) None  Grid-point #24 at 2.37e+22 	Fold #5 Rep #1 pred log like = 65.2571
#> Running at Laplace(9.1863e-12) None  Grid-point #24 at 2.37e+22 	Fold #6 Rep #1 pred log like = 57.278
#> Running at Laplace(9.1863e-12) None  Grid-point #24 at 2.37e+22 	Fold #7 Rep #1 pred log like = 51.1934
#> Running at Laplace(9.1863e-12) None  Grid-point #24 at 2.37e+22 	Fold #8 Rep #1 pred log like = 104.578
#> Running at Laplace(9.1863e-12) None  Grid-point #24 at 2.37e+22 	Fold #9 Rep #1 pred log like = 65.1398
#> Running at Laplace(9.1863e-12) None  Grid-point #24 at 2.37e+22 	Fold #10 Rep #1 pred log like = 56.044
#> AvgPred = 62.3474 with stdev = 17.6113
#> Completed at 2.37e+22
#> Next point at 2.37e+23 with value 0 and continue = 1
#> search[ 0.237 ] = 62.3445(17.5833)
#> search[ 2.37 ] = 62.347(17.6025)
#> search[ 23.7 ] = 62.3474(17.6085)
#> search[ 237 ] = 62.3474(17.6104)
#> search[ 2370 ] = 62.3474(17.611)
#> search[ 23700 ] = 62.3474(17.6112)
#> search[ 237000 ] = 62.3474(17.6113)
#> search[ 2.37e+06 ] = 62.3474(17.6113)
#> search[ 2.37e+07 ] = 62.3474(17.6113)
#> search[ 2.37e+08 ] = 62.3474(17.6113)
#> search[ 2.37e+09 ] = 62.3474(17.6113)
#> search[ 2.37e+10 ] = 62.3474(17.6113)
#> search[ 2.37e+11 ] = 62.3474(17.6113)
#> search[ 2.37e+12 ] = 62.3474(17.6113)
#> search[ 2.37e+13 ] = 62.3474(17.6113)
#> search[ 2.37e+14 ] = 62.3474(17.6113)
#> search[ 2.37e+15 ] = 62.3474(17.6113)
#> search[ 2.37e+16 ] = 62.3474(17.6113)
#> search[ 2.37e+17 ] = 62.3474(17.6113)
#> search[ 2.37e+18 ] = 62.3474(17.6113)
#> search[ 2.37e+19 ] = 62.3474(17.6113)
#> search[ 2.37e+20 ] = 62.3474(17.6113)
#> search[ 2.37e+21 ] = 62.3474(17.6113)
#> search[ 2.37e+22 ] = 62.3474(17.6113)
#> 
#> Running at Laplace(2.90496e-12) None  Grid-point #25 at 2.37e+23 	Fold #1 Rep #1 pred log like = 49.2023
#> Running at Laplace(2.90496e-12) None  Grid-point #25 at 2.37e+23 	Fold #2 Rep #1 pred log like = 46.2629
#> Running at Laplace(2.90496e-12) None  Grid-point #25 at 2.37e+23 	Fold #3 Rep #1 pred log like = 45.9124
#> Running at Laplace(2.90496e-12) None  Grid-point #25 at 2.37e+23 	Fold #4 Rep #1 pred log like = 82.6066
#> Running at Laplace(2.90496e-12) None  Grid-point #25 at 2.37e+23 	Fold #5 Rep #1 pred log like = 65.2571
#> Running at Laplace(2.90496e-12) None  Grid-point #25 at 2.37e+23 	Fold #6 Rep #1 pred log like = 57.278
#> Running at Laplace(2.90496e-12) None  Grid-point #25 at 2.37e+23 	Fold #7 Rep #1 pred log like = 51.1934
#> Running at Laplace(2.90496e-12) None  Grid-point #25 at 2.37e+23 	Fold #8 Rep #1 pred log like = 104.578
#> Running at Laplace(2.90496e-12) None  Grid-point #25 at 2.37e+23 	Fold #9 Rep #1 pred log like = 65.1398
#> Running at Laplace(2.90496e-12) None  Grid-point #25 at 2.37e+23 	Fold #10 Rep #1 pred log like = 56.044
#> AvgPred = 62.3474 with stdev = 17.6113
#> Completed at 2.37e+23
#> Next point at 1.4504e+14 with value 62.3476 and continue = 0
#> search[ 0.237 ] = 62.3445(17.5833)
#> search[ 2.37 ] = 62.347(17.6025)
#> search[ 23.7 ] = 62.3474(17.6085)
#> search[ 237 ] = 62.3474(17.6104)
#> search[ 2370 ] = 62.3474(17.611)
#> search[ 23700 ] = 62.3474(17.6112)
#> search[ 237000 ] = 62.3474(17.6113)
#> search[ 2.37e+06 ] = 62.3474(17.6113)
#> search[ 2.37e+07 ] = 62.3474(17.6113)
#> search[ 2.37e+08 ] = 62.3474(17.6113)
#> search[ 2.37e+09 ] = 62.3474(17.6113)
#> search[ 2.37e+10 ] = 62.3474(17.6113)
#> search[ 2.37e+11 ] = 62.3474(17.6113)
#> search[ 2.37e+12 ] = 62.3474(17.6113)
#> search[ 2.37e+13 ] = 62.3474(17.6113)
#> search[ 2.37e+14 ] = 62.3474(17.6113)
#> search[ 2.37e+15 ] = 62.3474(17.6113)
#> search[ 2.37e+16 ] = 62.3474(17.6113)
#> search[ 2.37e+17 ] = 62.3474(17.6113)
#> search[ 2.37e+18 ] = 62.3474(17.6113)
#> search[ 2.37e+19 ] = 62.3474(17.6113)
#> search[ 2.37e+20 ] = 62.3474(17.6113)
#> search[ 2.37e+21 ] = 62.3474(17.6113)
#> search[ 2.37e+22 ] = 62.3474(17.6113)
#> search[ 2.37e+23 ] = 62.3474(17.6113)
#> 
#> 
#> Maximum predicted log likelihood (62.3476) estimated at:
#> 	1.4504e+14 (variance)
#> 	1.17428e-07 (lambda)
#> 
#> Fitting model at optimal hyperparameter
#> Using prior: Laplace(1.17428e-07) None 

#Find out what the optimal hyperparameter was:
getHyperParameter(fit)

#> [1] 1.450402e+14

#Extract the current log-likelihood, and coefficients
logLik(fit)

#> 'log Lik.' -1691.248 (df=3)

coef(fit)

#> (Intercept)           1           2 
#> -4.39464620 -0.04762332 -0.55968738 

#We can only retrieve the confidence interval for unregularized coefficients:
confint(fit, c(0))

#> Using 1 thread(s)

#>      covariate     2.5 %    97.5 % evaluations
#> [1,]         0 -4.439161 -4.350773          20