Extract log-likelihood

logLik returns the current log-likelihood of the fit in a Cyclops model fit object

# S3 method for cyclopsFit
logLik(object, ...)

Arguments

object
A Cyclops model fit object

...
Additional arguments

Examples

#Generate some simulated data:
sim <- simulateCyclopsData(nstrata = 1, nrows = 1000, ncovars = 2, eCovarsPerRow = 0.5, 
                           model = "poisson")

#> Sparseness = 76.3 %

cyclopsData <- convertToCyclopsData(sim$outcomes, sim$covariates, modelType = "pr", 
                                    addIntercept = TRUE)

#> Sorting covariates by covariateId and rowId

#Define the prior and control objects to use cross-validation for finding the 
#optimal hyperparameter:
prior <- createPrior("laplace", exclude = 0, useCrossValidation = TRUE)
control <- createControl(cvType = "auto", noiseLevel = "quiet")

#Fit the model
fit <- fitCyclopsModel(cyclopsData,prior = prior, control = control)  

#> Using cross-validation selector type byRow
#> Performing 10-fold cross-validation [seed = 1623936646] with data partitions of sizes 100 100 100 100 100 100 100 100 100 100
#> Using 1 thread(s)
#> Starting var = 0.237 (default)
#> Running at Laplace(2.90496) None  Grid-point #1 at 0.237 	Fold #1 Rep #1 pred log like = 81.7154
#> Running at Laplace(2.90496) None  Grid-point #1 at 0.237 	Fold #2 Rep #1 pred log like = 84.3058
#> Running at Laplace(2.90496) None  Grid-point #1 at 0.237 	Fold #3 Rep #1 pred log like = 44.7229
#> Running at Laplace(2.90496) None  Grid-point #1 at 0.237 	Fold #4 Rep #1 pred log like = 67.5691
#> Running at Laplace(2.90496) None  Grid-point #1 at 0.237 	Fold #5 Rep #1 pred log like = 102.612
#> Running at Laplace(2.90496) None  Grid-point #1 at 0.237 	Fold #6 Rep #1 pred log like = -1.50097
#> Running at Laplace(2.90496) None  Grid-point #1 at 0.237 	Fold #7 Rep #1 pred log like = 92.0468
#> Running at Laplace(2.90496) None  Grid-point #1 at 0.237 	Fold #8 Rep #1 pred log like = 37.2992
#> Running at Laplace(2.90496) None  Grid-point #1 at 0.237 	Fold #9 Rep #1 pred log like = 83.6135
#> Running at Laplace(2.90496) None  Grid-point #1 at 0.237 	Fold #10 Rep #1 pred log like = 30.8871
#> AvgPred = 62.327 with stdev = 31.3759
#> Completed at 0.237
#> Next point at 2.37 with value 0 and continue = 1
#> search[ 0.237 ] = 62.327(31.3759)
#> 
#> Running at Laplace(0.91863) None  Grid-point #2 at 2.37 	Fold #1 Rep #1 pred log like = 81.7226
#> Running at Laplace(0.91863) None  Grid-point #2 at 2.37 	Fold #2 Rep #1 pred log like = 84.2611
#> Running at Laplace(0.91863) None  Grid-point #2 at 2.37 	Fold #3 Rep #1 pred log like = 44.7842
#> Running at Laplace(0.91863) None  Grid-point #2 at 2.37 	Fold #4 Rep #1 pred log like = 67.5706
#> Running at Laplace(0.91863) None  Grid-point #2 at 2.37 	Fold #5 Rep #1 pred log like = 102.538
#> Running at Laplace(0.91863) None  Grid-point #2 at 2.37 	Fold #6 Rep #1 pred log like = -1.49342
#> Running at Laplace(0.91863) None  Grid-point #2 at 2.37 	Fold #7 Rep #1 pred log like = 91.9862
#> Running at Laplace(0.91863) None  Grid-point #2 at 2.37 	Fold #8 Rep #1 pred log like = 37.3588
#> Running at Laplace(0.91863) None  Grid-point #2 at 2.37 	Fold #9 Rep #1 pred log like = 83.6881
#> Running at Laplace(0.91863) None  Grid-point #2 at 2.37 	Fold #10 Rep #1 pred log like = 30.8629
#> AvgPred = 62.3279 with stdev = 31.3558
#> Completed at 2.37
#> Next point at 23.7 with value 0 and continue = 1
#> search[ 0.237 ] = 62.327(31.3759)
#> search[ 2.37 ] = 62.3279(31.3558)
#> 
#> Running at Laplace(0.290496) None  Grid-point #3 at 23.7 	Fold #1 Rep #1 pred log like = 81.7245
#> Running at Laplace(0.290496) None  Grid-point #3 at 23.7 	Fold #2 Rep #1 pred log like = 84.2464
#> Running at Laplace(0.290496) None  Grid-point #3 at 23.7 	Fold #3 Rep #1 pred log like = 44.8033
#> Running at Laplace(0.290496) None  Grid-point #3 at 23.7 	Fold #4 Rep #1 pred log like = 67.5705
#> Running at Laplace(0.290496) None  Grid-point #3 at 23.7 	Fold #5 Rep #1 pred log like = 102.514
#> Running at Laplace(0.290496) None  Grid-point #3 at 23.7 	Fold #6 Rep #1 pred log like = -1.49139
#> Running at Laplace(0.290496) None  Grid-point #3 at 23.7 	Fold #7 Rep #1 pred log like = 91.9666
#> Running at Laplace(0.290496) None  Grid-point #3 at 23.7 	Fold #8 Rep #1 pred log like = 37.3773
#> Running at Laplace(0.290496) None  Grid-point #3 at 23.7 	Fold #9 Rep #1 pred log like = 83.7113
#> Running at Laplace(0.290496) None  Grid-point #3 at 23.7 	Fold #10 Rep #1 pred log like = 30.8547
#> AvgPred = 62.3278 with stdev = 31.3495
#> Completed at 23.7
#> Next point at 5.15795 with value 62.328 and continue = 0
#> search[ 0.237 ] = 62.327(31.3759)
#> search[ 2.37 ] = 62.3279(31.3558)
#> search[ 23.7 ] = 62.3278(31.3495)
#> 
#> 
#> Maximum predicted log likelihood (62.328) estimated at:
#> 	5.15795 (variance)
#> 	0.622697 (lambda)
#> 
#> Fitting model at optimal hyperparameter
#> Using prior: Laplace(0.622697) None 

#Find out what the optimal hyperparameter was:
getHyperParameter(fit)

#> [1] 5.157948

#Extract the current log-likelihood, and coefficients
logLik(fit)

#> 'log Lik.' -1691.249 (df=3)

coef(fit)

#> (Intercept)           1           2 
#> -4.39513986 -0.04641266 -0.55810190 

#We can only retrieve the confidence interval for unregularized coefficients:
confint(fit, c(0))

#> Using 1 thread(s)

#>      covariate     2.5 %    97.5 % evaluations
#> [1,]         0 -4.439663 -4.351258          20