Simulation Cyclops dataset

simulateCyclopsData generates a simulated large, sparse data set for use by fitCyclopsSimulation.

simulateCyclopsData(
  nstrata = 200,
  nrows = 10000,
  ncovars = 20,
  effectSizeSd = 1,
  zeroEffectSizeProp = 0.9,
  eCovarsPerRow = ncovars/100,
  model = "survival"
)

Arguments

nstrata
Numeric: Number of strata

nrows
Numeric: Number of observation rows

ncovars
Numeric: Number of covariates

effectSizeSd
Numeric: Standard derivation of the non-zero simulated regression coefficients

zeroEffectSizeProp
Numeric: Expected proportion of zero effect size

eCovarsPerRow
Number: Effective number of non-zero covariates per data row

model
String: Simulation model. Choices are: logistic, poisson or survival

Value

A simulated data set

Examples

#Generate some simulated data:
sim <- simulateCyclopsData(nstrata = 1, nrows = 1000, ncovars = 2, eCovarsPerRow = 0.5, 
                           model = "poisson")

#> Sparseness = 77.4 %

cyclopsData <- convertToCyclopsData(sim$outcomes, sim$covariates, modelType = "pr", 
                                    addIntercept = TRUE)

#> Sorting covariates by covariateId and rowId

#Define the prior and control objects to use cross-validation for finding the 
#optimal hyperparameter:
prior <- createPrior("laplace", exclude = 0, useCrossValidation = TRUE)
control <- createControl(cvType = "auto", noiseLevel = "quiet")

#Fit the model
fit <- fitCyclopsModel(cyclopsData,prior = prior, control = control)  

#> Using cross-validation selector type byRow
#> Performing 10-fold cross-validation [seed = 1623936647] with data partitions of sizes 100 100 100 100 100 100 100 100 100 100
#> Using 1 thread(s)
#> Starting var = 0.226 (default)
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #1 Rep #1 pred log like = 657.252
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #2 Rep #1 pred log like = 695.827
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #3 Rep #1 pred log like = 565.49
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #4 Rep #1 pred log like = 658.746
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #5 Rep #1 pred log like = 796.882
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #6 Rep #1 pred log like = 585.307
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #7 Rep #1 pred log like = 823.083
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #8 Rep #1 pred log like = 702.257
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #9 Rep #1 pred log like = 606.283
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #10 Rep #1 pred log like = 578.22
#> AvgPred = 666.935 with stdev = 84.8616
#> Completed at 0.226
#> Next point at 2.26 with value 0 and continue = 1
#> search[ 0.226 ] = 666.935(84.8616)
#> 
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #1 Rep #1 pred log like = 657.25
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #2 Rep #1 pred log like = 695.793
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #3 Rep #1 pred log like = 565.477
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #4 Rep #1 pred log like = 658.699
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #5 Rep #1 pred log like = 796.848
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #6 Rep #1 pred log like = 585.3
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #7 Rep #1 pred log like = 823.096
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #8 Rep #1 pred log like = 702.252
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #9 Rep #1 pred log like = 606.293
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #10 Rep #1 pred log like = 578.204
#> AvgPred = 666.921 with stdev = 84.8612
#> Completed at 2.26
#> Next point at 0.0226 with value 0 and continue = 1
#> search[ 0.226 ] = 666.935(84.8616)
#> search[ 2.26 ] = 666.921(84.8612)
#> 
#> Running at Laplace(9.40721) None  Grid-point #3 at 0.0226 	Fold #1 Rep #1 pred log like = 657.255
#> Running at Laplace(9.40721) None  Grid-point #3 at 0.0226 	Fold #2 Rep #1 pred log like = 695.928
#> Running at Laplace(9.40721) None  Grid-point #3 at 0.0226 	Fold #3 Rep #1 pred log like = 565.524
#> Running at Laplace(9.40721) None  Grid-point #3 at 0.0226 	Fold #4 Rep #1 pred log like = 658.892
#> Running at Laplace(9.40721) None  Grid-point #3 at 0.0226 	Fold #5 Rep #1 pred log like = 796.982
#> Running at Laplace(9.40721) None  Grid-point #3 at 0.0226 	Fold #6 Rep #1 pred log like = 585.323
#> Running at Laplace(9.40721) None  Grid-point #3 at 0.0226 	Fold #7 Rep #1 pred log like = 823.038
#> Running at Laplace(9.40721) None  Grid-point #3 at 0.0226 	Fold #8 Rep #1 pred log like = 702.265
#> Running at Laplace(9.40721) None  Grid-point #3 at 0.0226 	Fold #9 Rep #1 pred log like = 606.248
#> Running at Laplace(9.40721) None  Grid-point #3 at 0.0226 	Fold #10 Rep #1 pred log like = 578.158
#> AvgPred = 666.961 with stdev = 84.8743
#> Completed at 0.0226
#> Next point at 0.00226 with value 0 and continue = 1
#> search[ 0.0226 ] = 666.961(84.8743)
#> search[ 0.226 ] = 666.935(84.8616)
#> search[ 2.26 ] = 666.921(84.8612)
#> 
#> Running at Laplace(29.7482) None  Grid-point #4 at 0.00226 	Fold #1 Rep #1 pred log like = 657.243
#> Running at Laplace(29.7482) None  Grid-point #4 at 0.00226 	Fold #2 Rep #1 pred log like = 696.195
#> Running at Laplace(29.7482) None  Grid-point #4 at 0.00226 	Fold #3 Rep #1 pred log like = 565.587
#> Running at Laplace(29.7482) None  Grid-point #4 at 0.00226 	Fold #4 Rep #1 pred log like = 659.291
#> Running at Laplace(29.7482) None  Grid-point #4 at 0.00226 	Fold #5 Rep #1 pred log like = 797.232
#> Running at Laplace(29.7482) None  Grid-point #4 at 0.00226 	Fold #6 Rep #1 pred log like = 585.335
#> Running at Laplace(29.7482) None  Grid-point #4 at 0.00226 	Fold #7 Rep #1 pred log like = 822.886
#> Running at Laplace(29.7482) None  Grid-point #4 at 0.00226 	Fold #8 Rep #1 pred log like = 702.347
#> Running at Laplace(29.7482) None  Grid-point #4 at 0.00226 	Fold #9 Rep #1 pred log like = 606.113
#> Running at Laplace(29.7482) None  Grid-point #4 at 0.00226 	Fold #10 Rep #1 pred log like = 578.005
#> AvgPred = 667.023 with stdev = 84.9109
#> Completed at 0.00226
#> Next point at 0.000226 with value 0 and continue = 1
#> search[ 0.00226 ] = 667.023(84.9109)
#> search[ 0.0226 ] = 666.961(84.8743)
#> search[ 0.226 ] = 666.935(84.8616)
#> search[ 2.26 ] = 666.921(84.8612)
#> 
#> Running at Laplace(94.0721) None  Grid-point #5 at 0.000226 	Fold #1 Rep #1 pred log like = 657.235
#> Running at Laplace(94.0721) None  Grid-point #5 at 0.000226 	Fold #2 Rep #1 pred log like = 696.232
#> Running at Laplace(94.0721) None  Grid-point #5 at 0.000226 	Fold #3 Rep #1 pred log like = 565.599
#> Running at Laplace(94.0721) None  Grid-point #5 at 0.000226 	Fold #4 Rep #1 pred log like = 659.49
#> Running at Laplace(94.0721) None  Grid-point #5 at 0.000226 	Fold #5 Rep #1 pred log like = 797.258
#> Running at Laplace(94.0721) None  Grid-point #5 at 0.000226 	Fold #6 Rep #1 pred log like = 585.335
#> Running at Laplace(94.0721) None  Grid-point #5 at 0.000226 	Fold #7 Rep #1 pred log like = 822.886
#> Running at Laplace(94.0721) None  Grid-point #5 at 0.000226 	Fold #8 Rep #1 pred log like = 702.377
#> Running at Laplace(94.0721) None  Grid-point #5 at 0.000226 	Fold #9 Rep #1 pred log like = 606.113
#> Running at Laplace(94.0721) None  Grid-point #5 at 0.000226 	Fold #10 Rep #1 pred log like = 578.005
#> AvgPred = 667.053 with stdev = 84.9141
#> Completed at 0.000226
#> Next point at 2.26e-05 with value 0 and continue = 1
#> search[ 0.000226 ] = 667.053(84.9141)
#> search[ 0.00226 ] = 667.023(84.9109)
#> search[ 0.0226 ] = 666.961(84.8743)
#> search[ 0.226 ] = 666.935(84.8616)
#> search[ 2.26 ] = 666.921(84.8612)
#> 
#> Running at Laplace(297.482) None  Grid-point #6 at 2.26e-05 	Fold #1 Rep #1 pred log like = 657.235
#> Running at Laplace(297.482) None  Grid-point #6 at 2.26e-05 	Fold #2 Rep #1 pred log like = 696.232
#> Running at Laplace(297.482) None  Grid-point #6 at 2.26e-05 	Fold #3 Rep #1 pred log like = 565.599
#> Running at Laplace(297.482) None  Grid-point #6 at 2.26e-05 	Fold #4 Rep #1 pred log like = 659.49
#> Running at Laplace(297.482) None  Grid-point #6 at 2.26e-05 	Fold #5 Rep #1 pred log like = 797.258
#> Running at Laplace(297.482) None  Grid-point #6 at 2.26e-05 	Fold #6 Rep #1 pred log like = 585.335
#> Running at Laplace(297.482) None  Grid-point #6 at 2.26e-05 	Fold #7 Rep #1 pred log like = 822.886
#> Running at Laplace(297.482) None  Grid-point #6 at 2.26e-05 	Fold #8 Rep #1 pred log like = 702.377
#> Running at Laplace(297.482) None  Grid-point #6 at 2.26e-05 	Fold #9 Rep #1 pred log like = 606.113
#> Running at Laplace(297.482) None  Grid-point #6 at 2.26e-05 	Fold #10 Rep #1 pred log like = 578.005
#> AvgPred = 667.053 with stdev = 84.9141
#> Completed at 2.26e-05
#> Next point at 1.40708e-18 with value 667.235 and continue = 0
#> search[ 2.26e-05 ] = 667.053(84.9141)
#> search[ 0.000226 ] = 667.053(84.9141)
#> search[ 0.00226 ] = 667.023(84.9109)
#> search[ 0.0226 ] = 666.961(84.8743)
#> search[ 0.226 ] = 666.935(84.8616)
#> search[ 2.26 ] = 666.921(84.8612)
#> 
#> 
#> Maximum predicted log likelihood (667.235) estimated at:
#> 	1.40708e-18 (variance)
#> 	1.19222e+09 (lambda)
#> 
#> Fitting model at optimal hyperparameter
#> Using prior: Laplace(1.19222e+09) None 

#Find out what the optimal hyperparameter was:
getHyperParameter(fit)

#> [1] 1.407076e-18

#Extract the current log-likelihood, and coefficients
logLik(fit)

#> 'log Lik.' -2176.569 (df=3)

coef(fit)

#> (Intercept)           1           2 
#>   -3.668132    0.000000    0.000000 

#We can only retrieve the confidence interval for unregularized coefficients:
confint(fit, c(0))

#> Using 1 thread(s)

#>      covariate    2.5 %    97.5 % evaluations
#> [1,]         0 -3.69286 -3.643592          24