Simulation Cyclops dataset

simulateCyclopsData generates a simulated large, sparse data set for use by fitCyclopsSimulation.

simulateCyclopsData(
  nstrata = 200,
  nrows = 10000,
  ncovars = 20,
  effectSizeSd = 1,
  zeroEffectSizeProp = 0.9,
  eCovarsPerRow = ncovars/100,
  model = "survival"
)

Arguments

nstrata

Numeric: Number of strata

nrows

Numeric: Number of observation rows

ncovars

Numeric: Number of covariates

effectSizeSd

Numeric: Standard derivation of the non-zero simulated regression coefficients

zeroEffectSizeProp

Numeric: Expected proportion of zero effect size

eCovarsPerRow

Number: Effective number of non-zero covariates per data row

model

String: Simulation model. Choices are: logistic, poisson or survival

Value

A simulated data set

Examples

#Generate some simulated data:
sim <- simulateCyclopsData(nstrata = 1, nrows = 1000, ncovars = 2, eCovarsPerRow = 0.5, 
                           model = "poisson")
#> Sparseness = 76.25 %
cyclopsData <- convertToCyclopsData(sim$outcomes, sim$covariates, modelType = "pr", 
                                    addIntercept = TRUE)
#> Sorting covariates by covariateId and rowId

#Define the prior and control objects to use cross-validation for finding the 
#optimal hyperparameter:
prior <- createPrior("laplace", exclude = 0, useCrossValidation = TRUE)
control <- createControl(cvType = "auto", noiseLevel = "quiet")

#Fit the model
fit <- fitCyclopsModel(cyclopsData,prior = prior, control = control)  
#> Using cross-validation selector type byRow
#> Performing 10-fold cross-validation [seed = 1698757941] with data partitions of sizes 100 100 100 100 100 100 100 100 100 100
#> Using 1 thread(s)
#> Starting var = 0.2375 (default)
#> Running at Laplace(2.90191) None  Grid-point #1 at 0.2375 	Fold #1 Rep #1 pred log like = 102.616
#> Running at Laplace(2.90191) None  Grid-point #1 at 0.2375 	Fold #2 Rep #1 pred log like = 194.806
#> Running at Laplace(2.90191) None  Grid-point #1 at 0.2375 	Fold #3 Rep #1 pred log like = 198.183
#> Running at Laplace(2.90191) None  Grid-point #1 at 0.2375 	Fold #4 Rep #1 pred log like = 145.178
#> Running at Laplace(2.90191) None  Grid-point #1 at 0.2375 	Fold #5 Rep #1 pred log like = 75.747
#> Running at Laplace(2.90191) None  Grid-point #1 at 0.2375 	Fold #6 Rep #1 pred log like = 78.3006
#> Running at Laplace(2.90191) None  Grid-point #1 at 0.2375 	Fold #7 Rep #1 pred log like = 123.887
#> Running at Laplace(2.90191) None  Grid-point #1 at 0.2375 	Fold #8 Rep #1 pred log like = 192.631
#> Running at Laplace(2.90191) None  Grid-point #1 at 0.2375 	Fold #9 Rep #1 pred log like = 153.328
#> Running at Laplace(2.90191) None  Grid-point #1 at 0.2375 	Fold #10 Rep #1 pred log like = 101.303
#> AvgPred = 136.598 with stdev = 45.0972
#> Completed at 0.2375
#> Next point at 2.375 with value 0 and continue = 1
#> search[ 0.2375 ] = 136.598(45.0972)
#> 
#> Running at Laplace(0.917663) None  Grid-point #2 at 2.375 	Fold #1 Rep #1 pred log like = 102.586
#> Running at Laplace(0.917663) None  Grid-point #2 at 2.375 	Fold #2 Rep #1 pred log like = 194.788
#> Running at Laplace(0.917663) None  Grid-point #2 at 2.375 	Fold #3 Rep #1 pred log like = 198.074
#> Running at Laplace(0.917663) None  Grid-point #2 at 2.375 	Fold #4 Rep #1 pred log like = 145.144
#> Running at Laplace(0.917663) None  Grid-point #2 at 2.375 	Fold #5 Rep #1 pred log like = 75.6951
#> Running at Laplace(0.917663) None  Grid-point #2 at 2.375 	Fold #6 Rep #1 pred log like = 78.2926
#> Running at Laplace(0.917663) None  Grid-point #2 at 2.375 	Fold #7 Rep #1 pred log like = 123.843
#> Running at Laplace(0.917663) None  Grid-point #2 at 2.375 	Fold #8 Rep #1 pred log like = 192.573
#> Running at Laplace(0.917663) None  Grid-point #2 at 2.375 	Fold #9 Rep #1 pred log like = 153.292
#> Running at Laplace(0.917663) None  Grid-point #2 at 2.375 	Fold #10 Rep #1 pred log like = 101.237
#> AvgPred = 136.553 with stdev = 45.0876
#> Completed at 2.375
#> Next point at 0.02375 with value 0 and continue = 1
#> search[ 0.2375 ] = 136.598(45.0972)
#> search[ 2.375 ] = 136.553(45.0876)
#> 
#> Running at Laplace(9.17663) None  Grid-point #3 at 0.02375 	Fold #1 Rep #1 pred log like = 102.69
#> Running at Laplace(9.17663) None  Grid-point #3 at 0.02375 	Fold #2 Rep #1 pred log like = 194.857
#> Running at Laplace(9.17663) None  Grid-point #3 at 0.02375 	Fold #3 Rep #1 pred log like = 198.387
#> Running at Laplace(9.17663) None  Grid-point #3 at 0.02375 	Fold #4 Rep #1 pred log like = 145.273
#> Running at Laplace(9.17663) None  Grid-point #3 at 0.02375 	Fold #5 Rep #1 pred log like = 75.8784
#> Running at Laplace(9.17663) None  Grid-point #3 at 0.02375 	Fold #6 Rep #1 pred log like = 78.3257
#> Running at Laplace(9.17663) None  Grid-point #3 at 0.02375 	Fold #7 Rep #1 pred log like = 124.02
#> Running at Laplace(9.17663) None  Grid-point #3 at 0.02375 	Fold #8 Rep #1 pred log like = 192.801
#> Running at Laplace(9.17663) None  Grid-point #3 at 0.02375 	Fold #9 Rep #1 pred log like = 153.42
#> Running at Laplace(9.17663) None  Grid-point #3 at 0.02375 	Fold #10 Rep #1 pred log like = 101.43
#> AvgPred = 136.708 with stdev = 45.1178
#> Completed at 0.02375
#> Next point at 0.002375 with value 0 and continue = 1
#> search[ 0.02375 ] = 136.708(45.1178)
#> search[ 0.2375 ] = 136.598(45.0972)
#> search[ 2.375 ] = 136.553(45.0876)
#> 
#> Running at Laplace(29.0191) None  Grid-point #4 at 0.002375 	Fold #1 Rep #1 pred log like = 102.69
#> Running at Laplace(29.0191) None  Grid-point #4 at 0.002375 	Fold #2 Rep #1 pred log like = 194.874
#> Running at Laplace(29.0191) None  Grid-point #4 at 0.002375 	Fold #3 Rep #1 pred log like = 198.387
#> Running at Laplace(29.0191) None  Grid-point #4 at 0.002375 	Fold #4 Rep #1 pred log like = 145.273
#> Running at Laplace(29.0191) None  Grid-point #4 at 0.002375 	Fold #5 Rep #1 pred log like = 75.988
#> Running at Laplace(29.0191) None  Grid-point #4 at 0.002375 	Fold #6 Rep #1 pred log like = 78.3279
#> Running at Laplace(29.0191) None  Grid-point #4 at 0.002375 	Fold #7 Rep #1 pred log like = 124.209
#> Running at Laplace(29.0191) None  Grid-point #4 at 0.002375 	Fold #8 Rep #1 pred log like = 192.831
#> Running at Laplace(29.0191) None  Grid-point #4 at 0.002375 	Fold #9 Rep #1 pred log like = 153.42
#> Running at Laplace(29.0191) None  Grid-point #4 at 0.002375 	Fold #10 Rep #1 pred log like = 101.43
#> AvgPred = 136.743 with stdev = 45.1033
#> Completed at 0.002375
#> Next point at 0.0002375 with value 0 and continue = 1
#> search[ 0.002375 ] = 136.743(45.1033)
#> search[ 0.02375 ] = 136.708(45.1178)
#> search[ 0.2375 ] = 136.598(45.0972)
#> search[ 2.375 ] = 136.553(45.0876)
#> 
#> Running at Laplace(91.7663) None  Grid-point #5 at 0.0002375 	Fold #1 Rep #1 pred log like = 102.69
#> Running at Laplace(91.7663) None  Grid-point #5 at 0.0002375 	Fold #2 Rep #1 pred log like = 194.874
#> Running at Laplace(91.7663) None  Grid-point #5 at 0.0002375 	Fold #3 Rep #1 pred log like = 198.387
#> Running at Laplace(91.7663) None  Grid-point #5 at 0.0002375 	Fold #4 Rep #1 pred log like = 145.273
#> Running at Laplace(91.7663) None  Grid-point #5 at 0.0002375 	Fold #5 Rep #1 pred log like = 75.988
#> Running at Laplace(91.7663) None  Grid-point #5 at 0.0002375 	Fold #6 Rep #1 pred log like = 78.3279
#> Running at Laplace(91.7663) None  Grid-point #5 at 0.0002375 	Fold #7 Rep #1 pred log like = 124.209
#> Running at Laplace(91.7663) None  Grid-point #5 at 0.0002375 	Fold #8 Rep #1 pred log like = 192.831
#> Running at Laplace(91.7663) None  Grid-point #5 at 0.0002375 	Fold #9 Rep #1 pred log like = 153.42
#> Running at Laplace(91.7663) None  Grid-point #5 at 0.0002375 	Fold #10 Rep #1 pred log like = 101.43
#> AvgPred = 136.743 with stdev = 45.1033
#> Completed at 0.0002375
#> Next point at 0.00014571 with value 136.751 and continue = 0
#> search[ 0.0002375 ] = 136.743(45.1033)
#> search[ 0.002375 ] = 136.743(45.1033)
#> search[ 0.02375 ] = 136.708(45.1178)
#> search[ 0.2375 ] = 136.598(45.0972)
#> search[ 2.375 ] = 136.553(45.0876)
#> 
#> 
#> Maximum predicted log likelihood (136.751) estimated at:
#> 	0.00014571 (variance)
#> 	117.158 (lambda)
#> 
#> Fitting model at optimal hyperparameter
#> Using prior: Laplace(117.158) None 

#Find out what the optimal hyperparameter was:
getHyperParameter(fit)
#> [1] 0.0001457099

#Extract the current log-likelihood, and coefficients
logLik(fit)
#> 'log Lik.' -1805.625 (df=3)
coef(fit)
#> (Intercept)           1           2 
#>   -4.287453    0.000000    0.000000 

#We can only retrieve the confidence interval for unregularized coefficients:
confint(fit, c(0))
#> Using 1 thread(s)
#> 
#> Warning: BLR gradient is ill-conditioned
#> Enforcing convergence!
#> 
#> Warning: BLR gradient is ill-conditioned
#> Enforcing convergence!
#> 
#> Warning: BLR gradient is ill-conditioned
#> Enforcing convergence!
#> 
#> Warning: BLR gradient is ill-conditioned
#> Enforcing convergence!
#> 
#> Warning: BLR gradient is ill-conditioned
#> Enforcing convergence!
#> 
#> Warning: BLR gradient is ill-conditioned
#> Enforcing convergence!
#> 
#> Warning: BLR gradient is ill-conditioned
#> Enforcing convergence!
#> 
#> Warning: BLR gradient is ill-conditioned
#> Enforcing convergence!
#> 
#> Warning: BLR gradient is ill-conditioned
#> Enforcing convergence!
#> 
#> Warning: BLR gradient is ill-conditioned
#> Enforcing convergence!
#> 
#> Warning: BLR gradient is ill-conditioned
#> Enforcing convergence!
#> 
#> Warning: BLR gradient is ill-conditioned
#> Enforcing convergence!
#> 
#> Warning: BLR gradient is ill-conditioned
#> Enforcing convergence!
#> 
#> Warning: BLR gradient is ill-conditioned
#> Enforcing convergence!
#> 
#> Warning: BLR gradient is ill-conditioned
#> Enforcing convergence!
#>             covariate     2.5 %    97.5 % evaluations
#> (Intercept)         0 -4.321798 -4.253487          23