Simulation Cyclops dataset

simulateCyclopsData generates a simulated large, sparse data set for use by fitCyclopsSimulation.

simulateCyclopsData(
  nstrata = 200,
  nrows = 10000,
  ncovars = 20,
  effectSizeSd = 1,
  zeroEffectSizeProp = 0.9,
  eCovarsPerRow = ncovars/100,
  model = "survival"
)

Arguments

nstrata
Numeric: Number of strata

nrows
Numeric: Number of observation rows

ncovars
Numeric: Number of covariates

effectSizeSd
Numeric: Standard derivation of the non-zero simulated regression coefficients

zeroEffectSizeProp
Numeric: Expected proportion of zero effect size

eCovarsPerRow
Number: Effective number of non-zero covariates per data row

model
String: Simulation model. Choices are: logistic, poisson or survival

Value

A simulated data set

Examples

#Generate some simulated data:
sim <- simulateCyclopsData(nstrata = 1, nrows = 1000, ncovars = 2, eCovarsPerRow = 0.5,
                           model = "poisson")

#> Sparseness = 77.4 %

cyclopsData <- convertToCyclopsData(sim$outcomes, sim$covariates, modelType = "pr",
                                    addIntercept = TRUE)

#> Sorting covariates by covariateId and rowId

#Define the prior and control objects to use cross-validation for finding the 
#optimal hyperparameter:
prior <- createPrior("laplace", exclude = 0, useCrossValidation = TRUE)
control <- createControl(cvType = "auto", noiseLevel = "quiet")

#Fit the model
fit <- fitCyclopsModel(cyclopsData,prior = prior, control = control)

#> Using cross-validation selector type byRow
#> Performing 10-fold cross-validation [seed = 1591596888] with data partitions of sizes 100 100 100 100 100 100 100 100 100 100
#> Using 1 thread(s)
#> Starting var = 0.226 (default)
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #1 Rep #1 pred log like = 746.822
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #2 Rep #1 pred log like = 729.749
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #3 Rep #1 pred log like = 485.962
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #4 Rep #1 pred log like = 615.833
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #5 Rep #1 pred log like = 710.062
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #6 Rep #1 pred log like = 658.988
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #7 Rep #1 pred log like = 621.009
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #8 Rep #1 pred log like = 700.555
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #9 Rep #1 pred log like = 684.812
#> Running at Laplace(2.97482) None  Grid-point #1 at 0.226 	Fold #10 Rep #1 pred log like = 713.505
#> AvgPred = 666.73 with stdev = 73.0347
#> Completed at 0.226
#> Next point at 2.26 with value 0 and continue = 1
#> search[ 0.226 ] = 666.73(73.0347)
#> 
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #1 Rep #1 pred log like = 746.75
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #2 Rep #1 pred log like = 729.75
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #3 Rep #1 pred log like = 485.964
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #4 Rep #1 pred log like = 615.811
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #5 Rep #1 pred log like = 710.107
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #6 Rep #1 pred log like = 658.972
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #7 Rep #1 pred log like = 620.997
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #8 Rep #1 pred log like = 700.527
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #9 Rep #1 pred log like = 684.774
#> Running at Laplace(0.940721) None  Grid-point #2 at 2.26 	Fold #10 Rep #1 pred log like = 713.506
#> AvgPred = 666.716 with stdev = 73.0293
#> Completed at 2.26
#> Next point at 0.0226 with value 0 and continue = 1
#> search[ 0.226 ] = 666.73(73.0347)
#> search[ 2.26 ] = 666.716(73.0293)
#> 
#> Running at Laplace(9.40721) None  Grid-point #3 at 0.0226 	Fold #1 Rep #1 pred log like = 747.043
#> Running at Laplace(9.40721) None  Grid-point #3 at 0.0226 	Fold #2 Rep #1 pred log like = 729.743
#> Running at Laplace(9.40721) None  Grid-point #3 at 0.0226 	Fold #3 Rep #1 pred log like = 485.974
#> Running at Laplace(9.40721) None  Grid-point #3 at 0.0226 	Fold #4 Rep #1 pred log like = 615.9
#> Running at Laplace(9.40721) None  Grid-point #3 at 0.0226 	Fold #5 Rep #1 pred log like = 709.915
#> Running at Laplace(9.40721) None  Grid-point #3 at 0.0226 	Fold #6 Rep #1 pred log like = 659.062
#> Running at Laplace(9.40721) None  Grid-point #3 at 0.0226 	Fold #7 Rep #1 pred log like = 621.043
#> Running at Laplace(9.40721) None  Grid-point #3 at 0.0226 	Fold #8 Rep #1 pred log like = 700.639
#> Running at Laplace(9.40721) None  Grid-point #3 at 0.0226 	Fold #9 Rep #1 pred log like = 684.924
#> Running at Laplace(9.40721) None  Grid-point #3 at 0.0226 	Fold #10 Rep #1 pred log like = 713.497
#> AvgPred = 666.774 with stdev = 73.0455
#> Completed at 0.0226
#> Next point at 0.00226 with value 0 and continue = 1
#> search[ 0.0226 ] = 666.774(73.0455)
#> search[ 0.226 ] = 666.73(73.0347)
#> search[ 2.26 ] = 666.716(73.0293)
#> 
#> Running at Laplace(29.7482) None  Grid-point #4 at 0.00226 	Fold #1 Rep #1 pred log like = 747.067
#> Running at Laplace(29.7482) None  Grid-point #4 at 0.00226 	Fold #2 Rep #1 pred log like = 729.734
#> Running at Laplace(29.7482) None  Grid-point #4 at 0.00226 	Fold #3 Rep #1 pred log like = 485.99
#> Running at Laplace(29.7482) None  Grid-point #4 at 0.00226 	Fold #4 Rep #1 pred log like = 616.081
#> Running at Laplace(29.7482) None  Grid-point #4 at 0.00226 	Fold #5 Rep #1 pred log like = 709.829
#> Running at Laplace(29.7482) None  Grid-point #4 at 0.00226 	Fold #6 Rep #1 pred log like = 659.346
#> Running at Laplace(29.7482) None  Grid-point #4 at 0.00226 	Fold #7 Rep #1 pred log like = 621.162
#> Running at Laplace(29.7482) None  Grid-point #4 at 0.00226 	Fold #8 Rep #1 pred log like = 700.986
#> Running at Laplace(29.7482) None  Grid-point #4 at 0.00226 	Fold #9 Rep #1 pred log like = 685.192
#> Running at Laplace(29.7482) None  Grid-point #4 at 0.00226 	Fold #10 Rep #1 pred log like = 713.648
#> AvgPred = 666.904 with stdev = 73.0475
#> Completed at 0.00226
#> Next point at 0.000226 with value 0 and continue = 1
#> search[ 0.00226 ] = 666.904(73.0475)
#> search[ 0.0226 ] = 666.774(73.0455)
#> search[ 0.226 ] = 666.73(73.0347)
#> search[ 2.26 ] = 666.716(73.0293)
#> 
#> Running at Laplace(94.0721) None  Grid-point #5 at 0.000226 	Fold #1 Rep #1 pred log like = 747.057
#> Running at Laplace(94.0721) None  Grid-point #5 at 0.000226 	Fold #2 Rep #1 pred log like = 729.728
#> Running at Laplace(94.0721) None  Grid-point #5 at 0.000226 	Fold #3 Rep #1 pred log like = 485.99
#> Running at Laplace(94.0721) None  Grid-point #5 at 0.000226 	Fold #4 Rep #1 pred log like = 616.192
#> Running at Laplace(94.0721) None  Grid-point #5 at 0.000226 	Fold #5 Rep #1 pred log like = 709.829
#> Running at Laplace(94.0721) None  Grid-point #5 at 0.000226 	Fold #6 Rep #1 pred log like = 659.548
#> Running at Laplace(94.0721) None  Grid-point #5 at 0.000226 	Fold #7 Rep #1 pred log like = 621.217
#> Running at Laplace(94.0721) None  Grid-point #5 at 0.000226 	Fold #8 Rep #1 pred log like = 701.269
#> Running at Laplace(94.0721) None  Grid-point #5 at 0.000226 	Fold #9 Rep #1 pred log like = 685.227
#> Running at Laplace(94.0721) None  Grid-point #5 at 0.000226 	Fold #10 Rep #1 pred log like = 713.666
#> AvgPred = 666.972 with stdev = 73.0482
#> Completed at 0.000226
#> Next point at 2.26e-05 with value 0 and continue = 1
#> search[ 0.000226 ] = 666.972(73.0482)
#> search[ 0.00226 ] = 666.904(73.0475)
#> search[ 0.0226 ] = 666.774(73.0455)
#> search[ 0.226 ] = 666.73(73.0347)
#> search[ 2.26 ] = 666.716(73.0293)
#> 
#> Running at Laplace(297.482) None  Grid-point #6 at 2.26e-05 	Fold #1 Rep #1 pred log like = 747.057
#> Running at Laplace(297.482) None  Grid-point #6 at 2.26e-05 	Fold #2 Rep #1 pred log like = 729.728
#> Running at Laplace(297.482) None  Grid-point #6 at 2.26e-05 	Fold #3 Rep #1 pred log like = 485.99
#> Running at Laplace(297.482) None  Grid-point #6 at 2.26e-05 	Fold #4 Rep #1 pred log like = 616.192
#> Running at Laplace(297.482) None  Grid-point #6 at 2.26e-05 	Fold #5 Rep #1 pred log like = 709.829
#> Running at Laplace(297.482) None  Grid-point #6 at 2.26e-05 	Fold #6 Rep #1 pred log like = 659.548
#> Running at Laplace(297.482) None  Grid-point #6 at 2.26e-05 	Fold #7 Rep #1 pred log like = 621.217
#> Running at Laplace(297.482) None  Grid-point #6 at 2.26e-05 	Fold #8 Rep #1 pred log like = 701.269
#> Running at Laplace(297.482) None  Grid-point #6 at 2.26e-05 	Fold #9 Rep #1 pred log like = 685.227
#> Running at Laplace(297.482) None  Grid-point #6 at 2.26e-05 	Fold #10 Rep #1 pred log like = 713.666
#> AvgPred = 666.972 with stdev = 73.0482
#> Completed at 2.26e-05
#> Next point at 1.22035e+61 with value 664.91 and continue = 0
#> search[ 2.26e-05 ] = 666.972(73.0482)
#> search[ 0.000226 ] = 666.972(73.0482)
#> search[ 0.00226 ] = 666.904(73.0475)
#> search[ 0.0226 ] = 666.774(73.0455)
#> search[ 0.226 ] = 666.73(73.0347)
#> search[ 2.26 ] = 666.716(73.0293)
#> 
#> 
#> Maximum predicted log likelihood (664.91) estimated at:
#> 	1.22035e+61 (variance)
#> 	4.0483e-31 (lambda)
#> 
#> Fitting model at optimal hyperparameter
#> Using prior: Laplace(4.0483e-31) None 

#Find out what the optimal hyperparameter was:
getHyperParameter(fit)

#> [1] 1.220352e+61

#Extract the current log-likelihood, and coefficients
logLik(fit)

#> 'log Lik.' -2175.875 (df=3)

coef(fit)

#> (Intercept)           1           2 
#> -3.65807193 -0.03022405 -0.01370368 

#We can only retrieve the confidence interval for unregularized coefficients:
confint(fit, c(0))

#> Using 1 thread(s)

#>      covariate     2.5 %    97.5 % evaluations
#> [1,]         0 -3.688344 -3.628078          22