Simulation Cyclops dataset

simulateCyclopsData generates a simulated large, sparse data set for use by fitCyclopsSimulation.

simulateCyclopsData(nstrata = 200, nrows = 10000, ncovars = 20,
  effectSizeSd = 1, zeroEffectSizeProp = 0.9,
  eCovarsPerRow = ncovars/100, model = "survival")

Arguments

nstrata
Numeric: Number of strata

nrows
Numeric: Number of observation rows

ncovars
Numeric: Number of covariates

effectSizeSd
Numeric: Standard derivation of the non-zero simulated regression coefficients

zeroEffectSizeProp
Numeric: Expected proportion of zero effect size

eCovarsPerRow
Number: Effective number of non-zero covariates per data row

model
String: Simulation model. Choices are: logistic, poisson or survival

Value

A simulated data set

Examples

#Generate some simulated data:
sim <- simulateCyclopsData(nstrata = 1, nrows = 1000, ncovars = 2, eCovarsPerRow = 0.5,
                           model = "poisson")

#> Sparseness = 75.45 %

cyclopsData <- convertToCyclopsData(sim$outcomes, sim$covariates, modelType = "pr",
                                    addIntercept = TRUE)

#> Sorting covariates by covariateId and rowId

#Define the prior and control objects to use cross-validation for finding the 
#optimal hyperparameter:
prior <- createPrior("laplace", exclude = 0, useCrossValidation = TRUE)
control <- createControl(cvType = "auto", noiseLevel = "quiet")

#Fit the model
fit <- fitCyclopsModel(cyclopsData,prior = prior, control = control)

#> Using cross-validation selector type byRow
#> Performing 10-fold cross-validation [seed = 1543260567] with data partitions of sizes 100 100 100 100 100 100 100 100 100 100
#> Using 1 thread(s)
#> Starting var = 0.2455 (default)
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #1 Rep #1 pred log like = -208.19
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #2 Rep #1 pred log like = -218.354
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #3 Rep #1 pred log like = -208.694
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #4 Rep #1 pred log like = -200.976
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #5 Rep #1 pred log like = -220.347
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #6 Rep #1 pred log like = -204.021
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #7 Rep #1 pred log like = -212.409
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #8 Rep #1 pred log like = -211.977
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #9 Rep #1 pred log like = -206.164
#> Running at Laplace(2.85423) None  Grid-point #1 at 0.2455 	Fold #10 Rep #1 pred log like = -211.902
#> AvgPred = -210.303 with stdev = 5.7109
#> Completed at 0.2455
#> Next point at 2.455 with value 0 and continue = 1
#> search[ 0.2455 ] = -210.303(5.7109)
#> 
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #1 Rep #1 pred log like = -208.194
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #2 Rep #1 pred log like = -218.356
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #3 Rep #1 pred log like = -208.7
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #4 Rep #1 pred log like = -200.992
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #5 Rep #1 pred log like = -220.37
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #6 Rep #1 pred log like = -204.11
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #7 Rep #1 pred log like = -212.42
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #8 Rep #1 pred log like = -211.991
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #9 Rep #1 pred log like = -206.184
#> Running at Laplace(0.902587) None  Grid-point #2 at 2.455 	Fold #10 Rep #1 pred log like = -211.949
#> AvgPred = -210.327 with stdev = 5.70313
#> Completed at 2.455
#> Next point at 0.02455 with value 0 and continue = 1
#> search[ 0.2455 ] = -210.303(5.7109)
#> search[ 2.455 ] = -210.327(5.70313)
#> 
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #1 Rep #1 pred log like = -208.201
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #2 Rep #1 pred log like = -218.372
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #3 Rep #1 pred log like = -208.675
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #4 Rep #1 pred log like = -200.949
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #5 Rep #1 pred log like = -220.301
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #6 Rep #1 pred log like = -203.746
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #7 Rep #1 pred log like = -212.394
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #8 Rep #1 pred log like = -211.894
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #9 Rep #1 pred log like = -206.118
#> Running at Laplace(9.02587) None  Grid-point #3 at 0.02455 	Fold #10 Rep #1 pred log like = -211.761
#> AvgPred = -210.241 with stdev = 5.73701
#> Completed at 0.02455
#> Next point at 0.002455 with value 0 and continue = 1
#> search[ 0.02455 ] = -210.241(5.73701)
#> search[ 0.2455 ] = -210.303(5.7109)
#> search[ 2.455 ] = -210.327(5.70313)
#> 
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #1 Rep #1 pred log like = -208.272
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #2 Rep #1 pred log like = -218.423
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #3 Rep #1 pred log like = -208.646
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #4 Rep #1 pred log like = -200.942
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #5 Rep #1 pred log like = -220.191
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #6 Rep #1 pred log like = -203.071
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #7 Rep #1 pred log like = -212.461
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #8 Rep #1 pred log like = -211.551
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #9 Rep #1 pred log like = -206.167
#> Running at Laplace(28.5423) None  Grid-point #4 at 0.002455 	Fold #10 Rep #1 pred log like = -211.515
#> AvgPred = -210.124 with stdev = 5.78769
#> Completed at 0.002455
#> Next point at 0.0002455 with value 0 and continue = 1
#> search[ 0.002455 ] = -210.124(5.78769)
#> search[ 0.02455 ] = -210.241(5.73701)
#> search[ 0.2455 ] = -210.303(5.7109)
#> search[ 2.455 ] = -210.327(5.70313)
#> 
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #1 Rep #1 pred log like = -208.272
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #2 Rep #1 pred log like = -218.423
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #3 Rep #1 pred log like = -208.646
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #4 Rep #1 pred log like = -200.942
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #5 Rep #1 pred log like = -220.187
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #6 Rep #1 pred log like = -202.726
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #7 Rep #1 pred log like = -212.461
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #8 Rep #1 pred log like = -211.551
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #9 Rep #1 pred log like = -206.167
#> Running at Laplace(90.2587) None  Grid-point #5 at 0.0002455 	Fold #10 Rep #1 pred log like = -211.515
#> AvgPred = -210.089 with stdev = 5.82972
#> Completed at 0.0002455
#> Next point at 2.455e-005 with value 0 and continue = 1
#> search[ 0.0002455 ] = -210.089(5.82972)
#> search[ 0.002455 ] = -210.124(5.78769)
#> search[ 0.02455 ] = -210.241(5.73701)
#> search[ 0.2455 ] = -210.303(5.7109)
#> search[ 2.455 ] = -210.327(5.70313)
#> 
#> Running at Laplace(285.423) None  Grid-point #6 at 2.455e-005 	Fold #1 Rep #1 pred log like = -208.272
#> Running at Laplace(285.423) None  Grid-point #6 at 2.455e-005 	Fold #2 Rep #1 pred log like = -218.423
#> Running at Laplace(285.423) None  Grid-point #6 at 2.455e-005 	Fold #3 Rep #1 pred log like = -208.646
#> Running at Laplace(285.423) None  Grid-point #6 at 2.455e-005 	Fold #4 Rep #1 pred log like = -200.942
#> Running at Laplace(285.423) None  Grid-point #6 at 2.455e-005 	Fold #5 Rep #1 pred log like = -220.187
#> Running at Laplace(285.423) None  Grid-point #6 at 2.455e-005 	Fold #6 Rep #1 pred log like = -202.726
#> Running at Laplace(285.423) None  Grid-point #6 at 2.455e-005 	Fold #7 Rep #1 pred log like = -212.461
#> Running at Laplace(285.423) None  Grid-point #6 at 2.455e-005 	Fold #8 Rep #1 pred log like = -211.551
#> Running at Laplace(285.423) None  Grid-point #6 at 2.455e-005 	Fold #9 Rep #1 pred log like = -206.167
#> Running at Laplace(285.423) None  Grid-point #6 at 2.455e-005 	Fold #10 Rep #1 pred log like = -211.515
#> AvgPred = -210.089 with stdev = 5.82972
#> Completed at 2.455e-005
#> Next point at 7.60222e-010 with value -209.989 and continue = 0
#> search[ 2.455e-005 ] = -210.089(5.82972)
#> search[ 0.0002455 ] = -210.089(5.82972)
#> search[ 0.002455 ] = -210.124(5.78769)
#> search[ 0.02455 ] = -210.241(5.73701)
#> search[ 0.2455 ] = -210.303(5.7109)
#> search[ 2.455 ] = -210.327(5.70313)
#> 
#> 
#> Maximum predicted log likelihood (-209.989) estimated at:
#> 	7.60222e-010 (variance)
#> 	51291.4 (lambda)
#> 
#> Fitting model at optimal hyperparameter
#> Using prior: Laplace(51291.4) None 

#Find out what the optimal hyperparameter was:
getHyperParameter(fit)

#> [1] 7.602223e-10

#Extract the current log-likelihood, and coefficients
logLik(fit)

#> 'log Lik.' -2099.802 (df=3)

coef(fit)

#> (Intercept)           1           2 
#>    -3.80275     0.00000     0.00000 

#We can only retrieve the confidence interval for unregularized coefficients:
confint(fit, c(0))

#> Using 1 thread(s)

#>      covariate     2.5 %    97.5 % evaluations
#> [1,]         0 -3.829242 -3.776476          24