vcov.cyclopsFit {Cyclops}	R Documentation

Calculate variance-covariance matrix for a fitted Cyclops model object

Description

vcov.cyclopsFit returns the variance-covariance matrix for all covariates of a Cyclops model object

Usage

## S3 method for class 'cyclopsFit'
vcov(object, control, overrideNoRegularization = FALSE, ...)

Arguments

object
A fitted Cyclops model object

control
A "cyclopsControl" object constructed by createControl

overrideNoRegularization
Logical: Enable variance-covariance estimation for regularized parameters

...
Additional argument(s) for methods

Value

A matrix of the estimates covariances between all covariate estimates.

[Package Cyclops version 3.2.0 Index]